% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptLearnerToTask.R
\name{adaptLearnerToTask}
\alias{adaptLearnerToTask}
\title{Update learner according to the task.}
\usage{
adaptLearnerToTask(lrn, task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Update learner according to the task.
}

