% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult_operators.R
\name{addRRMeasure}
\alias{addRRMeasure}
\title{Compute new measures for existing ResampleResult}
\usage{
addRRMeasure(res, measures)
}
\arguments{
\item{res}{([ResampleResult])\cr
The result of [resample] run with `keep.pred = TRUE`.}

\item{measures}{([Measure] | list of [Measure])\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here [getDefaultMeasure].}
}
\value{
([ResampleResult]).
}
\description{
Adds new measures to an existing `ResampleResult`.
}
\seealso{
Other resample: \code{\link{ResamplePrediction}},
  \code{\link{ResampleResult}},
  \code{\link{getRRPredictionList}},
  \code{\link{getRRPredictions}},
  \code{\link{getRRTaskDescription}},
  \code{\link{getRRTaskDesc}},
  \code{\link{makeResampleDesc}},
  \code{\link{makeResampleInstance}},
  \code{\link{resample}}
}

