% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_train.R
\name{aggregateBaseLearners}
\alias{aggregateBaseLearners}
\title{Train function for simple aggregation of base learner predictions without weights.}
\usage{
aggregateBaseLearners(learner, task)
}
\description{
Train function for simple aggregation of base learner predictions without weights.
}

