% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{aggregatePredictions}
\alias{aggregatePredictions}
\title{Aggregate predictions}
\usage{
aggregatePredictions(pred.list, sm.pt = NULL, pL = FALSE)
}
\arguments{
\item{pred.list}{[list of \code{Predictions}]\cr}

\item{sm.pt}{Final predict type, "prob" or "response"}

\item{pL}{FALSE if Predictions with truth (test data), TRUE for truth=NA (new data)}
}
\description{
Aggregate predicitons results by averaging (for \code{regr}, and  \code{classif} with prob) or mode ( \code{classif} with response). 
(works for regr, classif, multiclass)
}

