% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{agri.task}
\alias{agri.task}
\title{European Union Agricultural Workforces clustering task.}
\description{
Contains the task (\code{agri.task}).
}
\references{
See \link[cluster:agriculture]{cluster::agriculture}.
}
\keyword{data}
