\name{analyzeFeatSelResult}
\alias{analyzeFeatSelResult}
\title{Shows the detailed steps of the feature selection.}
\usage{
analyzeFeatSelResult(fs.obj, reduce = TRUE)
}
\arguments{
  \item{fs.obj}{[\code{FeatSelResult(1)}]\cr The result of
  one feature selection generated by
  \code{\link{selectFeatures}}.}

  \item{reduce}{[\code{logical(1)}]\cr \code{TRUE} if the
  results of all calculations for each step should be
  printed. \cr \code{FALSE} only the calculation leading to
  the improving feature set is printed. \cr Default is
  \code{FALSE}.}
}
\value{
Nothing.
}
\description{
This function prints the steps \code{\link{selectFeatures}}
took to find it's optimal set of features and the reason
why it stopped. It can also give information about all
calculations done on each intermediate step.
}

