% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{applyCPO}
\alias{applyCPO}
\title{CPO Apply}
\usage{
applyCPO(cpo, task)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The CPO representing the operation to perform.}

\item{cpo1}{[\code{\link{Task}}]\cr
The task to operate on.}
}
\description{
The given transformation will be applied to the data in the given \code{link{Task}}.

See the preferred \code{\link{\%>>\%}} for more info.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{attachCPO}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{makeCPOFunctional}},
  \code{\link{makeCPOObject}}, \code{\link{setHyperPars}}
}
