% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_train.R
\name{applyEnsembleSelection}
\alias{applyEnsembleSelection}
\title{Ensemble selection algorithm}
\usage{
applyEnsembleSelection(pred.list = pred.list,
  bls.performance = bls.performance, parset = list(replace = TRUE, init = 1,
  bagprob = 1, bagtime = 1, maxiter = NULL, tolerance = 1e-08, metric = NULL))
}
\arguments{
\item{pred.list}{A named list of predictions.}

\item{bls.performance}{Named vector of performance results from training 
(note that this should be results from resampled predictions to overcome overfitting issues).}

\item{parset}{list of parameters. See /code{/{link{makeStackedLearner}}}.}
}
\description{
Ensemble selection algorithm
}
\references{
Caruana, Rich, et al. "Ensemble selection from libraries of models." 
  Proceedings of the twenty-first international conference on Machine learning. 
  ACM, 2004. \url{http://www.cs.cornell.edu/~caruana/ctp/ct.papers/caruana.icml04.icdm06long.pdf}
}

