\name{as.ROCR.preds}
\alias{as.ROCR.preds-methods}
\alias{as.ROCR.preds}
\title{Convert to ROCR format.}
\description{Converts predictions to a format package ROCR can handle.}
\arguments{\item{x}{[character] \cr
Predictions.}
}
\alias{as.ROCR.preds,prediction-method}
\alias{as.ROCR.preds}
\alias{as.ROCR.preds,resample.prediction-method}
\alias{as.ROCR.preds}
