\name{base.wrapper-class}
\alias{base.wrapper-class}
\alias{base.wrapper}
\title{Abstract base class to wrap something around a learner.}
\description{Abstract base class to wrap something around a learner.}
\section{Extends}{\code{\linkS4class{learner}}}
\section{Methods}{\describe{\item{\code{\link[=[,base.wrapper-method]{[}}}{}\item{\code{\link[=train.learner,base.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=pred.learner,base.wrapper,wrapped.model,data.frame,ANY-method]{pred.learner}}}{}\item{\code{\link[=set.hyper.pars,base.wrapper,list-method]{set.hyper.pars}}}{}}}
\alias{[,base.wrapper-method}

