% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchmark.r
\name{batchmark}
\alias{batchmark}
\title{Run Machine Learning Benchmarks as Experiments}
\usage{
batchmark(learners, tasks, resamplings, measures = NULL, repls = 1L,
  save.models = FALSE, reg = getDefaultRegistry())
}
\arguments{
\item{learners}{[(list of) \code{\link{Learner}}]\cr
Learning algorithms which should be compared.}

\item{tasks}{[(list of) \code{\link{Task}}]\cr
Tasks that learners should be run on.}

\item{resamplings}{[(list of) \code{\link{ResampleDesc}} | \code{\link{ResampleInstance}}]\cr
Resampling strategy for each tasks.
If only one is provided, it will be replicated to match the number of tasks.
If missing, a 10-fold cross validation is used.}

\item{measures}{[(list of) \code{\link{Measure}}]\cr
Performance measures for all tasks.
If missing, the default measure of the first task is used.}

\item{repls}{[code{integer(1)}]\cr
Number of replications for each experiment.}

\item{save.models}{[code{logical(1)}]\cr
Store (potentially large) models?}

\item{reg}{[\code{\link[batchtools]{Registry}}]\cr
Registry, created by \code{\link[bacthtools]{makeExperimentRegistry}}. If not explicitly passed, 
uses the last created registry.}
}
\value{
[\code{data.table}]. Generated job ids are stored in the column \dQuote{job.id}.
}
\description{
This function is a very parallel version of \code{\link{benchmark}}.
Experiments are created in the provided registry for each combination of
learners, tasks and resamplings. The experiments are then stored in a registry and the 
runs can be started via \code{\link[batchtools]{submitJobs}]}. For details on the usage
and support backends have a look at the batchtools tutorial page: 
\url{https://github.com/mllg/batchtools}.
}

