% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bc.task}
\alias{bc.task}
\title{Wisconsin Breast Cancer classification task}
\description{
Contains the task (\code{bc.task}).
}
\references{
See \code{\link[mlbench]{BreastCancer}}.
  The column \code{"Id"} and all incomplete cases have been removed from the task.
}
\keyword{data}

