% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{bc.task.spatial}
\alias{bc.task.spatial}
\title{Basque pathogen infection spatial binary classification task.}
\description{
Contains the task (\code{bc.task.spatial}).
}
\references{
Iturritxa, E., Mesanza, N., & Brenning, A. (2014). Spatial
analysis of the risk of major forest diseases in Monterey pine plantations.
Plant Pathology, 64(4), 880-889. doi:10.1111/ppa.12328
}
