% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Result of a benchmark run.}
\description{
Container for results of benchmarked experiments using \code{\link{benchmark}}.
The structure of the object itself is rather complicated, it is recommended to
retrive required information via the \code{getBMR*} getter functions.
You can also convert the object using \code{\link[base]{as.data.frame}}.
}
\seealso{
Other benchmark: \code{\link{benchmark}};
  \code{\link{convertBMRToRankMatrix}};
  \code{\link{friedmanPostHocTestBMR}};
  \code{\link{friedmanTestBMR}};
  \code{\link{generateBenchmarkSummaryData}};
  \code{\link{generateCritDifferencesData}};
  \code{\link{generateRankMatrixAsBarData}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRLearnerIds}};
  \code{\link{getBMRLearners}};
  \code{\link{getBMRMeasureIds}};
  \code{\link{getBMRMeasures}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}};
  \code{\link{plotBenchmarkResult}};
  \code{\link{plotBenchmarkSummary}};
  \code{\link{plotCritDifferences}};
  \code{\link{plotRankMatrixAsBar}}
}

