% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bh.task}
\alias{bh.task}
\title{Boston Housing regression task}
\description{
Contains the task (\code{bh.task}).
}
\references{
See \code{\link[mlbench]{BostonHousing}}.
}
\keyword{data}

