\name{bs.desc-class}
\alias{bs.desc-class}
\alias{bs.desc}
\title{bs.desc}
\description{Description class for bootstrapping.}
\seealso{\code{\link{make.res.desc}}}
\section{Extends}{\code{\linkS4class{resample.desc}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,bs.desc-method]{initialize}}}{\code{signature( = "bs.desc")}: Create description object for bootstrapping.}}}
