% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_helpers.R
\name{cache_helpers}
\alias{cache_helpers}
\alias{getCacheDir}
\alias{deleteCacheDir}
\title{Get or delete mlr cache directory}
\usage{
getCacheDir()

deleteCacheDir()
}
\description{
Helper functions to deal with mlr caching.
}
\details{
\code{getCacheDir()} returns the default mlr cache directory \cr
\code{deleteCacheDir()} clears the default mlr cache directory. Custom cache
directories must be deleted by hand.
}
