% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateROCMeasures.R
\name{calculateROCMeasures}
\alias{calculateROCMeasures}
\alias{print.ROCMeasures}
\title{Calculate receiver operator measures.}
\usage{
calculateROCMeasures(pred)

\method{print}{ROCMeasures}(x, abbreviations = TRUE, digits = 2, ...)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}

\item{x}{[\code{\link{ROCMeasures}}]\cr
Created by \code{\link{calculateROCMeasures}}.}

\item{abbreviations}{[\code{logical(1)}]\cr
If \code{TRUE} a short paragraph with explainations of the used measures is printed additionally.}

\item{digits}{[\code{integer(1)}]\cr
Number of digits the measures are rounded to.}

\item{...}{\code{[any]}\cr
Currently not used.}
}
\value{
[\code{ROCMeasures}]. 
   A list containing two elements \code{confusion.matrix} which is 
   the 2 times 2 confusion matrix of relative frequencies and \code{measures}, a list of the above mentioned measures.
}
\description{
Calculate the relative number of correct/incorrect classifications and the following evaluation measures:


\itemize{
 \item \code{TPR} True positive rate (Sensisivity, Recall)
 \item \code{FPR} False positve rate (Fall-out)
 \item \code{FNR} False negative rate (Miss rate)
 \item \code{TNR} True negative rate (Specifity)
 \item \code{PPV} Positive predictive value (Precision)
 \item \code{FOR} False omission rate
 \item \code{LRP} Positive likelihood ratio (LR+)
 \item \code{FDR} False discovery rate
 \item \code{NPV} Negative predictive value
 \item \code{ACC} Accuracy
 \item \code{LRM} Nevative likelihood ratio (LR-)
 \item \code{DOR} Diagnostic odds ratio)
}

For details on the used measures see \code{\link{measures}} and also
\url{https://en.wikipedia.org/wiki/Receiver_operating_characteristic}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}
\examples{
lrn = makeLearner("classif.rpart", predict.type = "prob")
fit = train(lrn, sonar.task)
pred = predict(fit, task = sonar.task)
calculateROCMeasures(pred)

}

