% Generated by roxygen2 (4.0.1): do not edit by hand
\name{capLargeValues}
\alias{capLargeValues}
\title{Convert large/infinite numeric values in a data.frame or task.}
\usage{
capLargeValues(obj, cols = NULL, threshold = Inf, impute = threshold,
  what = "abs")
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{Task}}]\cr
Input data.}

<<<<<<< HEAD
\item{cols}{[\code{character} | \code{integer}]
Which columns to convert.
Default is all numeric columns}
=======
\item{cols}{[\code{character}]
Which columns to convert.
Default is all numeric columns.}
>>>>>>> mlrEDA

\item{threshold}{[\code{numeric(1)}]\cr
Threshold for capping.
Every entry whose absolute value is equal or larger is converted.
Default is \code{Inf}.}

\item{impute}{[\code{numeric(1)}]\cr
Replacement value for large entries.
Large negative entries are converted to \code{-impute}.
Default is \code{threshold}.}

\item{what}{[character(1)]
What kind of entries are affected?
\dQuote{abs} means \code{abs(x) > threshold},
<<<<<<< HEAD
\dQuote{high} means \code{x > threshold},
\dQuote{low} means \code{x < -threshold}
=======
\dQuote{pos} means \code{abs(x) > threshold && x > 0},
\dQuote{neg} means \code{abs(x) > threshold && x < 0},
>>>>>>> mlrEDA
Default is \dQuote{abs}}
}
\value{
[\code{data.frame}]
}
\description{
Convert numeric entries which large/infinite (absolute) values in a data.frame
Only numeric/integer columns are affected.
}
\examples{
capLargeValues(iris, threshold = 5, impute = 5)
}
\seealso{
<<<<<<< HEAD
Other eda_and_preprocess: \code{\link{summarizeColumns}}
=======
Other eda_and_preprocess: \code{\link{mergeSmallFactorLevels}};
  \code{\link{summarizeColumns}}
>>>>>>> mlrEDA
}

