% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{chainCPO}
\alias{chainCPO}
\title{Turn a list of preprocessing operators into a single chained one}
\usage{
chainCPO(pplist)
}
\arguments{
\item{pplist}{[\code{list} of \code{CPO} | \code{list} of \code{CPORetrafo}]\cr
A list of \code{CPO} or \code{CPORetrafo} objects.}
}
\description{
Chain a list of preprocessing operators, or retrafo objects, turning \code{list(a, b, c)} into
\code{a \%>>\% b \%>>\% c}. This is the inverse operation of \code{as.list},
applied on a \code{CPO} chain.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{makeCPOFunctional}},
  \code{\link{makeCPOObject}}, \code{\link{retrafo<-}},
  \code{\link{retrafo}}, \code{\link{setHyperPars}}
}
