% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLearner.R
\name{checkLearner}
\alias{checkLearner}
\title{Exported for internal use only.}
\usage{
checkLearner(learner, type = NULL, props = NULL)
}
\arguments{
\item{learner}{([Learner] | `character(1)`)\cr
The learner to check, or the name of the learner to create}

\item{type}{(`character(1)`)\cr
What type of learner to require.}

\item{props}{(`character(1)`)\cr
What properties to require.}
}
\description{
Exported for internal use only.
}
\keyword{internal}

