% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMeasures.R
\name{checkMeasures}
\alias{checkMeasures}
\title{Checks for correct strucutre of selected measures}
\usage{
checkMeasures(measures, obj, aggr = NULL, na.rm = FALSE)
}
\arguments{
\item{measures}{[\code{list}]\cr
list of measures}

\item{obj}{[\code{data.frame}]\cr
task}

\item{na.rm}{[\code{logical(1)}]\cr
Should `NA` values be removed? Default `FALSE`.
This applies to all selected measures.}
}
\keyword{internal}
