% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner.R
\name{classif.bs.optimal}
\alias{classif.bs.optimal}
\title{choose weights in a weighted sum to minimize the brier score}
\usage{
classif.bs.optimal(learner, task)
}
\description{
internal helper function
}
\details{
importFrom quadprog solve.QP
}
