% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_classif_featureless.R
\name{classif.featureless}
\alias{classif.featureless}
\title{Featureless classification learner.}
\description{
A very basic baseline method which is useful for model comparisons (if you
don't beat this, you very likely have a problem).
Does not consider any features of the task and only uses the target feature
of the training data to make predictions.
Using observation weights is currently not supported.

Method \dQuote{majority} predicts always the majority class for each new
observation. In case of ties, the predictions will depend on the selected
\dQuote{ties.method}, i.e., either always the first factor level, always the
last factor level or a randomly sampled class for each observation is predicted.

Method \dQuote{sample-prior} always samples a random class for each new
observation according to the prior probabilities observed in the training data.

The default method is \dQuote{majority} with \dQuote{ties.method = "first"}
which corresponds to the ZeroR classifier from WEKA, see
\url{https://weka.wikispaces.com/ZeroR}.

If you opt to predict probabilities, the class probabilities always
correspond to the prior probabilities observed in the training data.
}

