% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateConfusionMatrix.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{Confusion matrix}
\description{
The result of [calculateConfusionMatrix].

Object members:
\describe{
\item{result ([matrix])}{Confusion matrix of absolute values and marginals. Can also contain
  row and column sums of observations.}
\item{task.desc ([TaskDesc])}{Additional information about the task.}
\item{sums (`logical(1)`)}{Flag if marginal sums of observations are calculated.}
\item{relative (`logical(1)`)}{Flag if the relative confusion matrices are calculated.}
\item{relative.row ([matrix])}{Confusion matrix of relative values and marginals normalized by row.}
\item{relative.col ([matrix])}{Confusion matrix of relative values and marginals normalized by column.}
\item{relative.error (`numeric(1)`)}{Relative error overall.}
}
}
\seealso{
Other performance: \code{\link{calculateConfusionMatrix}},
  \code{\link{calculateROCMeasures}},
  \code{\link{estimateRelativeOverfitting}},
  \code{\link{makeCostMeasure}},
  \code{\link{makeCustomResampledMeasure}},
  \code{\link{makeMeasure}}, \code{\link{measures}},
  \code{\link{performance}}, \code{\link{setAggregation}},
  \code{\link{setMeasurePars}}
}

