% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constantPrediction.R
\name{constantPrediction}
\alias{constantPrediction}
\title{Make a constant prediction}
\usage{
constantPrediction(task, measure, predict.type = "response")
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{measure}{[\code{\link{Measure}}]\cr
Performance measure.
Default is the first measure used in the benchmark experiment.}

\item{predict.type}{[\code{character(1)}]\cr
Only meaningful for classification tasks. Either \code{predict.type = "response"} (the default)
if the label of the best class is returned, or \code{predict.type = "prob"} for a vector of
probabilities.}
}
\value{
[\code{factor(1)} | \code{numeric(1)} | \code{numeric}]\cr
 For regression task the best constant numeric prediction is returned. For classification see \code{predict.type}.
}
\description{
This function is used to make the optimal constant prediction for a given task based on a measure.

In case of a regression task the measure is optimized with a line search between the min and max
value of the response. For classification the measure is evaluated for every class
of the response. If \code{predict.type = "prob"} the results over
all classes are scaled to 1 and the probability for every class is returned, if \code{predict.type = "response"}
only the best class label is returned.

Mostly for internal use. Featureleass learners that use this function can be created with 
\code{makeLearner("classif.featurelesss")} or \code{makeLearner("regr.featureless")}.
}

