% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertBMRToRankMatrix.R
\name{convertBMRToRankMatrix}
\alias{convertBMRToRankMatrix}
\title{Convert BenchmarkResult to a Rankmatrix}
\usage{
convertBMRToRankMatrix(bmr, measure = NULL, aggregation = "default")
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{aggregation}{[\code{character](1)}] \cr
 Aggregation method for resampling strategy. \cr
 Can be \code{default} or \code{mean}.
 See \link{getAggrPerformances} for details.}
}
\value{
[\code{data.frame}] \cr
Matrix, with measure ranks as entries. \cr
The columns of the matrix correspond to tasks, rows correspond to learners.
}
\description{
Computes a matrix of all the ranks of different algorithms
over different datasets(tasks). Ranks are computed from aggregated
measures.
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
convertBMRToRankMatrix(res,acc)
}

