% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{convertModelNameToBlsName}
\alias{convertModelNameToBlsName}
\title{Convert models names (when model was saved on disc) to base learner name}
\usage{
convertModelNameToBlsName(base.model.id, stack.id)
}
\arguments{
\item{base.model.id}{Unique ID used to save model on disc}

\item{stack.id}{ID from makeStackedLearner}
}
\description{
Convert models names (when model was saved on disc) to base learner name
}

