% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSfeatures.R
\name{convertTSTaskToNormalTask}
\alias{convertTSTaskToNormalTask}
\title{Create task on time series features.}
\usage{
convertTSTaskToNormalTask(task, method, pars = NULL)
}
\arguments{
\item{task}{[\code{TimeSeriesClassifTask}]\cr
Time series classification task.}

\item{method}{[\code{character(1)}]\cr
Which method is used to create time series features. Methods available.
Wavelet transformation: \dQuote{wavelets}.
Fourier transformation: \dQuote{fourier}.
Shapelet model learning: \dQuote{shapelets}.}

\item{pars}{\cr
Further parameters passed as argument e.g., for feature representation
methods. See the methods' man pages.}
}
\value{
Either [\code{ClassifTask}] based on the transformed data or the
  learned shapelet model [\code{ShapeletModel}].
}
\description{
The function transform a task of type
  \code{TimeSeriesClassifTask} into a standard \code{ClassifTask}. For this,
  it creates a feature representation of the raw time series. The method used
  to create this feature representation must be specified by the user in
  \code{method}. For \code{wavelets} or \code{fourier} features, the
  resulting data does not contain temporal structure anymore, so the returned
  task is a \code{ClassifTask}. For \code{shapelets}, the learned shapelet
  model is returned. See \code{{getTSShapeletFeatures}}.
}

