% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copyTask}
\alias{copyTask}
\title{Copy a task.}
\usage{
copyTask(task, ignore = character(0L))
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{ignore}{[\code{character}]\cr
Items in task to ignore on copy.}
}
\value{
[\code{\link{Task}}].
}
\description{
Copy a task.
}
\seealso{
Other Task: \code{\link{ClassifTask}},
  \code{\link{ClusterTask}}, \code{\link{CostSensTask}},
  \code{\link{RegrTask}}, \code{\link{SurvTask}},
  \code{\link{Task}}, \code{\link{makeClassifTask}},
  \code{\link{makeClusterTask}},
  \code{\link{makeCostSensTask}},
  \code{\link{makeRegrTask}},
  \code{\link{makeSupervisedTask}},
  \code{\link{makeSurvTask}}, \code{\link{makeTask}},
  \code{\link{makeUnsupervisedTask}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFeatures}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTarget}}
}

