% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{costiris.task}
\alias{costiris.task}
\title{Iris cost-sensitive classification task.}
\description{
Contains the task (\code{costiris.task}).
}
\references{
See \link[datasets:iris]{datasets::iris}.
The cost matrix was generated artificially following

Tu, H.-H. and Lin, H.-T. (2010), One-sided support vector regression for multiclass cost-sensitive classification.
In ICML, J. Fürnkranz and T. Joachims, Eds., Omnipress, 1095--1102.
}
\keyword{data}
