% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cov_selected.R
\name{Cov_selected}
\alias{Cov_selected}
\title{Cov_selected}
\usage{
Cov_selected(output, alpha = 0.05)
}
\arguments{
\item{output}{the result object from mlr function}

\item{alpha}{the given threshold of p-value. If not provided, 0.05 is used.}
}
\value{
A vector of selected covariates names
}
\description{
Select covariates using the significance level
}
\examples{
data(mtcars)
output <- mlr("mpg", c("wt","hp","vs"), data=mtcars)
Cov_selected(output)

}
