% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{CPO}
\alias{CPO}
\title{CPO: Composable Preprocessing Operators}
\seealso{
Other CPO: \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{makeCPOFunctional}},
  \code{\link{makeCPOObject}}, \code{\link{setHyperPars}}
}
