% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOCbind.R
\name{cpoCbind}
\alias{cpoCbind}
\title{\dQuote{cbind} the Result of multiple CPOs}
\usage{
cpoCbind(..., .cpos = list())
}
\arguments{
\item{...}{[\code{CPO}]\cr
The CPOs to cbind. Named arguments will result in the respective
columns being prefixed with the name. This is highly recommended if there
is any chance of name collision otherwise. it is possible to use the same
name multiple times.}

\item{.cpos}{[\code{list} of \code{CPO}]\cr
Alternatively, give the CPOs to cbind as a list. Default is \code{list()}.}
}
\description{
Build a CPO that represents the operations of its input parameters,
performed in parallel and put together column wise.

For example, to construct a \code{\link{Task}} that contains the original
data, as well as the data after scaling, one could do

\code{task \%>>\% cpoCbind(NULLCPO, cpoScale())}

The result of \code{cpoCbind} is itself a CPO which exports its constituents'
hyperparameters. CPOs with the same type / ID get combined automatically.
To get networks, e.g. of the form
\preformatted{
      ,-C--E-.
     /    /   \
A---B ---D-----F---G
}
one coul use the code
\preformatted{
initcpo = A \%>>\% B
route1 = initcpo \%>>\% D
route2 = cpoCbind(route1, initcpo \%>>\% C) \%>>\% E
result = cpoCbind(route1, route2) \%>>\% F \%>>\% G
}

cpoCbind finds common paths among its arguments and combines them into one operation.
This saves computation and makes it possible for one exported hyperparameter to
influence multiple of \code{cpoCbind}'s inputs. However, if you want to use the same
operation with different parameters on different parts of \code{cpoCbind} input,
you must give these operations different IDs. If CPOs that could represent an identical CPO,
with the same IDs (or both with IDs absent) but different parameter settings or different
parameter exportations occur, an error will be thrown.
}
