% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_concrete.R
\name{cpoDropConstants}
\alias{cpoDropConstants}
\title{Drop constant or near-constant Features.}
\usage{
cpoDropConstants(rel.tol = 1e-08, abs.tol = 1e-08, ignore.na = FALSE,
  id = NULL, affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\arguments{
\item{rel.tol}{[\code{numeric(1)}]\cr
Relative tolerance within which to consider a feature constant.
Set to \code{Inf} to disregard relative tolerance.
Default is \code{1e-8}.}

\item{abs.tol}{[\code{numeric(1)}]\cr
Absolute tolerance within which to consider a feature constant.
Set to \code{Inf} to disregard absolute tolerance.
Default is \code{1e-8}.}

\item{ignore.na}{[\code{logical(1)}]\cr
Whether to ignore \code{NA} and \code{NaN} values. If this is
\code{TRUE}, values that are \code{NA} or \code{NaN} will not
be counted as different from any other value. If this is
\code{FALSE}, columns with \code{NA} or \code{NaN} in them will
only count as constant if they are entirely made up of \code{NA},
or entirely made up of \code{NaN}.
Default is \code{FALSE}.}

\item{id}{[\code{character(1)}]\cr
id to use as prefix for the CPO's hyperparameters. this
must be used to avoid name clashes when composing two
CPOs of the same type, or with learners or other CPOS
with hyperparameters with clashing names.}
}
\description{
Drop all columns that are either constant, or close to constant for numerics,
and columns that have only one value for factors or ordered columns.

This function creates a CPO object, which can be applied to
\code{\link{Task}}s, \code{data.frame}s, \code{link{Learner}}s
and other CPO objects using the \code{\link{\%>>\%}} operator.

The parameters of this object can be changed after creation
using the function \code{\link{setHyperPars}}. The other
hyper-parameter manipulating functins, \code{\link{getHyperPars}}
and \code{\link{getParamSet}} similarly work as one expects.

If the \dQuote{id} parameter is given, the hyperparameters
will have this id as aprefix; this will, however, not change
the parameters of the creator function.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{CPO}},
  \code{\link{NULLCPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoApply}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}},
  \code{\link{cpoFixFactors}}, \code{\link{cpoImpute}},
  \code{\link{cpoMultiplex}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{cpoSelect}},
  \code{\link{getCPOAffect}}, \code{\link{getCPOId}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{hasTagInvert}}, \code{\link{inverter<-}},
  \code{\link{inverter}}, \code{\link{makeCPOTargetOp}},
  \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
