% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterAnova}
\alias{cpoFilterAnova}
\title{Filter \dQuote{anova.test} is based on the Analysis of Variance (ANOVA) between
feature and class. The value of the F-statistic is used as a measure of feature
importance.}
\usage{
cpoFilterAnova(perc = NULL, abs = NULL, threshold = NULL, id = NULL,
  export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{anova.test} is based on the Analysis of Variance (ANOVA) between
feature and class. The value of the F-statistic is used as a measure of feature
importance.
}
