% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterCarscore}
\alias{cpoFilterCarscore}
\title{Filter \dQuote{carscore} determines the \dQuote{Correlation-Adjusted (marginal) coRelation
scores} (short CAR scores). The CAR scores for a set of features are defined as the
correlations between the target and the decorrelated features.}
\usage{
cpoFilterCarscore(perc = NULL, abs = NULL, threshold = NULL,
  diagonal = FALSE, id = NULL, export = "export.default",
  affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{carscore} determines the \dQuote{Correlation-Adjusted (marginal) coRelation
scores} (short CAR scores). The CAR scores for a set of features are defined as the
correlations between the target and the decorrelated features.
}
