% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterChiSquared}
\alias{cpoFilterChiSquared}
\title{The chi-square test is a statistical test of independence to determine whether
two variables are independent. Filter \dQuote{chi.squared} applies this
test in the following way. For each feature the chi-square test statistic is
computed checking if there is a dependency between the feature and the target
variable. Low values of the test statistic indicate a poor relationship. High
values, i.e., high dependency identifies a feature as more important.}
\usage{
cpoFilterChiSquared(perc = NULL, abs = NULL, threshold = NULL,
  id = NULL, affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
The chi-square test is a statistical test of independence to determine whether
two variables are independent. Filter \dQuote{chi.squared} applies this
test in the following way. For each feature the chi-square test statistic is
computed checking if there is a dependency between the feature and the target
variable. Low values of the test statistic indicate a poor relationship. High
values, i.e., high dependency identifies a feature as more important.
}
