% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterGainRatio}
\alias{cpoFilterGainRatio}
\title{Filter \dQuote{gain.ratio} uses the entropy-based information gain ratio
between each feature and target individually as an importance measure.}
\usage{
cpoFilterGainRatio(perc = NULL, abs = NULL, threshold = NULL, id = NULL,
  export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{gain.ratio} uses the entropy-based information gain ratio
between each feature and target individually as an importance measure.
}
