% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterKruskal}
\alias{cpoFilterKruskal}
\title{Filter \dQuote{kruskal.test} applies a Kruskal-Wallis rank sum test of the
null hypothesis that the location parameters of the distribution of a feature
are the same in each class and considers the test statistic as an variable
importance measure: if the location parameters do not differ in at least one
case, i.e., the null hypothesis cannot be rejected, there is little evidence
that the corresponding feature is suitable for classification.}
\usage{
cpoFilterKruskal(perc = NULL, abs = NULL, threshold = NULL, id = NULL,
  affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{kruskal.test} applies a Kruskal-Wallis rank sum test of the
null hypothesis that the location parameters of the distribution of a feature
are the same in each class and considers the test statistic as an variable
importance measure: if the location parameters do not differ in at least one
case, i.e., the null hypothesis cannot be rejected, there is little evidence
that the corresponding feature is suitable for classification.
}
