% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterMrmr}
\alias{cpoFilterMrmr}
\title{Minimum redundancy, maximum relevance filter \dQuote{mrmr} computes the
mutual information between the target and each individual feature minus the
average mutual information of previously selected features and this feature
using the \pkg{mRMRe} package.}
\usage{
cpoFilterMrmr(perc = NULL, abs = NULL, threshold = NULL, id = NULL,
  affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Minimum redundancy, maximum relevance filter \dQuote{mrmr} computes the
mutual information between the target and each individual feature minus the
average mutual information of previously selected features and this feature
using the \pkg{mRMRe} package.
}
