% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterOneR}
\alias{cpoFilterOneR}
\title{Filter \dQuote{oneR} makes use of a simple \dQuote{One-Rule} (OneR) learner to
determine feature importance. For this purpose the OneR learner generates one
simple association rule for each feature in the data individually and computes
the total error. The lower the error value the more important the correspoding
feature.}
\usage{
cpoFilterOneR(perc = NULL, abs = NULL, threshold = NULL, id = NULL,
  affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{oneR} makes use of a simple \dQuote{One-Rule} (OneR) learner to
determine feature importance. For this purpose the OneR learner generates one
simple association rule for each feature in the data individually and computes
the total error. The lower the error value the more important the correspoding
feature.
}
