% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterPermutationImportance}
\alias{cpoFilterPermutationImportance}
\title{Filter \dQuote{permutation.importance} computes a loss function between predictions made by a
learner before and after a feature is permuted. Special arguments to the filter function are
\code{imp.learner}, a [\code{\link{Learner}} or \code{character(1)}] which specifies the learner
to use when computing the permutation importance, \code{contrast}, a \code{function} which takes two
numeric vectors and returns one (default is the difference), \code{aggregation}, a \code{function} which
takes a \code{numeric} and returns a \code{numeric(1)} (default is the mean), \code{nmc},
an \code{integer(1)}, and \code{replace}, a \code{logical(1)} which determines whether the feature being
permuted is sampled with or without replacement.}
\usage{
cpoFilterPermutationImportance(perc = NULL, abs = NULL, threshold = NULL,
  imp.learner, measure = NULL, contrast = function (x, y)  x - y,
  aggregation = function (x, ...)  UseMethod("mean"), nmc = 50,
  replace = FALSE, id = NULL, export = "export.default",
  affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{permutation.importance} computes a loss function between predictions made by a
learner before and after a feature is permuted. Special arguments to the filter function are
\code{imp.learner}, a [\code{\link{Learner}} or \code{character(1)}] which specifies the learner
to use when computing the permutation importance, \code{contrast}, a \code{function} which takes two
numeric vectors and returns one (default is the difference), \code{aggregation}, a \code{function} which
takes a \code{numeric} and returns a \code{numeric(1)} (default is the mean), \code{nmc},
an \code{integer(1)}, and \code{replace}, a \code{logical(1)} which determines whether the feature being
permuted is sampled with or without replacement.
}
