% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterRelief}
\alias{cpoFilterRelief}
\title{Filter \dQuote{relief} is based on the feature selection algorithm \dQuote{ReliefF}
by Kononenko et al., which is a generalization of the orignal \dQuote{Relief}
algorithm originally proposed by Kira and Rendell. Feature weights are initialized
with zeros. Then for each instance \code{sample.size} instances are sampled,
\code{neighbours.count} nearest-hit and nearest-miss neighbours are computed
and the weight vector for each feature is updated based on these values.}
\usage{
cpoFilterRelief(perc = NULL, abs = NULL, threshold = NULL, id = NULL,
  export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{relief} is based on the feature selection algorithm \dQuote{ReliefF}
by Kononenko et al., which is a generalization of the orignal \dQuote{Relief}
algorithm originally proposed by Kira and Rendell. Feature weights are initialized
with zeros. Then for each instance \code{sample.size} instances are sampled,
\code{neighbours.count} nearest-hit and nearest-miss neighbours are computed
and the weight vector for each feature is updated based on these values.
}
\references{
Kira, Kenji and Rendell, Larry (1992). The Feature Selection Problem: Traditional
Methods and a New Algorithm. AAAI-92 Proceedings.

Kononenko, Igor et al. Overcoming the myopia of inductive learning algorithms
with RELIEFF (1997), Applied Intelligence, 7(1), p39-55.
}
