% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterRfCImportance}
\alias{cpoFilterRfCImportance}
\title{Permutation importance of random forests fitted in package \pkg{party}.
The implementation follows the principle of mean decrese in accuracy used
by the \pkg{randomForest} package (see description of \dQuote{randomForest.importance})
filter.}
\usage{
cpoFilterRfCImportance(perc = NULL, abs = NULL, threshold = NULL,
  mtry = 5, id = NULL, export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
Permutation importance of random forests fitted in package \pkg{party}.
The implementation follows the principle of mean decrese in accuracy used
by the \pkg{randomForest} package (see description of \dQuote{randomForest.importance})
filter.
}
