% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterRfImportance}
\alias{cpoFilterRfImportance}
\title{Filter \dQuote{randomForest.importance} makes use of the \code{\link[randomForest]{importance}}
from package \pkg{randomForest}. The importance measure to use is selected via
the \code{method} parameter:
\describe{
  \item{oob.accuracy}{Permutation of Out of Bag (OOB) data.}
  \item{node.impurity}{Total decrease in node impurity.}
}}
\usage{
cpoFilterRfImportance(perc = NULL, abs = NULL, threshold = NULL,
  id = NULL, affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{randomForest.importance} makes use of the \code{\link[randomForest]{importance}}
from package \pkg{randomForest}. The importance measure to use is selected via
the \code{method} parameter:
\describe{
  \item{oob.accuracy}{Permutation of Out of Bag (OOB) data.}
  \item{node.impurity}{Total decrease in node impurity.}
}
}
