% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterRfSRCImportance}
\alias{cpoFilterRfSRCImportance}
\title{Filter \dQuote{randomForestSRC.rfsrc} computes the importance of random forests
fitted in package \pkg{randomForestSRC}. The concrete method is selected via
the \code{method} parameter. Possible values are \code{permute} (default), \code{random},
\code{anti}, \code{permute.ensemble}, \code{random.ensemble}, \code{anti.ensemble}.
See the VIMP section in the docs for \code{\link[randomForestSRC]{rfsrc}} for
details.}
\usage{
cpoFilterRfSRCImportance(perc = NULL, abs = NULL, threshold = NULL,
  id = NULL, export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{randomForestSRC.rfsrc} computes the importance of random forests
fitted in package \pkg{randomForestSRC}. The concrete method is selected via
the \code{method} parameter. Possible values are \code{permute} (default), \code{random},
\code{anti}, \code{permute.ensemble}, \code{random.ensemble}, \code{anti.ensemble}.
See the VIMP section in the docs for \code{\link[randomForestSRC]{rfsrc}} for
details.
}
