% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterRfSRCMinDepth}
\alias{cpoFilterRfSRCMinDepth}
\title{Filter \dQuote{randomForestSRC.var.select} uses the minimal depth variable
selection proposed by Ishwaran et al. (2010) (\code{method = "md"}) or a
variable hunting approach (\code{method = "vh"} or \code{method = "vh.vimp"}).
The minimal depth measure is the default.}
\usage{
cpoFilterRfSRCMinDepth(perc = NULL, abs = NULL, threshold = NULL,
  id = NULL, export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{randomForestSRC.var.select} uses the minimal depth variable
selection proposed by Ishwaran et al. (2010) (\code{method = "md"}) or a
variable hunting approach (\code{method = "vh"} or \code{method = "vh.vimp"}).
The minimal depth measure is the default.
}
