% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterSymmetricalUncertainty}
\alias{cpoFilterSymmetricalUncertainty}
\title{Filter \dQuote{symmetrical.uncertainty} uses the entropy-based symmetrical uncertainty
between each feature and target individually as an importance measure.}
\usage{
cpoFilterSymmetricalUncertainty(perc = NULL, abs = NULL, threshold = NULL,
  id = NULL, affect.type = NULL, affect.index = integer(0),
  affect.names = character(0), affect.pattern = NULL,
  affect.invert = FALSE, affect.pattern.ignore.case = FALSE,
  affect.pattern.perl = FALSE, affect.pattern.fixed = FALSE)
}
\description{
Filter \dQuote{symmetrical.uncertainty} uses the entropy-based symmetrical uncertainty
between each feature and target individually as an importance measure.
}
