% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_filterFeatures.R
\name{cpoFilterUnivariate}
\alias{cpoFilterUnivariate}
\title{The \dQuote{univariate.model.score} feature filter resamples an \pkg{mlr}
learner specified via \code{perf.learner} for each feature individually
with randomForest from package \pkg{rpart} being the default learner.
Further parameter are the resamling strategey \code{perf.resampling} and
the performance measure \code{perf.measure}.}
\usage{
cpoFilterUnivariate(perc = NULL, abs = NULL, threshold = NULL,
  perf.learner = NULL, perf.measure = NULL, perf.resampling = NULL,
  id = NULL, export = "export.default", affect.type = NULL,
  affect.index = integer(0), affect.names = character(0),
  affect.pattern = NULL, affect.invert = FALSE,
  affect.pattern.ignore.case = FALSE, affect.pattern.perl = FALSE,
  affect.pattern.fixed = FALSE)
}
\description{
The \dQuote{univariate.model.score} feature filter resamples an \pkg{mlr}
learner specified via \code{perf.learner} for each feature individually
with randomForest from package \pkg{rpart} being the default learner.
Further parameter are the resamling strategey \code{perf.resampling} and
the performance measure \code{perf.measure}.
}
