% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_meta.R
\name{cpoMeta}
\alias{cpoMeta}
\title{Build data-dependent CPOs}
\usage{
cpoMeta(..., .cpo.name = "meta", .par.set = NULL, .par.vals = list(),
  .export = list(), .datasplit = c("target", "most", "all", "no", "task",
  "factor", "onlyfactor", "ordered", "numeric"), .properties = NULL,
  .properties.adding = NULL, .properties.needed = NULL,
  .properties.target = NULL, cpo.build)
}
\description{
The Meta-CPO determines what CPO to apply to a data depending on
a provided function
}
