% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPO_concrete.R
\name{cpoPca}
\alias{cpoPca}
\title{Construct a CPO for PCA preprocessing}
\usage{
cpoPca(center = TRUE, scale = TRUE, id = NULL)
}
\arguments{
\item{center}{[\code{logical(1)}]\cr
Whether to center the data before performing PCA.
Default is \code{TRUE}.}

\item{scale}{[\code{logical(1)}]\cr
Whether to scale the data before performing PCA. The centering / scaling algorithm
of R's dQuote{scale} is used.
Default is \code{TRUE}.}

\item{id}{[\code{character(1)}]\cr
id to use as prefix for the CPO's hyperparameters. this
must be used to avoid name clashes when composing two
CPOs of the same type, or with learners or other CPOS
with hyperparameters with clashing names.}
}
\description{
This function creates a CPO object, which can be applied to
\code{\link{Task}}s, \code{data.frame}s, \code{link{Learner}}s
and other CPO objects using the \code{\link{\%>>\%}} operator.

The parameters of this object can be changed after creation
using the function \code{\link{setHyperPars}}. The other
hyper-parameter manipulating functins, \code{\link{getHyperPars}}
and \code{\link{getParamSet}} similarly work as one expects.

If the \dQuote{id} parameter is given, the hyperparameters
will have this id as aprefix; this will, however, not change
the parameters of the creator function.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{makeCPOFunctional}},
  \code{\link{makeCPOObject}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}
}
