% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDummyFeatures}
\alias{createDummyFeatures}
\title{Generate dummy variables for factor features}
\usage{
createDummyFeatures(obj, target = NULL, intercept = TRUE, exclude = NULL)
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{SupervisedTask}}]\cr Input data.}

\item{target}{[\code{character(1)}]\cr Name of the column specifying the
response. Only used when \code{obj} is a data.frame, otherwise ignored.}

\item{intercept}{[\code{boolean(1)}]\cr \code{TRUE} if for n levels you want
n dummy variables. Otherwise \code{FALSE} gives you n-1 dummy variables.}

\item{exclude}{[\code{character}]\cr Names of the columns to exclude.}
}
\value{
[\code{data.frame} | \code{\link{SupervisedTask}}]. Same type as
  \code{obj}.
}
\description{
Replace all factor features with their dummy variables.
}
\seealso{
\code{\link{model.matrix}}
}

