% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_resample.R
\name{createNewEsParVals}
\alias{createNewEsParVals}
\title{Create new \code{es.par.vals}.}
\usage{
createNewEsParVals(org.es.par.vals, new.es.par.vals)
}
\arguments{
\item{org.es.par.vals}{orginal/old es.par.vals from obj.}

\item{new.es.par.vals}{parameters which should be updated.}
}
\description{
Create new \code{es.par.vals}.
}

