% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_resample.R
\name{createNewParset}
\alias{createNewParset}
\title{Create new \code{parset}.}
\usage{
createNewParset(org.parset, new.parset)
}
\arguments{
\item{org.parset}{orginal/old parset from obj.}

\item{new.parset}{parameters which should be updated.}
}
\description{
Create new \code{parset}.
}

