% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_resample.R
\name{createPreds}
\alias{createPreds}
\title{Create predictions for training or testing set (depends on idx) (with base model or saved one in RDS).}
\usage{
createPreds(fold.i, bls, idx, task, save.on.disc)
}
\arguments{
\item{fold.i}{Current fold number.}

\item{bls}{base.learner to use.}

\item{idx}{idx for subsetting.}

\item{task}{task.}

\item{save.on.disc}{wether model are present in \code{bls} or must be loaded using readRDS.}
}
\description{
Create predictions for training or testing set (depends on idx) (with base model or saved one in RDS).
}

