% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_resample.R
\name{createTask}
\alias{createTask}
\title{Create a classif or regr Task.}
\usage{
createTask(type, data, target, id = deparse(substitute(data)))
}
\arguments{
\item{type}{[\code{character(1)}]\cr Use "classif" for classification and "regr" for regression.}

\item{data}{[\code{data.frame}]\cr  Data to use.}

\item{target}{[\code{character(1)}]\cr Target to use.}

\item{id}{[\code{character(1)}] Task id.}
}
\description{
Create a classif or regr Task.
}

