\name{cv.desc-class}
\alias{cv.desc-class}
\alias{cv.desc}
\title{cv.desc}
\description{Description class for cross-validation.}
\seealso{\code{\link{make.res.desc}}}
\section{Extends}{\code{\linkS4class{resample.desc}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,cv.desc-method]{initialize}}}{\code{signature( = "cv.desc")}: Create description object for cross-validation.}}}
