% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{doTrainPredict}
\alias{doTrainPredict}
\title{Training and prediction in one function (used for parallelMap)}
\usage{
doTrainPredict(bls, task, show.info, id, save.on.disc)
}
\description{
Training and prediction in one function (used for parallelMap)
}

