% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{doTrainResample}
\alias{doTrainResample}
\title{Resampling and prediction in one function (used for parallelMap)}
\usage{
doTrainResample(bls, task, rin, measures, show.info, id, save.on.disc)
}
\description{
Resampling and prediction in one function (used for parallelMap)
}

