% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dropFeatures}
\alias{dropFeatures}
\title{Drop some features of task.}
\usage{
dropFeatures(task, features)
}
\arguments{
\item{task}{[\code{\link{SupervisedTask}}]\cr
The task.}

\item{features}{[\code{character}]\cr
Features to drop.}
}
\value{
[\code{\link{SupervisedTask}}].
}
\description{
Drop some features of task.
}

