% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateResidualVariance.R
\name{estimateResidualVariance}
\alias{estimateResidualVariance}
\title{Estimate the residual variance.}
\usage{
estimateResidualVariance(x, task, data, target)
}
\arguments{
\item{x}{([Learner] or [WrappedModel])\cr
Learner or wrapped model.}

\item{task}{([RegrTask])\cr
Regression task.
If missing, `data` and `target` must be supplied.}

\item{data}{([data.frame])\cr
A data frame containing the features and target variable.
If missing, `task` must be supplied.}

\item{target}{(`character(1)`)\cr
Name of the target variable.
If missing, `task` must be supplied.}
}
\description{
Estimate the residual variance of a regression model on a given task.
If a regression learner is provided instead of a model, the model is
trained (see [train]) first.
}

