% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateLearner.R
\name{evaluateLearner}
\alias{evaluateLearner}
\alias{evaluateParset}
\title{Evaluates expressions within a learner or parameter set according to the task.}
\usage{
evaluateLearner(learner, task)

evaluateParset(par.set, task)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set of (hyper)parameters and their constraints.
Dependent parameters with a \code{requires} field must use \code{quote} and not
\code{expression} to define it.}
}
\value{
[\code{\link{Learner}} | \code{\link[ParamHelpers]{ParamSet}}].
}
\description{
Updates learners and/or parameter sets by evaluating their expressions
based on a specific task. An overview of the possible expressions can be found in the details.
}
\details{
The expressions can be based on any information provided by the task. For convenience,
the most often used keys are available directly
\itemize{
  \item{\code{task}:} the task itself, allowing to access any of its elements
  \item{\code{p}:} the number of features in the task
  \item{\code{n}:} the number of observations in the task
  \item{\code{type}:} the task type, i.e. "classif", "regr", "surv", "cluster", "costcens" or "multilabel"
  \item{\code{k}:} the number of classes of the target variable (only available for classification tasks)
}
However, if one wants to access any other parts of the \code{task}, one can do so. For instance, one could
access the "blocking" via \code{task$task.desc$has.blocking}.
}
\examples{
## (1) evaluation of a learner's hyperparameters
task = makeClassifTask(data = iris, target = "Species")
lrn1 = makeLearner("classif.rpart", minsplit = expression(k * p),
  minbucket = expression(3L + 4L * task$task.desc$has.blocking))
lrn2 = evaluateLearner(learner = lrn1, task = task)

lrn1$par.vals
lrn2$par.vals

## (2) evaluation of a learner's entire parameter set
task = makeClassifTask(data = iris, target = "Species")
lrn1 = makeLearner("classif.randomForest")
lrn2 = evaluateLearner(learner = lrn1, task = task)

## focus on the parameters 'mtry', 'classwt' and 'cutoff'
lrn1$par.set
lrn2$par.set

## (3) evaluation of a parameter set
task = makeClassifTask(data = iris, target = "Species")
ps1 = makeParamSet(
  makeNumericParam("C", lower = expression(k), upper = expression(n), trafo = function(x) 2^x),
  makeDiscreteParam("sigma", values = expression(list(k, p)))
)
ps2 = evaluateParset(par.set = ps1, task = task)
}

