% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{expandPredList}
\alias{expandPredList}
\title{Expand Predictions according to frequency argument}
\usage{
expandPredList(pred.list, freq)
}
\arguments{
\item{pred.list}{[\code{list} of \code{Predictions}]\cr
List of Predictions which should be expanded.}

\item{freq}{[\code{named vector}]\cr
Named vector containing the frequency of the chosen predictions. 
Vector names must be set to the model names.}
}
\description{
Expand Predictions according to frequency argument
}

