% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDAFeaturesMethods.R
\name{extractFDABsignal}
\alias{extractFDABsignal}
\title{Bspline mlq features}
\usage{
extractFDABsignal(bsignal.knots = 10L, bsignal.df = 3)
}
\arguments{
\item{bsignal.knots}{(\code{integer(1)})\cr
The number of knots for bspline.}

\item{bsignal.df}{(\code{numeric(1)})\cr
The effective degree of freedom of penalized bspline.}
}
\value{
(\link{data.frame}).
}
\description{
The function extracts features from functional data based on the Bspline fit.
For more details refer to \code{\link[FDboost:bsignal]{FDboost::bsignal()}}.
}
\seealso{
Other fda_featextractor: 
\code{\link{extractFDADTWKernel}()},
\code{\link{extractFDAFPCA}()},
\code{\link{extractFDAFourier}()},
\code{\link{extractFDAMultiResFeatures}()},
\code{\link{extractFDATsfeatures}()},
\code{\link{extractFDAWavelets}()}
}
\concept{fda_featextractor}
