% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDAFeaturesMethods.R
\name{extractFDAFourier}
\alias{extractFDAFourier}
\title{Fast Fourier transform features.}
\usage{
extractFDAFourier(trafo.coeff = "phase")
}
\arguments{
\item{trafo.coeff}{[\code{character}]\cr
Specifies which transformation of the complex frequency domain
representation should be calculated as a feature representation.
Must be one of \dQuote{amplitude} or \dQuote{phase}.
Default is \dQuote{phase}.}
}
\value{
[\code{data.frame}].
}
\description{
The function extracts features from functional data based on the fast fourier
transform. For more details refer to \code{\link[stats]{fft}}.
}
\seealso{
Other fda_featextractor: \code{\link{extractFDAFPCA}},
  \code{\link{extractFDAMultiResFeatures}},
  \code{\link{extractFDAWavelets}}
}
\concept{fda_featextractor}
