% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDAFeaturesMethods.R
\name{extractFDAWavelets}
\alias{extractFDAWavelets}
\title{Discrete Wavelet transform features.}
\usage{
extractFDAWavelets(filter = "la8", boundary = "periodic")
}
\arguments{
\item{boundary}{\link{character(1)}\cr
Boundary to be used.
\dQuote{periodic} assumes circular time series,
for \dQuote{reflection} the series is extended to twice its length.
Default is \dQuote{periodic}.}

\item{n.levels}{\link{\code{integer(1)}}\cr
Level of decomposition. See \code{\link[wavelets]{dwt}} for details.}
}
\value{
(\link{data.frame}).
}
\description{
The function extracts discrete wavelet transform coefficients from the raw
functional data.
See \link[wavelets:dwt]{wavelets::dwt} for more information.
}
\seealso{
Other fda_featextractor: \code{\link{extractFDABsignal}},
  \code{\link{extractFDADTWKernel}},
  \code{\link{extractFDAFPCA}},
  \code{\link{extractFDAFourier}},
  \code{\link{extractFDAMultiResFeatures}},
  \code{\link{extractFDATsfeatures}}
}
