% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_recombine.R
\name{extractLevel1Task}
\alias{extractLevel1Task}
\title{Extract level 1 data from stacking resample and create a task}
\usage{
extractLevel1Task(obj, as.task = TRUE)
}
\arguments{
\item{obj}{[\code{ResampleResult}]\n Object using \code{StackedLearner} as Learner.}

\item{as.task}{[\code{logical(1)}]\n Specify if task or data.frame should be returned.}
}
\description{
Extract level 1 data from stacking resample and create a task
}

