\name{filter.wrapper-class}
\alias{filter.wrapper-class}
\alias{filter.wrapper}
\title{Wrapper class for learners to filter variables.}
\description{Wrapper class for learners to filter variables. Experimental. Can currently 
only filter to manually selected variables.}
\section{Extends}{\code{\linkS4class{base.wrapper}}}
\section{Methods}{\describe{\item{\code{\link[=train.learner,filter.wrapper,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}}}
