% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecasting generic}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A model to be used for forecasting}

\item{...}{Used to pass package specific parameters}
}
\description{
Forecasts using a fitted model
}
