% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fuelsubset.task}
\alias{fuelsubset.task}
\title{FuelSubset functional data regression task with 2 functional covariate and 1
scalar covariate. Regression #'about the heat value of some fuel based on the
ultraviolet radiation spectrum and infrared ray radiation and one scalar
column called h2o.}
\description{
Contains the task (\code{fuelsubset.task}).
}
\references{
See Brockhaus, S., Scheipl, F., Hothorn, T., & Greven, S. (2015). The functional linear array model. Statistical Modelling, 15(3), 279–300.
}
\keyword{data}
