% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBenchmarkSummary.R
\name{generateBenchmarkSummaryData}
\alias{generateBenchmarkSummaryData}
\title{Generate data for a BenchmarkSummary plot.}
\usage{
generateBenchmarkSummaryData(bmr, measure = NULL, fill = "best",
  order.Lrns = NULL, order.Tsks = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc). \cr
Defaults to first.}

\item{fill}{[\code{character(0)}] \cr
 Fill bars proportional to [\code{fill}] \cr
 Can be \code{best} or \code{worst}.
 \code{"best"} compares all performances to the best performance accross all
 classifiers. Bars are filled proportional to the best performance. \cr
\code{"worst"} compares all performances to the worst performance accross
 all classifiers. Bars are filled proportional to the worst performance. \cr}

\item{order.Lrns}{[\code{character(nLearners)}] or \cr
                 [\code{integer(nLearners)}] \cr
Character vector with \code{learner.ids} in new order , or integer
vector refering to the positions in the new order.}

\item{order.Tsks}{[\code{character(nTasks)}] or \cr
                 [\code{integer(nTasks)}] \cr
Character vector with \code{task.ids} in new order, or an integer
vector refering to the positions in the new order.}
}
\value{
\code{BenchmarkSummaryData}, contains: \cr
$\code{data}: [\code{data.frame}]. \cr
$\code{fill}: selected \code{fill} option.
$\code{measure}: selected \code{measure}.
}
\description{
Generate data used to plot a BenchmarkSummary plot.
}
\examples{
lrns = list(makeLearner("classif.randomForest"), makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc, meas)
generateBenchmarkSummaryData(res, acc)
}
\seealso{
Other generate_plot_data: \code{\link{generateFilterValuesData}};
  \code{\link{generateLearningCurveData}};
  \code{\link{generateROCRCurvesData}};
  \code{\link{generateRankMatrixAsBarData}};
  \code{\link{generateThreshVsPerfData}};
  \code{\link{getFilterValues}}
}

