% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateCalibration.R
\name{generateCalibrationData}
\alias{generateCalibrationData}
\title{Generate classifier calibration data.}
\usage{
generateCalibrationData(obj, breaks = "Sturges", groups = NULL,
  task.id = NULL)
}
\arguments{
\item{obj}{[(list of) \code{\link{Prediction}} | (list of) \code{\link{ResampleResult}} | \code{\link{BenchmarkResult}}]\cr
Single prediction object, list of them, single resample result, list of them, or a benchmark result.
In case of a list probably produced by different learners you want to compare, then
name the list with the names you want to see in the plots, probably
learner shortnames or ids.}

\item{breaks}{[\code{character(1)}] or [\code{numeric}]\cr
If \code{character(1)}, the algorithm to use in generating probability bins.
See \code{\link{hist}} for details.
If \code{numeric}, the cut points for the bins.
Default is \dQuote{Sturges}.}

\item{groups}{[\code{integer(1)}]\cr
The number of bins to construct.
If specified, \code{breaks} is ignored.
Default is \code{NULL}.}

\item{task.id}{[\code{character(1)}]\cr
Selected task in \code{\link{BenchmarkResult}} to do plots for, ignored otherwise.
Default is first task.}
}
\description{
Generate classifier calibration data.
}
\seealso{
Other calibration: \code{\link{plotCalibration}}

Other generate_plot_data: \code{\link{generateBenchmarkSummaryData}};
  \code{\link{generateCritDifferencesData}};
  \code{\link{generateFilterValuesData}};
  \code{\link{generateLearningCurveData}};
  \code{\link{generatePartialPredictionData}};
  \code{\link{generateROCRCurvesData}};
  \code{\link{generateRankMatrixAsBarData}};
  \code{\link{generateThreshVsPerfData}};
  \code{\link{getFilterValues}}
}

