% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRankMatrixAsBar.R
\name{generateRankMatrixAsBarData}
\alias{generateRankMatrixAsBarData}
\title{Generate Data for RankMatrix as Barplot}
\usage{
generateRankMatrixAsBarData(bmr, measure = NULL, pos = "tile",
  order.Lrns = NULL, order.Tsks = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
\link{Measure} for which ranks should be calculated (e.g: acc). \cr
Defaults to first.}

\item{pos}{[\code{character(1)}]
Optionally set how the bars are positioned in [\link{ggplot2}]. \cr
 \code{"tile"} plots a heatmap with \link{task} as the y-axis.\cr
 Allows identification of the performance in a special task.\cr
 \code{"stack"} plots a stacked barplot. \cr
 Allows for comparison of learners within and and accross ranks.\cr
 \code{"dodge"} plots a barplot of with dodged instead of stacked bars.\cr}

\item{order.Lrns}{[\code{character(nLearners)}] or \cr
                 [\code{integer(nLearners)}] \cr
Character vector with \code{learner.ids} in new order , or integer
vector refering to the positions in the new order.}

\item{order.Tsks}{[\code{character(nTasks)}] or \cr
                 [\code{integer(nTasks)}] \cr
Character vector with \code{task.ids} in new order, or an integer
vector refering to the positions in the new order.}
}
\value{
[\code{RankMatrixAsBarData}] \cr
List which contains the following info: \cr
$\code{data}: \code{data.frame} containing the data for plotting\cr
$\code{measure}: \link{Measure} the ranks are calculated on\cr
$\code{pos}: Positioning info for the plot\cr
}
\description{
Generate Data for \link{plotRankMatrixAsBar}.
}
\examples{
lrns = list(makeLearner("classif.randomForest"), makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc, meas)
# Tasks sortet Backwards:
generateRankMatrixAsBarData(res, acc, order.Tsks = c(3L,2L,1L))
}
\seealso{
Other generate_plot_data: \code{\link{generateBenchmarkSummaryData}};
  \code{\link{generateFilterValuesData}};
  \code{\link{generateLearningCurveData}};
  \code{\link{generateROCRCurvesData}};
  \code{\link{generateThreshVsPerfData}};
  \code{\link{getFilterValues}}
}

