% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateValidation.R
\name{generateValidationData}
\alias{generateValidationData}
\title{Generate the Validation data}
\usage{
generateValidationData(TuneResult, include.diagnostics = FALSE)
}
\arguments{
\item{TuneResult}{}

\item{include.diagnostics}{}
}
\value{
what does it return
}
\description{
Generate the Validation data
}

