\name{get.test.set}
\alias{get.test.set,bs632.instance,integer-method}
\alias{get.test.set}
\title{get.test.set,bs632.instance,integer-method}
\alias{get.test.set,bs632plus.instance,integer-method}
\alias{get.test.set}
\alias{get.test.set-methods}
\alias{get.test.set}
\description{Implement this for your own sequential resampling.}
\alias{get.test.set,resample.instance,integer-method}
\alias{get.test.set}
