% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_95CI.R
\name{get_95CI}
\alias{get_95CI}
\title{Get 95 percent confidence interval for multiple linear regression model
parameters.}
\usage{
get_95CI(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A matrix that contains 95 percent C.I for model parameters.
}
\description{
Get 95 percent confidence interval for multiple linear regression model
parameters.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_95CI(X,y)
}
