% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anova.R
\name{get_anova}
\alias{get_anova}
\title{ANOVA table values}
\usage{
get_anova(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A matrix that contains values for anova table. The first column
is the degree of freedom. The second column is the Sum of Squares.
The third column is the Mean Sum of Squares. The fourth column is the F
values. The last column is the p value.
}
\description{
Get the anova table values
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_anova(X,y)
}
