% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta.R
\name{get_beta}
\alias{get_beta}
\title{Get parameter estimates for multiple linear regression model.}
\usage{
get_beta(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A vector that contains parameters estimates.
}
\description{
Get parameter estimates for multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_beta(X,y)
}
