% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef_table.R
\name{get_coef_table}
\alias{get_coef_table}
\title{Summary table values for multiple linear regression}
\usage{
get_coef_table(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A matrix that contains coefficient values for summary table.
The first column is the beta. The second column is the Standard Error for beta.
The third column is the t value for beta. The fourth column is the p value
for beta.
}
\description{
Get the summary table values for coefficient
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_coef_table(X,y)
}
