% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F.R
\name{get_F}
\alias{get_F}
\title{Get F statistics with degrees of freedom
for multiple linear regression model.}
\usage{
get_F(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A vector that contains F statistics with degrees of freedom.
}
\description{
Get F statistics with degrees of freedom
for multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_F(X,y)
}
