% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MSE.R
\name{get_MSE}
\alias{get_MSE}
\title{Get Mean Square Error for multiple linear regression model.}
\usage{
get_MSE(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
The value of mean square error.
}
\description{
Get Mean Square Error for multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_MSE(X,y)
}
