% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalue.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Get p values for parameter estimates in multiple linear regression model.}
\usage{
get_pvalue(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A vector that contains p values for parameters estimates.
}
\description{
Get p values for parameter estimates in multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_pvalue(X,y)
}
