% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Radj.R
\name{get_Radj}
\alias{get_Radj}
\title{Get R^2 adjusted value for multiple linear regression model.}
\usage{
get_Radj(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
The value of R^2 adjusted.
}
\description{
Get R^2 adjusted value for multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_Radj(X,y)
}
