% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Rsquare.R
\name{get_Rsquare}
\alias{get_Rsquare}
\title{Get R^2 for multiple linear regression model.}
\usage{
get_Rsquare(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
The value of R^2
}
\description{
Get R^2 for multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_Rsquare(X,y)
}
