% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Std.R
\name{get_Std}
\alias{get_Std}
\title{Get Standard Errors for parameter estimates for
multiple linear regression model.}
\usage{
get_Std(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A vector contains Standard Errors for parameter estimates.
}
\description{
Get Standard Errors for parameter estimates for
multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_Std(X,y)
}
