% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tvalue.R
\name{get_tvalue}
\alias{get_tvalue}
\title{Get t values for parameter estimates in multiple linear regression model.}
\usage{
get_tvalue(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A vector contains t values for parameter estimates.
}
\description{
Get t values for parameter estimates in multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_tvalue(X,y)
}
