% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yhat.R
\name{get_yhat}
\alias{get_yhat}
\title{Get fitted values for multiple linear regression model.}
\usage{
get_yhat(X, y)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response values.}
}
\value{
A vector contains fitted values.
}
\description{
Get fitted values for multiple linear regression model.
}
\examples{
X = unname(model.matrix(hp~mpg+wt,data = mtcars))
y = mtcars$hp
get_yhat(X,y)
}
