% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAggrMeasures}
\alias{getAggrMeasures}
\title{Extract the aggregated measures of an object.}
\usage{
getAggrMeasures(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Object which contains the aggregated measures.}
}
\value{
[\code{data.frame}].
}
\description{
Extract the aggregated measures of an object.
}

