% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAggrPerformances}
\alias{getAggrPerformances}
\title{Extract the aggregated measures of a benchmark result.}
\usage{
getAggrPerformances(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{data.frame}].
}
\description{
Extract the aggregated measures of a benchmark result.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{FilterResult}},
  \code{\link{getFilterResult}}; \code{\link{benchmark}};
  \code{\link{getFeatSelResult}};
  \code{\link{getPerformances}};
  \code{\link{getPredictions}}; \code{\link{getTuneResult}}
}

