% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBaseLearnerPredictions}
\alias{getBaseLearnerPredictions}
\title{Get predictions of base learners.}
\usage{
getBaseLearnerPredictions(model, newdata = NULL)
}
\arguments{
\item{model}{[\code{WrappedModel}]\cr Wrapped model, result of train.}

\item{newdata}{[\code{data.frame}]\cr New observations which should be predicted.}
}
\description{
Get predictions of base learners.
}
\details{


None.
}

