% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBenchmarkSummary.R
\name{getBenchmarkSummaryData}
\alias{getBenchmarkSummaryData}
\title{Get data for a BenchmarkSummary plot.}
\usage{
getBenchmarkSummaryData(bmr, measure = NULL, fill = "best")
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc). \cr
Defaults to first.}

\item{fill}{[\code{character(0)}] \cr
 Fill bars proportional to [\code{fill}] \cr
 Can be \code{best} or \code{worst}. See \code{Details}.}
}
\value{
\code{data.frame}
}
\description{
Get dataset used to plot a BenchmarkSummary plot.
}
\details{
'best' compares all performances to the best performance accross all \cr
 classifiers. Bars are filled proportional to the best performance. \cr
'worst' compares all performances to the worst performance accross all \cr
 classifiers. Bars are filled proportional to the worst performance. \cr
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
getBenchmarkSummaryData(res,acc)
}

