% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBMRFilteredFeatures}
\alias{getBMRFilteredFeatures}
\title{Extract the feature selection results from a benchmark result.}
\usage{
getBMRFilteredFeatures(bmr, task.ids = NULL, learner.ids = NULL,
  as.df = FALSE)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}

\item{task.ids}{[\code{character(1)}]\cr
Restrict result to certain tasks.
Default is all.}

\item{learner.ids}{[\code{character(1)}]\cr
Restrict result to certain learners.
Default is all.}

\item{as.df}{[\code{character(1)}]\cr
Return one data.frame as result - or a list of lists of objects?.
Default is \code{FALSE}}
}
\value{
[\code{list} | \code{data.frame}]. See above.
}
\description{
Returns a list of lists of \dQuote{} data.frames, as returned by
\code{\link{resample}}, or these objects are rbind-ed with extra columns
\dQuote{task.id} and \dQuote{learner.id}.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRLearnerIds}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}}
}

