% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BenchmarkResult_operators.R
\name{getBMRLearnerIds}
\alias{getBMRLearnerIds}
\title{Return learner ids used in benchmark.}
\usage{
getBMRLearnerIds(bmr)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{list} | \code{character}]. See above.
}
\description{
Return learner ids used in benchmark.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}}
}

