% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BenchmarkResult_operators.R
\name{getBMRLearners}
\alias{getBMRLearners}
\title{Return learners used in benchmark.}
\usage{
getBMRLearners(bmr)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{list}].
}
\description{
Return learners used in benchmark.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRLearnerIds}};
  \code{\link{getBMRMeasureIds}};
  \code{\link{getBMRMeasures}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}}
}

