% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BenchmarkResult_operators.R
\name{getBMRPerformances}
\alias{getBMRPerformances}
\title{Extract the test performance values from a benchmark result.}
\usage{
getBMRPerformances(bmr, task.ids = NULL, learner.ids = NULL,
  as.df = FALSE)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}

\item{task.ids}{[\code{character(1)}]\cr
Restrict result to certain tasks.
Default is all.}

\item{learner.ids}{[\code{character(1)}]\cr
Restrict result to certain learners.
Default is all.}

\item{as.df}{[\code{character(1)}]\cr
Return one data.frame as result - or a list of lists of objects?.
Default is \code{FALSE}}
}
\value{
[\code{list} | \code{data.frame}]. See above.
}
\description{
Either a list of lists of \dQuote{measure.test} data.frames, as returned by
\code{\link{resample}}, or these objects are rbind-ed with extra columns
\dQuote{task.id} and \dQuote{learner.id}.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRLearnerIds}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}}
}

