% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMEnsembleWrapper.R
\name{getBMRPHMEnsembleAccuracy}
\alias{getBMRPHMEnsembleAccuracy}
\title{Compute accuracy table of bmr PHMEnsemble prediction being within given probability bounds
Assumes a Gaussian distribution of predictions}
\usage{
getBMRPHMEnsembleAccuracy(bmr, task, p_range = 0.9, only_last = TRUE)
}
\description{
Compute accuracy table of bmr PHMEnsemble prediction being within given probability bounds
Assumes a Gaussian distribution of predictions
}
