% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMEnsembleWrapper.R
\name{getBMRPHMEnsemblePredictions}
\alias{getBMRPHMEnsemblePredictions}
\title{Helper function to get prediction from each ensemble learner in a benchmark test
WARNING: May take a lot of time, consider passing a small task}
\usage{
getBMRPHMEnsemblePredictions(bmr, learner.ids = getBMRLearnerIds(bmr), task)
}
\description{
Helper function to get prediction from each ensemble learner in a benchmark test
WARNING: May take a lot of time, consider passing a small task
}
