% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult_operators.R
\name{getBMRTaskDescriptions}
\alias{getBMRTaskDescriptions}
\title{Extract all task descriptions from benchmark result (DEPRECATED).}
\usage{
getBMRTaskDescriptions(bmr)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{list}].
}
\description{
A list containing all \code{\link{TaskDesc}}s for each task contained in the benchmark experiment.
}
