% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BenchmarkResult_operators.R
\name{getBMRTaskIds}
\alias{getBMRTaskIds}
\title{Return task ids used in benchmark.}
\usage{
getBMRTaskIds(bmr)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{character}].
}
\description{
Return task ids used in benchmark.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRLearnerIds}};
  \code{\link{getBMRLearners}};
  \code{\link{getBMRMeasureIds}};
  \code{\link{getBMRMeasures}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTuneResults}}
}

