% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCostSensClassifModel}
\alias{getCostSensClassifModel}
\title{Returns the underlying classification model.}
\usage{
getCostSensClassifModel(model, learner.model = TRUE)
}
\arguments{
\item{model}{[\code{\link[mlr]{WrappedModel}}]\cr
Model produced by training a cost-sensitive classification learner.}

\item{learner.model}{[\code{logical(1)}]\cr
Return underlying R model or wrapped
mlr model (\code{\link[mlr]{WrappedModel}}).
Default is \code{FALSE}.}
}
\value{
[\code{list}].
}
\description{
Returns the underlying classification model.
}

