% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getCPOAffect}
\alias{getCPOAffect}
\title{Get the Selection Arguments for affected CPOs}
\usage{
getCPOAffect(cpo, drop.defaults = TRUE)
}
\arguments{
\item{cpo}{[\code{CPO}]\cr
The CPO.}

\item{drop.defaults}{[\code{logical(1)}]\cr
Whether to only return the arguments that deviate from the default.
Default is \code{TRUE}.}
}
\description{
Get the \code{affected.*} arguments from when the CPO was constructed.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{CPO}},
  \code{\link{NULLCPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoApply}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}}, \code{\link{cpoImpute}},
  \code{\link{cpoMultiplex}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{cpoSelect}},
  \code{\link{getCPOId}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{hasTagInvert}}, \code{\link{inverter<-}},
  \code{\link{inverter}}, \code{\link{makeCPOTargetOp}},
  \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
