% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getCPOBound}
\alias{getCPOBound}
\title{determine the bound of a CPO or Retrafo}
\usage{
getCPOBound(cpo)
}
\arguments{
\item{cpo}{[\code{CPO} | \code{CPORetrafo}]\cr
The CPO or Retrafo to inspect.}
}
\description{
Gives the bound of a CPO or Retrafo. These can be
\dQuote{targetbound} for a CPO / Retrafo that
manipulates target columns, \dQuote{databound} for
a CPO / Retrafo that manipulates non-target columns.

For a CPO that affects both, it is a \code{character(2)}
with both values; for one that affects neither, it is
\code{character(0)}.
}
