% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getCPOName}
\alias{getCPOName}
\title{Get the CPO object's Name}
\usage{
getCPOName(cpo)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The CPO object.}
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getHyperPars}},
  \code{\link{makeCPOFunctional}},
  \code{\link{makeCPOObject}}, \code{\link{setHyperPars}}
}
