% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getCPOPredictType}
\alias{getCPOPredictType}
\title{Get the CPO predict.type}
\usage{
getCPOPredictType(cpo)
}
\arguments{
\item{cpo}{[\code{CPO}]\cr
The CPO.}
}
\description{
Get the possible predict.types this CPO is able to handle.
}
