% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getCPOProperties}
\alias{getCPOProperties}
\title{Get the Properties of the given CPO object}
\usage{
getCPOProperties(cpo)
}
\arguments{
\item{cpo}{[\code{CPO}]\cr
The CPO to query.}
}
\description{
The properties of a CPO object determine the kind of data the CPO will be able to handle, and how
it transforms data. Each entry can be one of: \dQuote{numerics}, \dQuote{factors}, \dQuote{ordered},
\dQuote{missings}.

This function returns a list of three values: \dQuote{properties}, \dQuote{properties.adding}, and
\dQuote{properties.needed}.

The \dQuote{properties} determines what data the CPO handles. If a property of a given data set is absent,
the preproc operator will reject the data.

\dQuote{properties.adding} can be one or many of the same values as \dQuote{properties}. These properties
get added to a Learner (or CPO) coming after / behind this CPO. When a CPO imputes missing values, for example,
this is \dQuote{missings}. This is always a subset of \dQuote{properties}.

\dQuote{properties.needed} can be one or many of the same values as \dQuote{properties}. These properties
are required from a Learner (or CPO) coming after / behind this CPO. E.g., when a CPO converts factors to
numerics, this is \dQuote{numerics} (and \dQuote{properties.adding} is \dQuote{factors}).
}
