% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCritDifferences.R
\name{getCritDifferencesData}
\alias{getCritDifferencesData}
\title{Get Data for critical-differences Plot}
\usage{
getCritDifferencesData(bmr, measure)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
 Measure for which ranks should be calculated (e.g: acc).
 Defaults to first.}
}
\value{
[\code{data.frame}] containing the meanRanks of all
learners and infos on positioning of the descriptive lines.
}
\description{
Get a \code{data.frame} that can be used to plot a critical differences
plot. P-values and critical differences can be obtained from
\link{friedmanTestBMR} and \link{posthocNemenyiTestBMR}
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
plotCritDifferences(res,acc,0.05,"classif.svm")
}

