% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{getData}
\alias{getData}
\title{Retrive data from a DataSource.}
\usage{
getData(data, rows = NULL, cols = NULL)
}
\arguments{
\item{data}{[\code{\link{DataSource}}]\cr
\code{\link{DataSource}}.}

\item{rows}{[\code{integer}]\cr
Subset of rows to retrieve.}

\item{cols}{[\code{character}]\cr
Names of features to retrieve.}
}
\value{
[\code{data.frame}].
}
\description{
\code{\link{getData}} takes a \code{\link{DataSource}} and retrieves the data
in a data.frame.
}

