% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrappedModel.R
\name{getFailureModelDump}
\alias{getFailureModelDump}
\title{Return the error dump of FailureModel.}
\usage{
getFailureModelDump(model)
}
\arguments{
\item{model}{([WrappedModel])\cr
The model.}
}
\value{
(`last.dump`).
}
\description{
Returns the error dump that can be used with `debugger()` to evaluate errors.
If [configureMlr] configuration `on.error.dump` is `FALSE`, this returns
`NULL`.
}

