% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFailureModelMsg}
\alias{getFailureModelMsg}
\title{Return error message of FailureModel.}
\usage{
getFailureModelMsg(model)
}
\arguments{
\item{model}{[\code{\link{WrappedModel}}]\cr
The model.}
}
\value{
[\code{character(1)}].
}
\description{
Such a model is created when one sets the corresponding option in \code{\link{configureMlr}}.
If no failure occurred, \code{NA} is returned.

For complex wrappers this getter returns the first error essage encountered in ANY model that failed.
}

