% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFeatSelResult}
\alias{getFeatSelResult}
\title{Returns the selected feature set and optimization path after training.}
\usage{
getFeatSelResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}}]\cr
Trained Model created with \code{\link{makeFeatSelWrapper}}.}
}
\value{
[\code{\link{FeatSelResult}}].
}
\description{
Returns the selected feature set and optimization path after training.
}
\seealso{
Other featsel: \code{\link{FeatSelControl}},
  \code{\link{FeatSelControlExhaustive}},
  \code{\link{FeatSelControlGA}},
  \code{\link{FeatSelControlRandom}},
  \code{\link{FeatSelControlSequential}},
  \code{\link{makeFeatSelControlExhaustive}},
  \code{\link{makeFeatSelControlGA}},
  \code{\link{makeFeatSelControlRandom}},
  \code{\link{makeFeatSelControlSequential}};
  \code{\link{analyzeFeatSelResult}};
  \code{\link{makeFeatSelWrapper}};
  \code{\link{selectFeatures}}
}

