% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFilteredFeatures}
\alias{getFilteredFeatures}
\title{Returns the filtered features.}
\usage{
getFilteredFeatures(model)
}
\arguments{
\item{model}{[\code{\link{WrappedModel}}]\cr
Trained Model created with \code{\link{makeFilterWrapper}}.}
}
\value{
[\code{character}].
}
\description{
Returns the filtered features.
}
\seealso{
Other filter: \code{\link{FilterResult}},
  \code{\link{getFilterResult}};
  \code{\link{filterFeatures}};
  \code{\link{getFilterMethods}};
  \code{\link{getFilterValues}};
  \code{\link{makeFilterWrapper}}
}

