% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFilterMethods}
\alias{getFilterMethods}
\title{Returns all available feature filter methods.}
\usage{
getFilterMethods()
}
\value{
[\code{character(1)}].
}
\description{
Returns all available feature filter methods.
}
\seealso{
Other filter: \code{\link{filterFeatures}};
  \code{\link{getFilterValues}}
}

