% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFilterResult}
\alias{FilterResult}
\alias{getFilterResult}
\title{Returns a filter result}
\usage{
getFilterResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}} | \code{BenchMarkResult}]\cr
Trained Model created with \code{\link{makeFilterWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{FilterResult}} or list of \code{\link{FilterResult}}s].
}
\description{
This is a which adds a class attribute to the result of \code{\link{getFilteredFeatures}}
which is required for \code{\link{benchmark}}.
}

