% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHomogeneousEnsembleModels}
\alias{getHomogeneousEnsembleModels}
\title{Returns the list of fitted models.}
\usage{
getHomogeneousEnsembleModels(model, learner.models = FALSE)
}
\arguments{
\item{model}{[\code{\link[mlr]{WrappedModel}}]\cr
Model produced by training a learner of homogeneous models.}

\item{learner.models}{[\code{logical(1)}]\cr
Return underlying R models or wrapped
mlr models (\code{\link[mlr]{WrappedModel}}).
Default is \code{FALSE}.}
}
\value{
[\code{list}].
}
\description{
Returns the list of fitted models.
}

