% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getLearnerBare}
\alias{getLearnerBare}
\title{Get the learner with the reachable CPOs removed}
\usage{
getLearnerBare(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner to strip.}
}
\description{
Get the bare Learner without the CPOs that were previously added.

It is still possible for the result to be a wrapped learner, e.g. a
TuningWrapper wrapped learner. It is also possible that below the
tuning wrapper, there are more CPOs. These will not be removed.
}
