% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getLearnerCPO}
\alias{getLearnerCPO}
\title{Get the CPO associated with a learner}
\usage{
getLearnerCPO(learner, warn.buried = TRUE)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner to query}

\item{warn.buried}{[\code{logical(1)}]\cr
Whether to warn about CPOs that could not be retrieved.}
}
\description{
Returns the (outermost) chain of CPOs that are part of a learner. This is useful to inspect the
preprocessing done by a learner object.

If there are hidden CPOs (e.g. if \dQuote{learner} has CPOs, but is wrapped by a \code{TuneWrapper}),
this function can not retrieve these CPOs, but it will emit a warning if \dQuote{warn.buried} is \dQuote{TRUE}.
}
