% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMeasures}
\alias{getMeasures}
\title{Extract the measure results of an object}
\usage{
getMeasures(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Object which contains the measure results.}
}
\value{
[\code{data.frame}].
}
\description{
Extract the measure results of an object
}

