% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getMlrOptions}
\alias{getMlrOptions}
\title{Returns a list of mlr's options.}
\usage{
getMlrOptions()
}
\value{
(\link{list}).
}
\description{
Gets the options for mlr.
}
\seealso{
Other configure: 
\code{\link{configureMlr}()}
}
\concept{configure}
