% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsOptPathDf}
\alias{getNestedTuneResultsOptPathDf}
\title{Get the \code{opt.path}s from each tuning step from the outer resampling.}
\usage{
getNestedTuneResultsOptPathDf(r)
}
\arguments{
\item{r}{[\code{\link{ResampleResult}}] \cr
The result of a resampling of a tuneWrapped learner.}
}
\value{
\code{data.frame}
}
\description{
After you have conducted a resampling on a wrapped learner using \code{\link{makeTuneWrapper}} with \code{resample(..., extract = getTuneresult)} this helper gives you a \code{data.frame} with all \code{opt.path}s combined by \code{rbind}.
  An additional column \code{iter} indicates to what resampling iteration the column belongs.
}
\examples{
## see example of makeTuneWrapper
}

