% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsX}
\alias{getNestedTuneResultsX}
\title{Get the tuned hyperparameter settings from a nested tuning.}
\usage{
getNestedTuneResultsX(r)
}
\arguments{
\item{r}{[\code{\link{ResampleResult}}] \cr
The result of a resampling of a tuneWrapped learner.}
}
\value{
\code{data.frame}
  \code{data.frame} with one column for each tuned hyperparameter and one row for each outer resampling iteration.
}
\description{
After you have conducted a resampling on a wrapped learner using \code{\link{makeTuneWrapper}} with \code{resample(..., extract = getTuneresult)} this helper gives you a \code{data.frame} with the hyperparameter settings which were found to be the best.
}
\examples{
## see example of makeTuneWrapper
}

