% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutOfBag.R
\name{getOutOfBag}
\alias{getOutOfBag}
\title{Extracts out of bag data from trained models}
\usage{
getOutOfBag(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}}]\cr
Wrapped model, result of \code{\link{train}}.}
}
\value{
A list with slots \code{response}, containing the out of bag
predictions, and \code{err} with the error rate.
}
\description{
Learners like \code{randomForest} produce out of bag predictions and error
rates. \code{getOutOfBag} extracts this information from trained models.
See \sQuote{Details} for a list of learners for which this is implemented.
}
\details{
The way how the out of bag error is computed depends on the underlying task.
For classification it is the \sQuote{mean misclassification error}
on oob data. For regression and survival tasks
it is the mean of the residuals.

The following learners support out of bag predictions:
\itemize{
\item{randomForest} \cr
{Support for classification and regression.}
\item{randomForestSRC} \cr
{Regression, classification as well as a survival method exists.}
\item{ranger} \cr
{Full support for classification and regression. For survival
only the overall out of bag error is available.}
\item{rFerns} \cr
{Only for classification.}
}
}

