% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamSet.R
\name{getParamSet}
\alias{getParamSet}
\title{Get a description of all possible parameter settings for a learner.}
\value{
[\code{\link[ParamHelpers]{ParamSet}}].
}
\description{
Returns the \code{\link[ParamHelpers]{ParamSet}} from a \code{\link{Learner}}.
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getClassWeightParam}},
  \code{\link{getHyperPars}}, \code{\link{makeLearner}},
  \code{\link{removeHyperPars}},
  \code{\link{setHyperPars}}, \code{\link{setId}},
  \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

