\name{getPerformances}
\alias{getPerformances}
\title{Extract performance measures of bechmark result.}
\usage{
getPerformances(object)
}
\arguments{
  \item{object}{[\code{\link{BenchmarkResult}}]\cr
  Benchmark result.}
}
\value{
[\code{data.frame}].
}
\description{
Extract performance measures of bechmark result.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{FilterResult}},
  \code{\link{getFilterResult}}; \code{\link{benchmark}};
  \code{\link{getAggrPerformances}};
  \code{\link{getFeatSelResult}};
  \code{\link{getPredictions}}; \code{\link{getTuneResult}}
}

