% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_resample.R
\name{getPreciseTaskType}
\alias{getPreciseTaskType}
\title{Get precise type of the task, i.e. distinguish between "classif" (binary target), "multiclassif" and all other types.}
\usage{
getPreciseTaskType(x)
}
\arguments{
\item{x}{\code{Task} or \code{TaskDesc}.}
}
\description{
Get precise type of the task, i.e. distinguish between "classif" (binary target), "multiclassif" and all other types.
}

