% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{getPredictionDataNonMulticoll}
\alias{getPredictionDataNonMulticoll}
\title{Returns response or probabilites from Prediction with the speciality that the 
first feature of a multiclass classification prediction will be removed to 
overcome multicollinearity problems}
\usage{
getPredictionDataNonMulticoll(pred)
}
\arguments{
\item{pred}{Prediction from predict or resample}
}
\description{
Returns response or probabilites from Prediction with the speciality that the 
first feature of a multiclass classification prediction will be removed to 
overcome multicollinearity problems
}

