% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_operators.R
\name{getPredictionDump}
\alias{getPredictionDump}
\title{Return the error dump of a failed Prediction.}
\usage{
getPredictionDump(pred)
}
\arguments{
\item{pred}{([Prediction])\cr
Prediction object.}
}
\value{
(`last.dump`).
}
\description{
Returns the error dump that can be used with `debugger()` to evaluate errors.
If [configureMlr] configuration `on.error.dump` is `FALSE` or if the
prediction did not fail, this returns `NULL`.
}
\seealso{
Other debug: \code{\link{FailureModel}},
  \code{\link{ResampleResult}}, \code{\link{getRRDump}}
}

