% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Prediction_operators.R
\name{getPredictionResponse}
\alias{getPredictionResponse}
\alias{getPredictionTruth}
\title{Get response / truth from prediction object.}
\usage{
getPredictionResponse(pred)

getPredictionTruth(pred)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}
}
\value{
See above.
}
\description{
The following types are returned, depending on task type:
 \tabular{ll}{
   classif     \tab factor\cr
   regr        \tab numeric\cr
   cluster     \tab factor\cr
   surv        \tab numeric\cr
   multilabel  \tab logical matrix, columns named with labels\cr
}
}
\seealso{
Other predict: \code{\link{asROCRPrediction}};
  \code{\link{generateROCRCurvesData}};
  \code{\link{getPredictionProbabilities}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotViperCharts}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}
}

