% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPredictions}
\alias{getPredictions}
\title{Extract the predictions from a benchmark result.}
\usage{
getPredictions(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{data.frame}].
}
\description{
Extract the predictions from a benchmark result.
}
\seealso{
Other benchmark: \code{\link{benchmark}};
  \code{\link{getAggrMeasures}}; \code{\link{getMeasures}}
}

