% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_operators.R
\name{getPredictionTaskDesc}
\alias{getPredictionTaskDesc}
\title{Get summarizing task description from prediction.}
\usage{
getPredictionTaskDesc(pred)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}
}
\value{
ret_taskdesc
}
\description{
See title.
}
\seealso{
Other predict: \code{\link{asROCRPrediction}},
  \code{\link{forecast.WrappedModel}},
  \code{\link{getPredictionProbabilities}},
  \code{\link{getPredictionResponse}},
  \code{\link{plotViperCharts}},
  \code{\link{predict.WrappedModel}},
  \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}
