% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Prediction_operators.R
\name{getProbabilities}
\alias{getProbabilities}
\title{Deprecated, use \code{getPredictionProbabilities} instead.}
\usage{
getProbabilities(pred, cl)
}
\arguments{
\item{pred}{}

\item{cl}{Order of columns is defined by \code{cl}.}
}
\description{
Deprecated, use \code{getPredictionProbabilities} instead.
}

