% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRankMatrixAsBar.R
\name{getRankMatrixAsBarData}
\alias{getRankMatrixAsBarData}
\title{Get Data for RankMatrix as Barplot}
\usage{
getRankMatrixAsBarData(bmr, measure)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
 Measure for which ranks should be calculated (e.g: acc).
 Defaults to first.}
}
\value{
[\code{data.frame}]
}
\description{
Get the \code{data.frame} to plot a RanMatrix as a Barplot.
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
getRankMatrixAsBarData(res,acc)
getRankMatrixAsBarData(res,acc)
}

