% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{getResponse}
\alias{getResponse}
\title{Returns response from Prediction object}
\usage{
getResponse(pred, full.matrix = NULL)
}
\arguments{
\item{pred}{Prediction}

\item{full.matrix}{Wether all n prediction values should be returned or in case of binary classification only one}
}
\description{
Returns response from Prediction object
}

