% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{getRetrafoState}
\alias{getRetrafoState}
\title{Get the internal state of a Retrafo object}
\usage{
getRetrafoState(retrafo.object)
}
\arguments{
\item{retrafo.object}{[\code{CPORetrafo}]\cr
The object to get the state of.}
}
\value{
a list.
}
\description{
A retrafo function always has access to some kind of state
that represents information gotten from the training data,
as well as the parameters it was called with.

The structure of the internal state depends on the CPO backend
used. For Functional CPO, the state is the environment of the
retrafo function, turned into a list. For Objectbased CPO,
the state is a list containing the parameters, as well as the
control object generated by the trafo function.

The object can be slightly modified and used to create a new
CPO retrafo object using \code{\link{makeRetrafoFromState}}.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{hasTagInvert}}, \code{\link{inverter}},
  \code{\link{makeCPOTargetOp}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
