% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult_operators.R
\name{getRRPredictionList}
\alias{getRRPredictionList}
\title{Get list of predictions for train and test set of each single resample iteration.}
\usage{
getRRPredictionList(res, ...)
}
\arguments{
\item{res}{([ResampleResult])\cr
The result of [resample] run with `keep.pred = TRUE`.}

\item{...}{(any)\cr
Further options passed to [makePrediction].}
}
\value{
[list].
}
\description{
This function creates a list with two slots `train` and `test` where
each slot is again a list of [Prediction] objects for each single
resample iteration.
In case that `predict = "train"` was used for the resample description
(see [makeResampleDesc]), the slot `test` will be `NULL`
and in case that `predict = "test"` was used, the slot `train` will be
`NULL`.
}
\seealso{
Other resample: \code{\link{ResamplePrediction}},
  \code{\link{ResampleResult}}, \code{\link{addRRMeasure}},
  \code{\link{getRRPredictions}},
  \code{\link{getRRTaskDescription}},
  \code{\link{getRRTaskDesc}},
  \code{\link{makeResampleDesc}},
  \code{\link{makeResampleInstance}},
  \code{\link{resample}}
}

