% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResampleResult_operators.R
\name{getRRPredictions}
\alias{getRRPredictions}
\title{Get predictions from resample results.}
\usage{
getRRPredictions(res)
}
\arguments{
\item{res}{[\code{ResampleResult}]\cr
The result of \code{\link{resample}} run with \code{keep.pred = TRUE}.}
}
\value{
[\code{ResamplePrediction}]
}
\description{
Get predictions from resample results.
}
\seealso{
Other resample: \code{\link{ResampleDesc}},
  \code{\link{makeResampleDesc}};
  \code{\link{ResampleInstance}},
  \code{\link{makeResampleInstance}};
  \code{\link{ResamplePrediction}};
  \code{\link{ResampleResult}};
  \code{\link{bootstrapB632}},
  \code{\link{bootstrapB632plus}},
  \code{\link{bootstrapOOB}}, \code{\link{crossval}},
  \code{\link{holdout}}, \code{\link{repcv}},
  \code{\link{resample}}, \code{\link{subsample}}
}

