% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_predict.R
\name{getStackedBaseLearnerPredictions}
\alias{getStackedBaseLearnerPredictions}
\title{Returns the predictions for each base learner.}
\usage{
getStackedBaseLearnerPredictions(model, newdata = NULL)
}
\arguments{
\item{model}{[\code{BaseEnsembleModel}]\cr BaseEnsembleModel, result of train.}

\item{newdata}{[\code{data.frame}]\cr
New observations, for which the predictions using the specified base learners should be returned.
Default is \code{NULL} and extracts the base learner predictions that were made during the training.}
}
\description{
Returns the predictions for each base learner based on \code{newdata}. 
  If \code{newdata} is not supported, training prediction will be returned. 
  For \code{ensembleselection} prediction is only applyed to base models which were selected. 
  Predictions are made based on \code{model$learner.model$base.models}. If \code{save.on.disc = TRUE} 
  only a name file is saved which direct to a RDS containing the model.
}

