% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskClassLevels}
\alias{getTaskClassLevels}
\title{Get the class levels for classification and multilabel tasks.}
\usage{
getTaskClassLevels(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character}].
}
\description{
Get the class levels for classification and multilabel tasks.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskDescription}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFormula}}; \code{\link{getTaskId}};
  \code{\link{getTaskNFeats}}; \code{\link{getTaskSize}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}};
  \code{\link{subsetTask}}
}

