% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskCosts}
\alias{getTaskCosts}
\title{Extract costs from task.}
\usage{
getTaskCosts(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{matrix} | \code{NULL}].
}
\description{
Retuns \dQuote{NULL} if the task is not of type \dQuote{costsens}.
}
\seealso{
Other task: \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskNFeats}}; \code{\link{getTaskWeights}}
}

