% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskDescription}
\alias{getTaskDescription}
\title{Get the type of the task.}
\usage{
getTaskDescription(x)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}
}
\description{
See title.
Deprecated, use \code{\link{getTaskDesc}} instead.
}
