% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskFeatures}
\alias{getTaskFeatures}
\title{Get feature column(s) of a task.}
\usage{
getTaskFeatures(task, subset)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{subset}{[\code{logical} | \code{integer}]\cr
Subset of Features to extract.}
}
\value{
[\code{\link{Task}}].

[\code{data.frame}] of features.
}
\description{
Get feature column(s) of a task.
}
\seealso{
Other Task: \code{\link{ClassifTask}},
  \code{\link{ClusterTask}}, \code{\link{CostSensTask}},
  \code{\link{RegrTask}}, \code{\link{SurvTask}},
  \code{\link{Task}}, \code{\link{makeClassifTask}},
  \code{\link{makeClusterTask}},
  \code{\link{makeCostSensTask}},
  \code{\link{makeRegrTask}},
  \code{\link{makeSupervisedTask}},
  \code{\link{makeSurvTask}}, \code{\link{makeTask}},
  \code{\link{makeUnsupervisedTask}};
  \code{\link{copyTask}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTarget}}
}

