% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskFeatureTypes}
\alias{getTaskFeatureTypes}
\title{Get variable types of features.}
\usage{
getTaskFeatureTypes(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{factor}] with levels \dQuote{numeric}, \dQuote{factor} and \dQuote{ordered},
 same length as feature columns.
}
\description{
Get variable types of features.
}

