% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskFormula}
\alias{getTaskFormula}
\alias{getTaskFormulaAsString}
\title{Get formula of a task.}
\usage{
getTaskFormula(task, target = getTaskTargetNames(task),
  env = parent.frame())

getTaskFormulaAsString(task, target = getTaskTargetNames(task))
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{target}{[\code{character(1)}]\cr
Left hand side of formula.
Default is defined by task \code{x}.}

\item{env}{[\code{environment}]\cr
Environment of the formula. Set this to \code{parent.frame()}
for the default behaviour.
Default is \code{NULL} which deletes the environment.}
}
\value{
[\code{formula}].
}
\description{
This is simply \dQuote{<target> ~ .}.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskNFeats}}; \code{\link{getTaskWeights}}
}

