% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskId}
\alias{getTaskId}
\title{Get the id of the task.}
\usage{
getTaskId(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character(1)}].
}
\description{
Get the id of the task.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskDescription}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskNFeats}}; \code{\link{getTaskSize}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}};
  \code{\link{subsetTask}}
}

