\name{getTaskNFeats}
\alias{getTaskNFeats}
\title{Get number of feature in task.}
\usage{
getTaskNFeats(task)
}
\arguments{
  \item{task}{[\code{\link{Task}}]\cr The task.}
}
\value{
[\code{integer(1)}].
}
\description{
Get number of feature in task.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskTargets}}; \code{\link{subsetTask}}
}

