% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskRows}
\alias{getTaskCols}
\alias{getTaskRows}
\title{Get number of rows or columns for a task}
\usage{
getTaskRows(task)

getTaskCols(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[integer(1)].
}
\description{
Get number of rows or columns for a task
}

