% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskTarget}
\alias{getTaskTarget}
\title{Get target column(s) of a task.}
\usage{
getTaskTarget(task, drop = TRUE)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{drop}{[\code{logical}]\cr
Drop unused dimensions?}
}
\value{
[\code{data.frame} | \code{vector}] of target column(s).
}
\description{
Get target column(s) of a task.
}
\seealso{
Other Task: \code{\link{ClassifTask}},
  \code{\link{ClusterTask}}, \code{\link{CostSensTask}},
  \code{\link{RegrTask}}, \code{\link{SurvTask}},
  \code{\link{Task}}, \code{\link{makeClassifTask}},
  \code{\link{makeClusterTask}},
  \code{\link{makeCostSensTask}},
  \code{\link{makeRegrTask}},
  \code{\link{makeSupervisedTask}},
  \code{\link{makeSurvTask}}, \code{\link{makeTask}},
  \code{\link{makeUnsupervisedTask}};
  \code{\link{copyTask}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFeatures}};
  \code{\link{getTaskTargetNames}}
}

