% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskTargetNames}
\alias{getTaskTargetNames}
\title{Get names of target column(s) of a task.}
\usage{
getTaskTargetNames(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character}] of target name(s).
}
\description{
Get names of target column(s) of a task.
}
\seealso{
Other Task: \code{\link{ClassifTask}},
  \code{\link{ClusterTask}}, \code{\link{CostSensTask}},
  \code{\link{RegrTask}}, \code{\link{SurvTask}},
  \code{\link{Task}}, \code{\link{makeClassifTask}},
  \code{\link{makeClusterTask}},
  \code{\link{makeCostSensTask}},
  \code{\link{makeRegrTask}},
  \code{\link{makeSupervisedTask}},
  \code{\link{makeSurvTask}}, \code{\link{makeTask}},
  \code{\link{makeUnsupervisedTask}};
  \code{\link{copyTask}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFeatures}};
  \code{\link{getTaskTarget}}
}

