% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskWeights}
\alias{getTaskWeights}
\title{Extract weights from task.}
\usage{
getTaskWeights(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{numeric} | \code{NULL}].
}
\description{
Retuns \dQuote{NULL} if the task does not have weights.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskNFeats}}
}

