\name{getter,classif.task-method}
\alias{[,classif.task-method}

\title{Getter for classif.task}
\description{Getter.}
\alias{classif.task.getter}
\alias{getter,classif.task-method}
\seealso{\code{\link{getter,learn.task-method}}}
\arguments{\item{x}{classif.task object}
\item{i}{[character]
\describe{
\item{class.levels}{All possible class values.}
\item{class.nr}{Number of different classes.}
}}
}
