\name{getter,learn.task-method}
\alias{[,bench.result-method}

\title{Getter for learn.task}
\description{Getter.}
\alias{learn.task.getter}
\alias{getter,learn.task-method}
\arguments{\item{x}{bench.result object}
\item{i}{[character]
\describe{
\item{perf}{Performance matrix.}
\item{tuned.pars}{Values of tuned paramters. NA if no tuning was done.}
\item{conf.mats}{Confusion matrices - only for classification.}
}}
\item{j}{[integer or character] \cr Select subset of learners.}
}
\alias{[,learn.task-method}
\alias{[}
\alias{learn.task.getter}
\alias{getter,learn.task-method}
