\name{getter,resample.desc-method}
\alias{[,resample.desc-method}

\title{Getter for resample.desc}
\description{Getter.}
\alias{resample.desk.getter}
\alias{getter,resample.desc-method}
\arguments{\item{x}{resample.desc object}
\item{i}{\describe{
\item{instance.class}{S4 class name of the corresponding resample.instance}
\item{name}{Name of this resampling algorithm}
\item{iters}{Number of iterations}
}}
}
