\name{getter,resample.result-method}
\alias{[,resample.result-method}

\title{Getter for resample.result}
\description{Getter.}
\alias{resample.result.getter}
\alias{getter,resample.result-method}
\arguments{\item{x}{resample.result object}
\item{i}{[character]
\describe{
\item{iters}{Number of predictions.}
\item{fitted}{If j is missing all fitted values are returned. Otherwise they are indexed by j.}
}}
\item{j}{[integer] \cr See above, i == "fitted".}
}
