\name{getter,wrapped.model-method}
\alias{[,wrapped.model-method}

\title{Getter for wrapped.model}
\description{Getter.}
\alias{wrapped.model.getter}
\alias{getter,wrapped.model-method}
\arguments{\item{x}{wrapped.model object}
\item{i}{[\code{\link{character}}]
\describe{
\item{task.class}{Learning task class}
\item{learner.class}{Learner class}
\item{learner.name}{Name of the learning method}
\item{vars}{Variables used to build the model.} 
\item{subset}{Indices of used training cases in fit.}
\item{parset}{Selected hyperparameters for model.}
\item{learner.model}{External model from existing R packages like lda, rpart, etc.}
}}
}
