% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSfeatures_fourier.R
\name{getTSFourierFeatures}
\alias{getTSFourierFeatures}
\title{Fast Fourier transform features}
\usage{
getTSFourierFeatures(data, target, include.target = FALSE, fft.coeff = NULL)
}
\arguments{
\item{data}{[\code{data.frame},\code{matrix}]\cr
Time series curve data.}

\item{target}{[\code{character}]\cr
Name of the target variable.}

\item{include.target}{[\code{logical}]\cr
Should the target variable (i.e. the label) be added in the returned
data.frame? Default is \code{FALSE}.}

\item{fft.coeff}{[\code{character}]\cr
Optional, specifies which 'transformation' of the complex frequency domain
representation should be calculated as feature representation. Must be one
of \dQuote{amplitude} or \dQuote{Phase}. Default: \dQuote{amplitude}.}
}
\value{
Returns an \code{data.frame} object containing the fourier
  coefficients.
}
\description{
The function creates time series features based on the fast
  discrete fourier transform.
}

