% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSfeatures_wavelets.R
\name{getTSWaveletFeatures}
\alias{getTSWaveletFeatures}
\title{Discrete Wavelet transform features}
\usage{
getTSWaveletFeatures(data, target, include.target = FALSE, filter = NULL,
  boundary = NULL)
}
\arguments{
\item{data}{[\code{data.frame},\code{matrix}]\cr
Time series curve data.}

\item{target}{[\code{character}]\cr
Name of the target variable.}

\item{include.target}{[\code{logical}]\cr
Should the target variable (i.e. the label) be added in the returned
data.frame? Default is \code{FALSE}.}

\item{filter, boundary}{[\code{character}]\cr
Optional, specifies which filter or boundary should be used. Default:
\code{filter} = \dQuote{la8}, \code{boundary} = \dQuote{periodic}. See
package \link{\code{wavelets}} for more information.}
}
\value{
Returns an \code{data.frame} object containing the wavelet
  coefficients.
}
\description{
The function extracts discrete wavelet transform coefficients
  from the raw time series curve data.
}

