% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gunpoint.task}
\alias{gunpoint.task}
\title{Gunpoint functional data classification task. Classify whether the person
raise up a gun or just the empty hand.}
\description{
Contains the task (\code{gunpoint.task}).
}
\references{
See Ratanamahatana, C. A. & Keogh. E. (2004). Everything you know
  about Dynamic Time Warping is Wrong. Proceedings of SIAM International
  Conference on Data Mining (SDM05), 506-510.
}
\keyword{data}
