% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasFunctionalFeatures.R
\name{hasFunctionalFeatures}
\alias{hasFunctionalFeatures}
\title{Check whether the object conatins functional features.}
\usage{
hasFunctionalFeatures(obj)
}
\arguments{
\item{obj}{(\code{Task} | \code{TaskDesc} | \code{data.frame})\cr
Object to check.}
}
\value{
(\code{logical(1)})
}
\description{
See title.
}
