% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{hasTagInvert}
\alias{hasTagInvert}
\title{Get the Inverter Tag}
\usage{
hasTagInvert(data)
}
\arguments{
\item{data}{[\code{data.frame} | \code{\link{Task}}]\cr
The result of a \code{\link{\%>>\%}} chain applied to a data set.}
}
\description{
Check whether the data is tagged for inverter saving
when it is sent through a \code{\link{\%>>\%}} chain.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoApply}}, \code{\link{cpoDummyEncode}},
  \code{\link{cpoMultiplex}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}}, \code{\link{inverter<-}},
  \code{\link{inverter}}, \code{\link{makeCPOTargetOp}},
  \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
