% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpLearner.R
\name{helpLearnerParam}
\alias{helpLearnerParam}
\title{Get specific help for a learner's parameters.}
\usage{
helpLearnerParam(learner, param = NULL)
}
\arguments{
\item{learner}{(\link{Learner} | \code{character(1)})\cr
The learner.
If you pass a string the learner will be created via \link{makeLearner}.}

\item{param}{(\code{character} | NULL)\cr
Parameter(s) to describe. Defaults to NULL, which prints information on the documentation
status of all parameters.}
}
\description{
Print the description of parameters of a given learner. The description
is automatically extracted from the help pages of the learner, so it may be incomplete.
}
\seealso{
Other learner: 
\code{\link{LearnerProperties}},
\code{\link{getClassWeightParam}()},
\code{\link{getHyperPars}()},
\code{\link{getLearnerId}()},
\code{\link{getLearnerNote}()},
\code{\link{getLearnerPackages}()},
\code{\link{getLearnerParVals}()},
\code{\link{getLearnerParamSet}()},
\code{\link{getLearnerPredictType}()},
\code{\link{getLearnerShortName}()},
\code{\link{getLearnerType}()},
\code{\link{getParamSet}()},
\code{\link{helpLearner}()},
\code{\link{makeLearners}()},
\code{\link{makeLearner}()},
\code{\link{removeHyperPars}()},
\code{\link{setHyperPars}()},
\code{\link{setId}()},
\code{\link{setLearnerId}()},
\code{\link{setPredictThreshold}()},
\code{\link{setPredictType}()}

Other help: 
\code{\link{helpLearner}()}
}
\concept{help}
\concept{learner}
