% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleInstances.R
\name{instantiateResampleInstance}
\alias{instantiateResampleInstance}
\title{instantiateResampleInstance}
\usage{
instantiateResampleInstance(desc, size, task)
}
\arguments{
\item{desc}{(\link{ResampleDesc})\cr
A resample description.}

\item{size}{(\code{integer(1)})\cr
The size of the resample.}

\item{task}{(\link{Task} | \code{NULL})\cr
A task, necessary only for some resampling methods.}
}
\value{
(\link{ResampleInstance})\cr
An instantiated resample instance.
}
\description{
Instantiate a resample instance, i.e. prepare a
\link{ResampleInstance} that specifically refers to a given \link{Task}.

Used internally.
}
\keyword{internal}
