% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{invert}
\alias{invert}
\title{Invert Target Preprocessing}
\usage{
invert(inverter, prediction, predict.type = "response")
}
\arguments{
\item{inverter}{[\code{CPORetrafo}]\cr
The retrafo or inverter to apply}

\item{prediction}{[\code{\link{Prediction}} | \code{matrix} | \code{data.frame}]\cr
The prediction to invert}
}
\value{
A transformed \code{\link{Prediction}} if a prediction was given,
  or a \code{data.frame}. The 'truth' column(s) of the prediction will be dropped.
}
\description{
Invert the transformation, done on the target column(s)
of a data set, after prediction.

Use either a retrafo object, or an inverter retrieved with
\code{\link{inverter}} from a data object that was fed through a retrafo
chain with \code{\link{tagInvert}} set to \code{TRUE}.

If the retrafo object used had no target-bound transformations,
this is mostly a no-op, except that it possibly changes the task description
of the prediction.
}
