% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{inverter<-}
\alias{inverter<-}
\title{Set the prediction inverse function}
\usage{
inverter(data) <- value
}
\arguments{
\item{data}{[\code{data.frame} | \code{\link{Task}}]\cr
Something to be applied to a \code{\link{\%>>\%}} chain.}

\item{value}{[\code{CPORetrafo}]\cr
An inverter chain.}
}
\description{
Sets the retrafo function that can be applied to the prediction.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoApply}}, \code{\link{cpoDummyEncode}},
  \code{\link{cpoMultiplex}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{hasTagInvert}}, \code{\link{inverter}},
  \code{\link{makeCPOTargetOp}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
