% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{iris.mlr}
\alias{iris.mlr}
\title{Exemplary iris classification task}
\description{
Contains the following objects:
\describe{
 \item{iris.task}{Classification task.}
 \item{iris.lrn}{Classification learner using \code{classif.rpart}.}
 \item{iris.train}{Training subset of \code{iris.task}.}
 \item{iris.test}{Test subset of \code{iris.task}.}
}
}
\references{
See \code{\link[datasets]{iris}}.
}
\keyword{data}

