% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iris.oneclass.task}
\alias{iris.oneclass.task}
\title{Iris oneclass task.}
\description{
Contains the task (\code{iris.oneclass.task}).
}
\references{
See \code{\link[datasets]{iris}}.
}
\keyword{data}
