% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isFailureModel}
\alias{isFailureModel}
\title{Is the model a FailureModel?}
\usage{
isFailureModel(model)
}
\arguments{
\item{model}{[\code{\link{WrappedModel}}]\cr
The model.}
}
\value{
[\code{logical(1)}].
}
\description{
Such a model is created when one sets the corresponding option in \code{\link{configureMlr}}.

For complex wrappers this getter returns \code{TRUE} if ANY model contained in it failed.
}

