\name{kknn.classif-class}
\alias{kknn.classif-class}
\alias{kknn.classif}
\title{kknn.classif}
\description{Wrapped learner for k-Nearest Neighbor from package \code{kknn} for classification problems.}
\details{\emph{Common hyperparameters:}
\describe{
\item{\code{k}}{Number of neighbors considered.} 	
\item{\code{distance}}{Parameter of Minkowski distance.}
}}
\seealso{\code{\link[kknn]{kknn}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{{}{}}}
