\name{kknn.regr-class}
\alias{kknn.regr-class}
\alias{kknn.regr}
\title{kknn.regr}
\description{Wrapped learner for k-Nearest Neighbor from package \code{kknn} for regression problems.}
\details{\emph{Common hyperparameters:}
\describe{
\item{\code{k}}{Number of neighbors considered.} 	
\item{\code{distance}}{Parameter of Minkowski distance.}
}}
\seealso{\code{\link[kknn]{kknn}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.regr}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,kknn.regr-method]{initialize}}}{\code{signature( = "kknn.regr")}: Constructor.}}}

