% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnerArgsToControl.R
\name{learnerArgsToControl}
\alias{learnerArgsToControl}
\title{Convert arguments to control structure.}
\usage{
learnerArgsToControl(control, ...)
}
\arguments{
\item{control}{(`function`)\cr
Function that creates control structure.}

\item{...}{(any)\cr
Arguments for control structure function.}
}
\value{
Control structure for learner.
}
\description{
Find all elements in `...` which are not missing and
call `control` on them.
}

