% Generated by roxygen2 (4.0.2): do not edit by hand
\name{learners}
\alias{learners}
\title{List of supported learning algorithms.}
\description{
All supported learners can be found by \code{\link{listLearners}} or as a table
in the tutorial appendix: \url{http://berndbischl.github.io/mlr/tutorial/html/integrated_learners/}.
}

