% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMHIWrapper.R
\name{linearHI}
\alias{linearHI}
\title{HI model functions for PHMHIWrapper}
\usage{
linearHI(type = "linear", start.n = 1, end.n = 1, scaled = TRUE,
  multiplier = NULL)
}
\arguments{
\item{type}{[\code{character(1)}]\cr
Either "linear" or "exponential"}

\item{start.n}{[\code{numeric(1)}]\cr
Number of points at begining to be modelled as HI 1. Default 1}

\item{end.n}{[\code{numeric(1)}]\cr
Number of points at end to be modelled as HI 0. Default 1}

\item{scaled}{[\code{logical(1)}]\cr
Should the features be scaled. Default True}

\item{multiplier}{[\code{numeric(1)}]
Multiplier to be used during prediction. If NULL mean of train RUL will be used. Defaults to NULL}
}
\description{
HI from features on a window at start and end of each seq.id
}
