% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A multiple regression model (RC)}
\arguments{
\item{formula}{Contains dependent and independent variables for linear regression}

\item{data}{A data.frame to conduct linear regression}
}
\description{
Class for the convenient multiple linear regression
}
\section{Methods}{

\describe{
\item{\code{coef()}}{Return the coefficients as a named vector}

\item{\code{plot()}}{Plotting Residuals vs Fitted graph & Scale-Location graph}

\item{\code{pred()}}{Return the predicted values y_hat}

\item{\code{print()}}{Print out the coefficients and coefficient names}

\item{\code{resid()}}{Return the vector of residuals e}

\item{\code{summary()}}{Simplified summary of the linear model}
}}

