% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{listCPO}
\alias{listCPO}
\title{List all built-in CPOs}
\usage{
listCPO()
}
\description{
Return a \code{data.frame} with the columns \dQuote{name},
\dQuote{cponame}, \dQuote{category}, \dQuote{subcategory},
\dQuote{description}.
}
