% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterEnsemble.R
\name{listFilterEnsembleMethods}
\alias{listFilterEnsembleMethods}
\title{List ensemble filter methods.}
\usage{
listFilterEnsembleMethods(desc = TRUE)
}
\arguments{
\item{desc}{(\code{logical(1)})\cr
Provide more detailed information about filters.
Default is \code{TRUE}.}
}
\value{
(\link{data.frame}).
}
\description{
Returns a subset-able dataframe with filter information.
}
\seealso{
Other filter: 
\code{\link{filterFeatures}()},
\code{\link{generateFilterValuesData}()},
\code{\link{getFilteredFeatures}()},
\code{\link{listFilterMethods}()},
\code{\link{makeFilterEnsemble}()},
\code{\link{makeFilterWrapper}()},
\code{\link{makeFilter}()},
\code{\link{plotFilterValues}()}
}
\concept{filter}
