% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_properties.R
\name{listLearnerProperties}
\alias{listLearnerProperties}
\title{List the supported learner properties}
\usage{
listLearnerProperties(type = "any")
}
\arguments{
\item{type}{(\code{character(1)})\cr
Only return properties for a specified task type. Default is \dQuote{any}.}
}
\value{
(\link{character}).
}
\description{
This is useful for determining which learner properties are available.
}
