% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listLearners}
\alias{listLearners}
\alias{listLearnersForTask}
\alias{listMeasuresForTask}
\title{Find matching learning algorithms.}
\usage{
listLearners(type = NA_character_, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

listLearnersForTask(task, properties = character(0L),
  warn.missing.packages = TRUE)

listMeasuresForTask(task, properties = character(0L))
}
\arguments{
\item{type}{[\code{character(1)}]\cr
Type of the learning algorithm, one of \dQuote{classif}, \dQuote{regr}
or \dQuote{surv}. Default is \code{NA}, matching all types.}

\item{properties}{[\code{character)}]\cr
Set of required properties to filter for. Default is \code{character(0)}.}

\item{quiet}{[\code{logical(1)}]\cr
Construct learners quietly to check their properties, shows no package startup messages.
Turn off if you suspect errors.
Default is code{TRUE}.}

\item{warn.missing.packages}{[\code{logical(1)}]\cr
If some learner cannot be constructed because its package is missing,
should a warning be shown?
Default is code{TRUE}.}

\item{task}{[\code{\link{SupervisedTask}}]\cr
The task. Learners are returned that are applicable.}

\item{task}{[\code{\link{SupervisedTask}}]\cr
The task. Learners are returned that are applicable.}
}
\value{
[\code{character}]. Class names of matching learners.
}
\description{
Returns the class names of learning algorithms which have specific characteristics, e.g.
whether they supports missing values, case weights, etc.
}
\details{
Note that the packages of all learners are loaded during the search.
}

