% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_properties.R
\name{listMeasureProperties}
\alias{listMeasureProperties}
\title{List the supported measure properties.}
\usage{
listMeasureProperties()
}
\value{
(\link{character}).
}
\description{
This is useful for determining which measure properties are available.
}
