% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{listTaskTypes}
\alias{listTaskTypes}
\title{List the supported task types in mlr}
\usage{
listTaskTypes()
}
\value{
(\link{character}).
}
\description{
Returns a character vector with each of the supported task types in mlr.
}
