\name{loclda-class}
\alias{loclda-class}
\alias{loclda}
\title{loclda}
\description{Wrapped learner for Localized Linear Discriminant Analysis from package \code{MASS}.}
\seealso{\code{\link[klaR]{loclda}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{{}{}}}
