\name{logreg-class}
\alias{logreg-class}
\alias{logreg}
\title{logreg}
\description{Wrapped learner for Logistic Regression from package \code{stats} for classification problems.}
\details{\emph{Common hyperparameters:}
\describe{
\item{\code{start}}{Starting values for the parameters in the linear predictor.}	
\item{\code{etastart}}{Starting values for the linear predictor.}
\item{\code{mustart}}{Starting values for the vector of means.}
}}
\seealso{\code{\link[stats]{glm}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{}}
