% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{lung.task}
\alias{lung.task}
\title{NCCTG Lung Cancer survival task}
\description{
Contains the task (\code{lung.task}).
}
\references{
See \code{\link[survival]{lung}}.
 Incomplete cases have been removed from the task.
}
\keyword{data}

