% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_measure_AMVhd.R
\name{makeAMVhdMeasure}
\alias{makeAMVhdMeasure}
\title{Creates the measure Area under Mass-Volume Curve (AMV) for Anomaly detection (oneclass) for high dimensional data}
\usage{
makeAMVhdMeasure(id = "AMVhd", minimize = TRUE, alphas = c(0.9, 0.99),
  n.alpha = 50, n.sim = 1e+05, best = 0, worst = NULL, name = id,
  note = "")
}
\arguments{
\item{id}{[\code{character(1)}]\cr
Name of measure.}

\item{minimize}{[\code{logical(1)}]\cr
Should the measure be minimized?
Default is \code{TRUE}.}

\item{alphas}{[\code{numeric}] \cr
Numeric vector of alphas, which lies in [0, 1), representing the computed quantiles.
Default: lower quantile alpha1 = 0.9, upper quantile alpha2 = 0.99 as we are interested in the performance of the scoring function in the the low density regions.}

\item{n.alpha}{[\code{numeric}] \cr
Numeric discretization parameter greater than one, which splits the intervall of alpha1 and alpha2 as follows: {alpha1 + j * (alpha2-alpha1)/(n.alpha-1), j element of {0,...,n.alpha-1}}, Default: n.alpha = 50.}

\item{n.sim}{[\code{numeric(1)}] \cr
Number of Monte-Carlo Samples, Default: 10^4.}

\item{best}{[\code{numeric(1)}]\cr
Best obtainable value for measure.
Default is -\code{Inf} or \code{Inf}, depending on \code{minimize}.}

\item{worst}{[\code{numeric(1)}]\cr
Worst obtainable value for measure.
Default is \code{Inf} or -\code{Inf}, depending on \code{minimize}.}

\item{name}{[\code{character}] \cr
Name of the measure. Default is \code{id}.}

\item{note}{[\code{character}] \cr
Description and additional notes for the measure. Default is \dQuote{}.}
}
\value{
[\code{numeric(1)}]
  Area under Mass-Volume Curve (AMV) for high dimensional data.

[\code{\link{Measure}}].
}
\description{
Creates a measure for oneclass classification on high dimensional data
(recommend for dimension greater than 8) called AMVhd, which is based on the
Area under the Mass-Volume Curve (AMV) (see \code{makeAMVMeasure}).
The basic idea is to do several feature sub-samplings (of dimension less than 8)
to reduce the dimension of the data, therefore AMV can be applied on each
subsamples, yielding partial scores AMV_k. The mean of the partial scores is
the new performancecriteria AMVhd.
}
\references{
Nicolas, G. How to Evaluate the Quality of Unsupervised Anomaly Detection Algorithms,
arXiv preprint arXiv:1607.01152
}
\seealso{
Other performance.: \code{\link{makeAMVMeasure}}
}
