% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMVhdWrapper.R
\name{makeAMVhdWrapper}
\alias{makeAMVhdWrapper}
\title{Preprocessing for using the anomaly detecion performance measure
Area under the Mass Volume Curve for high dimensional data (AMVhd).}
\usage{
makeAMVhdWrapper(learner, amv.iters, amv.feats)
}
\arguments{
\item{learner}{([Learner] | `character(1)`)\cr
The learner.
If you pass (`character(1)`) string the learner will be created via [makeLearner].}

\item{amv.iters}{\link{\code{integer(1)}}\cr
Iterations = number of fitted sub models.
Default is 10.}

\item{amv.feats}{\link{\code{numeric(1)}}\cr
Size of randomly selected features for the sub models.
Default is 3.}
}
\value{
[Learner].
}
\description{
For prediction on data with dimension higher than eight the AMV should not be
used directly as performance measure (see \code{makeAMVMeasure}).
The basic idea is that the wrapper does several feature sub-samplings
(of dimension less than 8) to reduce the dimension of the subsets and applying
the model and the prediction on each subsample.
Afterwards AMV can be applied on each subsamples, yielding partial scores AMV_k.
The mean of the partial scores is the new performance criteria AMVhd (see \code{makeAMVhdMeasure}).

The wrapper is used within the AMVhd measure.

Training is implemented as follows:
For each iteration a data subset with random feature subsample is drawn
(without replacement). On each subset the model is trained with the base learner.
Addtionally, the model on the full data set is also trained. The training with
the wrapper returns a list of all models. The first element of the list contrains
the model on the full data set.

Prediction works as follows:
For every model from the training with the wrapper the prediction is calculated
on the test set with the corresponding feature sample. The prediction object
returns the prediction of the model, which is trained on the full data set and
additionally returns an attribute 'AMVhdSubpredict' which contains the prediction
of all subsamples. In \code{makeAMVhdMeasure} AMV is calculated for each subsample
and aggregated to AMVhd.

Note that the passed base learner must always have \code{predict.type = 'prob'}.
}
\note{
see example of \code{makeAMVhdMeasure}
}
\references{
Nicolas, G. How to Evaluate the Quality of Unsupervised Anomaly Detection Algorithms,
arXiv preprint arXiv:1607.01152
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeExtractFDAFeatsWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
