% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainModel.R
\name{makeChainModel}
\alias{makeChainModel}
\title{Only exported for internal use.}
\usage{
makeChainModel(next.model, cl)
}
\arguments{
\item{next.model}{(\link{WrappedModel})\cr
The next model.}

\item{cl}{(\link{character})\cr
Subclass to assign to the resulting model.}
}
\description{
Only exported for internal use.
}
\keyword{internal}
