% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationViaRegressionWrapper.R
\name{makeClassificationViaRegressionWrapper}
\alias{makeClassificationViaRegressionWrapper}
\title{Classification via regression wrapper.}
\usage{
makeClassificationViaRegressionWrapper(learner, predict.type = "response")
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{predict.type}{[\code{character(1)}]\cr
\dQuote{response} (= labels) or \dQuote{prob} (= probabilities and labels by selecting the ones with maximal probability).}

\item{...}{[any]\cr
Additional parameters passed down to the wrapped learner.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Builds regression models that predict for each class separately whether a particular example belongs to it (1) or not (-1). Predictions probabilities are aggregated by selecting the class with the largest predicted value.

Probabilities are generated by transforming the predictions with a softmax.
}
\examples{
lrn = makeLearner("regr.rpart")
lrn = makeClassificationViaRegressionWrapper(lrn)
mod = train(lrn, iris.task, subset = 1:100)
predictions = predict(mod, newdata = iris[101:150, 1:4])
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
