% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOFunctional.R
\name{makeCPOFunctional}
\alias{makeCPOFunctional}
\title{Create a custom CPO constructor}
\usage{
makeCPOFunctional(.cpo.name, ..., .par.set = NULL, .par.vals = list(),
  cpo.trafo)
}
\arguments{
\item{.cpo.name}{[\code{character(1)}]\cr
The name of the resulting CPO constructor / CPO. This is used for identification in output.}

\item{...}{Parameters of the CPO, in the format of \code{\link{paramSetSugar}}.}

\item{.par.set}{[\code{ParamSet}]\cr
Optional parameter set. If this is not \code{NULL}, the \dQuote{...} parameters are ignored.}

\item{.par.vals}{[\code{list}]\cr
Named list of default parameter values for the CPO. These are used additionally to the
parameter default values in \dQuote{...} and \code{.par.set}. It is preferred to use
these default values, and not \code{.par.vals}.}

\item{cpo.trafo}{[\code{language} | \code{function}]\cr
This can either be a function, just the expressions to perform wrapped in curly braces.
If this is a function, it must have the parameters \dQuote{data} and \dQuote{target},
as well as the parameters specified in \dQuote{...} or \dQuote{.par.set}. (Alternatively,
the function may have a dotdotdot argument). It must return a \dQuote{data.frame} object
with an attribute \dQuote{retrafo}. This attribute must be a function with the argument
\dQuote{data} and return another \dQuote{data.frame}.\cr
If \dQuote{cpo.trafo} is a list of expressions (preferred), it is turned into a function
by mlr, with the above mentioned criteria.}
}
\description{
Create a function-based CPO constructor. This has the advantage of being possibly more
elegant and less boilerplate-heavy than object-based CPOs, but it uses more memory
and is a bit harder to debug. For object-based CPOs, see \code{\link{makeCPOFunctional}}.
}
\examples{
noop = makeCPOFunctional("noop", dummy: logical, cpo.trafo = {
  attr(data, "retrafo") = function(data) data
  data
})

}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{makeCPOObject}}, \code{\link{setHyperPars}}
}
