% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOTarget.R
\name{makeCPOTargetOp}
\alias{makeCPOTargetOp}
\title{Create a custom target-bound CPO constructor}
\usage{
makeCPOTargetOp(.cpo.name, ..., .par.set = NULL, .par.vals = list(),
  .datasplit = c("target", "most", "all", "no", "task", "factor",
  "onlyfactor", "ordered", "numeric"), .properties = character(0),
  .properties.adding = character(0), .properties.needed = character(0),
  .properties.data = c("numerics", "factors", "ordered", "missings"),
  .type = c("cluster", "classif", "multilabel", "regr", "surv"),
  .type.out = .type, cpo.trafo, cpo.retrafo)
}
\description{
Create a CPO constructor.

Most of the parameters are as in \link{makeCPO}.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}
}
