% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{makeDataSource}
<<<<<<< 6873e5499211c3bb3aca201d005b6732cb56fcbb
\alias{DataSource}
\alias{makeDataSource}
\title{Create a Data Source backend for mlr Tasks.}
\usage{
makeDataSource(src, rows = NULL, cols = NULL, ...)
=======
\alias{makeDataSource}
\title{Create a Data Source backend for mlr Tasks.}
\usage{
makeDataSource(src, cols = NULL, ...)
>>>>>>> tests, man pages, ...
}
\arguments{
\item{src}{[\code{\link[base]{data.frame}} | \code{\link[dplyr]{tbl}}]\cr
  Source of the data. Currently supported
  \itemize{
    \item[\code{\link[base]{data.frame}}]{Regular data.frames.}
    \item[\code{\link[dplyr]{tbl}}]{Tables as used in \pkg{dplyr}. Can use a SQL backend.}
  }}

<<<<<<< 6873e5499211c3bb3aca201d005b6732cb56fcbb
\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}

=======
>>>>>>> tests, man pages, ...
\item{cols}{[\code{character} | \code{NULL}]\cr
Subset of columns to use in the task. Additional columns in \code{src} are ignored.}

\item{...}{[\code{any}]\cr
Additional arguments, depending on the type of \code{src}.}

<<<<<<< 6873e5499211c3bb3aca201d005b6732cb56fcbb
\item{key}{[\code{character}]\cr
Required for database backends: Name of the primary key column.}
=======
\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}
>>>>>>> tests, man pages, ...
}
\value{
[\code{DataSource}].
}

