% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDAClassifTask.R
\name{makeFDAClassifTask}
\alias{makeFDAClassifTask}
\alias{FDAClassifTask}
\title{Functional data classification task.}
\usage{
makeFDAClassifTask(id = deparse(substitute(data)), data, target,
  weights = NULL, blocking = NULL, positive = NA_character_,
  fixup.data = "warn", check.data = TRUE, fd.features = NULL,
  fd.grids = NULL)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
Id string for object.
Default is the name of the R variable passed to \code{data}.}

\item{data}{[\code{data.frame}]\cr
A data frame containing the features and target variable(s).}

\item{target}{[\code{character(1)} | \code{character(2)} | \code{character(n.classes)}]\cr
Name(s) of the target variable(s).
For survival analysis these are the names of the survival time and event columns,
so it has length 2. For multilabel classification it contains the names of the logical
columns that encode whether a label is present or not and its length corresponds to the
number of classes.}

\item{weights}{[\code{numeric}]\cr
Optional, non-negative case weight vector to be used during fitting.
Cannot be set for cost-sensitive learning.
Default is \code{NULL} which means no (= equal) weights.}

\item{blocking}{[\code{factor}]\cr
An optional factor of the same length as the number of observations.
Observations with the same blocking level \dQuote{belong together}.
Specifically, they are either put all in the training or the test set
during a resampling iteration.
Default is \code{NULL} which means no blocking.}

\item{positive}{[\code{character(1)}]\cr
Positive class for binary classification (otherwise ignored and set to NA).
Default is the first factor level of the target attribute.}

\item{fixup.data}{[\code{character(1)}]\cr
Should some basic cleaning up of data be performed?
Currently this means removing empty factor levels for the columns.
Possible coices are:
\dQuote{no} = Don't do it.
\dQuote{warn} = Do it but warn about it.
\dQuote{quiet} = Do it but keep silent.
Default is \dQuote{warn}.}

\item{check.data}{[\code{logical(1)}]\cr
Should sanity of data be checked initially at task creation?
You should have good reasons to turn this off (one might be speed).
Default is \code{TRUE}.}

\item{fd.features}{[\code{list}]\cr Optional. Named list of column indices or
column names of the passed data. Each list entry defines one functional covariate through
either a character vector of column names or an integer vector of column
indices. The list names specify the names of the functional covariates. All
columns that are not referenced in \code{fd.features} are scalar variables.
Default is \dQuote{fd1} which means we assume that all columns form one
functional covariate.}

\item{fd.grids}{[\code{list}]\cr Optional. Named list of grids over which the
functional variables are observed. Each grid of observation points must be
provided as a numerical vector. Default is \code{NULL} then all functional
variables are observed on equidistant observation grids 1, 2, …, <number of
observations per function>.}
}
\value{
[\code{\link{FDAClassifTask}}].
}
\description{
Create a functional data classification task. This means that some features 
in the task will be so-called functional covariates / functional features, 
measured on a grid or time scale. Different covariates might come from
different sensors for example.
}
\examples{
dat = data.frame(matrix(rnorm(20), nrow = 2))
dat$target = as.factor(c(0,1))
# X1 to X5 is functional covariate 1 and X6 to X10 functional covariate 2
# grd specifies the time points the curves were sampled at.
grd = list(fd_1 = 1:5, fd_2 = 1:5)
# One row per Observation
tsk = makeFDAClassifTask(data = dat, fd.features = list(fd_1 = 1:5, fd_2 = 6:10),
  target = "target", fd.grid = grd, positive = "1")
}
