% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFixedHoldoutInstances.R
\name{makeFixedHoldoutInstances}
\alias{makeFixedHoldoutInstances}
\title{Make Fixed Holdout Instances for a number of data splits}
\usage{
makeFixedHoldoutInstances(train.inds, test.inds, size = NULL)
}
\arguments{
\item{train.inds}{[list] \cr
List of the training id vectors}

\item{test.inds}{[list] \cr
List of the test id vectors}

\item{size}{[integer] \cr
Size of the data set to resample. 
The function needs to know the largest possible index of the whole data set.}
}
\description{
Make Fixed Holdout Instances for a number of data splits
}
