% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMHIWrapper.R
\name{makeHIFunction}
\alias{makeHIFunction}
\title{HI model functions for PHMHIWrapper}
\usage{
makeHIFunction(type = "rul_scaling", start.n = 1, end.n = 1,
  scaled = TRUE, multiplier = NULL, exp_tol = 0.3, exp_max_iter = 100)
}
\arguments{
\item{type}{[\code{character(1)}]\cr
rul_scaling: Scales RUL to [0, 1] for each unit
rul_limiting: Limits RUL to multiplier and then scales to [0, 1]
linear: Extrapolates linear model to features fitted on ends for each unit
exponential: Fits a[exp(-b*RUL+c)-exp(c)] on result of linear}

\item{start.n}{[\code{numeric(1)}]\cr
Number of points at begining to be modelled as HI 1. Default 1}

\item{end.n}{[\code{numeric(1)}]\cr
Number of points at end to be modelled as HI 0. Default 1}

\item{scaled}{[\code{logical(1)}]\cr
Should the features be scaled. Default True}

\item{multiplier}{[\code{numeric(1)}]
Multiplier to be used during prediction. If NULL mean of train RUL will be used. Defaults to NULL}

\item{exp_tol}{[\code{numeric(1)}]\cr
Warning will be produced if sse error while fitting exponential model is larger than this value}

\item{exp_max_iter}{[\code{numeric(1)}]\cr
Maximum number of iteration for exponential function with random initializations}
}
\description{
HI from features on a window at start and end of each seq.id
}
