% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedStacking_helper.R
\name{makeLearnerFromMMTuneResult}
\alias{makeLearnerFromMMTuneResult}
\title{Creates learner from TuneResult (from ModelMultiplexer) (veraltet)}
\usage{
makeLearnerFromMMTuneResult(tune.res = res, base.learners = base.learners)
}
\description{
Creates learner from TuneResult (from ModelMultiplexer) (veraltet)
}

