% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedStacking_helper.R
\name{makeLearnerFromTuneResult}
\alias{makeLearnerFromTuneResult}
\title{Creates learner from TuneResult (from ModelMultiplexer).}
\usage{
makeLearnerFromTuneResult(tune.res = res)
}
\description{
Creates learner from TuneResult (from ModelMultiplexer).
}

