% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLearners.R
\name{makeLearners}
\alias{Learner}
\alias{makeLearners}
\title{Create learner object.}
\usage{
makeLearners(cls, ids = cls, ...)
}
\arguments{
\item{...}{[any]\cr
Optional named (hyper)parameters.
Alternatively these can be given using the \code{par.vals} argument.}

\item{cl}{[\code{character(1)}]\cr
Class of learner. By convention, all classification learners
start with \dQuote{classif.}, all regression learners with
\dQuote{regr.}, all survival learners start with \dQuote{surv.},
all clustering learners with \dQuote{cluster.}, and all multilabel
classification learners start with \dQuote{multilabel.}.
A list of all integrated learners is available on the
\code{\link{learners}} help page.}

\item{id}{[\code{character(1)}]\cr
Id string for object. Used to display object.
Default is \code{cl}.}
}
\value{
[named list of \code{\link{Learner}}].
}
\description{
For a classification learner the \code{predict.type} can be set
to \dQuote{prob} to predict probabilities and the maximum
value selects the label. The threshold used to assign the label can later be changed using the
\code{\link{setThreshold}} function.

To see all possible properties of a learner, go to: \code{\link{LearnerProperties}}.
}
\examples{
makeLearner("classif.rpart")
makeLearner("classif.lda", predict.type = "prob")
lrn = makeLearner("classif.lda", method = "t", nu = 10)
print(lrn$par.vals)
}
\seealso{
[\code{\link{resample}}], [\code{\link{predict.WrappedModel}}]

Other learner: \code{\link{LearnerProperties}},
  \code{\link{getClassWeightParam}},
  \code{\link{getHyperPars}}, \code{\link{getParamSet}},
  \code{\link{makeLearner}}, \code{\link{removeHyperPars}},
  \code{\link{setHyperPars}}, \code{\link{setId}},
  \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

