% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeModelMultiplexerParamSet}
\alias{makeModelMultiplexerParamSet}
\title{Creates a parameter set for model multiplexer tuning.}
\usage{
makeModelMultiplexerParamSet(multiplexer, ..., .check = TRUE)
}
\arguments{
\item{multiplexer}{[\code{\link{ModelMultiplexer}}]\cr
The muliplexer learner.}

\item{...}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Named param sets. Names must correspond to base learners.
You only need to enter the parameters you want to tune without reference
to the \code{selected.learner} field in any way.}

\item{.check}{[\code{logical}]\cr
Check that for each param in \code{...} one param in found in the base learners.
Default is \code{TRUE}}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}].
}
\description{
Handy way to create the param set with too much typing.

The following is done automatically:
\itemize{
\item{The \code{selected.learner} param is created}
\item{Parameter names are prefixed.}
\item{The \code{requires} field of each param is set.
  This makes all parameters subordinate to \code{selected.learner}}
}
}
\examples{
# See makeModelMultiplexer
}
\seealso{
Other multiplexer, tune: \code{\link{ModelMultiplexer}},
  \code{\link{makeModelMultiplexer}}
}

