% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMARMAWrapper.R
\name{makePHMARMAWrapper}
\alias{makePHMARMAWrapper}
\title{PHM with ARMA features}
\usage{
makePHMARMAWrapper(learner, lag.ar = 0L, lag.ma = 0L, fill.arma = "mean",
  ar.stage = "single-step")
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{lag.ar}{[\code{integer(1)}]\cr
Auto-regressive (target) lag to be incorporated
Default in 0}

\item{lag.ma}{[\code{integer(1)}]\cr
Moving Average (feature) lag to be incorporated
Default is 0}

\item{fill.arma}{[\code{character(1)}]\cr
Value to be filled at begining of lagged columns
\code{"zero"} for 0 and \code{"mean"} for mean value}

\item{ar.stage}{[\code{character(1)}]\cr
The stage of using auto-regressive features
\code{"single-step"} for using auto-regressive \code{target} values
\code{"two-step"} for using auto-regressive features of \code{target} predicted in first step using only \code{features}}
}
\value{
[\code{\link{Learner}}].
}
\description{
Wraps a PHM learner that trains and predicts on ARMA features
}
\examples{
lrn = makeLearner("regr.ksvm")
lrn = makePHMARMAWrapper(lrn)
mod = train(lrn, phm.task)
predictions = predict(mod, newdata = getTaskData(phm.task))
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePHMViaRegressionWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
