% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMEnsembleWrapper.R
\name{makePHMEnsembleWrapper}
\alias{makePHMEnsembleWrapper}
\title{Extend learner as an ensemble trained on subset of seq.ids.}
\usage{
makePHMEnsembleWrapper(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Learner is trained on subsets of sequences for the PHM task to form an ensemble of models
Models can easily be accessed via \code{\link{getLearnerModel}}.

Models are trained either on each seq.id individually or a group of seqs

For prediction the mean value and the standard deviations across predictions is computed.

Note that the passed base learner must always have \code{predict.type = 'response'},
while the PHMEnsembleWrapper can estimate probabilities and standard errors, so it can
be set, e.g., to \code{predict.type = 'prob'}. For this reason, when you call
\code{\link{setPredictType}}, the type is only set for the PHMEnsembleWrapper, not passed
down to the inner learner.

FIXME: Add option to group sequences based on length?
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePHMARMAWrapper}},
  \code{\link{makePHMHIWrapper}},
  \code{\link{makePHMOpNormalizationWrapper}},
  \code{\link{makePHMViaRegressionWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
