% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMHIWrapper.R
\name{makePHMHIWrapper}
\alias{makePHMHIWrapper}
\title{PHM with Health Index Transformation}
\usage{
makePHMHIWrapper(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{hi_function}{[\code{function(task)}]\cr
Function that returns hi: computed HI for task and hi2rulFn: function to compute RUL from HI}
}
\value{
[\code{\link{Learner}}].
}
\description{
Wraps a PHM learner that trains to target a given health index that is scaled back during prediction
}
\examples{
lrn = makePHMViaRegressionWrapper(makeLearner("regr.ksvm"))
lrn = makePHMHIWrapper(lrn)
mod = train(lrn, phm.task)
predictions = predict(mod, newdata = getTaskData(phm.task))
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePHMARMAWrapper}},
  \code{\link{makePHMEnsembleWrapper}},
  \code{\link{makePHMViaRegressionWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
