% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMStageClassifierWrapper.R
\name{makePHMStageClassifierWrapper}
\alias{makePHMStageClassifierWrapper}
\title{PHM with stage classification}
\usage{
makePHMStageClassifierWrapper(learner)
}
\description{
Wraps a PHM learner that initially trains a simple classifier to detect normal/abnormal stage and trains PHM learner only on abnormal data
}
