% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMViaRegressionWrapper.R
\name{makePHMViaRegressionWrapper}
\alias{makePHMViaRegressionWrapper}
\title{PHM via regression wrapper}
\usage{
makePHMViaRegressionWrapper(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Wraps a regression learner to handle PHM tasks by removing the seq.id and order.by columns from the data. 
Prediction will re-assign original seq.id and order.by
}
\examples{
lrn = makeLearner("regr.ksvm")
lrn = makePHMViaRegressionWrapper(lrn)
mod = train(lrn, phm.task)
predictions = predict(mod, newdata = getTaskData(phm.task))
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePHMARMAWrapper}},
  \code{\link{makePHMEnsembleWrapper}},
  \code{\link{makePHMHIWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
