% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocWrapper.R
\name{makePreprocWrapper}
\alias{makePreprocWrapper}
\title{Fuse learner with preprocessing.}
\usage{
makePreprocWrapper(learner, train, predict, par.set = makeParamSet(),
  par.vals = list())
}
\arguments{
\item{learner}{([Learner] | `character(1)`)\cr
The learner.
If you pass a string the learner will be created via [makeLearner].}

\item{train}{(`function(data, target, args)`)\cr
Function to preprocess the data before training.
`target` is a string and denotes the target variable in `data`.
`args` is a list of further arguments and parameters to influence the
preprocessing.
Must return a `list(data, control)`, where `data` is the preprocessed
data and `control` stores all information necessary to do the preprocessing
before predictions.}

\item{predict}{(`function(data, target, args, control)`)\cr
Function to preprocess the data before prediction.
`target` is a string and denotes the target variable in `data`.
`args` are the args that were passed to `train`.
`control` is the object you returned in `train`.
Must return the processed data.}

\item{par.set}{([ParamHelpers::ParamSet])\cr
Parameter set of [ParamHelpers::LearnerParam] objects to describe the
parameters in `args`.
Default is empty set.}

\item{par.vals}{([list])\cr
Named list of default values for params in `args` respectively `par.set`.
Default is empty list.}
}
\value{
([Learner]).
}
\description{
Fuses a base learner with a preprocessing method. Creates a learner object, which can be
used like any other learner object, but which internally preprocesses the data as requested.
If the train or predict function is called on data / a task, the preprocessing is always performed automatically.
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeExtractFDAFeatsWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}

