% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PreprocWrapperCaret.R
\name{makePreprocWrapperCaret}
\alias{makePreprocWrapperCaret}
\title{Fuse learner with preprocessing}
\usage{
makePreprocWrapperCaret(learner, ...)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{...}{[any]\cr
See \code{\link[caret]{preProcess}} for parameters not listed above.
If you use them you might want to define them in the \code{add.par.set} so that they can be tuned.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Fuses a learner with preprocessing methods provided by \code{\link[caret]{preProcess}}.
}
\seealso{
Other wrapper: \code{\link{CostSensClassifModel}},
  \code{\link{CostSensClassifWrapper}},
  \code{\link{makeCostSensClassifWrapper}};
  \code{\link{CostSensRegrModel}},
  \code{\link{CostSensRegrWrapper}},
  \code{\link{makeCostSensRegrWrapper}};
  \code{\link{makeBaggingWrapper}};
  \code{\link{makeDownsampleWrapper}};
  \code{\link{makeFeatSelWrapper}};
  \code{\link{makeFilterWrapper}};
  \code{\link{makeImputeWrapper}};
  \code{\link{makeMulticlassWrapper}};
  \code{\link{makeOverBaggingWrapper}};
  \code{\link{makeOversampleWrapper}},
  \code{\link{makeUndersampleWrapper}};
  \code{\link{makePreprocWrapper}};
  \code{\link{makeSMOTEWrapper}};
  \code{\link{makeTuneWrapper}};
  \code{\link{makeWeightedClassesWrapper}}
}

