% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PreprocWrapperICA.R
\name{makePreprocWrapperICA}
\alias{makePreprocWrapperICA}
\title{Perform a ICA on all numeric columns in the data set.}
\usage{
makePreprocWrapperICA(learner, lrate, ...)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{lrate}{\code{numeric(1)} \cr
learning rate}

\item{...}{\cr
see {\code{?e1071::ica}} for further arguments.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Before training a ICA (Independent Component Analysis) will be
performed on all numeric columns in the trainings dataset. Internally uses
\code{\link{e1071::ica}} before training.
}
\seealso{
Other wrapper: \code{\link{CostSensClassifModel}},
  \code{\link{CostSensClassifWrapper}},
  \code{\link{makeCostSensClassifWrapper}};
  \code{\link{CostSensRegrModel}},
  \code{\link{CostSensRegrWrapper}},
  \code{\link{makeCostSensRegrWrapper}};
  \code{\link{makeBaggingWrapper}};
  \code{\link{makeDownsampleWrapper}};
  \code{\link{makeFeatSelWrapper}};
  \code{\link{makeFilterWrapper}};
  \code{\link{makeImputeWrapper}};
  \code{\link{makeMulticlassWrapper}};
  \code{\link{makeOverBaggingWrapper}};
  \code{\link{makeOversampleWrapper}},
  \code{\link{makeUndersampleWrapper}};
  \code{\link{makePreprocWrapperPCA}};
  \code{\link{makePreprocWrapper}};
  \code{\link{makeSMOTEWrapper}};
  \code{\link{makeTuneWrapper}};
  \code{\link{makeWeightedClassesWrapper}}
}

