% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LambertWFeaturesWrapper.R
\name{makePreprocWrapperLambert}
\alias{makePreprocWrapperLambert}
\title{Gaussianize Numerical Data with a Lambert W Transform}
\usage{
makePreprocWrapperLambert(learner, type = c("h", "hh", "s"),
  methods = c("IGMM", "MLE"), verbose = FALSE)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{type}{[\code{character(1)}] \cr
What type of non-normality: symmetric heavy-tails "h" (default),
 skewed heavy-tails "hh", or just skewed "s".}

\item{methods}{[\code{character(1)}] \cr
 What estimator should be used: "MLE" or "IGMM". "IGMM" gives exactly
Gaussian characteristics (kurtosis = 3 for "h" or skewness = 0 for "s"),
"MLE" comes close to this. Default: "IGMM" since it is much faster than "MLE".}

\item{verbose}{[\code{logical(1)}] \cr
If TRUE, it prints out progress information in the console. Default: FALSE.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Takes numeric data and "Gaussianizes" it using an h, hh, or s Lambert W transform.
}
\seealso{
Other wrapper: \code{\link{makeBaggingWrapper}},
  \code{\link{makeClassificationViaRegressionWrapper}},
  \code{\link{makeConstantClassWrapper}},
  \code{\link{makeCostSensClassifWrapper}},
  \code{\link{makeCostSensRegrWrapper}},
  \code{\link{makeDownsampleWrapper}},
  \code{\link{makeDummyFeaturesWrapper}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeImputeWrapper}},
  \code{\link{makeMulticlassWrapper}},
  \code{\link{makeMultilabelBinaryRelevanceWrapper}},
  \code{\link{makeMultilabelClassifierChainsWrapper}},
  \code{\link{makeMultilabelDBRWrapper}},
  \code{\link{makeMultilabelNestedStackingWrapper}},
  \code{\link{makeMultilabelStackingWrapper}},
  \code{\link{makeOverBaggingWrapper}},
  \code{\link{makePreprocWrapperCaret}},
  \code{\link{makePreprocWrapper}},
  \code{\link{makeRemoveConstantFeaturesWrapper}},
  \code{\link{makeSMOTEWrapper}},
  \code{\link{makeTuneWrapper}},
  \code{\link{makeUndersampleWrapper}},
  \code{\link{makeWeightedClassesWrapper}}
}
