% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{makeRetrafoFromState}
\alias{makeRetrafoFromState}
\title{Set the internal state of a Retrafo object}
\usage{
makeRetrafoFromState(constructor, state)
}
\arguments{
\item{constructor}{A cpo constructor}

\item{state}{A state gotten from another CPO retrafo object using
\code{\link{getRetrafoState}}}
}
\value{
a \code{CPORetrafo}.
}
\description{
This creates a new \code{Retrafo} object which will
behave according to \dQuote{state}.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}}, \code{\link{makeCPO}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}
}
