% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_classif_fdausc.glm.R
\name{makeRLearner.classif.fdausc.glm}
\alias{makeRLearner.classif.fdausc.glm}
\title{Classification of functional data by Generalized Linear Models.}
\usage{
\method{makeRLearner}{classif.fdausc.glm}()
}
\description{
Learner for classification using Generalized Linear Models.
}
