% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_classif_fdausc.kernel.R
\name{makeRLearner.classif.fdausc.kernel}
\alias{makeRLearner.classif.fdausc.kernel}
\title{Learner for kernel classification for functional data.}
\usage{
\method{makeRLearner}{classif.fdausc.kernel}()
}
\description{
Learner for kernel Classification.
}
