% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_classif_fdausc.np.R
\name{makeRLearner.classif.fdausc.np}
\alias{makeRLearner.classif.fdausc.np}
\title{Learner for nonparametric classification for functional data.}
\usage{
\method{makeRLearner}{classif.fdausc.np}()
}
\description{
Learner for Nonparametric Supervised Classification.
}
