% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_regr_fregre.glm.R
\name{makeRLearner.regr.fregre.glm}
\alias{makeRLearner.regr.fregre.glm}
\title{Classification of functional data by Generalized Linear Models.}
\usage{
\method{makeRLearner}{regr.fregre.glm}()
}
\description{
Learner for classification using Generalized Linear Models.
}
