% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeSmoteWrapper}
\alias{makeSmoteWrapper}
\title{Fuse learner with SMOTE oversampling for imbalancy correction in binary classification.}
\usage{
makeSmoteWrapper(learner, sw.rate)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{sw.rate}{[\code{numeric(1)}]\cr
Factor to oversample the smaller class. Must be between 1 and \code{Inf},
where 1 means no oversampling and 2 would mean doubling the class size.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Creates a learner object, which can be
used like any other learner object.
Internally uses \code{\link{smote}} before every model fit.

Note that observation weights do not influence the sampling and are simply passed
down to the next learner.
}

