% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{makeSuperLearnerTask}
\alias{makeSuperLearnerTask}
\title{Create a super learner task}
\usage{
makeSuperLearnerTask(type, data, target)
}
\arguments{
\item{type}{"classif" or "regr"}

\item{data}{data}

\item{target}{target as character}
}
\description{
Create a super learner task
}

