% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifTask.R, R/ClusterTask.R,
%   R/CostSensTask.R, R/MultilabelTask.R, R/RegrTask.R, R/SurvTask.R, R/Task.R
\name{makeClassifTaskDesc}
\alias{makeClassifTaskDesc}
\alias{makeClusterTaskDesc}
\alias{makeCostSensTaskDesc}
\alias{makeMultilabelTaskDesc}
\alias{makeRegrTaskDesc}
\alias{makeSurvTaskDesc}
\alias{makeTaskDesc}
\title{Exported for internal use.}
\usage{
makeClassifTaskDesc(id, data, target, weights, blocking, positive, spatial)

makeClusterTaskDesc(id, data, weights, blocking, spatial)

makeCostSensTaskDesc(id, data, target, blocking, costs, spatial)

makeMultilabelTaskDesc(id, data, target, weights, blocking, spatial)

makeRegrTaskDesc(id, data, target, weights, blocking, spatial)

makeSurvTaskDesc(id, data, target, weights, blocking, spatial)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
task id}

\item{data}{[\code{data.frame}]\cr
data}

\item{target}{[\code{character}]\cr
target columns}

\item{weights}{[\code{numeric}]\cr
weights}

\item{blocking}{[\code{numeric}\cr
task data blocking}

\item{spatial}{[\code{logical(1)}]\cr
whether data is spatial}
}
\description{
Exported for internal use.
}
\keyword{internal}
