% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskDesc.R
\name{makeTaskDescInternal}
\alias{makeTaskDescInternal}
\title{Exported for internal use.}
\usage{
makeTaskDescInternal(type, id, data, target, weights, blocking, spatial)
}
\arguments{
\item{type}{[\code{character(1)}]\cr
Task type.}

\item{id}{[\code{character(1)}]\cr
task id}

\item{data}{[\code{data.frame}]\cr
data}

\item{target}{[\code{character}]\cr
target columns}

\item{weights}{[\code{numeric}]\cr
weights}

\item{blocking}{[\code{numeric}\cr
task data blocking}

\item{spatial}{[\code{logical(1)}]\cr
whether data is spatial}
}
\description{
Exported for internal use.
}
\keyword{internal}
