% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_make_tdAUC.R
\name{makeTimeDependentAUCMeasure}
\alias{makeTimeDependentAUCMeasure}
\title{Creates a measure for time dependent AUC.}
\usage{
makeTimeDependentAUCMeasure(id = "td.AUC", method = "km", time = NA,
  name = "Time dependent AUC for survival", note = "")
}
\arguments{
\item{id}{[\code{character(1)}]\cr
Name of measure.
Default is \dQuote{Time dependent AUC for survival}.}

\item{method}{[\code{character(1)}]\cr
Method that should be used to calculate the time dependent AUC for survival predictions.
Should be one of the following:
\code{"kw"} for using kernel weighting method,
\code{"km"} for using Kaplan Meier method,
\code{"nne"} for using nearest neigbour method,
\code{"ipcw"} for using inverse probability of censoring weighting.
Default is \code{"kw"}.}

\item{time}{[\code{numeric(1)}]\cr
For what time should be the time dependent AUC be calculated?
If the value exceeds the maximum value of the true values, it is reset to this maximum value.
Default is \code{NA}.}

\item{name}{[\code{character}] \cr
Name of the measure. Default is \code{id}.}

\item{note}{[\code{character}] \cr
Description and additional notes for the measure. Default is \dQuote{}.}
}
\description{
Creates one of several possible measures for time dependent AUC in the survival setting.
}
\seealso{
Other performance: \code{\link{ConfusionMatrix}},
  \code{\link{calculateConfusionMatrix}},
  \code{\link{calculateROCMeasures}},
  \code{\link{estimateRelativeOverfitting}},
  \code{\link{makeCostMeasure}},
  \code{\link{makeCustomResampledMeasure}},
  \code{\link{makeMeasure}}, \code{\link{measures}},
  \code{\link{performance}}, \code{\link{setAggregation}},
  \code{\link{setMeasurePars}}
}
