% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSPreprocWrapper.R
\name{makeTSPreprocWrapper}
\alias{makeTSPreprocWrapper}
\title{Extract time series wrapper}
\usage{
makeTSPreprocWrapper(learner)
}
\description{
Extract time series features like wavelets, fourier,etc.
}

