% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneControlCMAES.R
\name{makeTuneControlCMAES}
\alias{TuneControlCMAES}
\alias{makeTuneControlCMAES}
\title{Create control object for hyperparameter tuning with CMAES.}
\usage{
makeTuneControlCMAES(same.resampling.instance = TRUE, impute.val = NULL,
  start = NULL, tune.threshold = FALSE, tune.threshold.args = list(),
  log.fun = "default", final.dw.perc = NULL, budget = NULL, ...)
}
\arguments{
\item{budget}{(`integer(1)`)\cr
Maximum budget for tuning. This value restricts the number of function
evaluations. The `budget` corresponds to the product of the number of generations
(`maxit`) and the number of offsprings per generation
(`lambda`).}
}
\value{
([TuneControlCMAES])
}
\description{
CMA Evolution Strategy with method [cmaes::cma_es].
Can handle numeric(vector) and integer(vector) hyperparameters, but no dependencies.
For integers the internally proposed numeric values are automatically rounded.
The sigma variance parameter is initialized to 1/4 of the span of box-constraints per
parameter dimension.
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlMBO}},
  \code{\link{makeTuneControlRandom}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}},
  \code{\link{tuneThreshold}}
}

