% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneControlGenSA.R
\name{makeTuneControlGenSA}
\alias{TuneControlGenSA}
\alias{makeTuneControlGenSA}
\title{Create control object for hyperparameter tuning with GenSA.}
\usage{
makeTuneControlGenSA(same.resampling.instance = TRUE, impute.val = NULL,
  start = NULL, tune.threshold = FALSE, tune.threshold.args = list(),
  log.fun = "default", final.dw.perc = NULL, budget = NULL, ...)
}
\arguments{
\item{budget}{(`integer(1)`)\cr
Maximum budget for tuning. This value restricts the number of function
evaluations. [GenSA::GenSA] defines the `budget` via
the argument `max.call`. However, one should note that this algorithm
does not stop its local search before its end. This behavior might lead
to an extension of the defined budget and will result in a warning.}
}
\value{
([TuneControlGenSA]).
}
\description{
Generalized simulated annealing with method [GenSA::GenSA].
Can handle numeric(vector) and integer(vector) hyperparameters, but no dependencies.
For integers the internally proposed numeric values are automatically rounded.
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlMBO}},
  \code{\link{makeTuneControlRandom}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}},
  \code{\link{tuneThreshold}}
}

