% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneControlGrid.R
\name{makeTuneControlGrid}
\alias{TuneControlGrid}
\alias{makeTuneControlGrid}
\title{Create control object for hyperparameter tuning with grid search.}
\usage{
makeTuneControlGrid(same.resampling.instance = TRUE, impute.val = NULL,
  resolution = 10L, tune.threshold = FALSE, tune.threshold.args = list(),
  log.fun = "default", final.dw.perc = NULL, budget = NULL)
}
\arguments{
\item{resolution}{([integer])\cr
Resolution of the grid for each numeric/integer parameter in `par.set`.
For vector parameters, it is the resolution per dimension.
Either pass one resolution for all parameters, or a named vector.
See [ParamHelpers::generateGridDesign].
Default is 10.}

\item{budget}{(`integer(1)`)\cr
Maximum budget for tuning. This value restricts the number of function
evaluations. If set, must equal the size of the grid.}
}
\value{
([TuneControlGrid])
}
\description{
A basic grid search can handle all kinds of parameter types.
You can either use their correct param type and `resolution`,
or discretize them yourself by always using [ParamHelpers::makeDiscreteParam]
in the `par.set` passed to [tuneParams].
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlMBO}},
  \code{\link{makeTuneControlRandom}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}},
  \code{\link{tuneThreshold}}
}

