% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneControlIrace.R
\name{makeTuneControlIrace}
\alias{TuneControlIrace}
\alias{makeTuneControlIrace}
\title{Create control object for hyperparameter tuning with Irace.}
\usage{
makeTuneControlIrace(impute.val = NULL, n.instances = 100L,
  show.irace.output = FALSE, tune.threshold = FALSE,
  tune.threshold.args = list(), log.fun = "default", final.dw.perc = NULL,
  budget = NULL, ...)
}
\arguments{
\item{n.instances}{(`integer(1)`)\cr
Number of random resampling instances for irace, see details.
Default is 100.}

\item{show.irace.output}{(`logical(1)`)\cr
Show console output of irace while tuning?
Default is `FALSE`.}

\item{budget}{(`integer(1)`)\cr
Maximum budget for tuning. This value restricts the number of function
evaluations. It is passed to `maxExperiments`.}
}
\value{
([TuneControlIrace])
}
\description{
Tuning with iterated F-Racing with method [irace::irace].
All kinds of parameter types can be handled. We return the best of the final elite
candidates found by irace in the last race. Its estimated performance is the mean of all
evaluations ever done for that candidate. More information on irace can be found in the TR at
<http://iridia.ulb.ac.be/IridiaTrSeries/link/IridiaTr2011-004.pdf>.

For resampling you have to pass a [ResampleDesc],
not a [ResampleInstance].
The resampling strategy is randomly instantiated `n.instances` times and
these are the instances in the sense of irace (`instances` element of `tunerConfig`
in [irace::irace]). Also note that irace will always
store its tuning results in a file on disk, see the package documentation for details on this
and how to change the file path.
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlMBO}},
  \code{\link{makeTuneControlRandom}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}},
  \code{\link{tuneThreshold}}
}

