% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneControlMBO.R
\name{makeTuneControlMBO}
\alias{TuneControlMBO}
\alias{makeTuneControlMBO}
\title{Create control object for hyperparameter tuning with MBO.}
\usage{
makeTuneControlMBO(same.resampling.instance = TRUE, impute.val = NULL,
  learner = NULL, mbo.control = NULL, tune.threshold = FALSE,
  tune.threshold.args = list(), continue = FALSE, log.fun = "default",
  final.dw.perc = NULL, budget = NULL, mbo.design = NULL)
}
\arguments{
\item{learner}{([Learner] | `NULL`)\cr
 The surrogate learner: A regression learner to model performance landscape.
For the default, `NULL`, \pkg{mlrMBO} will automatically create a suitable learner based on the rules described in [mlrMBO::makeMBOLearner].}

\item{mbo.control}{([mlrMBO::MBOControl] | `NULL`)\cr
Control object for model-based optimization tuning.
For the default, `NULL`, the control object will be created with all the defaults as described in [mlrMBO::makeMBOControl].}

\item{continue}{(`logical(1)`)\cr
Resume calculation from previous run using [mlrMBO::mboContinue]?
Requires \dQuote{save.file.path} to be set.
Note that the [ParamHelpers::OptPath] in the [mlrMBO::OptResult]
will only include the evaluations after the continuation.
The complete [OptPath] will be found in the slot `$mbo.result$opt.path`.}

\item{budget}{(`integer(1)`)\cr
Maximum budget for tuning. This value restricts the number of function evaluations.}

\item{mbo.design}{([data.frame] | `NULL`)\cr
Initial design as data frame.
If the parameters have corresponding trafo functions,
the design must not be transformed before it is passed!
For the default, `NULL`, a default design is created like described in [mlrMBO::mbo].}
}
\value{
([TuneControlMBO])
}
\description{
Model-based / Bayesian optimization with the function
[mlrMBO::mbo] from the \pkg{mlrMBO} package.
Please refer to <https://github.com/mlr-org/mlrMBO> for further info.
}
\references{
Bernd Bischl, Jakob Richter, Jakob Bossek, Daniel Horn, Janek Thomas and Michel Lang; mlrMBO: A Modular Framework for Model-Based Optimization of Expensive Black-Box Functions, Preprint: <https://arxiv.org/abs/1703.03373> (2017).
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlRandom}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}},
  \code{\link{tuneThreshold}}
}

