% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuneControlRandom.R
\name{makeTuneControlRandom}
\alias{TuneControlRandom}
\alias{makeTuneControlRandom}
\title{Create control object for hyperparameter tuning with random search.}
\usage{
makeTuneControlRandom(same.resampling.instance = TRUE, maxit = NULL,
  tune.threshold = FALSE, tune.threshold.args = list(),
  log.fun = "default", final.dw.perc = NULL, budget = NULL)
}
\arguments{
\item{maxit}{(`integer(1)` | NULL)\cr
Number of iterations for random search.
Default is 100.}

\item{budget}{(`integer(1)`)\cr
Maximum budget for tuning. This value restricts the number of function
evaluations. The `budget` equals the number of iterations (`maxit`) performed by
the random search algorithm.}
}
\value{
([TuneControlRandom])
}
\description{
Random search. All kinds of parameter types can be handled.
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlMBO}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}},
  \code{\link{tuneThreshold}}
}

