% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedStacking.R
\name{makeXBestLearnersFromMMTuneResult}
\alias{makeXBestLearnersFromMMTuneResult}
\title{Creates x best learners from ModelMultiplexer TuneResult.}
\usage{
makeXBestLearnersFromMMTuneResult(tune.result, model.multiplexer,
  mm.ps = mm.ps, x.best = 5, measure = mmce)
}
\arguments{
\item{tune.result}{[\code{\link{TuneResult}}]\cr
\code{TuneResult} from \code{ModelMultiplexer}}

\item{model.multiplexer}{[\code{\link{ModelMultiplexer}}]\cr
Needed to exract fix parameters.
@param mm.ps [\code{list containing \link{ParamSet}}]\cr
Needed to exract fix parameters.}

\item{x.best}{[\code{integer(1)}]
Number of best learners to extract.}

\item{measure}{[\code{\link{Measure}}]\cr
Measure to apply to "extract from best".}
}
\value{
[\code{list of x best learners}]
}
\description{
Creates x best learners from ModelMultiplexer TuneResult.
}
\examples{
tsk = pid.task
lrns = list(
  makeLearner("classif.gbm", distribution = "bernoulli"),
  makeLearner("classif.randomForest", ntree = 3))
bpt = spt = "prob"
lrns = lapply(lrns, setPredictType, bpt)
mm = makeModelMultiplexer(lrns)
ctrl = makeTuneControlRandom(maxit = 3L)
ps = makeModelMultiplexerParamSet(mm,
  makeIntegerParam("n.trees", lower = 1L, upper = 3L),
  makeIntegerParam("interaction.depth", lower = 1L, upper = 2L),
  makeIntegerParam("ntree", lower = 1L, upper = 3L),
  makeIntegerParam("mtry", lower = 1L, upper = getTaskNFeats(tsk)))
tune.res = tuneParams(learner = mm, task = tsk, resampling = cv2, par.set = ps, control = ctrl)
best = makeXBestLearnersFromMMTuneResult(tune.result = tune.res, model.multiplexer = mm, mm.ps = ps, x.best = 2)
best
}

