\name{mda-class}
\alias{mda-class}
\alias{mda}
\title{mda}
\description{Wrapped learner for Mixture Discriminant Analysis from package \code{mda} for classification problems.}
\details{\emph{Common hyperparameters:}}
\seealso{\code{\link[mda]{mda}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{}}
