% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_properties.R
\name{MeasureProperties}
\alias{MeasureProperties}
\alias{getMeasureProperties}
\alias{hasMeasureProperties}
\title{Query properties of measures.}
\usage{
getMeasureProperties(measure)

hasMeasureProperties(measure, props)
}
\arguments{
\item{measure}{([Measure])\cr
Performance measure.
Default is the first measure used in the benchmark experiment.}

\item{props}{([character])\cr
Vector of properties to query.}
}
\value{
`getMeasureProperties` returns a character vector with measure properties.
 `hasMeasureProperties` returns a logical vector of the same length as `props`.
}
\description{
Properties can be accessed with `getMeasureProperties(measure)`, which returns a
character vector.

The measure properties are defined in [Measure].
}

