% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkResult_mergers.R
\name{mergeBenchmarkResultLearner}
\alias{mergeBenchmarkResultLearner}
\title{Merge different learners of BenchmarkResult objects.}
\usage{
mergeBenchmarkResultLearner(...)
}
\arguments{
\item{...}{[\code{\link{BenchmarkResult}}]\cr 
\code{BenchmarkResult} objects that should be merged.}
}
\description{
Combines the \code{\link{BenchmarkResult}} objects that were performed 
  with different learners on the same set of Task(s). 
  This can be helpful if you, e.g. forgot to run one learner on the set of tasks you used.
}

