% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBenchmarkResults.R
\name{mergeBenchmarkResults}
\alias{mergeBenchmarkResults}
\title{Merge different BenchmarkResult objects.}
\usage{
mergeBenchmarkResults(...)
}
\arguments{
\item{...}{[\code{\link{BenchmarkResult}}]\cr 
\code{BenchmarkResult} objects that should be merged.}
}
\value{
\code{\link{BenchmarkResult}}
}
\description{
Combines the \code{\link{BenchmarkResult}} objects that were performed 
  with different learners and tasks. 
  This can be helpful if you, e.g. forgot to run one learner on the set of tasks you used.
}
\details{
Note that if you want to merge several \code{\link{BenchmarkResult}}
  objects you must ensure that all possible learner and task combinations will be
  contained in the return object.\cr
  Furthermore all given objects must have been calculated on the same
  set of measures.
}

