% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkResult_mergers.R
\name{mergeBenchmarkResultTask}
\alias{mergeBenchmarkResultTask}
\title{Merge different tasks of BenchmarkResult objects.}
\usage{
mergeBenchmarkResultTask(...)
}
\arguments{
\item{...}{[\code{\link{BenchmarkResult}}]\cr 
\code{BenchmarkResult} objects that should be merged.}
}
\description{
Combines the \code{\link{BenchmarkResult}} objects that were performed
  on different tasks with the same set of learner(s). 
  This can be helpful if you, e.g. forgot to run the set of learners on a new task
}

