% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeFactorLevelsBySize}
\alias{mergeFactorLevelsBySize}
\title{Merges small levels of a factor into a new level.}
\usage{
mergeFactorLevelsBySize(x, min.perc = 0.01, new.name = "merged")
}
\arguments{
\item{x}{[\code{factor} | \code{character}]\cr
A character will be treated as a factor.}

\item{min.perc}{[\code{numeric(1)}]\cr
The smallest levels of the factor are merged until their combined proportion
w.r.t. the length of the factor exceeds \code{min.perc}.
Must be between 0 and 1. Default is 0.01.}

\item{new.name}{[\code{character(1)}]\cr
New name of merged level. Default is 'merged'.}
}
\value{
\code{factor}, where merged levels are comined into a new level of name \code{new.name}.
}
\description{
Merges smallest levels of a factor into a new level until it exceeds
a certain size.
}

