\name{mlr-package}
\alias{mlr-package}
\alias{mlr}
\docType{package}
\title{
\packageTitle{mlr}
}
\description{
\packageDescription{mlr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlr}
\packageIndices{mlr}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{mlr}

Maintainer: \packageMaintainer{mlr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
