% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mlr.agri}
\alias{agri.rin}
\alias{agri.task}
\alias{mlr.agri}
\title{European Union Agricultural Workforces clustering task}
\description{
Contains the task (\code{agri.task}) and a matching resample instance
(\code{agri.rin}).
}
\references{
See \code{\link[cluster]{agriculture}}.
}
\keyword{data}

