% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.bc}
\alias{bc.rin}
\alias{bc.task}
\alias{mlr.bc}
\title{Wisconsin Breast Cancer classification task}
\description{
Contains the task (\code{bc.task}) and a matching resample instance
(\code{bc.rin}).
}
\references{
See \code{\link[mlbench]{BreastCancer}}.
  The column \code{"Id"} and all incomplete cases have been removed from the task.
}
\keyword{data}

