% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.bh}
\alias{bh.lrn}
\alias{bh.task}
\alias{bh.test}
\alias{bh.train}
\alias{mlr.bh}
\title{BostonHousing regression task}
\description{
Contains the following objects:
\describe{
 \item{bh.task}{Regression task.}
 \item{bh.lrn}{Classification learner using \code{regr.lm}.}
 \item{bh.train}{Training subset of \code{bh.task}.}
 \item{bh.test}{Test subset of \code{bh.task}.}
}
}
\references{
See \code{\link[mlbench]{BostonHousing}}.
}
\keyword{data}

