% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.costIris}
\alias{costIris.rin}
\alias{costIris.task}
\alias{mlr.costIris}
\title{Iris cost-sensitive classification task}
\description{
Contains the task (\code{costIris.task}) and a resample instance
(\code{costIris.rin}).
}
\references{
See \code{\link[datasets]{iris}}.
  The cost matrix was generated artificially following

  Tu, H.-H. and Lin, H.-T. (2010), One-sided support vector regression for multiclass cost-sensitive classification.
  In ICML, J. Fürnkranz and T. Joachims, Eds., Omnipress, 1095--1102.
}
\keyword{data}

