% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.lung}
\alias{lung.rin}
\alias{lung.task}
\alias{mlr.lung}
\title{NCCTG Lung Cancer survival task}
\description{
Contains the task (\code{lung.task}) and a matching resample instance
(\code{lung.rin}).
}
\references{
See \code{\link[survival]{lung}}.
 Incomplete cases have been removed from the task.
}
\keyword{data}

