% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.mtcars}
\alias{mlr.mtcars}
\alias{mtcars.rin}
\alias{mtcars.task}
\title{Motor Trend Car Road Tests clustering task}
\description{
Contains the task (\code{mtcars.task}) and a matching resample instance
(\code{mtcars.rin}).
}
\references{
See \code{\link[datasets]{mtcars}}.
}
\keyword{data}

