% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{mlr}
\usage{
mlr(outcome, covs, data)
}
\arguments{
\item{outcome}{the name of the response variable in the model}

\item{covs}{a vector of covariate name(s) you'd like to involve in the model}

\item{data}{an object of class "data.frame" such as a dataset}
}
\value{
A list containing model details, y.fitted values, y.res, summary(coefficients), FP(F statistic and P value),
vcov.matrix(variance-covariance matrix), ANOVA.table, R_Square

If there exists collinearity between covariates, a warning message will be given instead of above results.
}
\description{
Conduct basic analyses with multiple linear regression
}
\examples{
data = mtcars
mlr("mpg", c("wt","hp","vs"), data = mtcars)

}
