% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.sonar}
\alias{mlr.sonar}
\alias{sonar.lrn}
\alias{sonar.task}
\alias{sonar.test}
\alias{sonar.train}
\title{Sonar classification task}
\description{
List with following named elements:
\describe{
 \item{sonar.task}{Classification task.}
 \item{sonar.lrn}{Classification learner using \code{classif.rpart}.}
 \item{sonar.train}{Training subset of \code{sonar.task}.}
 \item{sonar.test}{Test subset of \code{sonar.task}.}
}
}
\references{
See \code{\link[mlbench]{Sonar}}.
}
\keyword{data}

