% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.wpbc}
\alias{mlr.wpbc}
\alias{wpbc.lrn}
\alias{wpbc.task}
\alias{wpbc.test}
\alias{wpbc.train}
\title{Wisonsin Prognostic Breast Cancer (WPBC) survival task}
\description{
Contains the following objects:
\describe{
 \item{wpbc.task}{Regression task.}
 \item{wpbc.lrn}{Classification learner using \code{regr.lm}.}
 \item{wpbc.train}{Training subset of \code{wpbc.task}.}
 \item{wpbc.test}{Test subset of \code{wpbc.task}.}
}
}
\references{
See \code{\link[mboost]{wpbc}}. Four incomplete cases have been
  removed from the task.
}
\keyword{data}

