\name{multiclass.wrapper-class}
\alias{multiclass.wrapper-class}
\alias{multiclass.wrapper}
\title{Wrapper class for learners to handle multi-class problems.}
\description{Wrapper class for learners to handle multi-class problems.}
\section{Extends}{\code{\linkS4class{base.wrapper}}}
\section{Methods}{\describe{\item{\code{\link[=[,multiclass.wrapper-method]{[}}}{}\item{\code{\link[=train.learner,multiclass.wrapper,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=pred.learner,multiclass.wrapper,wrapped.model,data.frame,character-method]{pred.learner}}}{}}}
\alias{[,multiclass.wrapper-method}

