\name{naiveBayes-class}
\alias{naiveBayes-class}
\alias{naiveBayes}
\title{naiveBayes}
\description{Wrapped learner for Naive Bayes from package \code{e1071} for classification problems.}
\seealso{\code{\link[e1071]{naiveBayes}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,naiveBayes-method]{initialize}}}{\code{signature( = "naiveBayes")}: Constructor.}}}

