\name{nnet.multinom-class}
\alias{nnet.multinom-class}
\alias{nnet.multinom}
\title{nnet.multinom}
\description{Wrapped learner for Multinomial Regression from package \code{nnet} for classification problems.}
\details{\emph{Common hyperparameters:}}
\seealso{\code{\link[nnet]{multinom}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,nnet.multinom-method]{initialize}}}{\code{signature( = "nnet.multinom")}: Constructor.}{}{}}}
