% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\docType{data}
\name{NULLCPO}
\alias{NULLCPO}
\title{CPO composition neutral element}
\format{An object of class \code{NULLCPO} (inherits from \code{CPOPrimitive}, \code{PCORetrafo}, \code{CPO}) of length 0.}
\usage{
NULLCPO
}
\description{
FIXME to come
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}
}
\keyword{datasets}
