\name{object-class}
\alias{object-class}
\alias{object}
\title{Base class for everything.}
\description{Base class for all S4 classes in this package.}
\section{Methods}{\describe{\item{\code{\link[=[,object-method]{[}}}{}\item{\code{\link[=to.string,object-method]{to.string}}}{}\item{\code{\link[=print,object-method]{print}}}{}\item{\code{\link[=show,object-method]{show}}}{}}}
\alias{[,object-method}

