% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oneclass2d.task}
\alias{oneclass2d.task}
\title{Artificial oneclass toy task.}
\description{
Contains the task (\code{oneclass2d.task}).

The data for this class has two feature variables (V1, V2)
and is simulated with 1000 normal observations and 50 anomalous observations.
The features of the normal class were drawn from a multivariate normal distribution with zero mean
and a diagonal covariance matrix with variances (2,5).
The anomaly class was sampled with feature value uniformly distributed between 20 and 100.
}
\keyword{data}
