\name{opt.control-class}
\alias{opt.control-class}
\alias{opt.control}
\title{Base class for control objects for optimization.}
\description{Abstract base class for control objects for optimization. 
Cannot be instatiated.}
\details{\code{opt.control-class}: Getter.\cr

\describe{
\item{opt.type [string]}{'tune' or 'varsel'.}
}

}
\seealso{\code{\linkS4class{tune.control}}, \code{\linkS4class{varsel.control}}}
\section{Extends}{\code{\linkS4class{object}}}
\section{Methods}{\describe{\item{\code{\link[=[,opt.control-method]{[}}}{}}}
\alias{[,opt.control-method}

