\name{opt.wrapper-class}
\alias{opt.wrapper-class}
\alias{opt.wrapper}
\title{Abstract base class to wrap an optimization algorithm around a learner.}
\description{Abstract base class to wrap an optimization algorithm around a learner.}
\section{Extends}{\code{\linkS4class{base.wrapper}}}
\section{Methods}{\describe{\item{\code{\link[=[,opt.wrapper-method]{[}}}{}\item{\code{\link[=train.learner,opt.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}}}
\alias{[,opt.wrapper-method}

