% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{orderScore}
\alias{orderScore}
\title{Order a score vector and return the best init numbers}
\usage{
orderScore(scores, minimize, init)
}
\description{
Order a score vector and return the best init numbers
}

