% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePartialPrediction.R
\name{PartialPredictionData}
\alias{PartialPredictionData}
\title{Result of \code{\link{generatePartialPredictionData}}.}
\description{
Result of \code{\link{generatePartialPredictionData}}.
}

