% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phoneme.task}
\alias{phoneme.task}
\title{Phoneme functional data multilabel classification task.}
\description{
Contains the task (\code{phoneme.task}).
The task contains a single functional covariate and 5 equally big classes (aa, ao, dcl, iy, sh).
The aim is to predict the class of the phoneme in the functional.
The dataset is contained in the package fda.usc.
}
\references{
F. Ferraty and P. Vieu (2003) "Curve discrimination: a nonparametric functional approach", Computational Statistics and Data Analysis, 44(1-2), 161-173.
F. Ferraty and P. Vieu (2006) Nonparametric functional data analysis, New York: Springer.
T. Hastie and R. Tibshirani and J. Friedman (2009) The elements of statistical learning: Data mining, inference and prediction, 2nd edn, New York: Springer.
}
\keyword{data}
