% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pid.task}
\alias{pid.task}
\title{PimaIndiansDiabetes classification task}
\description{
Contains the task (\code{pid.task}).
}
\references{
See \code{\link[mlbench]{PimaIndiansDiabetes}}.
  Note that this is the uncorrected version from mlbench.
}
\keyword{data}

