% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBenchmarkResult.R
\name{plotBenchmarkResult}
\alias{plotBenchmarkResult}
\title{Create a Trellis-plot for a selected \link{Measure}}
\usage{
plotBenchmarkResult(bmr, measure = NULL, style = "box", order.Lrns = NULL,
  order.Tsks = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{style}{[\code{character(1)]}]\cr
Type of Plot, can be \code{"box"} for a boxplot or \code{"violin"} for
a violin-plot.}

\item{orderLrns}{[\code{character(nLearners)}] or \cr
                 [\code{integer(nLearners)}] \cr
Character vector with \code{learner.ids} in new order , or integer
vector refering to the positions in the new order.}

\item{orderLrns}{[\code{character(nTasks)}] or \cr
                 [\code{integer(nTasks)}] \cr
Character vector with \code{task.ids} in new order, or an integer
vector refering to the positions in the new order.}
}
\value{
[\link{ggplot2}] plot
}
\description{
Plots Boxplots for a selected \link{Measure} accross all iterations
of the resampling strategy, faceted by the \code{task.id}
}
\details{
Credit: This plot is similar to the one described in
Eugster, J.A. (2012)
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
plotBenchmarkResult(res,acc)
}
\seealso{
Other plot: \code{\link{plotBenchmarkSummary}};
  \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

