% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBenchmarkSummary.R
\name{plotBenchmarkSummary}
\alias{plotBenchmarkSummary}
\title{plotBenchmarkSummary}
\usage{
plotBenchmarkSummary(obj)
}
\arguments{
\item{obj}{[\code{BenchmarkSummaryData}]\cr
 Output of a \link{generateBenchmarkSummaryData} function.}
}
\value{
\link{ggplot2}] plot
}
\description{
Plots a BenchmarkSummaryPlot for a selected \link{Measure}.
Full tiles correspond
to the worst performance accross all \code{tasks.ids}
and \code{learner.ids}. The actuall fill
corresponds to the proportional performance of the \code{best}
or \code{worst} algorithm accross all tasks.
}
\details{
Credit: This plot is analogous to the one described in Eugster,J.A.(2012)
 but does not include any clustering.
}
\examples{
lrns = list(makeLearner("classif.randomForest"), makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc,meas)
g = generateBenchmarkSummaryData(res,ber,fill = "best")
plotBenchmarkSummary(g)
}
\seealso{
Other plot: \code{\link{plotBenchmarkResult}};
  \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

