% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMEnsembleWrapper.R
\name{plotBMRPHMEnsembleDistribution}
\alias{plotBMRPHMEnsembleDistribution}
\title{Plot BMRPHMEnsemble predictions as a distribution
Only last prediction for each seq.id will be plotted
WARNING: May take a lot of time, consider passing a small task}
\usage{
plotBMRPHMEnsembleDistribution(bmr, task, seq.ids = 1)
}
\description{
Plot BMRPHMEnsemble predictions as a distribution
Only last prediction for each seq.id will be plotted
WARNING: May take a lot of time, consider passing a small task
}
