% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMEnsembleWrapper.R
\name{plotBMRPHMEnsemblePrediction}
\alias{plotBMRPHMEnsemblePrediction}
\title{Plots the prediction density for one seq.id on BMRPHMEnsemble models}
\usage{
plotBMRPHMEnsemblePrediction(bmr, task, seq.id, n.regions = 4)
}
\description{
Plots the prediction density for one seq.id on BMRPHMEnsemble models
}
