% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCritDifferences.R
\name{plotCritDifferences}
\alias{plotCritDifferences}
\title{plotCritDifferences}
\usage{
plotCritDifferences(obj, baseline = NULL)
}
\arguments{
\item{obj}{[\code{critDifferencesData}]
Result of \link{generateCritDifferencesData} function.}

\item{baseline}{[\code{character(1)}]: [\code{learner.id}] \cr
Overwrites baseline from \link{generateCritDifferencesData}!\cr
Select a [\code{learner.id} as baseline for the critical difference
diagram, the critical difference will be positioned arround this learner.
Defaults to best performing algorithm.}
}
\value{
[\link{ggplot2}] plot
}
\description{
Plots a Critical-Differences Diagram for all classifiers and
a selected measure. If a baseline is selected, the Critical Difference
Interval will be positioned arround the baseline. If not, the best
performing algorithm will be chosen as baseline.
}
\details{
Credit: The output is a a critical differences plot, similar to
 the one proposed in Demsar(2006).
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.nnet"),
            makeLearner("classif.rpart"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5L)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc, meas)
g = generateCritDifferencesData(res,acc, p.value = 0.1)
plotCritDifferencesData(p)
}

