% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHyperParsEffect.R
\name{plotHyperParsEffect}
\alias{plotHyperParsEffect}
\title{Plot the hyperparameter effects data}
\usage{
plotHyperParsEffect(hyperpars.effect.data, x = NULL, y = NULL, z = NULL,
  plot.type = "scatter", loess.smooth = FALSE, facet = NULL,
  pretty.names = TRUE, global.only = TRUE, interpolate = TRUE,
  show.experiments = FALSE)
}
\arguments{
\item{hyperpars.effect.data}{[\code{HyperParsEffectData}]\cr
Result of \code{\link{generateHyperParsEffectData}}}

\item{x}{[\code{character(1)}]\cr
Specify what should be plotted on the x axis. Must be a column from
\code{HyperParsEffectData$data}}

\item{y}{[\code{character(1)}]\cr
Specify what should be plotted on the y axis. Must be a column from
\code{HyperParsEffectData$data}}

\item{z}{[\code{character(1)}]\cr
Specify what should be used as the extra axis for a particular geom. This
could be for the fill on a heatmap or color aesthetic for a line. Must be a 
column from \code{HyperParsEffectData$data}. Default is \code{NULL}.}

\item{plot.type}{[\code{character(1)}]\cr
Specify the type of plot: "scatter" for a scatterplot, "heatmap" for a 
heatmap, "line" for a scatterplot with a connecting line, or "contour" for a
contour plot layered ontop of a heatmap.
Default is scatter.}

\item{loess.smooth}{[\code{logical(1)}]\cr
If TRUE, will add loess smoothing line to plots where possible. Note that 
this is probably only useful when \code{plot.type} is set to either 
"scatter" or "line". Must be a column from \code{HyperParsEffectData$data}
Default is \code{FALSE}.}

\item{facet}{[\code{character(1)}]\cr
Specify what should be used as the facet axis for a particular geom. When 
using nested cross validation, set this to "nested_cv_run" to obtain a facet
for each outer loop. Must be a column from \code{HyperParsEffectData$data}
Default is \code{NULL}.}

\item{pretty.names}{[\code{logical{1}}]\cr
Whether to use the short name of the learner instead of its ID in labels. Defaults to \code{TRUE}.}

\item{global.only}{[\code{logical(1)}]\cr
If TRUE, will only plot the current global optima when setting 
x = "iteration" and y as a performance measure from 
\code{HyperParsEffectData$measures}. Set this to FALSE to always plot the 
performance of every iteration, even if it is not an improvement.
Default is \code{TRUE}.}

\item{interpolate}{[\code{logical(1)}]\cr
If TRUE, will interpolate non-complete grids in order to visualize a more 
complete path. Only meaningful when attempting to plot a heatmap or contour.
This will fill in "empty" cells in the heatmap or contour plot. Note that 
cases of irregular hyperparameter paths, you will most likely need to use
this to have a meaningful visualization.
Default is \code{TRUE}.}

\item{show.experiments}{[\code{logical(1)}]\cr
If TRUE, will overlay the plot with points indicating where an experiment
ran. This is only useful when creating a heatmap or contour plot with 
interpolation so that you can see which points were actually on the 
original path. Note: if any learner crashes occurred within the path, this
will become TRUE.
Default is \code{FALSE}.}
}
\value{
ggplot2 plot object.
}
\description{
Plot hyperparameter validation path. Automated plotting method for 
\code{HyperParsEffectData} object. Useful for determining the importance
or effect of a particular hyperparameter on some performance measure and/or
optimizer.
}
\note{
Any NAs incurred from learning algorithm crashes will be indicated in 
the plot and the NA values will be replaced with the column min/max depending
on the optimal values for the respective measure. Execution time will be
replaced with the max.
}
\examples{
# see generateHyperParsEffectData
}

