% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateLearningCurve.R
\name{plotLearningCurve}
\alias{plotLearningCurve}
\title{Plot learning curve data.}
\usage{
plotLearningCurve(obj, linesize = 1.5, pointsize = 1.5)
}
\arguments{
\item{obj}{[\code{LearningCurveData}]\cr
Result of \code{\link{generateLearningCurve}}.}

\item{linesize}{[\code{numeric(1)}]\cr
Linesize for ggplot2 \code{\link[ggplot2]{geom_line}} for graphs.
Default is 1.5.}

\item{pointsize}{[\code{numeric(1)}]\cr
Linesize for ggplot2 \code{\link[ggplot2]{geom_point}} for graphs.
Default is 1.5.}
}
\value{
ggplot2 plot object.
}
\description{
Visualizes data size (percentage used for model) vs. performance measure(s).
}

