% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateLearningCurve.R
\name{plotLearningCurveGGVIS}
\alias{plotLearningCurveGGVIS}
\title{Plot learning curve data using ggvis.}
\usage{
plotLearningCurveGGVIS(obj, color = "learner", interaction = "measure")
}
\arguments{
\item{obj}{[\code{LearningCurveData}]\cr
Result of \code{\link{generateLearningCurveData}}.}

\item{color}{[\code{character(1)}]\cr
Selects \dQuote{measure} or \dQuote{learner} to be used for coloring the lines and points in the plot.
A vector cannot be mapped to the color aesthetic and used for interaction.
The vector mapped to \code{color} must have more than one unique value, otherwise it will
be ignored. The default is \dQuote{learner}.}

\item{interaction}{[\code{character(1)}]\cr
Selects \dQuote{measure} or \dQuote{learner} to be used in a Shiny application
making the \code{interaction} variable selectable via a drop-down menu.
A vector cannot be mapped to \code{color} and used for interaction.
Note that if there are multiple learners and multiple measures interactivity is
necessary as ggvis does not currently support facetting or subplots.
The vector mapped to \code{interaction} must have more than one unique value, otherwise it will
be ignored. The default is \dQuote{measure}.}
}
\value{
a ggvis plot object.
}
\description{
Visualizes data size (percentage used for model) vs. performance measure(s).
}
\seealso{
Other learning_curve: \code{\link{generateLearningCurveData}};
  \code{\link{plotLearningCurve}}

Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

