% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePartialDependence.R
\name{plotPartialDependenceGGVIS}
\alias{plotPartialDependenceGGVIS}
\title{Plot a partial dependence using ggvis.}
\usage{
plotPartialDependenceGGVIS(obj, interact = NULL, p = 1)
}
\arguments{
\item{obj}{[\code{PartialDependenceData}]\cr
Generated by \code{\link{generatePartialDependenceData}}.}

\item{interact}{[\code{character(1)}]\cr
The name of a feature to be mapped to an interactive sidebar using Shiny.
This feature must have been an element of the \code{features} argument to
\code{\link{generatePartialDependenceData}} and is only applicable when said argument had length
greater than 1.
If \code{\link{generatePartialDependenceData}} is called with the \code{interaction} argument \code{FALSE}
(the default) with argument \code{features} of length greater than one, then \code{interact} is ignored and
the feature displayed is controlled by an interactive side panel.
Default is \code{NULL}.}

\item{p}{[\code{numeric(1)}]\cr
If \code{individual = TRUE} then \code{sample} allows the user to sample without replacement
from the output to make the display more readable. Each row is sampled with probability \code{p}.
Default is \code{1}.}
}
\value{
a ggvis plot object.
}
\description{
Plot partial dependence from \code{\link{generatePartialDependenceData}} using ggvis.
}
\seealso{
Other partial_dependence: \code{\link{generatePartialDependenceData}},
  \code{\link{plotPartialDependence}}

Other plot: \code{\link{plotBMRBoxplots}},
  \code{\link{plotBMRRanksAsBarChart}},
  \code{\link{plotBMRSummary}},
  \code{\link{plotCalibration}},
  \code{\link{plotCritDifferences}},
  \code{\link{plotFilterValuesGGVIS}},
  \code{\link{plotLearningCurveGGVIS}},
  \code{\link{plotLearningCurve}},
  \code{\link{plotPartialDependence}},
  \code{\link{plotROCCurves}}, \code{\link{plotResiduals}},
  \code{\link{plotThreshVsPerfGGVIS}},
  \code{\link{plotThreshVsPerf}}
}

