% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartialDependencePlotly.R
\name{plotPartialDependencePlotly}
\alias{plotPartialDependencePlotly}
\title{Plot a partial dependence with plotly.}
\usage{
plotPartialDependencePlotly(obj, p = 1, show.colorbar = TRUE,
  title = NULL)
}
\arguments{
\item{obj}{[\code{PartialDependenceData}]\cr
Generated by \code{\link{generatePartialDependenceData}}.}

\item{p}{[\code{numeric(1)}]\cr
If \code{individual = TRUE} then \code{sample} allows the user to sample without replacement
from the output to make the display more readable. Each row is sampled with probability \code{p}.
Default is \code{1}.}

\item{show.colorbar}{[\code{logical(1)}]\cr
Show the colorbar?
Default is \code{TRUE}.}

\item{title}{[\code{character(1)}]\cr
Set main title for plots.
Default is \code{NULL}.}
}
\description{
Plot a partial dependence from \code{\link{generatePartialDependenceData}} using plotly.
}
\seealso{
Other partial_dependence: \code{\link{generatePartialDependenceData}},
  \code{\link{plotPartialDependenceGGVIS}},
  \code{\link{plotPartialDependence}}

Other plot: \code{\link{plotBMRBoxplots}},
  \code{\link{plotBMRRanksAsBarChart}},
  \code{\link{plotBMRSummary}},
  \code{\link{plotCalibration}},
  \code{\link{plotCritDifferences}},
  \code{\link{plotFilterValuesGGVIS}},
  \code{\link{plotLearningCurveGGVIS}},
  \code{\link{plotLearningCurve}},
  \code{\link{plotPartialDependenceGGVIS}},
  \code{\link{plotPartialDependence}},
  \code{\link{plotROCCurves}},
  \code{\link{plotThreshVsPerfGGVIS}},
  \code{\link{plotThreshVsPerf}}
}

