% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePartialPrediction.R
\name{plotPartialPrediction}
\alias{plotPartialPrediction}
\title{Plot a partial prediction}
\usage{
plotPartialPrediction(obj, facet = NULL)
}
\arguments{
\item{obj}{\code{PartialPredictionData}\cr
Generated by \code{\link{generatePartialPredictionData}}.}

\item{facet}{\code{character(1)}\cr
The name of a feature to be used for facetting.
This feature must have been an element of the \code{features} argument to
\code{\link{generatePartialPredictionData}} and is only applicable when said argument had length
greater than 1.
Default is \code{NULL}.}
}
\value{
A ggplot2 object.
}
\description{
Plot a partial prediction from \code{\link{generatePartialPredictionData}} using ggplot2.
}
\seealso{
Other partial_prediction: \code{\link{generatePartialPredictionData}};
  \code{\link{plotPartialPredictionGGVIS}}

Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotPartialPredictionGGVIS}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

