% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePartialPrediction.R
\name{plotPartialPredictionGGVIS}
\alias{plotPartialPredictionGGVIS}
\title{Plot a partial prediction}
\usage{
plotPartialPredictionGGVIS(obj, interaction = NULL)
}
\arguments{
\item{obj}{\code{PartialPredictionData}\cr
Generated by \code{\link{generatePartialPredictionData}}.}

\item{interaction}{\code{character(1)}\cr
The name of a feature to be mapped to an interactive sidebar using Shiny.
This feature must have been an element of the \code{features} argument to
\code{\link{generatePartialPredictionData}} and is only applicable when said argument had length
greater than 1.
Default is \code{NULL}.}
}
\value{
A ggvis object.
}
\description{
Plot a partial prediction from \code{\link{generatePartialPredictionData}} using ggvis.
}
\seealso{
Other partial_prediction: \code{\link{generatePartialPredictionData}};
  \code{\link{plotPartialPrediction}}

Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotPartialPrediction}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

