% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMHIWrapper.R
\name{plotPHMHI}
\alias{plotPHMHI}
\title{PHM HI model exploratory plots}
\usage{
plotPHMHI(learner, task, hi_functions = "all", seq.ids = "all")
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{hi_functions}{Either a named list of functions or "all" for all functions in makeHIFunction}

\item{seq.ids}{[\code{character}]\cr
Sequence id to be plotted. Default "all"}
}
\description{
Plots the created synthtic HI and prediction made by learner on a given task
}
