% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrediction.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Visualize a prediction object. Right now supports only PHMRegr tasks}
\usage{
plotPrediction(prediction, linesize = 1, seq.ids = 1)
}
\arguments{
\item{prediction}{[\code{Prediction}]\cr
The prediction object to be plotted}

\item{linesize}{[\code{numeric(1)}]\cr
Linesize for ggplot2 \code{\link[ggplot2]{geom_point}} for data points.
Default is 1.}

\item{seq.ids}{[\code{character(1)} or \code{numeric(1)}]
For phm regression: Either a vector of seq.id to be plotted or a numeric fraction of seq.id to be randomly sampled
Default is 1.}
}
\value{
The ggplot2 object.
}
\description{
Plots the true value and the response in prediction via \code{\link[ggplot2]{ggplot}}.

FIXME: Implement for general classification & regression problems. Extend from plotLearnerPrediction.R

For phm regression, the true value and the predicted response of target is plotted for each seq.id in separate facets.

The plot title displays the learner name, its parameters, the training performance and the cross-validation performance.
}
