% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRankMatrixAsBar.R
\name{plotRankMatrixAsBar}
\alias{plotRankMatrixAsBar}
\title{Plot Rankmatrix as Barplot}
\usage{
plotRankMatrixAsBar(obj, pos = NULL)
}
\arguments{
\item{obj}{[\code{RankMatrixAsBarData}]\cr
Output of a \link{generateRankMatrixAsBarData} function.}

\item{pos}{[\code{character(1)}]
Optionally set how the bars are positioned in [\link{ggplot2}]. \cr
Overwrites the one created in \link{generateRankMatrixAsBar}. \cr
 \code{"tile"} plots a heatmap with \link{task} as the y-axis.\cr
 Allows identification of the performance in a special task.\cr
 \code{"stack"} plots a stacked barplot. \cr
 Allows for comparison of learners within and and accross ranks.\cr
 \code{"dodge"} plots a barplot of with dodged instead of stacked bars.\cr}
}
\value{
[\link{ggplot2}] plot
}
\description{
Plots a barchart from the ranks of algorithms. Alternatively
 tiles can be plotted for every rank-task combination, see \code{pos}
 for details. The x-axis accross all plots is the ranks of a learner.id.
 Areas are always coloured corresponding to the learner.
}
\examples{
lrns = list(makeLearner("classif.randomForest"), makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc, meas)
d = generateRankMatrixAsBarData(res, acc, order.Tsks = c(3L, 2L, 1L))
plotRankMatrixAsBar(d, "tile")
plotRankMatrixAsBar(d, "dodge")
}

