% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Create residual plots for prediction objects or benchmark results.}
\usage{
plotResiduals(obj, type = "scatterplot", loess.smooth = TRUE, rug = TRUE,
  pretty.names = TRUE)
}
\arguments{
\item{obj}{([Prediction] | [BenchmarkResult])\cr
Input data.}

\item{type}{Type of plot. Can be \dQuote{scatterplot}, the default. Or
\dQuote{hist}, for a histogram, or in case of classification problems
a barplot, displaying the residuals.}

\item{loess.smooth}{(`logical(1)`)\cr
Should a loess smoother be added to the plot? Defaults to `TRUE`.
Only applicable for regression tasks and if `type` is set to `scatterplot`.}

\item{rug}{(`logical(1)`)\cr
Should marginal distributions be added to the plot? Defaults to `TRUE`.
Only applicable for regression tasks and if `type` is set to `scatterplot`.}

\item{pretty.names}{(`logical(1)`)\cr
Whether to use the short name of the learner instead of its ID in labels.
Defaults to `TRUE`. \cr
Only applicable if a [BenchmarkResult]
is passed to `obj` in the function call, ignored otherwise.}
}
\value{
ggplot2 plot object.
}
\description{
Plots for model diagnostics. Provides scatterplots of true vs. predicted values
and histograms of the model's residuals.
}
\seealso{
Other plot: \code{\link{createSpatialResamplingPlots}},
  \code{\link{plotBMRBoxplots}},
  \code{\link{plotBMRRanksAsBarChart}},
  \code{\link{plotBMRSummary}},
  \code{\link{plotCalibration}},
  \code{\link{plotCritDifferences}},
  \code{\link{plotLearningCurve}},
  \code{\link{plotPartialDependence}},
  \code{\link{plotROCCurves}},
  \code{\link{plotThreshVsPerf}}
}

