% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotROCRCurves.R
\name{plotROCRCurves}
\alias{plotROCRCurves}
\title{Visualize binary classification predictions via ROCR ROC curves.}
\usage{
plotROCRCurves(obj, meas1 = "tpr", meas2 = "fpr", avg = "threshold",
  perf.args = list(), diagonal = FALSE, xlab = NULL, ylab = NULL,
  title = "", task.id = NULL)
}
\arguments{
\item{obj}{[(list of) \code{\link{Prediction}} | (list of) \code{\link{ResampleResult}} | \code{\link{BenchmarkResult}}]\cr
Single prediction object, list of them, single resample result, list of them, or a benchmark result.
In case of a list probably produced by different learners you want to compare, then
name the list with the names you want to see in the plots, probably
learner shortnames or ids.}

\item{meas1}{[\code{character(1)}]\cr
Measure on x-axis. Note that this is a measure name from *ROCR* and not from mlr!
Default is \dQuote{tpr}.}

\item{meas2}{[\code{character(1)}]\cr
Measure on y-axis. Note that this is a measure name from *ROCR* and not from mlr!
Default is \dQuote{fpr}.}

\item{avg}{[\code{chracter(1)}]\cr
If \code{obj} is of class \code{ResampleResult} or \code{BenchmarkResult}, how are
the predictions to be combined (by learner)?
If \code{obj} is not one of these classes, this argument is ignored.
Possibilities are \dQuote{threshold}, \dQuote{horizontal}, \dQuote{vertical}, and \dQuote{none}.
Default is \dQuote{threshold}.}

\item{perf.args}{[named \code{list}]\cr
Further arguments passed to ROCR's \code{\link[ROCR]{performance}}.
Usually not needed and \code{meas1} and \code{meas2} are set internally.
Default is an empty list.}

\item{diagonal}{[\code{logical(1)}]\cr
Whether to plot a dashed diagonal line.
Default is false.}

\item{xlab}{[\code{character(1)}]\cr
Label for x-axis.
Default is \code{meas1}.}

\item{ylab}{[\code{character(1)}]\cr
Label for y-axis
Default is \code{meas2}.}

\item{title}{[\code{character(1)}]\cr
Title for plot.
Default is an empty string.}

\item{task.id}{[\code{character(1)}]\cr
  Selected task in \code{\link{BenchmarkResult}} to do plots for, ignored otherwise.
  Default is first task.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object. To extract the raw data access the \code{data} element
  of the returned object.

[\code{invisible(NULL)}].
}
\description{
Plot is generated by calling \code{\link{asROCRPrediction}},
ROCR's \code{\link[ROCR]{performance}},
then ROCR's \code{plot} function.

See these methods in ROCR for further info.
}
\examples{
\dontrun{
lrn1 = makeLearner("classif.logreg", predict.type = "prob")
lrn2 = makeLearner("classif.rpart", predict.type = "prob")
b = benchmark(list(lrn1, lrn2), pid.task)
z = plotROCRCurves(b)
}
}
\seealso{
Other predict: \code{\link{asROCRPrediction}};
  \code{\link{getProbabilities}};
  \code{\link{plotViperCharts}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}

Other roc: \code{\link{asROCRPrediction}};
  \code{\link{plotViperCharts}}
}

