% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateROCRCurves.R
\name{plotROCRCurvesGGVIS}
\alias{plotROCRCurvesGGVIS}
\title{Plots results from generateROCRCurvesData using ggvis.}
\usage{
plotROCRCurvesGGVIS(obj, diagonal = FALSE, cutoffs = FALSE)
}
\arguments{
\item{obj}{[\code{ROCRCurvesData}]\cr
Result of \code{\link{generateROCRCurvesData}}.}

\item{diagonal}{[\code{logical(1)}]\cr
Whether to plot a dashed diagonal line.
Default is \code{FALSE}.}

\item{cutoffs}{[\code{logical(1)}]\cr
Whether to plot tooltips displaying threshold values using Shiny.
The plot will be opened in the default browser.}
}
\value{
a ggvis plot object.
}
\description{
Visualize how binary classification performs across different measures.
}
\seealso{
Other plot: \code{\link{plotBenchmarkResult}};
  \code{\link{plotBenchmarkSummary}};
  \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}

Other predict: \code{\link{asROCRPrediction}};
  \code{\link{generateROCRCurvesData}};
  \code{\link{getProbabilities}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotViperCharts}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}

Other roc: \code{\link{asROCRPrediction}};
  \code{\link{generateROCRCurvesData}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotViperCharts}}
}

