% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotThreshVsPerf.R
\name{plotThreshVsPerfGGVIS}
\alias{plotThreshVsPerfGGVIS}
\title{Plot threshold vs. performance(s) for 2-class classification using ggvis.}
\usage{
plotThreshVsPerfGGVIS(pred, measures, mark.th = NA_real_, gridsize = 100L)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}

\item{measures}{[\code{\link{Measure}} | list of \code{\link{Measure}}]\cr
Performance measure(s) to evaluate.}

\item{mark.th}{[\code{numeric(1)}]\cr
Mark given threshold with vertical line?
Default is \code{NA} which means not to do it.}

\item{gridsize}{[\code{integer(1)}]\cr
Grid resolution for x-axis (threshold).
Default is 100.}
}
\value{
a ggvis plot object.
}
\description{
Plot threshold vs. performance(s) for 2-class classification using ggvis.
}
\examples{
\dontrun{
lrn = makeLearner("classif.rpart", predict.type = "prob")
mod = train(lrn, sonar.task)
pred = predict(mod, sonar.task)
plotThreshVsPerf(pred, list(tpr, fpr))
}
}

