% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateValidation.R
\name{plotValidationData}
\alias{plotValidationData}
\title{Plot the Validation data}
\usage{
plotValidationData(ValidationData, x.axis, y.axis, plot.type)
}
\arguments{
\item{ValidationData}{}
}
\value{
what does it return
}
\description{
Plot the Validation data
}

