% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotViperCharts}
\alias{plotViperCharts}
\title{Visualize binary classification predictions via ViperCharts system.}
\usage{
plotViperCharts(obj, chart = "rocc", browse = TRUE, embedded = FALSE,
  legend = TRUE, fontsize = NULL)
}
\arguments{
\item{obj}{[(list of) \code{\link{Prediction}} | \code{\link{BenchmarkResult}}]\cr
Single prediction object, list of them, or a benchmark result.
In case of a list probably produced by different learners you want to compare, then
name the list with the names you want to see in the plots, probably
learner shortnames or ids.}

\item{chart}{[\code{character(1)}]\cr
First chart to display in focus in browser for non-embedded view,
or displayed chart for embedded mode.
Default is \dQuote{rocc}.}

\item{browse}{[\code{logical(1)}]\cr
Open ViperCharts plot in web browser? If not you simple get the URL returned.
Calls \code{\link{browseURL}}.
Default is \code{TRUE}.}

\item{embedded}{[\code{logical(1)}]\cr
Display a legend in plot?
Default is \code{FALSE}.}

\item{legend}{[\code{logical(1)}]\cr
Display a legend in plot?
Default is \code{TRUE}.}
}
\value{
[\code{character(1)}]. Invisibly returns the ViperCharts URL.
}
\description{
This includes ROC, lift charts, cost curves, and so on.
Please got to \url{http://viper.ijs.si} for further info.
}
\examples{
\dontrun{
lrn1 = makeLearner("classif.logreg", predict.type = "prob")
lrn2 = makeLearner("classif.rpart", predict.type = "prob")
b = benchmark(list(lrn1, lrn2), pid.task)
z = plotViperCharts(b, chart = "lift", browse = TRUE)
}
}
\seealso{
Other predict: \code{\link{asROCRPrediction}};
  \code{\link{getProbabilities}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictType}}

Other roc: \code{\link{asROCRPrediction}}
}

