% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/posthocNemenyiTestBMR.R
\name{posthocNemenyiTestBMR}
\alias{posthocNemenyiTestBMR}
\title{Perform a posthoc Nemenyi Test}
\usage{
posthocNemenyiTestBMR(bmr, measure = NULL, p.value = 0.05,
  aggregation = "default")
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{p.value}{[\code{numeric}(1)] \cr
 p-value for the tests.\cr  Default: 0.05}
}
\value{
A list of class \code{pairwise.htest}.\cr See
\link[PMCMR]{posthoc.friedman.nemenyi.test} for details. \cr
Additionally two components are added to the list: \cr
$\code{fRejNull:} whether the according friedman.test rejects the Null
hypothesis at the selected p.value.\cr
$\code{cDifference}: Minimal difference the mean ranks of two learners
need to have in order to be significantly different.
}
\description{
Performs a \link[PMCMR]{posthoc.friedman.nemenyi.test} for a
\link{BenchmarkResult} and a selected measure.
The null hypothesis is that apart from an effect of the different
[\code{tasks}], the location parameter (mean of aggregated
Perforamance Measure) is the same for each learner.
}
\details{
If the Null Hypthesis of the included ad-hoc friedman.test can be rejected a
\code{pairwise.htest} is returned. If not, the function returns the
corresponding \link{friedman.test}
}
\examples{
lrns = list(makeLearner("classif.randomForest"), makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc, meas)
posthocNemenyiTestBMR(res, acc)
}

