% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedStacking.R
\name{predictLearner.BoostedStackingLearner}
\alias{predictLearner.BoostedStackingLearner}
\title{Predict Boosted Stacking object.}
\usage{
\method{predictLearner}{BoostedStackingLearner}(.learner, .model, .newdata, ...)
}
\arguments{
\item{.learner}{Learner}

\item{.model}{Model.}

\item{.newdata}{New data.}

\item{...}{...}
}
\description{
Predict Boosted Stacking object.
}

