% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_predict.R
\name{predictLearner.StackedLearner}
\alias{predictLearner.StackedLearner}
\title{Predict a Stacked Learner.}
\usage{
\method{predictLearner}{StackedLearner}(.learner, .model, .newdata, ...)
}
\arguments{
\item{.learner}{[\code{StackedLearner}]}

\item{.model}{[\code{BaseEnsembleModel}]}

\item{.newdata}{[\code{data.frame}]\cr Data to predict on}

\item{...}{...}
}
\value{
Predictions are returned in matrix or vector.
}
\description{
Gets predictions from \code{getStackedBaseLearnerPredictions}. Then apply algo:
 \describe{
  \item{\code{aggregate}}{Aggregate predictions using \code{\link{aggregatePredictions}}.}
  \item{\code{ensembleselection}}{Expand prediction according to frequency chosen in training, then apply aggregation as for 'aggregate'.}
  \item{\code{superlearner}}{Convert list of \code{Prediction} to data.frame 
  (for classification with \code{predict.type = "prob"} the first feature 
   is removed to overcome mullticollinearity). Add original features if needed. 
   Apply meta model from training.}
 }
}

