% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMEnsembleWrapper.R
\name{predictPHMEnsemble}
\alias{predictPHMEnsemble}
\title{Helper function to get predictions from each ensemble members}
\usage{
predictPHMEnsemble(model, newdata, ...)
}
\description{
Helper function to get predictions from each ensemble members
}
