% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Properties}
\alias{Properties}
\alias{addProperties}
\alias{hasProperties}
\alias{removeProperties}
\alias{setProperties}
\title{Set, add, remove or query properties of mlr objects}
\usage{
setProperties(x, props)

addProperties(x, props)

removeProperties(x, props)

hasProperties(x, props)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{Learner}}]\cr
Object with properties.}

\item{props}{[\code{character}]\cr
Vector of properties to set, add, remove or query.}
}
\value{
\code{setProperties}, \code{addProperties} and \code{removeProperties}
 return an updated object.
 \code{hasProperties} returns a logical vector of the same length of \code{props}.
}
\description{
Set, add, remove or query properties of mlr objects
}

