\name{qda-class}
\alias{qda-class}
\alias{qda}
\title{qda}
\description{Wrapped learner for Quadratic Discriminant Analysis from package \code{MASS} for classification problems.}
\seealso{\code{\link[MASS]{qda}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{{}{}}}
