\name{randomForest.classif-class}
\alias{randomForest.classif-class}
\alias{randomForest.classif}
\title{randomForest.classif}
\description{Wrapped learner for Random Forests from package \code{randomForest} for classification problems.}
\details{\emph{Common hyperparameters:}
\describe{
\item{\code{ntree}}{Number of trees to grow.}
\item{\code{mtry}}{Number of variables randomly sampled as candidates at each split.}
\item{\code{nodesize}}{Minimum size of terminal nodes.}
}}
\seealso{\code{\link[randomForest]{randomForest}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{}}
