\name{rda-class}
\alias{rda-class}
\alias{rda}
\title{rda}
\description{Wrapped learner for Regularized Discriminant Analysis from package \code{klaR} for classification problems.}
\details{\emph{Common hyperparameters:}
\describe{
\item{\code{gamma}}{See details in \code{klaR}.}		
\item{\code{lambda}}{See details in \code{klaR}.}	
}}
\seealso{\code{\link[klaR]{rda}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{}}
