% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recodeTaskTarget}
\alias{recodeTaskTarget}
\title{Recode target column to a specific format.}
\usage{
recodeTaskTarget(task, subset, type = "no")
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{subset}{[\code{integer} | \code{logical}]\cr
Vector to optionally subset the target.}

\item{type}{[\code{character(1)}]\cr
Options are \dQuote{01} and \dQuote{-1+1} form classification tasks and
\dQuote{lcens}, \dQuote{rcens} and \dQuote{icens} for survival tasks.}
}
\value{
[data.frame].
}
\description{
Recode target column to a specific format.
}

