% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDAFeatures.R
\name{reextractFDAFeatures}
\alias{reextractFDAFeatures}
\title{Re-extract features from a data set}
\usage{
reextractFDAFeatures(obj, desc)
}
\arguments{
\item{obj}{([Task] | [data.frame])\cr
Task or data.frame to extract functional features from. Must contain functional features
as matrix columns.}

\item{desc}{(`extractFDAFeatDesc`)\cr
FDAFeature extraction description as returned by [extractFDAFeatures]}
}
\value{
[data.frame] or [Task] containing the extracted Features
}
\description{
This function accepts a data frame or a task and an extractFDAFeatDesc
(a FDA feature extraction description)
as returned by [extractFDAFeatures] to extract features
from previously unseen data.
}

