% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_regr_featureless.R
\name{regr.featureless}
\alias{regr.featureless}
\title{Featureless regression learner.}
\description{
Very basic baseline method, mainly useful in model comparisons (if you don't
beat this you very likely have a problem).
Does not consider features of the task and only looks at labels on training
data.

Methods \dQuote{mean} and \dQuote{median} always predict a constant value
for each new observation in the test set which is the mean or median of the
training data, respectively.

The predict.type \dQuote{se} always uses the standard deviation of the
training data as se-estimator for each new point in the test set.
}

