% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_regr_randomForest.R
\name{regr.randomForest}
\alias{regr.randomForest}
\title{regression using randomForest.}
\description{
a mlr learner for regrssion tasks using \code{\link[randomForest]{randomForest}}.
}
\details{
if \code{predict.type = "se"} the \code{se.method} (by default \dQuote{jackknife})
is estimated, using the methods described in Sexton and Laake (2009).
}
\references{
[Joseph Sexton] and [Petter Laake],; [Standard errors for bagged and random forest estimators], Computational Statistics and Data Analysis Volume 53, 2009, [801-811].
}

