\name{regr.task-class}
\alias{regr.task-class}
\alias{regr.task}
\title{classif.task}
\description{General description object for a regression experiment.  
Instantiate it by using its factory method.}
\seealso{learn.task make.regr.task}
\section{Extends}{\code{\linkS4class{learn.task}}}
\section{Methods}{\describe{\item{\code{\link[=default.loss,regr.task-method]{default.loss}}}{}\item{\code{\link[=initialize,regr.task-method]{initialize}}}{\code{signature( = "regr.task")}: Constructor.}
