% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_fit.R
\name{regression_fit}
\alias{regression_fit}
\title{regression_fit}
\usage{
regression_fit(formula, data, alpha = 0.05)
}
\arguments{
\item{formula}{an object of class "formula" or one that can be coerced by as.formula}

\item{data}{an object of class "data.frame" or one that can be coerced by as.data.frame.}

\item{alpha}{an optional argument indicating significance level of confidence intervals. If not provided, 0.05 is used.}
}
\value{
A list containing model details, estimated coefficients, confidence intervals,
fitted values, residuals, variance-covariance matrix, ANOVA test results, and R-squared.
}
\description{
Fits multiple linear regression model using ordinary least squares
}
\examples{
data(mtcars)
mlr_output <- regression_fit(formula = mpg ~ wt + hp + wt:hp, data = mtcars)

mlr_coefficients <- mlr_output$coefficients
wt_estimate <- mlr_coefficients["wt","Estimate"]

mlr_residuals <- mlr_output$residuals
mlr_vcov <- mlr_output$variances
}
