% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativeOverfitting.R
\name{relativeOverfitting}
\alias{relativeOverfitting}
\title{Estimate relative overfitting.}
\usage{
relativeOverfitting(pred, measures, task, model, feats = NULL)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}

\item{measures}{[\code{\link{Measure}} | list of \code{\link{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}

\item{task}{[\code{\link{Task}}]\cr
Learning task.}

\item{model}{[\code{\link{WrappedModel}}]\cr
Model built on training data.}

\item{feats}{[\code{data.frame}]\cr
Features of predicted data, usually not needed except for clustering.
If the prediction was generated from a \code{task}, you can also pass this instead and the features
are extracted from it.}
}
\value{
[named \code{numeric}]. Relative overfitting estimate(s), named by measure(s).
}
\description{
Estimates the relative overfitting of a model as the ratio of the difference in test and train performance to the difference of test performance in the no-information case and train performance.
In the no-information case the features carry no information with respect to the prediction. This is simulated by permuting features and predictions.
}
\details{
Currently only support for classification and regression tasks is implemented.
}
\examples{
training.set = seq(1, nrow(iris), by = 2)
test.set = seq(2, nrow(iris), by = 2)

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.lda")
mod = train(lrn, task, subset = training.set)
pred = predict(mod, newdata = iris[test.set, ])
relativeOverfitting(pred, measures = mmce, task = task, model = mod)

lrn2 = makeLearner("classif.knn")
mod2 = train(lrn2, task, subset = training.set)
pred2 = predict(mod2, newdata = iris[test.set, ])
relativeOverfitting(pred2, measures = mmce, task = task, model = mod2)
}
\references{
Bradley Efron and Robert Tibshirani; Improvements on Cross-Validation: The .632+ Bootstrap Method, Journal of the American Statistical Association, Vol. 92, No. 438. (Jun., 1997), pp. 548-560.
}
\seealso{
Other performance: \code{\link{G1}}, \code{\link{G2}},
  \code{\link{acc}}, \code{\link{auc}}, \code{\link{bac}},
  \code{\link{ber}}, \code{\link{brier}},
  \code{\link{cindex}}, \code{\link{db}},
  \code{\link{dunn}}, \code{\link{f1}}, \code{\link{fdr}},
  \code{\link{featperc}}, \code{\link{fn}},
  \code{\link{fnr}}, \code{\link{fp}}, \code{\link{fpr}},
  \code{\link{gmean}}, \code{\link{gpr}},
  \code{\link{hamloss}}, \code{\link{mae}},
  \code{\link{mcc}}, \code{\link{mcp}},
  \code{\link{meancosts}}, \code{\link{measureACC}},
  \code{\link{measureAUC}}, \code{\link{measureBAC}},
  \code{\link{measureBrier}}, \code{\link{measureFDR}},
  \code{\link{measureFN}}, \code{\link{measureFNR}},
  \code{\link{measureFP}}, \code{\link{measureFPR}},
  \code{\link{measureGMEAN}}, \code{\link{measureGPR}},
  \code{\link{measureHAMLOSS}}, \code{\link{measureMAE}},
  \code{\link{measureMCC}}, \code{\link{measureMEDAE}},
  \code{\link{measureMEDSE}}, \code{\link{measureMMCE}},
  \code{\link{measureMSE}}, \code{\link{measureNPV}},
  \code{\link{measurePPV}}, \code{\link{measureRMSE}},
  \code{\link{measureSAE}}, \code{\link{measureSSE}},
  \code{\link{measureTN}}, \code{\link{measureTNR}},
  \code{\link{measureTP}}, \code{\link{measureTPR}},
  \code{\link{measures}}, \code{\link{medae}},
  \code{\link{medse}}, \code{\link{mmce}},
  \code{\link{mse}}, \code{\link{multiclass.auc}},
  \code{\link{npv}}, \code{\link{ppv}}, \code{\link{rmse}},
  \code{\link{sae}}, \code{\link{silhouette}},
  \code{\link{sse}}, \code{\link{timeboth}},
  \code{\link{timepredict}}, \code{\link{timetrain}},
  \code{\link{tn}}, \code{\link{tnr}}, \code{\link{tp}},
  \code{\link{tpr}}; \code{\link{Measure}},
  \code{\link{makeMeasure}}; \code{\link{makeCostMeasure}};
  \code{\link{makeCustomResampledMeasure}};
  \code{\link{performance}}
}

