% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/removeHyperPars.R
\name{removeHyperPars}
\alias{removeHyperPars}
\title{Remove hyperparameters settings of a learner.}
\usage{
removeHyperPars(learner, ids = character(0L))
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{ids}{[\code{character}]\cr
Parameter names to remove settings for.
 Default is \code{character(0L)}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Remove settings (previously set through mlr) for some parameters.
Which means that the default behavior for that param will now be used.
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{addProperties}}, \code{\link{hasProperties}},
  \code{\link{removeProperties}},
  \code{\link{setProperties}}; \code{\link{Learner}},
  \code{\link{makeLearner}}; \code{\link{getHyperPars}};
  \code{\link{getParamSet}}; \code{\link{setHyperPars}};
  \code{\link{setId}}; \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}
}

