% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{removeModelsOnDisc}
\alias{removeModelsOnDisc}
\title{Remove models from disc}
\usage{
removeModelsOnDisc(stack.id = NULL, bls.ids = NULL)
}
\arguments{
\item{stack.id}{Name of stack}

\item{bls.ids.}{Vector of base learner names}
}
\description{
Remove models from disc
}

