% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{removeStackingModelsOnDisc}
\alias{removeStackingModelsOnDisc}
\title{Remove Stacking models from disc.}
\usage{
removeStackingModelsOnDisc(stack.id = NULL, bls.ids = NULL)
}
\arguments{
\item{stack.id}{Name of stack.}

\item{bls.ids}{Vector of base learner names.}
}
\description{
Remove Stacking models from disc.
}

