\name{resample.update}
\alias{resample.update-methods}
\alias{resample.update}
\title{Implement this for your own sequential resampling.}
\description{Implement this for your own sequential resampling.}
\alias{resample.update,resample.instance,learn.task,wrapped.model,prediction-method}
\alias{resample.update}
