% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_recombine.R
\name{retrainSuperLearner}
\alias{retrainSuperLearner}
\title{Retrain super learner on new test task}
\usage{
retrainSuperLearner(obj, super.learner)
}
\arguments{
\item{obj}{[\code{ResampleResult}]\n Object using \code{StackedLearner} as Learner.}

\item{super.learner}{[\code{Learner}]\n New \code{super.learner} to apply.}
}
\description{
Retrain super learner on new test task
}

