\name{rlearner-class}
\alias{rlearner-class}
\alias{rlearner}
\title{Base class for inducers.}
\description{Wraps an already implemented learning method from R to make it accessible to mlr.}
\details{\code{rlearner-class}: Getter.\cr

}
\section{Extends}{\code{\linkS4class{learner}}}
\section{Methods}{\describe{\item{\code{\link[=[,rlearner-method]{[}}}{\code{signature( = "rlearner")}: Getter.}}}
\alias{[,rlearner-method}

