% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{rowiseRatio}
\alias{rowiseRatio}
\title{Count the ratio (used if base.learner predict.type = "response" and 
super.learner predict.type is "prob")}
\usage{
rowiseRatio(pred.data, levels, model.weight = NULL)
}
\arguments{
\item{pred.data}{Prediction data}

\item{levels}{Target levels of classifiaction task}

\item{model.weight}{Model weights, default is 1/number of data points
export}
}
\description{
Count the ratio (used if base.learner predict.type = "response" and 
super.learner predict.type is "prob")
}

