\name{set.hyper.pars}
\alias{set.hyper.pars,base.wrapper,list-method}
\alias{set.hyper.pars}
\title{set.hyper.pars,base.wrapper,list-method}
\alias{set.hyper.pars-methods}
\alias{set.hyper.pars}
\description{Set hyperparameters of learner object.}
\value{\code{set.hyper.pars-methods}: \code{\linkS4class{learner}} with changed hyperparameters.

}
\arguments{\item{learner}{[\code{\linkS4class{learner}}]\cr 
Learner object.}
\item{...}{[any] \cr
Optional named (hyper)parameters. Alternatively, you can pass via the "par.vals" argument.}
\item{par.vals}{[list] \cr
Optional list of named (hyper)parameters. Alternatively, you can pass via the ... argument.}
}
\alias{set.hyper.pars,learner,list-method}
\alias{set.hyper.pars}
