\name{set.id}
\alias{set.id-methods}
\alias{set.id}
\title{Set id of learner object.}
\description{Set id of learner object.}
\value{\code{set.id-methods}: \code{\linkS4class{learner}} with changed id.

}
\arguments{\item{learner}{[\code{\linkS4class{learner}}]\cr 
Learner object.}
\item{id}{[string] \cr
New id.}
}
\alias{set.id,learner,character-method}
\alias{set.id}
