\name{set.label}
\alias{set.label-methods}
\alias{set.label}
\title{Set label of learner object.}
\description{Set label of learner object.}
\value{\code{set.label-methods}: \code{\linkS4class{learner}} with changed label.

}
\arguments{\item{learner}{[\code{\linkS4class{learner}}]\cr 
Learner object.}
\item{label}{[string] \cr
New label.}
}
\alias{set.label,learner,character-method}
\alias{set.label}
