% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{setAggregation}
\alias{setAggregation}
\title{Set aggregation function of measure.}
\usage{
setAggregation(measure, aggr)
}
\arguments{
\item{measure}{[\code{\link{Measure}}]\cr
Performance measure.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}

\item{aggr}{[\code{\link{Aggregation}}]\cr
Aggregation function.}
}
\value{
[\code{\link{Measure}}] with changed aggregation behaviour.
}
\description{
Set how this measure will be aggregated after resampling.
To see possible aggregation functions: \code{\link{aggregations}}.
}

