% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClassWeights.R
\name{setClassWeights}
\alias{setClassWeights}
\title{Sets class weights for classification learners that support the property.}
\usage{
setClassWeights(learner, w)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The classification learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
ret_learner
}

