% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{setCPOId}
\alias{setCPOId}
\title{the ID of a CPO object.}
\usage{
setCPOId(cpo, id)
}
\description{
Setting the ID of a CPO to a value will prefix all its
parameter names with this ID. This makes it possible to
compose CPOs that have clashing parameter names.
}
