% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHyperPars.R
\name{setHyperPars}
\alias{setHyperPars}
\title{Set the hyperparameters of a learner object.}
\usage{
setHyperPars(learner, ..., par.vals = list(), use.mlr.defaults = TRUE,
  update = TRUE, show.info = getMlrOption("show.info"))
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{...}{[any]\cr
Named (hyper)parameters with new setting. Alternatively these can be passed
using the \code{par.vals} argument.}

\item{par.vals}{[\code{list}]\cr
Optional list of named (hyper)parameter settings. The arguments in
\code{...} take precedence over values in this list.}

\item{use.mlr.defaults}{[\code{list}]\cr
Use the mlr.defaults for the given learner.
This is recommended as they guarantee that the learner works and the underlying function is called with all necessary arguments.
Default is \code{TRUE} which means that the mlr.defaults will be kept as long as they are feasible.}

\item{update}{Whether to update the existing parameter values of the learner or to just give respect to the new ones.
Default is \code{TRUE} which means that the old parameter values will be kept as long as they are feasible.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureMlr}}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Takes the possiby defined mlr defsault of a learners and inserts/overwrites
with the new settings.
}
\note{
If a named (hyper)parameter can't be found for the given learner, the 3
closest (hyper)parameter names will be output in case the user mistyped.
}
\examples{
cl1 = makeLearner("classif.ksvm", sigma = 1)
cl2 = setHyperPars(cl1, sigma = 10, par.vals = list(C = 2))
print(cl1)
# note the now set and altered hyperparameters:
print(cl2)
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getClassWeightParam}},
  \code{\link{getHyperPars}}, \code{\link{getParamSet}},
  \code{\link{makeLearner}}, \code{\link{removeHyperPars}},
  \code{\link{setId}}, \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

