% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_operators.R
\name{setLearnerId}
\alias{setLearnerId}
\title{Set the ID of a learner object.}
\usage{
setLearnerId(learner, id)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{id}{[\code{character(1)}]\cr
New ID for learner.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Set the ID of the learner.
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getClassWeightParam}},
  \code{\link{getHyperPars}}, \code{\link{getLearnerId}},
  \code{\link{getLearnerPackages}},
  \code{\link{getLearnerParVals}},
  \code{\link{getLearnerParamSet}},
  \code{\link{getLearnerPredictType}},
  \code{\link{getLearnerShortName}},
  \code{\link{getLearnerType}}, \code{\link{getParamSet}},
  \code{\link{makeLearners}}, \code{\link{makeLearner}},
  \code{\link{removeHyperPars}},
  \code{\link{setHyperPars}}, \code{\link{setId}},
  \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

