% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{setPredictType.StackedLearner}
\alias{setPredictType.StackedLearner}
\title{Sets the predict.type for the super learner of a stacked learner}
\usage{
\method{setPredictType}{StackedLearner}(learner, predict.type)
}
\arguments{
\item{learner}{StackedLearner.}

\item{predict.type}{"prob" or "response".}
}
\description{
Sets the predict.type for the super learner of a stacked learner
}

