% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showHyperPars}
\alias{showHyperPars}
\title{Display all possible hyperparameter settings for a learner that mlr knows.}
\usage{
showHyperPars(cl)
}
\arguments{
\item{cl}{[\code{character(1)}]\cr
Class of learner. By convention, all classification learners
start with \dQuote{classif.}, all regression learners with
\dQuote{regr.} and all survival learners start with \dQuote{surv.}.
A list of all learners is available on the
\code{\link{learners}} help page.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Useful for quick overview, also does not foce you to create the learner.
}

