% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simplifyMeasureNames}
\alias{simplifyMeasureNames}
\title{Simplify measure names}
\usage{
simplifyMeasureNames(xs)
}
\arguments{
\item{xs}{[\code{character}] Character vector that (possibly)
contains aggregated measure names.}
}
\value{
[\code{character}].
}
\description{
Clips aggregation names from character vector.\cr
  E.g: 'mmce.test.mean' becomes 'mmce'.\cr
  Elements that don't contain a measure name are ignored
  and returned unchanged.
}
