% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sonar.task}
\alias{sonar.task}
\title{Sonar classification task.}
\description{
Contains the task (\code{sonar.task}).
}
\references{
See \link[mlbench:Sonar]{mlbench::Sonar}.
}
\keyword{data}
