% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_stackcv.R
\name{stackCV}
\alias{stackCV}
\title{Stacking method stackCV uses super.learner to obtain level 1 data and uses inner cross-validation for prediction.}
\usage{
stackCV(learner, task)
}
\arguments{
\item{learner}{[\code{StackedLearner}]}

\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\description{
Stacking method stackCV uses super.learner to obtain level 1 data and uses inner cross-validation for prediction.
}

