\name{stats.lm-class}
\alias{stats.lm-class}
\alias{stats.lm}
\title{stats.lm}
\description{Wrapped learner for Linear Models from package \code{stats} for regression problems.}
\seealso{\code{\link[stats]{lm}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.regr}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,stats.lm-method]{initialize}}}{\code{signature( = "stats.lm")}: Constructor.}}}
