\name{subsample.desc-class}
\alias{subsample.desc-class}
\alias{subsample.desc}
\title{subsample.desc}
\description{Description class for subsampling.}
\seealso{\code{\link{make.subsample.desc}}}
\section{Extends}{\code{\linkS4class{resample.desc}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,subsample.desc-method]{initialize}}}{\code{signature( = "subsample.desc")}: Create description object for subsampling.}}}

