\name{subsample.instance-class}
\alias{subsample.instance-class}
\alias{subsample.instance}
\title{subsample.instance}
\description{A subsample.instance object is mainly a set of integer vectors generated by subsampling. 
They indicate the training part of data for different instances of subsample.}
\note{Normally you don't want to use the S4 constructor, but instead \code{\link{make.res.instance}}.}
\section{Extends}{\code{\linkS4class{resample.instance}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,subsample.instance-method]{initialize}}}{\code{signature( = "subsample.instance")}: Constructor.}}}
