% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarizeLevels}
\alias{summarizeLevels}
<<<<<<< HEAD
\title{Summarize factors of a data.frame.}
\usage{
summarizeLevels(data, which)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data to summarize.
Characters and logicals will be treated as factors.}

\item{which}{[\code{character}]\cr
Restrict result to columns in \code{which}.
Default is all factor, character and logical columns of \code{data}.}
}
\value{
A named list of tables.
}
\description{
Summarizes factors of a data.frame by tabling them.
=======
\title{Summarizes factors of a data.frame by tabling them.}
\usage{
summarizeLevels(obj, cols = NULL)
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{Task}}]\cr
Input data.}

\item{cols}{[\code{character}]\cr
Restrict result to columns in \code{cols}.
Default is all factor, character and logical columns of \code{obj}.}
}
\value{
[\code{list}]. Named list of tables.
}
\description{
Characters and logicals will be treated as factors.
>>>>>>> mlrEDA
}

