% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_train.R
\name{superlearnerBaseLearners}
\alias{superlearnerBaseLearners}
\title{Train function for stacking method "Super Learner", which uses meta learner to obtain level 1 data and uses inner cross-validation for prediction.}
\usage{
superlearnerBaseLearners(learner, task)
}
\arguments{
\item{learner}{[\code{StackedLearner}]}

\item{task}{[\code{Task}]}
}
\description{
Train function for stacking method "Super Learner", which uses meta learner to obtain level 1 data and uses inner cross-validation for prediction.
}

