% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_feritility}
\alias{swiss_feritility}
\title{swiss_feritility}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 1 columns.
}
\usage{
swiss_feritility
}
\description{
Standardized fertility measure for each of 47
French-speaking provinces of Switzerland at about 1888.
}
\details{
\describe{
 \item{Fertility}{Ig, 愼㸱愼㹥common standardized fertility measure愼㸱愼㹦}
}
}
\keyword{datasets}
