% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_predictors}
\alias{swiss_predictors}
\title{swiss_predictors}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 5 columns.
}
\usage{
swiss_predictors
}
\description{
Standardized socio-economic indicators for each of 47
French-speaking provinces of Switzerland at about 1888.
}
\details{
\describe{
 \item{Agriculture}{percentage of males involved in agriculture as occupation}
 \item{Examination}{percentage draftees receiving highest mark on army examination}
 \item{Education}{percentage education beyond primary school for draftees}
 \item{Catholic}{percentage 愼㸱愼㹥catholic愼㸱愼㹦 (as opposed to 愼㸱愼㹥protestant愼㸱愼㹦)}
 \item{Infant.Mortality}{live births who live less than 1 year}
}
}
\keyword{datasets}
