% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{tagInvert}
\alias{tagInvert}
\title{Set the Inverter Tag}
\usage{
tagInvert(data, set = TRUE)
}
\arguments{
\item{data}{[\code{data.frame} | \code{\link{Task}}]\cr
The data to tag}

\item{set}{[\code{logical(1)}]\cr
Whether to set the tag  or unset it. Default is TRUE.}
}
\description{
Tag the data that when it is sent through
a \code{\link{\%>>\%}} chain, the inverter
will be kept.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoApply}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{hasTagInvert}}, \code{\link{inverter<-}},
  \code{\link{inverter}}, \code{\link{makeCPOTargetOp}},
  \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}
}
