\name{task.desc-class}
\alias{task.desc-class}
\alias{task.desc}
\title{Description object for task.}
\description{Description object for task.}
\details{\code{task.desc-class}: Getter.\cr

\describe{
\item{id [string]}{Id string of task.}
\item{label [string]}{Label string of task.}
\item{is.classif [boolean]}{Classification task?}
\item{is.regr [boolean]}{Regression task?}
\item{has.weights [boolean]}{Are weights available in task for covariates?}
\item{has.blocking [boolean]}{Is blocking available in task for observations?}
\item{costs [matrix]}{Cost matrix, of dimension (0,0) if not available.}
\item{positive [string]}{Positive class label for binary classification, NA else.}
\item{negative [string]}{Negative class label for binary classification,, NA else.}
}

}
\section{Extends}{\code{\linkS4class{object}}}
\section{Methods}{\describe{\item{\code{\link[=train.learner,classif.ada,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.adaboost.M1,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,base.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.blackboost,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.blackboost,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.ctree,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.earth,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,filter.wrapper,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.gbm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.gbm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.glmboost,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.grplasso,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.J48,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.JRip,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.kknn,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.kknn,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.km,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.ksvm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.ksvm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.lasso,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.lm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.loclda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.logreg,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lssvm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lvq1,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.mars,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.mda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,multiclass.wrapper,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.multinom,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.naiveBayes,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.nnet,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.nnet,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.OneR,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,opt.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.PART,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.pcr,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lpsvm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,preproc.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.qda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.randomForest,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.randomForest,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.rda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.ridge,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.rpart,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.rpart,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.rsm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.rvm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.sda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.svm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=[,task.desc-method]{[}}}{}}}
\alias{[,task.desc-method}

