% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{Task}
\alias{Task}
\alias{ClassifTask}
\alias{RegrTask}
\alias{SurvTask}
\alias{CostSensTask}
\alias{ClusterTask}
\alias{MultilabelTask}
\title{Create a classification, regression, survival, cluster, cost-sensitive classification or
multilabel task.}
\description{
The task encapsulates the data and specifies - through its subclasses -
the type of the task.
It also contains a description object detailing further aspects of the data.
\itemize{
 \item{For classification see}{ \code{\link{makeClassifTask}}}
 \item{For regression see}{ \code{\link{makeRegrTask}}}
 \item{For survival analysis see}{ \code{\link{makeSurvTask}}}
 \item{For clustering see}{ \code{\link{makeClusterTask}}}
 \item{For cost sensitive classification see}{ \code{\link{makeCostSensTask}}}
 \item{For multilabel classification see}{ \code{\link{makeMultilabelTask}}}
}
}
\seealso{
Other make.task: \code{\link{makeClassifTask}},
  \code{\link{makeClusterTask}},
  \code{\link{makeCostSensTask}},
  \code{\link{makeMultilabelTask}},
  \code{\link{makeRegrTask}}, \code{\link{makeSurvTask}}
}
