\name{to.string}
\alias{to.string,bench.result-method}
\alias{to.string}
\title{to.string,bench.result-method}
\alias{to.string,repcv.desc-method}
\alias{to.string}
\alias{to.string,learner-method}
\alias{to.string}
\alias{to.string-methods}
\alias{to.string}
\description{General method to convert object to strings.}
\alias{to.string,object-method}
\alias{to.string}
\alias{to.string,opt.result-method}
\alias{to.string}
\alias{to.string,prediction-method}
\alias{to.string}
\alias{to.string,resample.prediction-method}
\alias{to.string}
\alias{to.string,resample.desc-method}
\alias{to.string}
\alias{to.string,resample.instance-method}
\alias{to.string}
\alias{to.string,classif.task-method}
\alias{to.string}
\alias{to.string,regr.task-method}
\alias{to.string}
\alias{to.string,wrapped.model-method}
\alias{to.string}
