\name{train.learner,kernlab.svm.classif,formula,data.frame,numeric,list-method}
\alias{train.learner,kernlab.svm.classif,formula,data.frame,numeric,list-method}
\alias{train.learner}
\title{Overwritten, to allow direct passing of kernel hyperparameters.}
\description{Overwritten, to allow direct passing of kernel hyperparameters.
Besides that, simply delegates to super method.}

