% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedStacking.R
\name{trainLearner.BoostedStackingLearner}
\alias{trainLearner.BoostedStackingLearner}
\title{Train Boosted Stacking object.}
\usage{
\method{trainLearner}{BoostedStackingLearner}(.learner, .task, .subset, ...)
}
\arguments{
\item{.learner}{Learner}

\item{.task}{[\code{Task}].}

\item{.subset}{Subset of task.}

\item{...}{...}
}
\description{
Train Boosted Stacking object.
}

