% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_train.R
\name{trainLearner.StackedLearner}
\alias{trainLearner.StackedLearner}
\title{Switch to train \code{StackedLearner}.}
\usage{
\method{trainLearner}{StackedLearner}(.learner, .task, .subset, ...)
}
\description{
Switch to train \code{StackedLearner}.
}

