\name{tune.control-class}
\alias{tune.control-class}
\alias{tune.control}
\title{Base class for control objects for tuning.}
\description{Abstract base class for control objects for tuning. 
Cannot be instatiated.}
\seealso{\code{\linkS4class{grid.control}}, \code{\linkS4class{optim.control}}, \code{\linkS4class{cmaes.control}}}
\section{Extends}{\code{\linkS4class{opt.control}}}
\section{Methods}{\describe{\item{\code{\link[=[,tune.control-method]{[}}}{}}}
\alias{[,tune.control-method}

