% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTuneMultiCritControlGrid}
\alias{TuneMultiCritControl}
\alias{TuneMultiCritControlGrid}
\alias{TuneMultiCritControlRandom}
\alias{makeTuneMultiCritControlGrid}
\alias{makeTuneMultiCritControlNSGA2}
\alias{makeTuneMultiCritControlRandom}
\title{Create control structures for tuning.}
\usage{
makeTuneMultiCritControlGrid(same.resampling.instance = TRUE,
  impute.val = Inf, resolution = 10L)

makeTuneMultiCritControlNSGA2(same.resampling.instance = TRUE,
  impute.val = Inf, ...)

makeTuneMultiCritControlRandom(same.resampling.instance = TRUE,
  impute.val = Inf, maxit = 100L)
}
\arguments{
\item{maxit}{[\code{integer(1)}]\cr
Number of iterations for random search.
Default is 100.}

\item{same.resampling.instance}{[\code{logical(1)}]\cr
Should the same resampling instance be used for all evaluations to reduce variance?
Default is \code{TRUE}.}

\item{impute.val}{[\code{numeric(1)}]\cr
If something goes wrong during optimization (e.g, the learner crashes),
this value is fed back to the tuner, so the tuning algorithm does not abort.
It is not stored in the optimization path, an NA and a corresponding error message are
logged instead.
Note that this value is later multiplied by -1 for maximization measures internally, so you
need to enter a larger positive value for maximization here as well.
Default is the worst obtainable value of the performance measure you optimize for when
you aggregate by mean value, or \code{Inf} instead.}

\item{resolution}{[\code{integer}]\cr
Resolution of the grid for each numeric/integer parameter in \code{par.set}.
For vector parameters, it is the resolution per dimension.
Either pass one resolution for all parameters, or a named vector.
See \code{\link[ParamHelpers]{generateGridDesign}}.
Default is 10.}

\item{...}{[any]\cr
Further control parameters passed to the \code{control} argument of \code{\link[stats]{optim}},
the \code{control} argument of \code{\link[cmaes]{cma_es}}, \code{tunerConfig}
argument of \code{\link[irace]{irace}}.}
}
\value{
[\code{\link{TuneControl}}]. The specific subclass is one of
  \code{\link{TuneControlGrid}}, \code{\link{TuneControlRandom}},
}
\description{
The following tuners are available:
\describe{
  \item{makeTuneControlGrid}{Grid search. All kinds of parameter types can be handled,
    but you have discretize them yourself by always using
    \code{\link[ParamHelpers]{makeDiscreteParam}} in the \code{par.set}
    passed to \code{\link{tuneParams}}.}
  \item{makeTuneControlRandom}{Random search. All kinds of parameter types can be handled.}
}

Dependent parameters can currently only be handled by random search and irace.
}
\seealso{
Other tune_multicrit: \code{\link{plotTuneMultiCritResult}};
  \code{\link{tuneParamsMultiCrit}}
}

