% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLagDiffFeatures.R
\name{updateLagDiff}
\alias{updateLagDiff}
\title{Update Lagged and Differenced Data}
\usage{
updateLagDiff(task, newdata, weights, ...)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{newdata}{[\code{data.frame}]\cr
New observations to update the model}

\item{weights}{[\code{numeric}]\cr
Optional, non-negative case weight vector to be used during fitting.
If given, must be of same length as \code{subset} and in corresponding order.
By default \code{NULL} which means no weights are used unless specified in the task (\code{\link{Task}}).
Weights from the task will be overwritten.}

\item{...}{[any]\cr
Currently ignored.}
}
\value{
[\code{\link{WrappedModel}}].
}
\description{
Update a task modified by lags and differences with new data
}

