\name{varsel.control-class}
\alias{varsel.control-class}
\alias{varsel.control}
\title{Base class for control objects for variable selection.}
\description{Abstract base class for control objects for variable selection. 
Cannot be instatiated.}
\seealso{\code{\link{sequential.control}}, \code{\link{randomvarsel.control}}}
\section{Extends}{\code{\linkS4class{opt.control}}}
\section{Methods}{\describe{}}
