% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{visualize_airport_delays}
\usage{
visualize_airport_delays()
}
\value{
A plot displaying the average delay for flights in 105 airports.
}
\description{
A ggplot visualisation of flight delays in New York city. Uses information from 336,776 flights from three airports(JFK, LGA, EWR).
}
\examples{
visualize_airport_delays()
}
