% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visualizeLearner}
\alias{visualizeLearner}
\title{Visualizes a learning algorithm on a 1D or 2D data set.}
\usage{
visualizeLearner(learner, task, features = NULL, ..., par.vals = list(),
  gridsize = 50L, pointsize = 2L)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{task}{[\code{\link{SupervisedTask}}]\cr
The task.}

\item{features}{[\code{character}]\cr
Selected features for model.
By default the first 2 features are used.}

\item{...}{[any]\cr
Parameters for \code{learner}.}

\item{...}{[\code{list}]\cr
Parameters for \code{learner}.}

\item{features}{[\code{character}]\cr
Selected features for model.
By default the first 2 features are used.}

\item{grid}{[\code{integer(1)}]\cr
Grid resolution per axis for 2D background.
Default is 50.}

\item{pointsize}{[\code{integer(1)}]\cr
Pointsize for ggplot2 geom_point for data points.
Default is 2.}
}
\value{
[\code{\link{WrappedModel}}].
}
\description{
Trains the model for 1 or 2 selected features, then displays it via \code{\link[ggplot2]{ggplot}}.
}

