% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{wpbc.task}
\alias{wpbc.task}
\title{Wisonsin Prognostic Breast Cancer (WPBC) survival task}
\description{
Contains the task (\code{wpbc.task}).
}
\references{
See \code{\link[TH.data]{wpbc}}.
 Incomplete cases have been removed from the task.
}
\keyword{data}

