\name{wrapped.learner.classif-class}
\alias{wrapped.learner.classif-class}
\alias{wrapped.learner.classif}
\title{wrapped.learner.classif}
\description{Wrapped.learner.classif ---text!----}
\section{Slots}{\describe{\item{\code{train.par.for.classes}:}{(\code{\link{list}})}\item{\code{train.par.for.probs}:}{(\code{\link{list}})}\item{\code{predict.par.for.classes}:}{(\code{\link{list}})}\item{\code{predict.par.for.probs}:}{(\code{\link{list}})}\item{\code{trafo.for.classes}:}{(\code{\link{function}})}\item{\code{trafo.for.probs}:}{(\code{\link{function}})}\item{\code{dummy.classes}:}{(\code{\link{logical}}) Does the predict function need a class column in the dataframe for prediction? 
If TRUE but no class column is avaible in the data a null column is generated 
in predict (default is FALSE).}}}
\section{Extends}{\code{\linkS4class{wrapped.learner}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,wrapped.learner.classif-method]{initialize}}}{\code{signature( = "wrapped.learner.classif")}: Constructor.}}}

