\name{wrapped.learner.regr-class}
\alias{wrapped.learner.regr-class}
\alias{wrapped.learner.regr}
\title{wrapped.learner.regr}
\description{Wraps an already implemented regression method from R to make it accesible to mlr.}
\section{Extends}{\code{\linkS4class{wrapped.learner}}}
\section{Methods}{\describe{\item{\code{\link[=train,wrapped.learner.regr,regr.task,numeric,list,character-method]{train}}}{}}}
