\name{wrapped.model-class}
\alias{wrapped.model-class}
\alias{wrapped.model}
\title{Induced model of learner}
\description{A classifier model is generated by \code{\link{train}}. It consists of the learn.task,
the classification model, the used hyperparameters and a vector specifying the cases of the training sample.}
\section{Slots}{\describe{\item{\code{task.class}:}{(\code{\link{character}}) Class of learn.task from which the model was generated.}\item{\code{wrapped.learner}:}{(\code{\link{wrapped.learner}}) Wrapped learner from which the model was generated.}\item{\code{learner.model}:}{(\code{\link{ANY}}) Underlying learner model from existing R packages.}\item{\code{target}:}{(\code{\link{character}}) Name of the target variable.}\item{\code{subset}:}{(\code{\link{numeric}}) An index vector specifying the cases of the training sample that were used for the model.}\item{\code{vars}:}{(\code{\link{character}}) Vector specifying the variables which were used to build the model.}\item{\code{parset}:}{(\code{\link{list}}) Contains the hyperparameters of the train function. If empty no parameters were used.}}}
\section{Methods}{\describe{{}
