% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{yeast.task}
\alias{yeast.task}
\title{Yeast multilabel classification task.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Yeast} (In long instead of wide format)
}
\description{
Contains the task (\code{yeast.task}).
}
\references{
Elisseeff, A., & Weston, J. (2001):
A kernel method for multi-labelled classification.
In Advances in neural information processing systems (pp. 681-687).
}
\keyword{data}
