% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult_operators.R
\name{addRRMeasure}
\alias{addRRMeasure}
\title{Compute new measures for existing ResampleResult}
\usage{
addRRMeasure(res, measures)
}
\arguments{
\item{res}{[\code{ResampleResult}]\cr
The result of \code{\link{resample}} run with \code{keep.pred = TRUE}.}

\item{measures}{[\code{\link{Measure}} | list of \code{\link{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}
}
\value{
[\code{\link{ResampleResult}}].
}
\description{
Adds new measures to an existing \code{ResampleResult}.
}
\seealso{
Other resample: \code{\link{ResamplePrediction}},
  \code{\link{ResampleResult}},
  \code{\link{getRRPredictionList}},
  \code{\link{getRRPredictions}},
  \code{\link{getRRTaskDescription}},
  \code{\link{makeResampleDesc}},
  \code{\link{makeResampleInstance}},
  \code{\link{resample}}
}

