% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatePredictions.R
\name{aggregatePredictions}
\alias{aggregatePredictions}
\title{Aggregate Predictions}
\usage{
aggregatePredictions(pred.list, spt = NULL)
}
\arguments{
\item{pred.list}{[list of \code{Predictions}]\cr}
}
\description{
Aggregate predicitons results by averaging (for \code{regr}, and  \code{classif} with prob) or mode ( \code{classif} with response). 
(works for regr, classif, multiclass)
}

