% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\docType{data}
\name{aggregations}
\alias{aggregations}
\alias{b632}
\alias{b632plus}
\alias{test.join}
\alias{test.max}
\alias{test.mean}
\alias{test.median}
\alias{test.min}
\alias{test.range}
\alias{test.rmse}
\alias{test.sd}
\alias{test.sum}
\alias{testgroup.mean}
\alias{testgroup.sd}
\alias{train.max}
\alias{train.mean}
\alias{train.median}
\alias{train.min}
\alias{train.range}
\alias{train.rmse}
\alias{train.sd}
\alias{train.sum}
\title{Aggregation methods.}
\format{None}
\usage{
test.mean

test.sd

test.median

test.min

test.max

test.sum

test.range

test.rmse

train.mean

train.sd

train.median

train.min

train.max

train.sum

train.range

train.rmse

b632

b632plus

testgroup.mean

testgroup.sd

test.join
}
\description{
\itemize{
  \item{**test.mean**}{\cr Mean of performance values on test sets.}
  \item{**test.sd**}{\cr Standard deviation of performance values on test sets.}
  \item{**test.median**}{\cr Median of performance values on test sets.}
  \item{**test.min**}{\cr Minimum of performance values on test sets.}
  \item{**test.max**}{\cr Maximum of performance values on test sets.}
  \item{**test.sum**}{\cr Sum of performance values on test sets.}
  \item{**train.mean**}{\cr Mean of performance values on training sets.}
  \item{**train.sd**}{\cr Standard deviation of performance values on training sets.}
  \item{**train.median**}{\cr Median of performance values on training sets.}
  \item{**train.min**}{\cr Minimum of performance values on training sets.}
  \item{**train.max**}{\cr Maximum of performance values on training sets.}
  \item{**train.sum**}{\cr Sum of performance values on training sets.}
  \item{**b632**}{\cr Aggregation for B632 bootstrap.}
  \item{**b632plus**}{\cr Aggregation for B632+ bootstrap.}
  \item{**testgroup.mean**}{\cr Performance values on test sets are grouped according
    to resampling method. The mean for every group is calculated, then the mean of those means.
    Mainly used for repeated CV.}
  \item{**testgroup.sd**}{\cr Similar to **testgroup.mean** - after
    the mean for every group is calculated, the standard deviation of those means is obtained.
    Mainly used for repeated CV.}
  \item{**test.join**}{\cr Performance measure on joined test sets.
    This is especially useful for small sample sizes where unbalanced group sizes have a significant impact
    on the aggregation, especially for cross-validation test.join might make sense now.
    For the repeated CV, the performance is calculated on each repetition and then aggregated
    with the arithmetic mean.}
}
}
\seealso{
[Aggregation]
}
\keyword{datasets}

