% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\docType{data}
\name{aggregations}
\alias{aggregations}
\alias{test.mean}
\alias{test.sd}
\alias{test.median}
\alias{test.min}
\alias{test.max}
\alias{test.sum}
\alias{test.range}
\alias{test.rmse}
\alias{train.mean}
\alias{train.sd}
\alias{train.median}
\alias{train.min}
\alias{train.max}
\alias{train.sum}
\alias{train.range}
\alias{train.rmse}
\alias{b632}
\alias{b632plus}
\alias{testgroup.mean}
\alias{testgroup.sd}
\alias{test.join}
\title{Aggregation methods.}
\description{
\describe{
\item{\strong{test.mean}}{Mean of performance values on test sets.}
\item{\strong{test.sd}}{Standard deviation of performance values on test sets.}
\item{\strong{test.median}}{Median of performance values on test sets.}
\item{\strong{test.min}}{Minimum of performance values on test sets.}
\item{\strong{test.max}}{Maximum of performance values on test sets.}
\item{\strong{test.sum}}{Sum of performance values on test sets.}
\item{\strong{train.mean}}{Mean of performance values on training sets.}
\item{\strong{train.sd}}{Standard deviation of performance values on training sets.}
\item{\strong{train.median}}{Median of performance values on training sets.}
\item{\strong{train.min}}{Minimum of performance values on training sets.}
\item{\strong{train.max}}{Maximum of performance values on training sets.}
\item{\strong{train.sum}}{Sum of performance values on training sets.}
\item{\strong{b632}}{Aggregation for B632 bootstrap.}
\item{\strong{b632plus}}{Aggregation for B632+ bootstrap.}
\item{\strong{testgroup.mean}}{Performance values on test sets are grouped according
to resampling method. The mean for every group is calculated, then the mean of those means.
Mainly used for repeated CV.}
\item{\strong{testgroup.sd}}{Similar to \strong{testgroup.mean} - after
the mean for every group is calculated, the standard deviation of those means is obtained.
Mainly used for repeated CV.}
\item{\strong{test.join}}{Performance measure on joined test sets.
This is especially useful for small sample sizes where unbalanced group sizes have a significant impact
on the aggregation, especially for cross-validation test.join might make sense now.
For the repeated CV, the performance is calculated on each repetition and then aggregated
with the arithmetic mean.}
}
}
\seealso{
\link{Aggregation}
}
\keyword{datasets}
