% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\docType{data}
\name{aggregations}
\alias{aggregations}
\alias{test.mean}
\alias{test.sd}
\alias{test.median}
\alias{test.min}
\alias{test.max}
\alias{test.sum}
\alias{test.range}
\alias{test.rmse}
\alias{train.mean}
\alias{train.sd}
\alias{train.median}
\alias{train.min}
\alias{train.max}
\alias{train.sum}
\alias{train.range}
\alias{train.rmse}
\alias{b632}
\alias{b632plus}
\alias{testgroup.mean}
\alias{testgroup.sd}
\alias{test.join}
\title{Aggregation methods.}
\format{None}
\usage{
test.mean

test.sd

test.median

test.min

test.max

test.sum

test.range

test.rmse

train.mean

train.sd

train.median

train.min

train.max

train.sum

train.range

train.rmse

b632

b632plus

testgroup.mean

testgroup.sd

test.join
}
\description{
\itemize{
\item{\strong{test.mean}}{\cr Mean of performance values on test sets.}
\item{\strong{test.sd}}{\cr Standard deviation of performance values on test sets.}
\item{\strong{test.median}}{\cr Median of performance values on test sets.}
\item{\strong{test.min}}{\cr Minimum of performance values on test sets.}
\item{\strong{test.max}}{\cr Maximum of performance values on test sets.}
\item{\strong{test.sum}}{\cr Sum of performance values on test sets.}
\item{\strong{train.mean}}{\cr Mean of performance values on training sets.}
\item{\strong{train.sd}}{\cr Standard deviation of performance values on training sets.}
\item{\strong{train.median}}{\cr Median of performance values on training sets.}
\item{\strong{train.min}}{\cr Minimum of performance values on training sets.}
\item{\strong{train.max}}{\cr Maximum of performance values on training sets.}
\item{\strong{train.sum}}{\cr Sum of performance values on training sets.}
\item{\strong{b632}}{\cr Aggregation for B632 bootstrap.}
\item{\strong{b632plus}}{\cr Aggregation for B632+ bootstrap.}
\item{\strong{testgroup.mean}}{\cr Performance values on test sets are grouped according
to resampling method. The mean for every group is calculated, then the mean of those means.
Mainly used for repeated CV.}
\item{\strong{testgroup.sd}}{\cr Similar to \strong{testgroup.mean} - after
the mean for every group is calculated, the standard deviation of those means is obtained.
Mainly used for repeated CV.}
\item{\strong{test.join}}{\cr Performance measure on joined test sets.
This is especially useful for small sample sizes where unbalanced group sizes have a significant impact
on the aggregation, especially for cross-validation test.join might make sense now.
For the repeated CV, the performance is calculated on each repetition and then aggregated
with the arithmetic mean.}
}
}
\seealso{
\link{Aggregation}
}
\keyword{datasets}
