% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeatSelResult.R
\name{analyzeFeatSelResult}
\alias{analyzeFeatSelResult}
\title{Show and visualize the steps of feature selection.}
\usage{
analyzeFeatSelResult(res, reduce = TRUE)
}
\arguments{
\item{res}{(\link{FeatSelResult})\cr
The result of of \link{selectFeatures}.}

\item{reduce}{(\code{logical(1)})\cr
Per iteration: Print only the selected feature (or all features that were evaluated)?
Default is \code{TRUE}.}
}
\value{
(\code{invisible(NULL)}).
}
\description{
This function prints the steps \link{selectFeatures} took to find its optimal set
of features and the reason why it stopped.
It can also print information about all calculations done in each intermediate step.

Currently only implemented for sequential feature selection.
}
\seealso{
Other featsel: 
\code{\link{FeatSelControl}},
\code{\link{getFeatSelResult}()},
\code{\link{makeFeatSelWrapper}()},
\code{\link{selectFeatures}()}
}
\concept{featsel}
