% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{applyCPO}
\alias{applyCPO}
\title{CPO Apply}
\usage{
applyCPO(cpo, task)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The CPO representing the operation to perform.}

\item{task}{[\code{\link{Task}}]\cr
The task to operate on.}
}
\description{
The given transformation will be applied to the data in the given \code{link{Task}}.

See the preferred \code{\link{\%>>\%}} for more info.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoMultiplex}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{hasTagInvert}}, \code{\link{inverter<-}},
  \code{\link{inverter}}, \code{\link{makeCPOTargetOp}},
  \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
