% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_hillclimb2.R
\name{applyEnsembleSelection}
\alias{applyEnsembleSelection}
\title{ensemble selection algo:}
\usage{
applyEnsembleSelection(pred.list = pred.list, bls.length = bls.length,
  bls.names = bls.names, bls.performance = bls.performance,
  parset = list(replace = TRUE, init = 1, bagprob = 1, bagtime = 1, metric =
  NULL, maxiter = NULL, tolerance = 1e-08))
}
\arguments{
\item{pred.list}{pred.list}

\item{bls.length}{bls.length}

\item{bls.names}{bls.names}

\item{bls.performance}{bls.performance}

\item{parset}{parset}

\item{parset}{list of parset}
}
\description{
ensemble selection algo:
}

