% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asROCRPrediction.R
\name{asROCRPrediction}
\alias{asROCRPrediction}
\title{Converts predictions to a format package ROCR can handle.}
\usage{
asROCRPrediction(pred)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}
}
\description{
Converts predictions to a format package ROCR can handle.
}
\seealso{
Other predict: \code{\link{generateROCRCurvesData}};
  \code{\link{getProbabilities}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotViperCharts}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}

Other roc: \code{\link{generateROCRCurvesData}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotViperCharts}}
}

