\name{asROCRPrediction}
\alias{asROCRPrediction}
\title{Converts predictions to a format package ROCR can handle.}
\usage{
asROCRPrediction(pred)
}
\arguments{
  \item{pred}{[\code{\link{Prediction}}]\cr Prediction
  object.}
}
\description{
Converts predictions to a format package ROCR can handle.
}

