% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asROCRPrediction.R
\name{asROCRPrediction}
\alias{asROCRPrediction}
\title{Converts predictions to a format package ROCR can handle.}
\usage{
asROCRPrediction(pred)
}
\arguments{
\item{pred}{(\link{Prediction})\cr
Prediction object.}
}
\description{
Converts predictions to a format package ROCR can handle.
}
\seealso{
Other roc: 
\code{\link{calculateROCMeasures}()}

Other predict: 
\code{\link{getPredictionProbabilities}()},
\code{\link{getPredictionResponse}()},
\code{\link{getPredictionTaskDesc}()},
\code{\link{predict.WrappedModel}()},
\code{\link{setPredictThreshold}()},
\code{\link{setPredictType}()}
}
\concept{predict}
\concept{roc}
