% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{attachCPO}
\alias{attachCPO}
\title{CPO Attachment}
\usage{
attachCPO(cpo, learner)
}
\arguments{
\item{cpo}{[\code{\link{CPO}}]\cr
The CPO object}

\item{learner}{[\code{\link{Learner}}]\cr
The learner.}
}
\description{
The second argument is a \code{\link{Learner}} and the CPO will be attached to
this learner. The same operation will be performed during the \dQuote{train} and
\dQuote{predict} phase; the behaviour during the predict phase may furthermore
be depend on the training data.

See the preferred \code{\link{\%>>\%}} for more info.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{makeCPOFunctional}},
  \code{\link{makeCPOObject}}, \code{\link{retrafo<-}},
  \code{\link{retrafo}}, \code{\link{setHyperPars}}
}
