% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bc.task}
\alias{bc.task}
\title{Wisconsin Breast Cancer classification task.}
\description{
Contains the task (\code{bc.task}).
}
\references{
See \link[mlbench:BreastCancer]{mlbench::BreastCancer}.
The column \code{"Id"} and all incomplete cases have been removed from the task.
}
\keyword{data}
