% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Result of a benchmark run.}
\description{
Container for results of benchmarked experiments using \code{\link{benchmark}}.
The structure of the object itself is rather complicated, it is recommended to
retrive required information via the \code{getBMR*} getter functions.
You can also convert the object using \code{\link[base]{as.data.frame}}.
}
\seealso{
Other benchmark: \code{\link{FilterResult}},
  \code{\link{getFilterResult}}; \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRLearnerIds}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}};
  \code{\link{getFeatSelResult}}
}

