% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Result of a benchmark run.}
\description{
Container for results of benchmarked experiments using \code{\link{benchmark}}.
The structure of the object itself is rather complicated, it is recommended to
retrive required information via \code{link{getAggrPerformances}}, \code{link{getPredictions}},
\code{\link{getPerformances}}, \code{\link{getFeatSelResult}}, \code{\link{getTuneResult}} or
\code{\link{getFilterResult}}. Alternatively, you can convert the object using
\code{\link[base]{as.data.frame}}
}
\seealso{
Other benchmark: \code{\link{FilterResult}},
  \code{\link{getFilterResult}}; \code{\link{benchmark}};
  \code{\link{getAggrPerformances}};
  \code{\link{getFeatSelResult}};
  \code{\link{getPerformances}};
  \code{\link{getPredictions}}; \code{\link{getTuneResult}}
}

