% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bh.task}
\alias{bh.task}
\title{Boston Housing regression task.}
\description{
Contains the task (\code{bh.task}).
}
\references{
See \link[mlbench:BostonHousing]{mlbench::BostonHousing}.
}
\keyword{data}
