% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateROCMeasures.R
\name{calculateROCMeasures}
\alias{calculateROCMeasures}
\alias{print.ROCMeasures}
\title{Calculate receiver operator measures.}
\usage{
calculateROCMeasures(pred)

\method{print}{ROCMeasures}(x, abbreviations = TRUE, digits = 2, ...)
}
\arguments{
\item{pred}{([Prediction])\cr
Prediction object.}

\item{x}{(`ROCMeasures`)\cr
Created by [calculateROCMeasures].}

\item{abbreviations}{(`logical(1)`)\cr
If `TRUE` a short paragraph with explanations of the used measures is printed additionally.}

\item{digits}{(`integer(1)`)\cr
Number of digits the measures are rounded to.}

\item{...}{`(any)`\cr
Currently not used.}
}
\value{
(`ROCMeasures`).
   A list containing two elements `confusion.matrix` which is
   the 2 times 2 confusion matrix of absolute frequencies and `measures`, a list of the above mentioned measures.
}
\description{
Calculate the absolute number of correct/incorrect classifications and the following evaluation measures:

* `tpr` True positive rate (Sensitivity, Recall)
* `fpr` False positive rate (Fall-out)
* `fnr` False negative rate (Miss rate)
* `tnr` True negative rate (Specificity)
* `ppv` Positive predictive value (Precision)
* `for` False omission rate
* `lrp` Positive likelihood ratio (LR+)
* `fdr` False discovery rate
* `npv` Negative predictive value
* `acc` Accuracy
* `lrm` Negative likelihood ratio (LR-)
* `dor` Diagnostic odds ratio

For details on the used measures see [measures] and also
<https://en.wikipedia.org/wiki/Receiver_operating_characteristic>.

The element for the false omission rate in the resulting object is not called `for` but
`fomr` since `for` should never be used as a variable name in an object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}
\examples{
lrn = makeLearner("classif.rpart", predict.type = "prob")
fit = train(lrn, sonar.task)
pred = predict(fit, task = sonar.task)
calculateROCMeasures(pred)

}
\seealso{
Other performance: \code{\link{ConfusionMatrix}},
  \code{\link{calculateConfusionMatrix}},
  \code{\link{estimateRelativeOverfitting}},
  \code{\link{makeCostMeasure}},
  \code{\link{makeCustomResampledMeasure}},
  \code{\link{makeMeasure}}, \code{\link{measures}},
  \code{\link{performance}}, \code{\link{setAggregation}},
  \code{\link{setMeasurePars}}

Other roc: \code{\link{asROCRPrediction}}
}

