% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capLargeValues.R
\name{capLargeValues}
\alias{capLargeValues}
\title{Convert large/infinite numeric values in a data.frame or task.}
\usage{
capLargeValues(obj, target = character(0L), cols = NULL, threshold = Inf,
  impute = threshold, what = "abs")
}
\arguments{
\item{obj}{([data.frame] | [Task])\cr
Input data.}

\item{target}{([character])\cr
Name of the column(s) specifying the response.
Target columns will not be capped.
Default is `character(0)`.}

\item{cols}{([character])\cr
Which columns to convert.
Default is all numeric columns.}

\item{threshold}{(`numeric(1)`)\cr
Threshold for capping.
Every entry whose absolute value is equal or larger is converted.
Default is `Inf`.}

\item{impute}{(`numeric(1)`)\cr
Replacement value for large entries.
Large negative entries are converted to `-impute`.
Default is `threshold`.}

\item{what}{(`character(1)`)\cr
What kind of entries are affected?
\dQuote{abs} means `abs(x) > threshold`,
\dQuote{pos} means `abs(x) > threshold && x > 0`,
\dQuote{neg} means `abs(x) > threshold && x < 0`.
Default is \dQuote{abs}.}
}
\value{
([data.frame])
}
\description{
Convert numeric entries which large/infinite (absolute) values
in a data.frame or task.
Only numeric/integer columns are affected.
}
\examples{
capLargeValues(iris, threshold = 5, impute = 5)
}
\seealso{
Other eda_and_preprocess: \code{\link{createDummyFeatures}},
  \code{\link{dropFeatures}},
  \code{\link{mergeSmallFactorLevels}},
  \code{\link{normalizeFeatures}},
  \code{\link{removeConstantFeatures}},
  \code{\link{summarizeColumns}}
}

