% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMeasures.R
\name{checkMeasures}
\alias{checkMeasures}
\title{Checks for correct strucutre of selected measures}
\usage{
checkMeasures(measures, obj, aggr = NULL, na.rm = FALSE)
}
\arguments{
\item{measures}{\code{\link{list}}\cr
list of measures}

\item{obj}{\code{\link{data.frame}}\cr
task}

\item{aggr}{\code{\link{function}}\cr
aggregation function}

\item{na.rm}{(\code{logical(1)})\cr
Should \code{NA} values be removed? Default \code{FALSE}.
This applies to all selected measures.}
}
\description{
Checks for correct strucutre of selected measures
}
\keyword{internal}
