% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLearner.R
\name{checkPredictLearnerOutput}
\alias{checkPredictLearnerOutput}
\title{Check output returned by predictLearner.}
\usage{
checkPredictLearnerOutput(learner, model, p)
}
\arguments{
\item{learner}{(\link{Learner})\cr
The learner.}

\item{model}{(\link{WrappedModel})]\cr
Model produced by training.}

\item{p}{(any)\cr
The prediction made by \code{learner}.}
}
\value{
(any). A sanitized version of \code{p}.
}
\description{
Check the output coming from a Learner's internal
\code{predictLearner} function.

This function is for internal use.
}
\keyword{internal}
