% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_classif_featureless.R
\name{classif.featureless}
\alias{classif.featureless}
\title{Featureless classification learner.}
\description{
Very basic baseline method, mainly useful in model comparisons (if you don't
beat this you very likely have a problem).
Does not consider features of the task and only looks at labels on training
data.

Method \dQuote{majority} predicts always the majority class for each new
observation. In case of ties, we pick a randomly sampled class and always
predict that class for each new observation.

Method \dQuote{sample-prior} samples a random label for each new observation,
where each label probability coincied with the proportion of that label
obsered in the training data.

If you opt to predict probabilities, the label is always computed by mlr
based on the prior probabilities obsered in the training data.
}

