% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_classif_featureless.R
\name{classif.featureless}
\alias{classif.featureless}
\title{Featureless classification learner.}
\description{
A very basic baseline method which is useful for model comparisons (if you
don't beat this, you very likely have a problem).
Does not consider any features of the task and only uses the target feature
of the training data to make predictions.
Using observation weights is currently not supported.

Method \dQuote{majority} predicts always the majority class for each new
observation. In case of ties, a randomly sampled class is predicted.
This method is very similar to the ZeroR classifier from WEKA (see
\url{https://weka.wikispaces.com/ZeroR}). The difference is that ZeroR
always predicts the first class of the tied class values instead of sampling
them randomly.

Method \dQuote{sample-prior} always samples a random class for each new
observation according to the prior probabilities observed in the training data.

If you opt to predict probabilities, the class probabilities always
correspond to the prior probabilities observed in the training data.
}

