\name{classif.task-class}
\alias{classif.task-class}
\alias{classif.task}
\title{Classification task.}
\description{General description object for a classification experiment.   
Instantiate it by using its factory method.}
\seealso{\code{\link{make.task}}}
\section{Extends}{\code{\linkS4class{learn.task}}}
\section{Methods}{\describe{\item{\code{\link[=to.string,classif.task-method]{to.string}}}{}}}
