\name{combine.ranges}
\alias{combine.ranges}
\title{Combine non-orthogonal ranges}
\usage{combine.ranges(...)
}
\description{Convenience function for \code{\link{tune}} to combine different \code{ranges} lists (i.e. lists of named vectors/lists 
of possible hyperparameter values).}
\value{A list of \code{ranges}}
\seealso{\code{\link{tune}}}
\arguments{\item{...}{Lists of ranges.}
}
