% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBMRToRankMatrix.R
\name{convertBMRToRankMatrix}
\alias{convertBMRToRankMatrix}
\title{Convert BenchmarkResult to a rank-matrix.}
\usage{
convertBMRToRankMatrix(
  bmr,
  measure = NULL,
  ties.method = "average",
  aggregation = "default"
)
}
\arguments{
\item{bmr}{(\link{BenchmarkResult})\cr
Benchmark result.}

\item{measure}{(\link{Measure})\cr
Performance measure.
Default is the first measure used in the benchmark experiment.}

\item{ties.method}{(\code{character(1)})\cr
See \link[base:rank]{base::rank} for details.}

\item{aggregation}{(\code{character(1)}) \cr
\dQuote{mean} or \dQuote{default}. See \link{getBMRAggrPerformances}
for details on \dQuote{default}.}
}
\value{
(\link{matrix}) with measure ranks as entries.
The matrix has one row for each \code{learner}, and one column for each \code{task}.
}
\description{
Computes a matrix of all the ranks of different algorithms
over different datasets (tasks). Ranks are computed from aggregated
measures.
Smaller ranks imply better methods, so for measures that are minimized, small ranks imply small scores.
for measures that are maximized, small ranks imply large scores.
}
\examples{
# see benchmark
}
\seealso{
Other benchmark: 
\code{\link{BenchmarkResult}},
\code{\link{batchmark}()},
\code{\link{benchmark}()},
\code{\link{friedmanPostHocTestBMR}()},
\code{\link{friedmanTestBMR}()},
\code{\link{generateCritDifferencesData}()},
\code{\link{getBMRAggrPerformances}()},
\code{\link{getBMRFeatSelResults}()},
\code{\link{getBMRFilteredFeatures}()},
\code{\link{getBMRLearnerIds}()},
\code{\link{getBMRLearnerShortNames}()},
\code{\link{getBMRLearners}()},
\code{\link{getBMRMeasureIds}()},
\code{\link{getBMRMeasures}()},
\code{\link{getBMRModels}()},
\code{\link{getBMRPerformances}()},
\code{\link{getBMRPredictions}()},
\code{\link{getBMRTaskDescs}()},
\code{\link{getBMRTaskIds}()},
\code{\link{getBMRTuneResults}()},
\code{\link{plotBMRBoxplots}()},
\code{\link{plotBMRRanksAsBarChart}()},
\code{\link{plotBMRSummary}()},
\code{\link{plotCritDifferences}()},
\code{\link{reduceBatchmarkResults}()}
}
\concept{benchmark}
