% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSfeatures.R
\name{convertTSTaskToNormalTask}
\alias{convertTSTaskToNormalTask}
\title{Create task on time series features.}
\usage{
convertTSTaskToNormalTask(task, method, pars = NULL)
}
\arguments{
\item{task}{[\code{TimeSeriesClassifTask}]\cr
Time series classification task.}

\item{method}{[\code{character(1)}]\cr
Which method is used to create time series features. Two methods available.
Wavelet transformation: \dQuote{wavelets}.
Fourier transformation: \dQuote{fourier}.}

\item{pars}{\cr
Further parameters passed as argument e.g., for feature representation
methods.
Wavelet transformation: \code{filter} and \code{boundary}.
Fourier transformation: \code{fft.coeff}.}
}
\value{
Either [\code{ClassifTask}] or the learned shapelet model.
}
\description{
The function creates a feature representation of raw time series
  for a time series classification task. The method used to create this
  feature representation must be specified by the user in \code{method}.
  For \code{wavelets} or \code{fourier} features, the resulting data does not
  contain temporal structure anymore, so the returned task is a \code{ClassifTask}.
  For \code{shapelets}, the learned shapelet model is returned. See
  \code{\link{getTSShapeletFeatures}}.
}

