% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_re2.R
\name{createNewParset}
\alias{createNewParset}
\title{Create new parset}
\usage{
createNewParset(org.parset, new.parset)
}
\arguments{
\item{org.parset}{orginal/old parset from obj}

\item{new.parset}{parameters which should be updated}
}
\description{
Create new parset
}

