% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_recombine.R
\name{createTask}
\alias{createTask}
\title{Create a Classif or Regr Task}
\usage{
createTask(type, data, target)
}
\arguments{
\item{type}{[\code{character(1)}]\n Use "classif" for Classification and "regr" for regression.}

\item{data}{[\code{data.frame}]\n  Data to use.}

\item{target}{[\code{character(1)}]\n Target to use.}
}
\description{
Create a Classif or Regr Task
}

