% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_re2.R
\name{createTestPreds}
\alias{createTestPreds}
\title{Create Predictions for Testing Set}
\usage{
createTestPreds(i, bls, test.idx, task, save.on.disc)
}
\arguments{
\item{i}{number of fold}

\item{bls}{base.learner to use}

\item{test.idx}{idx for subsetting}

\item{task}{task}

\item{save.on.disc}{wether model are presend in bls or must be loaded using readRDS}
}
\description{
Create Predictions for Testing Set
}

