% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover.}
\arguments{
\item{x}{(\link{logical})\cr
First parent string.}

\item{y}{(\link{logical})\cr
Second parent string.}

\item{rate}{(\code{numeric(1)})\cr
A number representing the probability of selecting an element of the first string.
Default is \code{0.5}.}
}
\value{
(\link{crossover}).
}
\description{
Takes two bit strings and creates a new one of the same size by selecting the items from the first string or
the second, based on a given rate (the probability of choosing an element from the first string).
}
