\name{data.desc-class}
\alias{data.desc-class}
\alias{data.desc}
\title{Description object for data.}
\description{Description object for data.}
\details{\code{data.desc-class}: Getter.\cr

\describe{
\item{target [string]}{Name of target variable.}
\item{excluded [character]}{Names of excluded covariates.}
\item{size [integer]}{Number of cases.}
\item{dim [integer]}{Number of covariates.}
\item{excluded [character]}{Names of excluded variables.}
\item{n.num [integer]}{Number of numerical covariates.}
\item{n.int [integer]}{Number of integer covariates.}
\item{n.fact [integer]}{Number of factor covariates.}
\item{n.char [integer]}{Number of character covariates.}
\item{has.missing [boolean]}{Are missing values present?}
\item{has.inf [boolean]}{Are infinite numerical values present?}
\item{is.classif [boolean]}{Factor target variable?}
\item{is.regr [boolean]}{Numerical target variable?}
\item{class.levels [character]}{Possible classes. NA if not classification.}
\item{class.nr [integer]}{Number of classes. NA if not classification.}
\item{class.dist [integer]}{Class distribution. Named vector. NA if not classification.}
\item{is.binary [boolean]}{Binary classification?. NA if not classification.}
\item{prepare.control [prepare.control]}{Control object used for preparing the original data.frame.}
}

}
\section{Extends}{\code{\linkS4class{object}}}
\section{Methods}{\describe{\item{\code{\link[=train.learner,classif.ada,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.adaboost.M1,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,base.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.blackboost,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.blackboost,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.ctree,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=[,data.desc-method]{[}}}{}\item{\code{\link[=train.learner,regr.earth,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,filter.wrapper,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.gbm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.gbm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.glmboost,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.grplasso,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.J48,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.JRip,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.kknn,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.kknn,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.km,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.ksvm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.ksvm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.lasso,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.lm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.loclda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.logreg,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lssvm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lvq1,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.mars,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.mda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,multiclass.wrapper,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.multinom,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.naiveBayes,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.nnet,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.nnet,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.OneR,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,opt.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.PART,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.pcr,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.lpsvm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,preproc.wrapper,character,data.frame,data.desc,task.desc,numeric,ANY-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.qda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.randomForest,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.randomForest,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.rda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.ridge,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.rpart,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.rpart,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.rsm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,regr.rvm,character,data.frame,data.desc,task.desc,numeric,missing-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.sda,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}\item{\code{\link[=train.learner,classif.svm,character,data.frame,data.desc,task.desc,numeric,matrix-method]{train.learner}}}{}}}
\alias{[,data.desc-method}

