% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{doTrainPredict}
\alias{doTrainPredict}
\title{Training and prediction in one function (used for parallelMap)}
\usage{
doTrainPredict(bls, task, show.info, id, save.on.disc)
}
\arguments{
\item{bls}{[list of base.learner]}

\item{task}{[Task]}

\item{show.info}{show.info}

\item{id}{Id needed to create unique model name}

\item{save.on.disc}{save.on.disc}
}
\description{
Training and prediction in one function (used for parallelMap)
}

