% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downsample}
\alias{downsample}
\title{Downsample (subsample) a task or a data.frame.}
\usage{
downsample(obj, target, percentage = NULL, n = NULL, stratify = FALSE)
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{SupervisedTask}} | \code{\link{ResampleInstance}}]\cr
Input data or a \code{ResampleInstance}.}

\item{target}{[\code{character(1)}]\cr
Name of the column specifying the response.
Only used when \code{obj} is a \code{data.frame} and \code{stratify = TRUE}, otherwise ignored.}

\item{percentage}{[\code{numeric(1)}]\cr
Percentage of the observations to keep in the downsampled data.}

\item{n}{[\code{numeric(1)}]\cr
Total number of observations to keep in the downsampled data.}

\item{stratify}{[\code{boolean(1)}]\cr
Should the downsample be stratified according to the target classes?}
}
\value{
[\code{data.frame} | \code{\link{SupervisedTask}} | \code{\link{ResampleInstance}}]. Same type as \code{obj}.
}
\description{
Decrease the observations in a \code{task} or a \code{data.frame}
  to a given percentage or absolute number.
}
\seealso{
\code{\link{makeResampleInstance}}
}

