% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample (subsample) a task or a data.frame.}
\usage{
downsample(obj, perc = 1, stratify = FALSE)
}
\arguments{
\item{obj}{([Task] | [ResampleInstance])\cr
Input data or a `ResampleInstance`.}

\item{perc}{(`numeric(1)`)\cr
Percentage from (0, 1).
Default is 1.}

\item{stratify}{(`logical(1)`)\cr
Only for classification:
Should the downsampled data be stratified according to the target classes?
Default is `FALSE`.}
}
\value{
([data.frame` | [Task] | [ResampleInstance]). Same type as `obj`.
}
\description{
Decrease the observations in a `task` or a `ResampleInstance`
to a given percentage of observations.
}
\seealso{
[makeResampleInstance]

Other downsample: \code{\link{makeDownsampleWrapper}}
}

