\name{dropFeatures}
\alias{dropFeatures}
\title{Drop some features of task.}
\usage{
dropFeatures(task, features)
}
\arguments{
  \item{task}{[\code{\link{Task}}]\cr The task.}

  \item{features}{[\code{character}]\cr Features to drop.}
}
\value{
[\code{\link{Task}}].
}
\description{
Drop some features of task.
}

