% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dropFeatures}
\alias{dropFeatures}
\title{Drop some features of task.}
\usage{
dropFeatures(task, features)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{features}{[\code{character}]\cr
Features to drop.}
}
\value{
[\code{\link{Task}}].
}
\description{
Drop some features of task.
}
\seealso{
Other eda_preproc: \code{\link{createDummyFeatures}};
  \code{\link{normalizeFeatures}};
  \code{\link{removeConstantFeatures}}
}

