% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropFeatures.R
\name{dropFeatures}
\alias{dropFeatures}
\title{Drop some features of task.}
\usage{
dropFeatures(task, features)
}
\arguments{
\item{task}{(\link{Task})\cr
The task.}

\item{features}{(\link{character})\cr
Features to drop.}
}
\value{
\link{Task}.
}
\description{
Drop some features of task.
}
\seealso{
Other eda_and_preprocess: 
\code{\link{capLargeValues}()},
\code{\link{createDummyFeatures}()},
\code{\link{mergeSmallFactorLevels}()},
\code{\link{normalizeFeatures}()},
\code{\link{removeConstantFeatures}()},
\code{\link{summarizeColumns}()},
\code{\link{summarizeLevels}()}
}
\concept{eda_and_preprocess}
