% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_train.R
\name{ensembleselectionBaseLearners}
\alias{ensembleselectionBaseLearners}
\title{Train function for "Ensemble Selection" method.}
\usage{
ensembleselectionBaseLearners(learner, task, replace = TRUE, init = 1,
  bagprob = 1, bagtime = 1, maxiter = NULL, tolerance = 1e-08,
  metric = NULL, ...)
}
\arguments{
\item{learner}{[list of base learners]}

\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{replace}{[\code{logical(1)}]}

\item{init}{[\code{integer(1)}] init >= 1}

\item{bagprob}{[\code{numeric(1)}] 0 < bagprob < 1}

\item{bagtime}{[\code{integer(1)}] bagtime >= 1}

\item{maxiter}{[\code{integer(1)}] maxiter >= 1}

\item{tolerance}{[\code{numeric(1)}] small numeric value.}

\item{measures}{[\code{\link{Measure}} | list of \code{\link{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}
}
\description{
Train function for "Ensemble Selection" method.
}

