% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativeOverfitting.R
\name{estimateRelativeOverfitting}
\alias{estimateRelativeOverfitting}
\alias{estimateRelativeOverfitting.ResampleDesc}
\title{Estimate relative overfitting.}
\usage{
estimateRelativeOverfitting(rdesc, measures, task, learner)

\method{estimateRelativeOverfitting}{ResampleDesc}(rdesc, measures, task,
  learner)
}
\arguments{
\item{rdesc}{[\code{\link{ResampleDesc}}\cr
Resampling strategy.}

\item{measures}{[\code{\link{Measure}} | list of \code{\link{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}

\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{data.frame}]. Relative overfitting estimate(s), named by measure(s), for each resampling iteration.
}
\description{
Estimates the relative overfitting of a model as the ratio of the difference in test and train performance to the difference of test performance in the no-information case and train performance.
In the no-information case the features carry no information with respect to the prediction. This is simulated by permuting features and predictions.
}
\details{
Currently only support for classification and regression tasks is implemented.
}
\examples{
task = makeClassifTask(data = iris, target = "Species")
rdesc = makeResampleDesc("CV", iters = 2)
estimateRelativeOverfitting(rdesc, acc, task, makeLearner("classif.knn"))
estimateRelativeOverfitting(rdesc, acc, task, makeLearner("classif.lda"))
}
\references{
Bradley Efron and Robert Tibshirani; Improvements on Cross-Validation: The .632+ Bootstrap Method, Journal of the American Statistical Association, Vol. 92, No. 438. (Jun., 1997), pp. 548-560.
}
\seealso{
Other performance: \code{\link{G1}}, \code{\link{G2}},
  \code{\link{acc}}, \code{\link{auc}}, \code{\link{bac}},
  \code{\link{ber}}, \code{\link{brier}},
  \code{\link{cindex}}, \code{\link{db}},
  \code{\link{dunn}}, \code{\link{f1}},
  \code{\link{f1mult}}, \code{\link{fdr}},
  \code{\link{featperc}}, \code{\link{fn}},
  \code{\link{fnr}}, \code{\link{fp}}, \code{\link{fpr}},
  \code{\link{gmean}}, \code{\link{gpr}},
  \code{\link{hamloss}}, \code{\link{jaccard}},
  \code{\link{mae}}, \code{\link{mcc}}, \code{\link{mcp}},
  \code{\link{meancosts}}, \code{\link{measureACC}},
  \code{\link{measureAUC}}, \code{\link{measureBAC}},
  \code{\link{measureBrier}}, \code{\link{measureF1MULT}},
  \code{\link{measureFDR}}, \code{\link{measureFN}},
  \code{\link{measureFNR}}, \code{\link{measureFP}},
  \code{\link{measureFPR}}, \code{\link{measureGMEAN}},
  \code{\link{measureGPR}}, \code{\link{measureHAMLOSS}},
  \code{\link{measureJACCARD}}, \code{\link{measureMAE}},
  \code{\link{measureMCC}}, \code{\link{measureMEDAE}},
  \code{\link{measureMEDSE}}, \code{\link{measureMMCE}},
  \code{\link{measureMSE}}, \code{\link{measureNPV}},
  \code{\link{measurePPV}}, \code{\link{measureRMSE}},
  \code{\link{measureSAE}}, \code{\link{measureSSE}},
  \code{\link{measureSUBSET01}}, \code{\link{measureTN}},
  \code{\link{measureTNR}}, \code{\link{measureTP}},
  \code{\link{measureTPR}}, \code{\link{measures}},
  \code{\link{medae}}, \code{\link{medse}},
  \code{\link{mmce}}, \code{\link{mse}},
  \code{\link{multiclass.auc}}, \code{\link{npv}},
  \code{\link{ppv}}, \code{\link{rmse}}, \code{\link{sae}},
  \code{\link{silhouette}}, \code{\link{sse}},
  \code{\link{subset01}}, \code{\link{timeboth}},
  \code{\link{timepredict}}, \code{\link{timetrain}},
  \code{\link{tn}}, \code{\link{tnr}}, \code{\link{tp}},
  \code{\link{tpr}}; \code{\link{Measure}},
  \code{\link{makeMeasure}}; \code{\link{makeCostMeasure}};
  \code{\link{makeCustomResampledMeasure}};
  \code{\link{performance}}
}

