% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDAFeaturesMethods.R
\name{extractFDAFPCA}
\alias{extractFDAFPCA}
\title{Extract functional principal component analysis features.}
\usage{
extractFDAFPCA(pve = 0.99, npc = NULL)
}
\arguments{
\item{pve}{[\code{numeric}]\cr
Fraction of variance explained for the functional principal components.
Default is 0.99.}

\item{npc}{[\code{integer}]\cr
Number of principal components to extract. Overrides \code{pve} param.
Default is \code{NULL}}
}
\value{
[\code{data.frame}].
}
\description{
The function extracts the functional principal components from a data.frame
containing functional features.
}
\seealso{
Other fda_featextractor: \code{\link{extractFDAFourier}},
  \code{\link{extractFDAMultiResFeatures}},
  \code{\link{extractFDAWavelets}}
}
\concept{fda_featextractor}
