% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFDAFeaturesMethods.R
\name{extractFDAWavelets}
\alias{extractFDAWavelets}
\title{Discrete Wavelet transform features.}
\usage{
extractFDAWavelets(filter = "la8", boundary = "periodic")
}
\arguments{
\item{filter}{[\code{character}]\cr
Specifies which filter should be used.
Default is \dQuote{la8}.}

\item{boundary}{[\code{character}]\cr
Boundary to be used.
\dQuote{periodic} assumes circular time series,
for \dQuote{reflection} the series is extended to twice its length.
Default is \dQuote{periodic}.}
}
\value{
[\code{data.frame}].
}
\description{
The function extracts discrete wavelet transform coefficients from the raw
functional data.
See \code{\link[wavelets]{dwt}} for more information.
}
\seealso{
Other fda_featextractor: \code{\link{extractFDAFPCA}},
  \code{\link{extractFDAFourier}},
  \code{\link{extractFDAMultiResFeatures}}
}
\concept{fda_featextractor}
