% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSelResult.R
\name{FeatSelResult}
\alias{FeatSelResult}
\title{Result of feature selection.}
\description{
Container for results of feature selection.
Contains the obtained features, their performance values
and the optimization path which lead there.  \cr
You can visualize it using [analyzeFeatSelResult].
}
\details{
Object members:
\describe{
\item{learner ([Learner])}{Learner that was optimized.}
\item{control ([FeatSelControl])}{ Control object from feature selection.}
\item{x ([character])}{Vector of feature names identified as optimal.}
\item{y ([numeric])}{Performance values for optimal `x`.}
\item{threshold ([numeric])}{Vector of finally found and used thresholds
  if `tune.threshold` was enabled in [FeatSelControl], otherwise not present and
  hence `NULL`.}
\item{opt.path ([ParamHelpers::OptPath])}{Optimization path which lead to `x`.}
}
}

