% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter features by thresholding importance values.}
\usage{
filterFeatures(obj, target, feat.importance, threshold = 0, n = NULL,
  percentage = NULL)
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{SupervisedTask}}]\cr
Input data.}

\item{target}{[\code{character(1)}]\cr
Name of the column specifying the response.
Only used when \code{obj} is a data.frame, otherwise ignored.}

\item{feat.importance}{[\code{numeric}]\cr
Importance values for all features. See \code{\link{getFeatureFilterValues}}.}

\item{threshold}{[\code{numeric(1)}]\cr
Information value as to be greater then the threshold. Default is 0.}

\item{n}{[\code{integer(1)}]\cr
Number of features ordered by the information value to select.}

\item{percentage}{[\code{numeric(1)}]\cr
Alternatively to \code{n} you can give a relative number of features.}
}
\value{
[\code{data.frame} | \code{\link{SupervisedTask}}]. Same type as \code{obj}.
}
\description{
Selects the features by importance values and according to a given threashold.
}
\seealso{
\code{\link{makeFilterWrapper}}
}

