% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter features by thresholding filter values.}
\usage{
filterFeatures(task, method = "rf.importance", fval = NULL, perc = NULL,
  abs = NULL, threshold = NULL, ...)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{method}{[\code{character(1)}]\cr
See \code{\link{listFilterMethods}}.
Default is \dQuote{rf.importance}.}

\item{fval}{[\code{\link{FilterValues}}]\cr
Result of \code{\link{getFilterValues}}.
If you pass this, the filter values in the object are used for feature filtering.
\code{method} and \code{...} are ignored then.
Default is \code{NULL} and not used.}

\item{perc}{[\code{numeric(1)}]\cr
If set, select \code{perc}*100 top scoring features.
Mutually exclusive with arguments \code{abs} and \code{threshold}.}

\item{abs}{[\code{numeric(1)}]\cr
If set, select \code{abs} top scoring features.
Mutually exclusive with arguments \code{perc} and \code{threshold}.}

\item{threshold}{[\code{numeric(1)}]\cr
If set, select features whose score exceeds \code{threshold}.
Mutually exclusive with arguments \code{perc} and \code{abs}.}

\item{...}{[any]\cr
Passed down to selected filter method.}
}
\value{
[\code{\link{Task}}].
}
\description{
First, calls \code{\link{getFilterValues}}.
Features are then selected via \code{select} and \code{val}.
}
\seealso{
Other filter: \code{\link{FilterResult}},
  \code{\link{getFilterResult}};
  \code{\link{FilterValues}};
  \code{\link{getFilterValues}};
  \code{\link{getFilteredFeatures}};
  \code{\link{makeFilterWrapper}}
}

