% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter features by thresholding importance values.}
\usage{
filterFeatures(obj, target, method = "random.forest.importance",
  select = "perc", val)
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{SupervisedTask}}]\cr
Input data.}

\item{target}{[\code{character(1)} | \code{character(2)}]\cr
Name of the target variable(s).
For survival analysis these are the names of the survival time and event columns,
so it has length 2.}

\item{method}{[\code{character(1)}]\cr
See \code{\link{getFilterValues}}.
Default is \dQuote{random.forest.importance}.}

\item{select}{[\code{character(1)}]\cr
How to select top features.
\dQuote{perc} = select top-scoring percentage, \dQuote{abd} = select number
of top-scoring features, \dQuote{threshold} = select all features whose
criterion value is >= \code{val}.
Default is \dQuote{perc}.}

\item{val}{[\code{numeric(1)}]\cr
Depends on \code{select}:
Either a percentage from [0, 1], a number of features or a threshold value for the criterion.}
}
\value{
[\code{data.frame} | \code{\link{SupervisedTask}}]. Same type as \code{obj}.
}
\description{
First, calls \code{\link{getFilterValues}}.
Features are then selected via one of: \code{fw.perc}, \code{fw.n} or \code{fw.threshold}.
}
\seealso{
Other filter: \code{\link{getFilterMethods}};
  \code{\link{getFilterValues}}
}

