% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateFilterValues.R
\name{FilterValues}
\alias{FilterValues}
\title{Result of \code{\link{generateFilterValuesData}}.}
\description{
Result of \code{\link{generateFilterValuesData}}.
}
\seealso{
Other filter: \code{\link{filterFeatures}};
  \code{\link{generateFilterValuesData}};
  \code{\link{getFilterValues}};
  \code{\link{getFilteredFeatures}};
  \code{\link{makeFilterWrapper}};
  \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}}
}

