% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFilterValues.R
\name{FilterValues}
\alias{FilterValues}
\title{Result of \code{\link{getFilterValues}}.}
\description{
\itemize{
  \item{task.desc [\code{\link{TaskDesc}}]}{Task description.}
  \item{method [\code{character}]}{Filter method.}
  \item{data [\code{data.frame}]}{Has columns: \code{name} = Names of features;
    \code{val} = Feature importance values; \code{type} = Feature column type.}
}
}
\seealso{
Other filter: \code{\link{filterFeatures}};
  \code{\link{getFilterValues}};
  \code{\link{getFilteredFeatures}};
  \code{\link{makeFilterWrapper}}
}

