% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FilterValues}
\alias{FilterValues}
\title{Result of \code{\link{getFilterValues}}.}
\description{
\itemize{
  \item{method [\code{character}]}{Filter method.}
  \item{data [\code{data.frame}]}{Has columns: \code{name} = Names of features;
    \code{val} = Feature importance values; \code{type} = Feature column type.}
}
}
\seealso{
Other filter: \code{\link{FilterResult}},
  \code{\link{getFilterResult}};
  \code{\link{filterFeatures}};
  \code{\link{getFilterMethods}};
  \code{\link{getFilterValues}};
  \code{\link{getFilteredFeatures}};
  \code{\link{makeFilterWrapper}}
}

