% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/friedmanTestBMR.R
\name{friedmanTestBMR}
\alias{friedmanTestBMR}
\title{Perform Friedman Test on BenchmarkResult}
\usage{
friedmanTestBMR(bmr, measure = NULL, aggregation = "default")
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{aggregation}{[\code{character](1)}] \cr
 Aggregation method for resampling strategy. \cr
 Can be \code{default} or \code{mean}.\cr
 See \link{getAggrPerformances} for details.}
}
\value{
A list of class \code{htest}. \cr
See \link[stats]{friedman.test} for details.\cr
}
\description{
Performs a \link{friedman.test} for a selected measure. \cr
The null hypothesis is that apart from an effect of the different
[\code{tasks}], the location parameter (aggregated performance-measure)
is the same for each [\code{learner}].
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
friedmanTestBMR(res, measure)
}

