% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCritDifferences.R
\name{generateCritDifferencesData}
\alias{generateCritDifferencesData}
\title{Generate Data for critical-differences Plot}
\usage{
generateCritDifferencesData(bmr, measure = NULL, p.value = 0.05,
  baseline = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
\link{Measure} for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{p.value}{[\code{numeric}(1)]\cr
P-value for the critical difference. Default: 0.05}

\item{baseline}{[\code{character(1)}]: [\code{learner.id}] \cr
Select a [\code{learner.id} as baseline for the critical difference
diagram, the critical difference will be positioned arround this learner.
Defaults to best performing algorithm.}
}
\value{
[\code{critDifferencesData}], containing: \cr
$\code{$data}: [\code{data.frame}] containing the info for the descriptive
               part of the plot.\cr
$\code{friedman.nemenyi.test}: [\code{list}] of class \code{pairwise.htest}
 \cr
                               \link[PMCMR]{friedman.nemenyi.test}
                               calculated. \cr
$\code{cdInfo}: [\code{list}] containing info on the critical difference
                 and its positioning.\cr
$\code{baseline}: \code{baseline} chosen for plotting.\cr
$\code{p.value}: p.value used for the \link[PMCMR]{friedman.nemenyi.test}
                 and for computation of the \code{Critical Difference}.
}
\description{
Generate data that can be used to plot a
critical differences plot.
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.nnet"),
            makeLearner("classif.rpart"), makeLearner("classif.svm"))
tasks = list(iris.task, sonar.task, pid.task)
rdesc = makeResampleDesc("CV", iters = 5L)
meas = list(acc, mmce, ber, featperc)
res = benchmark(lrns, tasks, rdesc, meas)
generateCritDifferencesData(res, acc)
}

