% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFilterValues.R
\name{generateFilterValuesData}
\alias{FilterValues}
\alias{generateFilterValuesData}
\title{Calculates feature filter values.}
\usage{
generateFilterValuesData(task, method = "randomForestSRC.rfsrc",
  nselect = getTaskNFeats(task), ..., more.args = list())
}
\arguments{
\item{task}{([Task])\cr
The task.}

\item{method}{([character])\cr
Filter method(s), see above.
Default is \dQuote{randomForestSRC.rfsrc}.}

\item{nselect}{(`integer(1)`)\cr
Number of scores to request. Scores are getting calculated for all features per default.}

\item{...}{(any)\cr
Passed down to selected method. Can only be use if `method` contains one element.}

\item{more.args}{(named [list])\cr
Extra args passed down to filter methods. List elements are named with the filter
`method` name the args should be passed down to.
A more general and flexible option than `...`.
Default is empty list.}
}
\value{
([FilterValues]). A `list` containing:
  \item{task.desc}{[[TaskDesc])\cr
    Task description.}
  \item{data}{([data.frame]) with columns:
    \itemize{
      \item `name`([character])\cr
        Name of feature.
      \item `type`([character])\cr
        Feature column type.
      \item `method`([numeric])\cr
        One column for each method with the feature importance values.
    }}
}
\description{
Calculates numerical filter values for features.
For a list of features, use [listFilterMethods].
}
\seealso{
Other filter: \code{\link{filterFeatures}},
  \code{\link{getFilterValues}},
  \code{\link{getFilteredFeatures}},
  \code{\link{listFilterMethods}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeFilter}}, \code{\link{plotFilterValues}}

Other generate_plot_data: \code{\link{generateCalibrationData}},
  \code{\link{generateCritDifferencesData}},
  \code{\link{generateFeatureImportanceData}},
  \code{\link{generateLearningCurveData}},
  \code{\link{generatePartialDependenceData}},
  \code{\link{generateThreshVsPerfData}},
  \code{\link{getFilterValues}},
  \code{\link{plotFilterValues}}
}

