% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameterValidation.R
\name{generateValidationData}
\alias{generateValidationData}
\title{Generate hyperparameter validation data from a tuning result.}
\usage{
generateValidationData(TuneResult, include.diagnostics = FALSE)
}
\arguments{
\item{TuneResult}{\code{\link{TuneResult}} & \code{OptResult}\cr
The tuning results, also containing the optimizer results}

\item{include.diagnostics}{\code{logical(1)}\cr
Should diagnostic info (eol and error msg) be included?}
}
\value{
\code{ValidationData}
 object containing the validation dataframe, the tuning performance measures
 used, the hyperparameters used, a flag for including diagnostic info, and 
 the optimization algorithm used.
}
\description{
Generate cleaned hyperparameter validation from a tuning result. The object 
returned can be used for custom visualization or passed downstream to an out 
of the box mlr method, \code{\link{plotValidation}}.
}
\examples{
\dontrun{
ps = makeParamSet(makeDiscreteParam("C", values = 2^(-4:4)))
ctrl = makeTuneControlRandom()
rdesc = makeResampleDesc("CV", iters = 3L)
res = tuneParams("classif.ksvm", task = pid.task, resampling = rdesc,
par.set = ps, control = ctrl)
valid = generateValidationData(res)
valid$data
}
}

