% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAggrMeasures}
\alias{getAggrMeasures}
\title{Extract the aggregated measures of a benchmark result.}
\usage{
getAggrMeasures(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{data.frame}].
}
\description{
Extract the aggregated measures of a benchmark result.
}
\seealso{
Other benchmark: \code{\link{benchmark}};
  \code{\link{getMeasures}}; \code{\link{getPredictions}}
}

