% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBaseLearnerPredictions}
\alias{getBaseLearnerPredictions}
\title{Returns the predictions for each base learner.}
\usage{
getBaseLearnerPredictions(model, newdata = NULL)
}
\arguments{
\item{model}{[\code{WrappedModel}]\cr Wrapped model, result of train.}

\item{newdata}{[\code{data.frame}]\cr
New observations, for which the predictions using the specified base learners should be returned.
Default is \code{NULL} and extracts the base learner predictions that were made during the training.}
}
\description{
Returns the predictions for each base learner.
}
\details{

}

