% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BenchmarkResult_operators.R
\name{getBMRLearnerIds}
\alias{getBMRLearnerIds}
\title{Return learner ids used in benchmark.}
\usage{
getBMRLearnerIds(bmr)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{character}].
}
\description{
Return learner ids used in benchmark.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{convertBMRToRankMatrix}};
  \code{\link{friedmanPostHocTestBMR}};
  \code{\link{friedmanTestBMR}};
  \code{\link{generateBenchmarkSummaryData}};
  \code{\link{generateCritDifferencesData}};
  \code{\link{generateRankMatrixAsBarData}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRLearners}};
  \code{\link{getBMRMeasureIds}};
  \code{\link{getBMRMeasures}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}};
  \code{\link{plotBenchmarkResult}};
  \code{\link{plotBenchmarkSummary}};
  \code{\link{plotCritDifferences}};
  \code{\link{plotRankMatrixAsBar}}
}

