% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBMRLearnerIds}
\alias{getBMRLearnerIds}
\title{Return learner ids used in benchmark.}
\usage{
getBMRLearnerIds(bmr, join = FALSE)
}
\arguments{
\item{bmr}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}

\item{join}{[\code{logical(1)}]\cr
For \code{FALSE}: Get list od character vectors, named by task.ids.
For \code{TRUE}: Join these char vectors and make them unique.
Default is \code{FALSE}.}
}
\value{
[\code{list} | \code{character}]. See above.
}
\description{
Return learner ids used in benchmark.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getBMRAggrPerformances}};
  \code{\link{getBMRFeatSelResults}};
  \code{\link{getBMRFilteredFeatures}};
  \code{\link{getBMRPerformances}};
  \code{\link{getBMRPredictions}};
  \code{\link{getBMRTaskIds}};
  \code{\link{getBMRTuneResults}}
}

