% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassWeightParam.R
\name{getClassWeightParam}
\alias{getClassWeightParam}
\title{Get the class weight parameter of a learner.}
\usage{
getClassWeightParam(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{numeric \link{LearnerParam}}].\cr
  A numeric parameter object, containing the class weight parameter of the
  given learner.
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getHyperPars}}, \code{\link{getParamSet}},
  \code{\link{makeLearner}}, \code{\link{removeHyperPars}},
  \code{\link{setHyperPars}}, \code{\link{setId}},
  \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

