% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{getData}
\alias{getData}
\title{Retrive data from a DataSource.}
\usage{
getData(src, subset, features)
}
\arguments{
\item{subset}{[\code{integer}]\cr
Subset of rows to retrieve.}

\item{features}{[\code{character}]\cr
Names of features to retrieve.}

\item{data}{[\code{\link{DataSource}}]\cr
\code{\link{DataSource}}.}
}
\value{
[\code{data.frame}].
}

