% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{getDefaultMeasure}
\alias{getDefaultMeasure}
\title{Get default measure.}
\usage{
getDefaultMeasure(x)
}
\arguments{
\item{x}{[\code{character(1)} | \code{\link{Task}} | \code{\link{TaskDesc}} | \code{\link{Learner}}]\cr
Task type, task, task description, learner name, a learner, or a type of learner (e.g. "classif").}
}
\value{
[\code{\link{Measure}}].
}
\description{
Get the default measure for a task type, task, task description or a learner.
Currently these are:
 \tabular{ll}{
   oneclass    \tab f1 \cr
   classif     \tab mmce\cr
   regr        \tab mse\cr
   cluster     \tab db\cr
   surv        \tab cindex\cr
   costsens    \tab mcp\cr
   multilabel  \tab multilabel.hamloss\cr
}
Note: default measure f1 for oneclass is an adhoc solution and not the optimal measurement for unsupervised learning.
}
