% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{getDefaultMeasure}
\alias{getDefaultMeasure}
\title{Get default measure.}
\usage{
getDefaultMeasure(x)
}
\arguments{
\item{x}{([character(1)` | \link{Task} | \link{TaskDesc} | \link{Learner})\cr
Task type, task, task description, learner name, a learner, or a type of learner (e.g. "classif").}
}
\value{
(\link{Measure}).
}
\description{
Get the default measure for a task type, task, task description or a learner.
Currently these are:
\tabular{ll}{
classif     \tab mmce\cr
regr        \tab mse\cr
cluster     \tab db\cr
surv        \tab cindex\cr
costsens    \tab mcp\cr
multilabel  \tab multilabel.hamloss\cr
}
}
