% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrappedModel.R
\name{getFailureModelDump}
\alias{getFailureModelDump}
\title{Return the error dump of FailureModel.}
\usage{
getFailureModelDump(model)
}
\arguments{
\item{model}{(\link{WrappedModel})\cr
The model.}
}
\value{
(\code{last.dump}).
}
\description{
Returns the error dump that can be used with \code{debugger()} to evaluate errors.
If \link{configureMlr} configuration \code{on.error.dump} is \code{FALSE}, this returns
\code{NULL}.
}
