% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrappedModel.R
\name{getFailureModelMsg}
\alias{getFailureModelMsg}
\title{Return error message of FailureModel.}
\usage{
getFailureModelMsg(model)
}
\arguments{
\item{model}{([WrappedModel])\cr
The model.}
}
\value{
(`character(1)`).
}
\description{
Such a model is created when one sets the corresponding option in [configureMlr].
If no failure occurred, `NA` is returned.

For complex wrappers this getter returns the first error message encountered in ANY model that failed.
}

