% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrappedModel.R
\name{getFailureModelMsg}
\alias{getFailureModelMsg}
\title{Return error message of FailureModel.}
\usage{
getFailureModelMsg(model)
}
\arguments{
\item{model}{(\link{WrappedModel})\cr
The model.}
}
\value{
(\code{character(1)}).
}
\description{
Such a model is created when one sets the corresponding option in \link{configureMlr}.
If no failure occurred, \code{NA} is returned.

For complex wrappers this getter returns the first error message encountered in ANY model that failed.
}
