% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFeatSelResult}
\alias{getFeatSelResult}
\title{Returns the selected feature set and optimization path after training or benchmarking.}
\usage{
getFeatSelResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}} | \code{BenchmarkResult}]\cr
Trained Model created with \code{\link{makeFeatSelWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{FeatSelResult}} or list of \code{\link{FeatSelResult}}s].
  \code{NULL}, if no feature selection was performed.
}
\description{
Returns the selected feature set and optimization path after training or benchmarking.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{FilterResult}},
  \code{\link{getFilterResult}}; \code{\link{benchmark}};
  \code{\link{getAggrPerformances}};
  \code{\link{getPerformances}};
  \code{\link{getPredictions}}; \code{\link{getTuneResult}}

Other featsel: \code{\link{FeatSelControl}},
  \code{\link{FeatSelControlExhaustive}},
  \code{\link{FeatSelControlGA}},
  \code{\link{FeatSelControlRandom}},
  \code{\link{FeatSelControlSequential}},
  \code{\link{makeFeatSelControlExhaustive}},
  \code{\link{makeFeatSelControlGA}},
  \code{\link{makeFeatSelControlRandom}},
  \code{\link{makeFeatSelControlSequential}};
  \code{\link{analyzeFeatSelResult}};
  \code{\link{makeFeatSelWrapper}};
  \code{\link{selectFeatures}}
}

