% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFeatSelResult}
\alias{getFeatSelResult}
\title{Returns the selected feature set and optimization path.}
\usage{
getFeatSelResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}} | \code{BenchMarkResult}]\cr
Trained Model created with \code{\link{makeFeatSelWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{FeatSelResult}} or list of \code{\link{FeatSelResult}}s].
}
\description{
Returns the selected feature set and optimization path.
}

