% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatSelResult.R
\name{getFeatSelResult}
\alias{getFeatSelResult}
\title{Returns the selected feature set and optimization path after training.}
\usage{
getFeatSelResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}}]\cr
Trained Model created with \code{\link{makeFeatSelWrapper}}.}
}
\value{
[\code{\link{FeatSelResult}}].
}
\description{
Returns the selected feature set and optimization path after training.
}
\seealso{
Other featsel: \code{\link{FeatSelControl}},
  \code{\link{analyzeFeatSelResult}},
  \code{\link{makeFeatSelWrapper}},
  \code{\link{selectFeatures}}
}

