\name{getFilteredFeatures}
\alias{getFilteredFeatures}
\title{Returns the filtered features.}
\usage{
  getFilteredFeatures(model)
}
\arguments{
  \item{model}{[\code{\link[mlr]{WrappedModel}}]\cr Trained
  Model created with \code{\link{makeFilterWrapper}}.}
}
\value{
  [\code{character}].
}
\description{
  Returns the filtered features.
}

