% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterWrapper.R
\name{getFilteredFeatures}
\alias{getFilteredFeatures}
\title{Returns the filtered features.}
\usage{
getFilteredFeatures(model)
}
\arguments{
\item{model}{(\link{WrappedModel})\cr
Trained Model created with \link{makeFilterWrapper}.}
}
\value{
(\link{character}).
}
\description{
Returns the filtered features.
}
\seealso{
Other filter: \code{\link{filterFeatures}},
  \code{\link{generateFilterValuesData}},
  \code{\link{getFilterValues}},
  \code{\link{listFilterMethods}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeFilter}}, \code{\link{plotFilterValues}}
}
