% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFilterResult}
\alias{FilterResult}
\alias{getFilterResult}
\title{Returns a filter result after training or benchmarking.}
\usage{
getFilterResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}} | \code{\link{BenchmarkResult}}]\cr
Trained Model created with \code{\link{makeFilterWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{FilterResult}} or list of \code{\link{FilterResult}}s].
}
\description{
Returns a filter result after training or benchmarking.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{benchmark}};
  \code{\link{getAggrPerformances}};
  \code{\link{getFeatSelResult}};
  \code{\link{getPerformances}};
  \code{\link{getPredictions}}; \code{\link{getTuneResult}}

Other filter: \code{\link{FilterValues}};
  \code{\link{filterFeatures}};
  \code{\link{getFilterValues}};
  \code{\link{getFilteredFeatures}};
  \code{\link{makeFilterWrapper}}
}

