% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFilterResult}
\alias{FilterResult}
\alias{getFilterResult}
\title{Returns a filter result after training or benchmarking.}
\usage{
getFilterResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}} | \code{\link{BenchmarkResult}}]\cr
Trained Model created with \code{\link{makeFilterWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{FilterResult}} or list of \code{\link{FilterResult}}s].
}
\description{
Returns a filter result after training or benchmarking.
}
\seealso{
Other filter: \code{\link{FilterValues}};
  \code{\link{filterFeatures}};
  \code{\link{getFilterValues}};
  \code{\link{getFilteredFeatures}};
  \code{\link{listFilterMethods}};
  \code{\link{makeFilterWrapper}}
}

