% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFilterValues.R
\name{getFilterValues}
\alias{getFilterValues}
\title{Calculates feature filter values.}
\usage{
getFilterValues(task, method = "rf.importance",
  nselect = getTaskNFeats(task), ...)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{method}{[\code{character}]\cr
Filter methods, see above.
Default is \dQuote{rf.importance}.}

\item{nselect}{[\code{integer(1)}]\cr
Number of scores to request. Scores are getting calculated for all features per default.}

\item{...}{[any]\cr
Passed down to selected method.}
}
\value{
[\code{\link{FilterValues}}].
}
\description{
Calculates numerical filter values for features.
For a list of features, use \code{\link{listFilterMethods}}.
}
\seealso{
Other filter: \code{\link{FilterValues}};
  \code{\link{filterFeatures}};
  \code{\link{getFilteredFeatures}};
  \code{\link{makeFilterWrapper}}
}

