% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HomogeneousEnsemble.R
\name{getHomogeneousEnsembleModels}
\alias{getHomogeneousEnsembleModels}
\title{Deprecated, use \code{getLearnerModel} instead.}
\usage{
getHomogeneousEnsembleModels(model, learner.models = FALSE)
}
\description{
Deprecated, use \code{getLearnerModel} instead.
}

