% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomogeneousEnsemble.R
\name{getHomogeneousEnsembleModels}
\alias{getHomogeneousEnsembleModels}
\title{Deprecated, use \code{getLearnerModel} instead.}
\usage{
getHomogeneousEnsembleModels(model, learner.models = FALSE)
}
\arguments{
\item{model}{Deprecated.}

\item{learner.models}{Deprecated.}
}
\description{
Deprecated, use \code{getLearnerModel} instead.
}
