% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMeasures}
\alias{getMeasures}
\title{Extract performance measures of bechmark result.}
\usage{
getMeasures(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Benchmark result.}
}
\value{
[\code{data.frame}].
}
\description{
Extract performance measures of bechmark result.
}
\seealso{
Other benchmark: \code{\link{benchmark}};
  \code{\link{getAggrMeasures}};
  \code{\link{getPredictions}}
}

