% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsOptPathDf}
\alias{getNestedTuneResultsOptPathDf}
\title{Get the \code{opt.path}s from each tuning step from the outer resampling.}
\usage{
getNestedTuneResultsOptPathDf(r)
}
\arguments{
\item{r}{[\code{\link{ResampleResult}}] \cr
The result of a resampling of a tuneWrapped learner.}
}
\value{
\code{data.frame}
}
\description{
After you have conducted a resampling on a wrapped learner using \code{\link{makeTuneWrapper}} with \code{resample(..., extract = getTuneresult)} this helper gives you a \code{data.frame} with all \code{opt.path}s combined by \code{rbind}.
An additional column \code{iter} indicates to what resampling iteration the column belongs.
}
\examples{
\\donotrun{
ps = makeParamSet(
  makeDiscreteParam("C", values = 2^(-2:2)),
  makeDiscreteParam("sigma", values = 2^(-2:2))
)
ctrl = makeTuneControlGrid()
inner = makeResampleDesc("Holdout")
outer = makeResampleDesc("CV", iters = 5)
lrn = makeTuneWrapper("classif.ksvm", inner, par.set = ps, control = ctrl)
r = resample(lrn, iris.task, resampling = outer, extract = getTuneResult)
getNestedTuneResultsOptPathDf(r)
}
}

