% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsOptPathDf}
\alias{getNestedTuneResultsOptPathDf}
\title{Get the \code{opt.path}s from each tuning step from the outer resampling.}
\usage{
getNestedTuneResultsOptPathDf(r)
}
\arguments{
\item{r}{[\code{\link{ResampleResult}}] \cr
The result of resampling of a tuning wrapper.}
}
\value{
[\code{data.frame}]. See above.
}
\description{
After you resampled a tuning wrapper (see \code{\link{makeTuneWrapper}})
with \code{resample(..., extract = getTuneresult)} this helper returns a \code{data.frame} with
with all \code{opt.path}s combined by \code{rbind}.
An additional column \code{iter} indicates to what resampling iteration the row belongs.
}
\examples{
# see example of makeTuneWrapper
}

