% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsX}
\alias{getNestedTuneResultsX}
\title{Get the tuned hyperparameter settings from a nested tuning.}
\usage{
getNestedTuneResultsX(r)
}
\arguments{
\item{r}{[\code{\link{ResampleResult}}] \cr
The result of resampling of a tuning wrapper.}
}
\value{
[\code{data.frame}]. One column for each tuned hyperparameter and one row for each outer resampling iteration.
}
\description{
After you resampled a tuning wrapper (see \code{\link{makeTuneWrapper}})
with \code{resample(..., extract = getTuneResult)} this helper returns a \code{data.frame} with
the the best found hyperparameter settings for each resampling iteration.
}
\examples{
# see example of makeTuneWrapper
}
\seealso{
Other tune: \code{\link{ModelMultiplexer}},
  \code{\link{makeModelMultiplexer}};
  \code{\link{TuneControl}},
  \code{\link{TuneControlCMAES}},
  \code{\link{TuneControlGenSA}},
  \code{\link{TuneControlGrid}},
  \code{\link{TuneControlIrace}},
  \code{\link{TuneControlRandom}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlRandom}};
  \code{\link{getNestedTuneResultsOptPathDf}};
  \code{\link{getTuneResult}};
  \code{\link{makeModelMultiplexerParamSet}};
  \code{\link{makeTuneWrapper}}; \code{\link{tuneParams}};
  \code{\link{tuneThreshold}}
}

