% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsX}
\alias{getNestedTuneResultsX}
\title{Get the tuned hyperparameter settings from a nested tuning.}
\usage{
getNestedTuneResultsX(r)
}
\arguments{
\item{r}{[\code{\link{ResampleResult}}] \cr
The result of resampling of a tuning wrapper.}
}
\value{
[\code{data.frame}]. One column for each tuned hyperparameter and one row for each outer resampling iteration.
}
\description{
After you resampled a tuning wrapper (see \code{\link{makeTuneWrapper}})
with \code{resample(..., extract = getTuneresult)} this helper returns a \code{data.frame} with
the the best found hyperparameter settings for each resampling iteration.
}
\examples{
# see example of makeTuneWrapper
}

