% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNestedTuneResults.R
\name{getNestedTuneResultsX}
\alias{getNestedTuneResultsX}
\title{Get the tuned hyperparameter settings from a nested tuning.}
\usage{
getNestedTuneResultsX(r)
}
\arguments{
\item{r}{(\link{ResampleResult}) \cr
The result of resampling of a tuning wrapper.}
}
\value{
(\link{data.frame}). One column for each tuned hyperparameter and one row for each outer resampling iteration.
}
\description{
After you resampled a tuning wrapper (see \link{makeTuneWrapper})
with \code{resample(..., extract = getTuneResult)} this helper returns a \code{data.frame} with
the best found hyperparameter settings for each resampling iteration.
}
\examples{
# see example of makeTuneWrapper
}
\seealso{
Other tune: 
\code{\link{TuneControl}},
\code{\link{getNestedTuneResultsOptPathDf}()},
\code{\link{getResamplingIndices}()},
\code{\link{getTuneResult}()},
\code{\link{makeModelMultiplexerParamSet}()},
\code{\link{makeModelMultiplexer}()},
\code{\link{makeTuneControlCMAES}()},
\code{\link{makeTuneControlDesign}()},
\code{\link{makeTuneControlGenSA}()},
\code{\link{makeTuneControlGrid}()},
\code{\link{makeTuneControlIrace}()},
\code{\link{makeTuneControlMBO}()},
\code{\link{makeTuneControlRandom}()},
\code{\link{makeTuneWrapper}()},
\code{\link{tuneParams}()},
\code{\link{tuneThreshold}()}
}
\concept{tune}
