% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getParamSet}
\alias{getParamSet}
\title{Get a description of all possible parameter settings for a learner.}
\usage{
getParamSet(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}].
}
\description{
Get a description of all possible parameter settings for a learner.
}

