% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getParamSet.R
\name{getParamSet}
\alias{getParamSet}
\title{Get a description of all possible parameter settings for a learner.}
\usage{
getParamSet(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{\link[ParamHelpers]{ParamSet}}].
}
\description{
Get a description of all possible parameter settings for a learner.
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{addProperties}}, \code{\link{hasProperties}},
  \code{\link{removeProperties}},
  \code{\link{setProperties}}; \code{\link{Learner}},
  \code{\link{makeLearner}}; \code{\link{getHyperPars}};
  \code{\link{removeHyperPars}};
  \code{\link{setHyperPars}}; \code{\link{setId}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}
}

