% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_recombine2.R
\name{getPreciseTaskType}
\alias{getPreciseTaskType}
\title{Get precise type of the task, which is "classif", "multiclassif" or "regr".}
\usage{
getPreciseTaskType(x)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}
}
\description{
Get precise type of the task, which is "classif", "multiclassif" or "regr".
}

