% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_operators.R
\name{getPredictionDump}
\alias{getPredictionDump}
\title{Return the error dump of a failed Prediction.}
\usage{
getPredictionDump(pred)
}
\arguments{
\item{pred}{(\link{Prediction})\cr
Prediction object.}
}
\value{
(\code{last.dump}).
}
\description{
Returns the error dump that can be used with \code{debugger()} to evaluate errors.
If \link{configureMlr} configuration \code{on.error.dump} is \code{FALSE} or if the
prediction did not fail, this returns \code{NULL}.
}
\seealso{
Other debug: 
\code{\link{FailureModel}},
\code{\link{ResampleResult}},
\code{\link{getRRDump}()}
}
\concept{debug}
