\name{getPredictions}
\alias{getPredictions}
\title{Extract the predictions from a benchmark result.}
\usage{
getPredictions(object)
}
\arguments{
  \item{object}{[\code{\link{BenchmarkResult}}]\cr
  Benchmark result.}
}
\value{
[\code{data.frame}].
}
\description{
Extract the predictions from a benchmark result.
}
\seealso{
Other benchmark: \code{\link{BenchmarkResult}};
  \code{\link{FilterResult}},
  \code{\link{getFilterResult}}; \code{\link{benchmark}};
  \code{\link{getAggrPerformances}};
  \code{\link{getFeatSelResult}};
  \code{\link{getPerformances}};
  \code{\link{getTuneResult}}
}

