% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPredictions}
\alias{getPredictions}
\title{Extract the prediction inforamation of an object}
\usage{
getPredictions(object)
}
\arguments{
\item{object}{[\code{\link{BenchmarkResult}}]\cr
Object which contains the predictions}
}
\value{
[\code{data.frame}].
}
\description{
Extract the prediction inforamation of an object
}

