% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_operators.R
\name{getProbabilities}
\alias{getProbabilities}
\title{Deprecated, use \code{getPredictionProbabilities} instead.}
\usage{
getProbabilities(pred, cl)
}
\arguments{
\item{pred}{Deprecated.}

\item{cl}{Deprecated.}
}
\description{
Deprecated, use \code{getPredictionProbabilities} instead.
}
