% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult_operators.R
\name{getRRPredictions}
\alias{getRRPredictions}
\title{Get predictions from resample results.}
\usage{
getRRPredictions(res)
}
\arguments{
\item{res}{[\code{ResampleResult}]\cr
The result of \code{\link{resample}} run with \code{keep.pred = TRUE}.}
}
\value{
[\code{ResamplePrediction}].
}
\description{
Very simple getter.
}
\seealso{
Other resample: \code{\link{ResamplePrediction}},
  \code{\link{ResampleResult}}, \code{\link{addRRMeasure}},
  \code{\link{getRRPredictionList}},
  \code{\link{getRRTaskDescription}},
  \code{\link{makeResampleDesc}},
  \code{\link{makeResampleInstance}},
  \code{\link{resample}}
}

