% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult_operators.R
\name{getRRTaskDescription}
\alias{getRRTaskDescription}
\title{Get task description from resample results.}
\usage{
getRRTaskDescription(res)
}
\arguments{
\item{res}{[\code{ResampleResult}]\cr
The result of \code{\link{resample}}.}
}
\value{
[\code{TaskDesc}].
}
\description{
Get a summarizing task description.
}
\seealso{
Other resample: \code{\link{ResamplePrediction}},
  \code{\link{ResampleResult}}, \code{\link{addRRMeasure}},
  \code{\link{getRRPredictionList}},
  \code{\link{getRRPredictions}},
  \code{\link{makeResampleDesc}},
  \code{\link{makeResampleInstance}},
  \code{\link{resample}}
}

