% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner.R
\name{getStackedBaseLearnerPredictions}
\alias{getStackedBaseLearnerPredictions}
\title{Returns the predictions for each base learner.}
\usage{
getStackedBaseLearnerPredictions(model, newdata = NULL)
}
\arguments{
\item{model}{(\link{WrappedModel})\cr Wrapped model, result of train.}

\item{newdata}{(\link{data.frame})\cr
New observations, for which the predictions using the specified base learners should be returned.
Default is \code{NULL} and extracts the base learner predictions that were made during the training.}
}
\description{
Returns the predictions for each base learner.
}
\details{
None.
}
