% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskCosts}
\alias{getTaskCosts}
\title{Extract costs in task.}
\usage{
getTaskCosts(task, subset)
}
\arguments{
\item{task}{[\code{\link{CostSensTask}}]\cr
The task.}

\item{subset}{[\code{integer}]\cr
Selected cases.
Default is all cases.}
}
\value{
[\code{matrix} | \code{NULL}].
}
\description{
Retuns \dQuote{NULL} if the task is not of type \dQuote{costsens}.
}
\seealso{
Other task: \code{\link{getTaskClassLevels}},
  \code{\link{getTaskData}},
  \code{\link{getTaskDescription}},
  \code{\link{getTaskFeatureNames}},
  \code{\link{getTaskFormula}}, \code{\link{getTaskId}},
  \code{\link{getTaskNFeats}}, \code{\link{getTaskSize}},
  \code{\link{getTaskTargetNames}},
  \code{\link{getTaskTargets}}, \code{\link{getTaskType}},
  \code{\link{subsetTask}}
}

