% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskCosts}
\alias{getTaskCosts}
\title{Extract costs in task.}
\usage{
getTaskCosts(task, subset = NULL)
}
\arguments{
\item{task}{(\link{CostSensTask})\cr
The task.}

\item{subset}{(\link{integer} | \link{logical} | \code{NULL})\cr
Selected cases. Either a logical or an index vector.
By default \code{NULL} if all observations are used.}
}
\value{
(\code{matrix} | \code{NULL}).
}
\description{
Returns \dQuote{NULL} if the task is not of type \dQuote{costsens}.
}
\seealso{
Other task: 
\code{\link{getTaskClassLevels}()},
\code{\link{getTaskData}()},
\code{\link{getTaskDesc}()},
\code{\link{getTaskFeatureNames}()},
\code{\link{getTaskFormula}()},
\code{\link{getTaskId}()},
\code{\link{getTaskNFeats}()},
\code{\link{getTaskSize}()},
\code{\link{getTaskTargetNames}()},
\code{\link{getTaskTargets}()},
\code{\link{getTaskType}()},
\code{\link{subsetTask}()}
}
\concept{task}
