% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskDescription}
\alias{getTaskDescription}
\title{Deprecated, use \code{\link{getTaskDesc}} instead.}
\usage{
getTaskDescription(x)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}
}
\value{
[\code{\link{TaskDesc}}].
}
\description{
Deprecated, use \code{\link{getTaskDesc}} instead.
}
