% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskDescription}
\alias{getTaskDescription}
\title{Deprecated, use \link{getTaskDesc} instead.}
\usage{
getTaskDescription(x)
}
\arguments{
\item{x}{(\link{Task} | \link{TaskDesc})\cr
Task or its description object.}
}
\description{
Deprecated, use \link{getTaskDesc} instead.
}
