% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskFeatureNames}
\alias{getTaskFeatureNames}
\title{Get feature names of task.}
\usage{
getTaskFeatureNames(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character}].
}
\description{
Target column name is not included.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}}; \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskNFeats}};
  \code{\link{getTaskTargets}}; \code{\link{subsetTask}}
}

