% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskFeatureNames}
\alias{getTaskFeatureNames}
\title{Get feature names of task.}
\usage{
getTaskFeatureNames(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character}].
}
\description{
Target column name is not included.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskDescription}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskId}}; \code{\link{getTaskNFeats}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}};
  \code{\link{subsetTask}}
}

