% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskFormulaAsString}
\alias{getTaskFormula}
\alias{getTaskFormulaAsString}
\title{Get formula of a task.}
\usage{
getTaskFormulaAsString(x, target = getTaskTargetNames(x))

getTaskFormula(x, target = getTaskTargetNames(x), env = NULL)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}

\item{target}{[\code{character(1)}]\cr
Left hand side of formula.
Default is defined by task \code{x}.}

\item{env}{[\code{environment}]\cr
Environment of the formula. Set this to \code{parent.frame()}
for the default behaviour.
Default is \code{NULL} which deletes the environment.}
}
\value{
[\code{formula} | \code{character(1)}].
}
\description{
This is simply \dQuote{<target> ~ .}.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskId}}; \code{\link{getTaskNFeats}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}};
  \code{\link{subsetTask}}
}

