% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskFormula}
\alias{getTaskFormula}
\title{Get formula of a task.}
\usage{
getTaskFormula(x, target = getTaskTargetNames(x), explicit.features = FALSE)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}

\item{target}{[\code{character(1)}]\cr
Left hand side of the formula.
Default is defined by task \code{x}.}

\item{explicit.features}{[\code{logical(1)}]\cr
Should the features (right hand side of the formula) be explicitly listed?
Default is \code{FALSE}, i.e., they will be represented as \code{"."}.}
}
\value{
[\code{formula} | \code{character(1)}].
}
\description{
This is simply \dQuote{<target> ~ .}.
}
\seealso{
Other task: \code{\link{getTaskClassLevels}};
  \code{\link{getTaskCosts}}; \code{\link{getTaskData}};
  \code{\link{getTaskDescription}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskId}}; \code{\link{getTaskNFeats}};
  \code{\link{getTaskSize}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}};
  \code{\link{subsetTask}}
}

