% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskNFeats}
\alias{getTaskNFeats}
\title{Get number of features in task.}
\usage{
getTaskNFeats(x)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}
}
\value{
[\code{integer(1)}].
}
\description{
Get number of features in task.
}
\seealso{
Other task: \code{\link{getTaskClassLevels}},
  \code{\link{getTaskCosts}}, \code{\link{getTaskData}},
  \code{\link{getTaskDescription}},
  \code{\link{getTaskFeatureNames}},
  \code{\link{getTaskFormula}}, \code{\link{getTaskId}},
  \code{\link{getTaskSize}},
  \code{\link{getTaskTargetNames}},
  \code{\link{getTaskTargets}}, \code{\link{getTaskType}},
  \code{\link{subsetTask}}
}

