% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskSize}
\alias{getTaskSize}
\title{Get number of observations in task.}
\usage{
getTaskSize(x)
}
\arguments{
\item{x}{[\code{\link{Task}} | \code{\link{TaskDesc}}]\cr
Task or its description object.}
}
\value{
[\code{integer(1)}].
}
\description{
See title.
}
\seealso{
Other task: \code{\link{getTaskClassLevels}},
  \code{\link{getTaskCosts}}, \code{\link{getTaskData}},
  \code{\link{getTaskDates}}, \code{\link{getTaskDesc}},
  \code{\link{getTaskFeatureNames}},
  \code{\link{getTaskFormula}}, \code{\link{getTaskId}},
  \code{\link{getTaskNFeats}},
  \code{\link{getTaskTargetNames}},
  \code{\link{getTaskTargets}}, \code{\link{getTaskType}},
  \code{\link{subsetTask}}
}
