% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskTargetNames}
\alias{getTaskTargetNames}
\title{Get the name(s) of the target column(s).}
\usage{
getTaskTargetNames(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character}].
}
\description{
Get the name(s) of the target column(s).
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskDescription}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskId}}; \code{\link{getTaskNFeats}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}};
  \code{\link{subsetTask}}
}

