% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{getTaskTargetNames}
\alias{getTaskTargetNames}
\title{Get the name(s) of the target column(s).}
\usage{
getTaskTargetNames(x)
}
\arguments{
\item{x}{(\link{Task} | \link{TaskDesc})\cr
Task or its description object.}
}
\value{
(\link{character}).
}
\description{
NB: For multilabel, \link{getTaskTargetNames} and \link{getTaskClassLevels}
actually return the same thing.
}
\seealso{
Other task: 
\code{\link{getTaskClassLevels}()},
\code{\link{getTaskCosts}()},
\code{\link{getTaskData}()},
\code{\link{getTaskDesc}()},
\code{\link{getTaskFeatureNames}()},
\code{\link{getTaskFormula}()},
\code{\link{getTaskId}()},
\code{\link{getTaskNFeats}()},
\code{\link{getTaskSize}()},
\code{\link{getTaskTargets}()},
\code{\link{getTaskType}()},
\code{\link{subsetTask}()}
}
\concept{task}
