% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTaskType}
\alias{getTaskType}
\title{Get the type of the task.}
\usage{
getTaskType(task)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}
}
\value{
[\code{character}].
}
\description{
Get the type of the task.
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskNFeats}};
  \code{\link{getTaskTargets}}; \code{\link{subsetTask}}
}

