\name{getter,learn.task-method}
\alias{[,learn.task-method}

\title{Getter for learn.task}
\description{Getter.}
\alias{learn.task.getter}
\alias{getter,learn.task-method}
\arguments{\item{x}{learn.task object}
\item{i}{[character]
\describe{
\item{target.name}{The name of the target variable.}
\item{target.col}{The column number of the target variable.}
\item{targets}{If j is missing all target values are returned. Otherwise they are indexed by j.}
\item{input.names}{The names of the input variables.}
}}
\item{j}{[integer] \cr See above, i == "targets".}
}
