\name{getter,wrapped.model-method}
\alias{[,wrapped.model-method}

\title{Getter for wrapped.model}
\description{Getter.}
\alias{wrapped.model.getter}
\alias{getter,wrapped.model-method}
\arguments{\item{x}{wrapped.model object}
\item{i}{[\code{\link{character}}]
\describe{
\item{<slot>}{A slot of the class.}
\item{fail}{Generally NULL but if the training failed, the error message of the underlying train function.}
}}
}
