% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTuneResult}
\alias{getTuneResult}
\title{Returns the optimal hyperparameters and optimization path.}
\usage{
getTuneResult(model)
}
\arguments{
\item{model}{[\code{\link{WrappedModel}} | \code{BenchMarkResult}]\cr
Trained Model created with \code{\link{makeTuneWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{TuneResult}} or list of \code{\link{TuneResult}}s].
}
\description{
Returns the optimal hyperparameters and optimization path.
}

