% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTuneResult}
\alias{getTuneResult}
\title{Returns the optimal hyperparameters and optimization path.}
\usage{
getTuneResult(object)
}
\arguments{
\item{object}{[\code{\link{WrappedModel}} | \code{BenchMarkResult}]\cr
Trained Model created with \code{\link{makeTuneWrapper}} or benchmark result created with \code{\link{benchmark}}.}
}
\value{
[\code{\link{TuneResult}} or list of \code{\link{TuneResult}}s].
}
\description{
Returns the optimal hyperparameters and optimization path.
}
\seealso{
Other tune: \code{\link{ModelMultiplexer}},
  \code{\link{makeModelMultiplexer}};
  \code{\link{TuneControl}},
  \code{\link{TuneControlCMAES}},
  \code{\link{TuneControlGrid}},
  \code{\link{TuneControlIrace}},
  \code{\link{TuneControlOptim}},
  \code{\link{TuneControlRandom}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlOptim}},
  \code{\link{makeTuneControlRandom}};
  \code{\link{makeModelMultiplexerParamSet}};
  \code{\link{makeTuneWrapper}}; \code{\link{tuneParams}};
  \code{\link{tuneThreshold}}
}

