% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTuneResult.R
\name{getTuneResult}
\alias{getTuneResult}
\title{Returns the optimal hyperparameters and optimization path after training.}
\usage{
getTuneResult(object)
}
\arguments{
\item{object}{(\link{WrappedModel})\cr
Trained Model created with \link{makeTuneWrapper}.}
}
\value{
(\link{TuneResult}).
}
\description{
Returns the optimal hyperparameters and optimization path after training.
}
\seealso{
Other tune: 
\code{\link{TuneControl}},
\code{\link{getNestedTuneResultsOptPathDf}()},
\code{\link{getNestedTuneResultsX}()},
\code{\link{getResamplingIndices}()},
\code{\link{makeModelMultiplexerParamSet}()},
\code{\link{makeModelMultiplexer}()},
\code{\link{makeTuneControlCMAES}()},
\code{\link{makeTuneControlDesign}()},
\code{\link{makeTuneControlGenSA}()},
\code{\link{makeTuneControlGrid}()},
\code{\link{makeTuneControlIrace}()},
\code{\link{makeTuneControlMBO}()},
\code{\link{makeTuneControlRandom}()},
\code{\link{makeTuneWrapper}()},
\code{\link{tuneParams}()},
\code{\link{tuneThreshold}()}
}
\concept{tune}
