% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams.R
\name{getTuneResultOptPath}
\alias{getTuneResultOptPath}
\title{Get the optimization path of a tuning result.}
\usage{
getTuneResultOptPath(tune.result, as.df = TRUE)
}
\arguments{
\item{tune.result}{(\link{TuneResult}) \cr
A tuning result of the (\link{tuneParams}) function.}

\item{as.df}{(\code{logical(1)})\cr
Should the optimization path be returned as a data frame?
Default is \code{TRUE}.}
}
\value{
(\link[ParamHelpers:OptPath]{ParamHelpers::OptPath}) or (\link{data.frame}).
}
\description{
Returns the opt.path from a (\link{TuneResult}) object.
}
