% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gunpoint.task}
\alias{gunpoint.task}
\title{Gunpoint functional data classification task.}
\description{
Contains the task (\code{gunpoint.task}).
You have to classify whether a person raises up a gun or just an empty hand.
}
\references{
See Ratanamahatana, C. A. & Keogh. E. (2004). Everything you know
about Dynamic Time Warping is Wrong. Proceedings of SIAM International
Conference on Data Mining (SDM05), 506-510.
}
\keyword{data}
