% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_properties.R
\name{hasProperties}
\alias{hasProperties}
\title{Deprecated, use \code{hasLearnerProperties} instead.}
\usage{
hasProperties(learner, props)
}
\arguments{
\item{learner}{Deprecated.}

\item{props}{Deprecated.}
}
\description{
Deprecated, use \code{hasLearnerProperties} instead.
}
