% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrappedModel.R
\name{isFailureModel}
\alias{isFailureModel}
\title{Is the model a FailureModel?}
\usage{
isFailureModel(model)
}
\arguments{
\item{model}{([WrappedModel])\cr
The model.}
}
\value{
(`logical(1)`).
}
\description{
Such a model is created when one sets the corresponding option in [configureMlr].

For complex wrappers this getter returns `TRUE` if ANY model contained in it failed.
}

