% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinClassLevels.R
\name{joinClassLevels}
\alias{joinClassLevels}
\title{Join some class existing levels to new, larger class levels for classification problems.}
\usage{
joinClassLevels(task, new.levels)
}
\arguments{
\item{task}{(\link{Task})\cr
The task.}

\item{new.levels}{(\code{list} of \code{character})\cr
Element names specify the new class levels to create, while the corresponding element
character vector specifies the existing class levels which will be joined to the new one.}
}
\value{
\link{Task}.
}
\description{
Join some class existing levels to new, larger class levels for classification problems.
}
\examples{
joinClassLevels(iris.task, new.levels = list(foo = c("setosa", "virginica")))
}
