\name{learner.props-class}
\alias{learner.props-class}
\alias{learner.props}
\title{Description object for the features of a learning algorithm.}
\description{Description object for the features of a learning algorithm.}
\section{Slots}{\describe{\item{\code{supports.missing}:}{(\code{\link{logical}}) [logical] Can missing values be dealt with?}\item{\code{supports.numerics}:}{(\code{\link{logical}}) [logical] Are numeric variables ok?}\item{\code{supports.factors}:}{(\code{\link{logical}}) [logical] Are factor variables ok?}\item{\code{supports.characters}:}{(\code{\link{logical}}) [logical] Are character variables ok?}\item{\code{supports.weights}:}{(\code{\link{logical}}) [logical] Can case weights be dealt with?}}}
\section{Methods}{\describe{}}
