% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Learner_properties.R
\name{LearnerProperties}
\alias{LearnerProperties}
\alias{getLearnerProperties}
\alias{hasLearnerProperties}
\title{Set, add, remove or query properties of learners}
\usage{
getLearnerProperties(learner)

hasLearnerProperties(learner, props)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{props}{[\code{character}]\cr
Vector of properties to set, add, remove or query.}
}
\value{
\code{getLearnerProperties} returns a character vector with learner properties.
 \code{hasLearnerProperties} returns a logical vector of the same length as \code{props}.
}
\description{
Properties can be accessed with \code{getLearnerProperties(learner)}, which returns a
character vector.
}
\seealso{
Other learner: \code{\link{Learner}},
  \code{\link{makeLearner}};
  \code{\link{getClassWeightParam}};
  \code{\link{getHyperPars}}; \code{\link{getParamSet}};
  \code{\link{removeHyperPars}};
  \code{\link{setHyperPars}}; \code{\link{setId}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}
}

