% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LearnerProperties}
\alias{LearnerProperties}
\alias{addProperties}
\alias{hasProperties}
\alias{removeProperties}
\alias{setProperties}
\title{Set, add, remove or query properties of learners}
\usage{
setProperties(learner, props)

addProperties(learner, props)

removeProperties(learner, props)

hasProperties(learner, props)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{props}{[\code{character}]\cr
Vector of properties to set, add, remove or query.}
}
\value{
\code{setProperties}, \code{addProperties} and \code{removeProperties}
 return an updated \code{\link{Learner}}.
 \code{hasProperties} returns a logical vector of the same length of \code{props}.
}
\description{
Set, add, remove or query properties of learners
}

