% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LearnerProperties}
\alias{LearnerProperties}
\alias{addProperties}
\alias{hasProperties}
\alias{removeProperties}
\alias{setProperties}
\title{Set, add, remove or query properties of learners}
\usage{
setProperties(lrn, props)

addProperties(lrn, props)

removeProperties(lrn, props)

hasProperties(lrn, props)
}
\arguments{
\item{lrn}{[\code{RLearner}]\cr
A \code{\link{RLearner}}.}

\item{props}{[\code{character}]\cr
Vector of properties to set, add, remove or query.}
}
\value{
\code{setProperties}, \code{addProperties} and \code{removeProperties}
 return an updated \code{\link{RLearner}}. \code{hasProperties} returns
 a logical vector of the same length of \code{props}.
}
\description{
Set, add, remove or query properties of learners
}

