% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{learners}
\alias{learners}
\title{List of supported learning algorithms.}
\description{
All supported learners can be found by \code{\link{listLearners}} or as a table
in the tutorial appendix: \url{http://mlr-org.github.io/mlr-tutorial/release/html/integrated_learners/}.
}

