% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{listFilterMethods}
\alias{listFilterMethods}
\title{List filter methods.}
\usage{
listFilterMethods(desc = TRUE, tasks = FALSE, features = FALSE,
  include.deprecated = FALSE)
}
\arguments{
\item{desc}{(`logical(1)`)\cr
Provide more detailed information about filters.
Default is `TRUE`.}

\item{tasks}{(`logical(1)`)\cr
Provide information on supported tasks.
Default is `FALSE`.}

\item{features}{(`logical(1)`)\cr
Provide information on supported features.
Default is `FALSE`.}

\item{include.deprecated}{(`logical(1)`)\cr
Should deprecated filter methods be included in the list.
Default is `FALSE`.}
}
\value{
([data.frame]).
}
\description{
Returns a subset-able dataframe with filter information.
}
\seealso{
Other filter: \code{\link{filterFeatures}},
  \code{\link{generateFilterValuesData}},
  \code{\link{getFilterValues}},
  \code{\link{getFilteredFeatures}},
  \code{\link{makeFilterWrapper}},
  \code{\link{makeFilter}}, \code{\link{plotFilterValues}}
}

