% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{listFilterMethods}
\alias{listFilterMethods}
\title{List filter methods.}
\usage{
listFilterMethods(desc = TRUE, tasks = FALSE, features = FALSE)
}
\arguments{
\item{desc}{[\code{logical(1)}]\cr
Provide more detailed information about filters.}

\item{tasks}{[\code{logical(1)}]\cr
Provide information on supported tasks.}

\item{features}{[\code{logical(1)}]\cr
Provide information on supported features.}
}
\value{
[\code{data.frame}].
}
\description{
Returns a subset-able dataframe with filter information.
}

