% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listLearners}
\alias{listLearners}
\alias{listLearners.SupervisedTask}
\alias{listLearners.character}
\alias{listLearners.default}
\title{Find matching learning algorithms.}
\usage{
listLearners(obj = NA_character_, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

\method{listLearners}{default}(obj, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

\method{listLearners}{character}(obj, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

\method{listLearners}{SupervisedTask}(obj, properties = character(0L),
  warn.missing.packages = TRUE)
}
\arguments{
\item{type}{[\code{character(1)}]\cr
Type of the learning algorithm, one of \dQuote{classif}, \dQuote{regr}
or \dQuote{surv}.
Default is \code{NA}, matching all types.}

\item{properties}{[\code{character)}]\cr
Set of required properties to filter for. Default is \code{character(0)}.}

\item{quiet}{[\code{logical(1)}]\cr
Construct learners quietly to check their properties, shows no package startup messages.
Turn off if you suspect errors.
Default is code{TRUE}.}

\item{warn.missing.packages}{[\code{logical(1)}]\cr
If some learner cannot be constructed because its package is missing,
should a warning be shown?
Default is code{TRUE}.}

\item{task}{[\code{\link{SupervisedTask}}]\cr
The task.}
}
\value{
[\code{character}]. Class names of matching learners.
}
\description{
Returns the class names of learning algorithms which have specific characteristics, e.g.
whether they supports missing values, case weights, etc.

\code{listLearnersForTask} returns all learners that are in principle applicable
for a given task.

Note that the packages of all learners are loaded during the search.
}

