% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listMeasures}
\alias{listMeasures}
\alias{listMeasures.SupervisedTask}
\alias{listMeasures.character}
\alias{listMeasures.default}
\title{Find matching measures.}
\usage{
listMeasures(obj, properties = character(0L))

\method{listMeasures}{default}(obj, properties = character(0L))

\method{listMeasures}{character}(obj, properties = character(0L))

\method{listMeasures}{SupervisedTask}(obj, properties = character(0L))
}
\arguments{
\item{obj}{[\code{character(1)} | \code{\link{SupervisedTask}}]\cr
Either a task or the type of the task, in the latter case one of:
\dQuote{classif}, \dQuote{regr},  \dQuote{surv}, dQuote{costsens}.
Default is \code{NA}, matching all types.}

\item{properties}{[\code{character)}]\cr
Set of required properties to filter for.
See \code{\link{Measure}} for some standardized properties.
Default is \code{character(0)}.}
}
\value{
[\code{character}]. Class names of matching learners.
}
\description{
Returns the matching measures which have specific characteristics, e.g.
whether they supports classification or regression.

\code{listMeasuresForTask} returns all learners that are in principle applicable
for a given task.
}

