% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listMeasures}
\alias{listMeasures}
\alias{listMeasuresForTask}
\title{Find matching measures.}
\usage{
listMeasures(properties = character(0L))

listMeasuresForTask(task, properties = character(0L))
}
\arguments{
\item{properties}{[\code{character)}]\cr
Set of required properties to filter for. Default is \code{character(0)}.
See \link{Measure} for some standardized properties.}

\item{task}{[\code{\link{SupervisedTask}}]\cr
The task.}
}
\value{
[\code{character}]. Class names of matching learners.
}
\description{
Returns the matching measures which have specific characteristics, e.g.
whether they supports classification or regression.

\code{listMeasuresForTask} returns all learners that are in principle applicable
for a given task.
}

