% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMeasures.R
\name{listMeasures}
\alias{listMeasures}
\alias{listMeasures.Task}
\alias{listMeasures.character}
\alias{listMeasures.default}
\title{Find matching measures.}
\usage{
listMeasures(obj, properties = character(0L), create = FALSE)

\method{listMeasures}{default}(obj, properties = character(0L),
  create = FALSE)

\method{listMeasures}{character}(obj, properties = character(0L),
  create = FALSE)

\method{listMeasures}{Task}(obj, properties = character(0L), create = FALSE)
}
\arguments{
\item{obj}{[\code{character(1)} | \code{\link{Task}}]\cr
Either a task or the type of the task, in the latter case one of:
\dQuote{classif}, \dQuote{regr},  \dQuote{surv}, \dQuote{costsens}, \dQuote{cluster}, \dQuote{multilabel}.
Default is \code{NA}, matching all types.}

\item{properties}{[\code{character}]\cr
Set of required properties to filter for.
See \code{\link{Measure}} for some standardized properties.
Default is \code{character(0)}.}

\item{create}{[\code{logical(1)}]\cr
Instantiate objects (or return strings)?
Default is \code{FALSE}.}
}
\value{
[\code{character} | \code{list} of \code{\link{Measure}}]. Class names of matching
  measures or instantiated objects.
}
\description{
Returns the matching measures which have specific characteristics, e.g.
whether they supports classification or regression.
}

