% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lung.task}
\alias{lung.task}
\title{NCCTG Lung Cancer survival task.}
\description{
Contains the task (\code{lung.task}).
}
\references{
See \link[survival:lung]{survival::lung}.
Incomplete cases have been removed from the task.
}
\keyword{data}
