% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_measure_AMVhd.R
\name{makeAMVhdMeasure}
\alias{makeAMVhdMeasure}
\title{Creates the measure Area under the Mass-Volume Curve (AMV) for Anomaly detection (oneclass) for high dimensional data}
\usage{
makeAMVhdMeasure(id = "AMVhd", minimize = TRUE, amv.iters = 50,
  amv.feats = 5, alphas = c(0.9, 0.99), n.alpha = 50, n.sim = 1000,
  best = 0, worst = NULL, name = id, note = "")
}
\arguments{
\item{id}{(`character(1)`)\cr
Name of measure.}

\item{minimize}{(`logical(1)`)\cr
Should the measure be minimized?
Default is `TRUE`.}

\item{amv.iters}{\link{\code{numeric}} \cr
Number of subsamples.
Default is 50.}

\item{amv.feats}{\link{\code{numeric}} \cr
Number of features to be drawn in the feature subsamples.
Default is 5.}

\item{alphas}{\link{\code{numeric}} \cr
Numeric vector of alphas, which lies in [0, 1), representing the computed quantiles.
Default: lower quantile alpha1 = 0.9, upper quantile alpha2 = 0.99 as we are interested in the performance of the scoring function in the the low density regions.}

\item{n.alpha}{\link{\code{numeric}} \cr
Numeric discretization parameter greater than one, which splits the intervall of alpha1 and alpha2 as follows: {alpha1 + j * (alpha2-alpha1)/(n.alpha-1), j element of {0,...,n.alpha-1}}, Default: n.alpha = 50.}

\item{n.sim}{\link{\code{numeric(1)}} \cr
Number of Monte-Carlo Samples, Default is 10^3.}

\item{best}{(`numeric(1)`)\cr
Best obtainable value for measure.
Default is -`Inf` or `Inf`, depending on `minimize`.}

\item{worst}{(`numeric(1)`)\cr
Worst obtainable value for measure.
Default is `Inf` or -`Inf`, depending on `minimize`.}

\item{name}{([character]) \cr
Name of the measure. Default is `id`.}

\item{note}{([character]) \cr
Description and additional notes for the measure. Default is \dQuote{}.}
}
\value{
\link{\code{numeric(1)}}
Area under the Mass-Volume Curve (AMV) for high dimensional data.

[Measure].
}
\description{
Creates a measure for oneclass classification on high dimensional data
(recommend for dimension greater than 8) called AMVhd, which is based on the
Area under the Mass-Volume Curve (AMV) (see \code{makeAMVMeasure}).
The basic idea is to do several feature subsamplings (of dimension less than 8)
to reduce the dimension of the data, therefore AMV can be applied on each
subsamples, yielding partial scores AMV_k. The mean of the partial scores is
the new performancecriteria AMVhd.
}
\note{
the ID of the learner have to to be the default id!
}
\examples{
# create anomaly data with feature size nine
sigma = matrix(0, 9, 9)
diag(sigma) = c(4, 5, 8, 3, 2, 6, 9, 3, 1)
normal = MASS::mvrnorm(n = 1000, rep(0, 9), sigma)
colnames(normal) = paste0("V", 1:9)
normal = as.data.frame(normal)
normal$normal = TRUE

anomaly = matrix(sample(size = 50 * 9, x = 20:100, replace = TRUE), 50, 9)
colnames(anomaly) = paste0("V", 1:9)
anomaly = as.data.frame(anomaly)
anomaly$normal = FALSE
data = rbind(normal, anomaly)
data = na.omit(data)

# create train and test sets
library(BBmisc)
inds.split = BBmisc::chunk(seq_len(nrow(data)), shuffle = TRUE, props = c(0.6, 0.4))
train.inds = inds.split[[1]]
test.inds = inds.split[[2]]

# create an AMVhd measure which calculates the Area under the Mass-Volume Curve between 0.8 and 0.99
# with 50 steps for high dimensional data.
AMVhd = makeAMVhdMeasure(id = "AMV", minimize = TRUE, alphas = c(0.8, 0.99),
n.alpha = 50, n.sim = 10e3, best = 0, worst = NULL)

task = makeOneClassTask(data = data, target = "normal", positive = "TRUE", negative = "FALSE")
lrn = makeLearner("oneclass.svm", predict.type = "prob")
mod = train(lrn, task, subset = train.inds)
pred = predict(mod, task, subset = test.inds)

# calculate AMVhd performance
set.seed(123)
performance(pred = pred, measures = list(AMVhd), model = mod,
task = task, feats = data[test.inds, 1:9])
}
\references{
Nicolas, G. How to Evaluate the Quality of Unsupervised Anomaly Detection Algorithms,
arXiv preprint arXiv:1607.01152
}
\seealso{
Other performance.: \code{\link{makeAMVMeasure}},
  \code{\link{makePrecisionMeasure}},
  \code{\link{makeWACMeasure}}
}
