% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CostSensRegrWrapper.R
\name{makeCostSensRegrWrapper}
\alias{CostSensRegrModel}
\alias{CostSensRegrWrapper}
\alias{makeCostSensRegrWrapper}
\title{Wraps a regression learner for use in cost-sensitive learning.}
\usage{
makeCostSensRegrWrapper(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The regression learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Creates a wrapper, which can be used like any other learner object.
Models can easily be accessed via \code{\link{getHomogeneousEnsembleModels}}.

For each class in the task, an individual regression model is fitted for the costs of that class.
During prediction, the class with the lowest predicted costs is selected.
}
\seealso{
Other costsens: \code{\link{ClassifTask}},
  \code{\link{ClusterTask}}, \code{\link{CostSensTask}},
  \code{\link{RegrTask}}, \code{\link{SurvTask}},
  \code{\link{Task}}, \code{\link{makeClassifTask}},
  \code{\link{makeClusterTask}},
  \code{\link{makeCostSensTask}},
  \code{\link{makeRegrTask}}, \code{\link{makeSurvTask}};
  \code{\link{CostSensClassifModel}},
  \code{\link{CostSensClassifWrapper}},
  \code{\link{makeCostSensClassifWrapper}};
  \code{\link{CostSensWeightedPairsModel}},
  \code{\link{CostSensWeightedPairsWrapper}},
  \code{\link{makeCostSensWeightedPairsWrapper}}

Other wrapper: \code{\link{CostSensClassifModel}},
  \code{\link{CostSensClassifWrapper}},
  \code{\link{makeCostSensClassifWrapper}};
  \code{\link{makeBaggingWrapper}};
  \code{\link{makeDownsampleWrapper}};
  \code{\link{makeFeatSelWrapper}};
  \code{\link{makeFilterWrapper}};
  \code{\link{makeImputeWrapper}};
  \code{\link{makeMulticlassWrapper}};
  \code{\link{makeOverBaggingWrapper}};
  \code{\link{makeOversampleWrapper}},
  \code{\link{makeUndersampleWrapper}};
  \code{\link{makePreprocWrapperPCA}};
  \code{\link{makePreprocWrapper}};
  \code{\link{makeSMOTEWrapper}};
  \code{\link{makeTuneWrapper}};
  \code{\link{makeWeightedClassesWrapper}}
}

