% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOFunctional.R
\name{makeCPOFunctional}
\alias{makeCPOFunctional}
\title{Create a custom CPO constructor}
\usage{
makeCPOFunctional(.cpo.name, ..., .par.set = NULL, .par.vals = list(),
  .datasplit = c("target", "most", "all", "no", "task"),
  .properties = c("numerics", "factors", "ordered", "missings"),
  .properties.adding = character(0), .properties.needed = character(0),
  cpo.trafo)
}
\arguments{
\item{.cpo.name}{[\code{character(1)}]\cr
The name of the resulting CPO constructor / CPO. This is used for identification in output.}

\item{...}{Parameters of the CPO, in the format of \code{\link{paramSetSugar}}.}

\item{.par.set}{[\code{ParamSet}]\cr
Optional parameter set. If this is not \code{NULL}, the \dQuote{...} parameters are ignored.}

\item{.par.vals}{[\code{list}]\cr
Named list of default parameter values for the CPO. These are used additionally to the
parameter default values in \dQuote{...} and \code{.par.set}. It is preferred to use
these default values, and not \code{.par.vals}.}

\item{.datasplit}{[\code{character(1)}]\cr
Indicate what format the data should be as seen by \dQuote{cpo.trafo}. Possibilities are:
\itemize{
  \item target the \dQuote{data} variable contains the data in a data.frame without
    the target column(s), the \dQuote{target} variable contains the target column(s) in
    a data.frame.
  \item most the \dQuote{data} is a list containing three data.frames: \dQuote{numeric}
    the numeric columns, \dQuote{factor} the factorial columns (ordered and unordered),
    \dQuote{other} the columns that are neither numeric nor factors. The \dQuote{target}
    variable contains the target column(s) in a data.frame.
  \item all similarly to \dQuote{most}, but factors are split up into \dQuote{factor}
    (unordered factors) and \dQuote{ordered}.
  \item no the \dQuote{data} variable contains a data.frame with all data, the \dQuote{target}
    variable is a \code{character} indicating the names of the target columns.
  \item task the \dQuote{data} variable contains the data as a \dQuote{\link{Task}}.
}
Note that the returned data must always be in the same format as the one requested.
Currently it is an error to change the target column(s) in the \dQuote{no} and \dQuote{task}
cases. Default is \dQuote{target}.}

\item{.properties}{[\code{character}]\cr
The kind if data that the CPO will be able to handle. This can be one or many of: \dQuote{numerics},
\dQuote{factors}, \dQuote{ordered}, \dQuote{missings}.
There should be a bias towards including properties. If a property is absent, the preproc
operator will reject the data. If an operation e.g. only works on numeric columns that have no
missings (like PCA), it is recommended to give all properties, ignore the columns that
are not numeric (using \dQuote{.datasplit} = \dQuote{most}), and giving an error when
there are missings in the numeric columns (since missings in factorial features are not a problem).
Defaults to the maximal set.}

\item{.properties.adding}{[\code{character}]\cr
Can be one or many of the same values as \dQuote{.properties}. These properties get added to
a Learner (or CPO) coming after / behind this CPO. When a CPO imputes missing values, for example,
this should be \dQuote{missings}. This must be a subset of \dQuote{.properties}. Default is
\code{character(0)}.}

\item{.properties.needed}{[\code{character}]\cr
Can be one or many of the same values as \dQuote{.properties}. These properties are required
from a Learner (or CPO) coming after / behind this CPO. E.g., when a CPO converts factors to
numerics, this should be \dQuote{numerics} (and \dQuote{.properties.adding} should be \dQuote{factors}).
Default is \code{character(0)}.}

\item{cpo.trafo}{[\code{language} | \code{function}]\cr
This can either be a function, just the expressions to perform wrapped in curly braces.
If this is a function, it must have the parameters \dQuote{data} and \dQuote{target},
as well as the parameters specified in \dQuote{...} or \dQuote{.par.set}. (Alternatively,
the function may have a dotdotdot argument). It must return a \dQuote{data.frame} object
with an added \dQuote{\link{retrafo}}. This must be a function with the argument
\dQuote{data} and return another \dQuote{data.frame}.\cr
If \dQuote{cpo.trafo} is a list of expressions (preferred), it is turned into a function
by mlr, with the above mentioned criteria.}
}
\description{
Create a function-based CPO constructor. This has the advantage of being possibly more
elegant and less boilerplate-heavy than object-based CPOs, but it uses more memory
and is a bit harder to debug. For object-based CPOs, see \code{\link{makeCPOFunctional}}.
}
\examples{
noop = makeCPOFunctional("noop", dummy: logical, cpo.trafo = {
  retrafo(data) = function(data) data
  data
})

}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{getCPOName}},
  \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{makeCPOObject}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}
}
