% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOTarget.R
\name{makeCPOTargetOp}
\alias{makeCPOTargetOp}
\title{Create a custom target-bound CPO constructor}
\usage{
makeCPOTargetOp(.cpo.name, ..., .par.set = NULL, .par.vals = list(),
  .datasplit = c("target", "most", "all", "no", "task", "factor",
  "onlyfactor", "ordered", "numeric"), .data.dependent = TRUE,
  .stateless = FALSE, .properties = character(0),
  .properties.adding = character(0), .properties.needed = character(0),
  .properties.data = c("numerics", "factors", "ordered", "missings"),
  .type = c("cluster", "classif", "multilabel", "regr", "surv"),
  .type.out = .type, .predict.type = c(response = "response"), cpo.trafo,
  cpo.retrafo)
}
\arguments{
\item{.predict.type}{[\code{character} | \code{list}]\cr
Must be a named \code{character}, or named \code{list} of \code{character(1)}, indicating
what \code{predict.type} (see \link{Prediction}) a prediction must have if the out put prediction
is to be of some type. E.g. if a CPO converts a \dQuote{regr} \code{Task} into a
\dQuote{classif} \code{Task}, and if for \dQuote{se} prediction it needs a classification
learner to give \dQuote{prob} type predictions, while for \dQuote{response} prediction it
also needs \dQuote{response} predictions, this would be \code{c(response = "response",
se = "prob")}. The names are the prediction types that are requested from this CPO, the
values are types that this CPO will request from an underlying learner. If a name is not
present, the \code{predict.type} is assumed not supported. Default is \code{c(response = "response")}.}
}
\description{
Create a CPO constructor.

Most of the parameters are as in \link{makeCPO}.
}
\seealso{
Other CPO: \code{\link{CPOImputer}}, \code{\link{CPO}},
  \code{\link{NULLCPO}}, \code{\link{\%>>\%}},
  \code{\link{applyCPO}}, \code{\link{attachCPO}},
  \code{\link{chainCPO}}, \code{\link{cpoApply}},
  \code{\link{cpoDropConstants}},
  \code{\link{cpoDummyEncode}},
  \code{\link{cpoFilterFeatures}}, \code{\link{cpoImpute}},
  \code{\link{cpoMultiplex}}, \code{\link{cpoPca}},
  \code{\link{cpoScale}}, \code{\link{cpoSelect}},
  \code{\link{getCPOAffect}}, \code{\link{getCPOId}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{hasTagInvert}}, \code{\link{inverter<-}},
  \code{\link{inverter}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo<-}}, \code{\link{retrafo}},
  \code{\link{setHyperPars}}, \code{\link{tagInvert}}
}
