% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{makeDataSource}
<<<<<<< HEAD
<<<<<<< 6873e5499211c3bb3aca201d005b6732cb56fcbb
=======
>>>>>>> 61b241de9b92c5f1f99e1aeb337792e8132e9f42
\alias{DataSource}
\alias{makeDataSource}
\title{Create a Data Source backend for mlr Tasks.}
\usage{
makeDataSource(src, rows = NULL, cols = NULL, ...)
<<<<<<< HEAD
=======
\alias{makeDataSource}
\title{Create a Data Source backend for mlr Tasks.}
\usage{
<<<<<<< 370bff17fe3845565c0f8fd1e4aa3f846e8b5564
makeDataSource(src, cols = NULL, ...)
>>>>>>> tests, man pages, ...
=======
makeDataSource(src, rows = NULL, cols = NULL, ...)
>>>>>>> removing direct data access
}
\arguments{
\item{src}{[\code{\link[base]{data.frame}} | \code{\link[dplyr]{tbl}}]\cr
  Source of the data. Currently supported
  \itemize{
    \item[\code{\link[base]{data.frame}}]{Regular data.frames.}
    \item[\code{\link[dplyr]{tbl}}]{Tables as used in \pkg{dplyr}. Can use a SQL backend.}
  }}

<<<<<<< 370bff17fe3845565c0f8fd1e4aa3f846e8b5564
<<<<<<< 6873e5499211c3bb3aca201d005b6732cb56fcbb
\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}

=======
>>>>>>> tests, man pages, ...
=======
\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}

>>>>>>> removing direct data access
=======
}
\arguments{
\item{src}{[\code{\link[base]{data.frame}} | \code{\link[dplyr]{tbl}}]\cr
Source of the data. Currently supported
\itemize{
  \item[\code{\link[base]{data.frame}}]{Regular data.frames.}
  \item[\code{\link[dplyr]{tbl}}]{Tables as used in \pkg{dplyr}. Can use a
  SQL backend.} .. }}

\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}

>>>>>>> 61b241de9b92c5f1f99e1aeb337792e8132e9f42
\item{cols}{[\code{character} | \code{NULL}]\cr
Subset of columns to use in the task. Additional columns in \code{src} are ignored.}

\item{...}{[\code{any}]\cr
<<<<<<< HEAD
Additional arguments, depending on the type of \code{src}.}

<<<<<<< 370bff17fe3845565c0f8fd1e4aa3f846e8b5564
<<<<<<< 6873e5499211c3bb3aca201d005b6732cb56fcbb
\item{key}{[\code{character}]\cr
Required for database backends: Name of the primary key column.}
=======
\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}
>>>>>>> tests, man pages, ...
=======
\item{key}{[\code{character}]\cr
Required for database backends: Name of the primary key column.}
>>>>>>> removing direct data access
=======
Additional arguments, depending on the type of \code{src}.
key [\code{character}]\cr Required for database backends: Name of the primary
key column.}
>>>>>>> 61b241de9b92c5f1f99e1aeb337792e8132e9f42
}
\value{
[\code{DataSource}].
}
<<<<<<< HEAD
=======
\description{
Creates a \code{\link{DataSource}} which will be stored in an task element
and the data will be retrieved via \code{\link{getData}}.
}
>>>>>>> 61b241de9b92c5f1f99e1aeb337792e8132e9f42

