% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{makeDataSource}
\alias{makeDataSource}
\title{Create a Data Source backend for mlr Tasks.}
\usage{
makeDataSource(src, cols = NULL, ...)
}
\arguments{
\item{src}{[\code{\link[base]{data.frame}} | \code{\link[dplyr]{tbl}}]\cr
  Source of the data. Currently supported
  \itemize{
    \item[\code{\link[base]{data.frame}}]{Regular data.frames.}
    \item[\code{\link[dplyr]{tbl}}]{Tables as used in \pkg{dplyr}. Can use a SQL backend.}
  }}

\item{cols}{[\code{character} | \code{NULL}]\cr
Subset of columns to use in the task. Additional columns in \code{src} are ignored.}

\item{...}{[\code{any}]\cr
Additional arguments, depending on the type of \code{src}.}

\item{rows}{[\code{integer} | \code{NULL}]\cr
Subset of rows to use in the task. Additional rows are ignored.}
}
\value{
[\code{DataSource}].
}

