% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDownsampleWrapper}
\alias{makeDownsampleWrapper}
\title{Fuse learner with simple downsampling (subsampling).}
\usage{
makeDownsampleWrapper(learner, sw.perc = NULL, sw.n = NULL,
  sw.stratify = FALSE)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{sw.perc}{[\code{numeric(1)}]\cr
Percentage of the observations to train the learner on.}

\item{sw.n}{[\code{numeric(1)}]\cr
Total number of observations to train the learner on.}

\item{sw.stratify}{[\code{boolean(1)}]\cr
Should the downsampled data be stratified according to the target classes? Default is \code{FALSE}.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Creates a learner object, which can be
used like any other learner object.
It will only be trained on a subset of the original data to save computational time.
}

