% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoldoutInstance_make_fixed.R
\name{makeFixedHoldoutInstance}
\alias{makeFixedHoldoutInstance}
\title{Generate a fixed holdout instance for resampling.}
\usage{
makeFixedHoldoutInstance(train.inds, test.inds, size)
}
\arguments{
\item{train.inds}{[\code{integer}]\cr
Indices for training set.}

\item{test.inds}{[\code{integer}]\cr
Indices for test set.}

\item{size}{[\code{integer(1)}]\cr
Size of the data set to resample.
The function needs to know the largest possible index of the whole data set.}
}
\value{
[\code{\link{ResampleInstance}}].
}
\description{
Generate a fixed holdout instance for resampling.
}

