% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeMethods.R
\name{makeImputeMethod}
\alias{makeImputeMethod}
\title{Create a custom imputation method.}
\usage{
makeImputeMethod(learn, impute, args = list())
}
\arguments{
\item{learn}{(`function(data, target, col, ...)`)\cr
Function to learn and extract information on column `col`
out of data frame `data`. Argument `target` specifies
the target column of the learning task.
The function has to return a named list of values.}

\item{impute}{(`function(data, target, col, ...)`)\cr
Function to impute missing values in `col` using information
returned by `learn` on the same column.
All list elements of the return values o `learn`
are passed to this function into `...`.}

\item{args}{([list])\cr
Named list of arguments to pass to `learn` via `...`.}
}
\description{
This is a constructor to create your own imputation methods.
}
\seealso{
Other impute: \code{\link{imputations}},
  \code{\link{impute}}, \code{\link{makeImputeWrapper}},
  \code{\link{reimpute}}
}

