% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLearners.R
\name{makeLearners}
\alias{makeLearners}
\title{Create multiple learners at once.}
\usage{
makeLearners(cls, ids = NULL, type = NULL, ...)
}
\arguments{
\item{cls}{([character])\cr
Classes of learners.}

\item{ids}{([character])\cr
Id strings. Must be unique.
Default is `cls`.}

\item{type}{(`character(1)`)\cr
Shortcut to prepend type string to `cls` so one can set `cls = "rpart"`.
Default is `NULL`, i.e., this is not used.}

\item{...}{(any)\cr
Optional named (hyper)parameters.
Alternatively these can be given using the `par.vals` argument.}
}
\value{
[named list of [Learner]). Named by `ids`.
}
\description{
Small helper function that can save some typing when creating mutiple learner objects.
Calls [makeLearner] multiple times internally.
}
\examples{
makeLearners(c("rpart", "lda"), type = "classif", predict.type = "prob")
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getClassWeightParam}},
  \code{\link{getHyperPars}}, \code{\link{getLearnerId}},
  \code{\link{getLearnerPackages}},
  \code{\link{getLearnerParVals}},
  \code{\link{getLearnerParamSet}},
  \code{\link{getLearnerPredictType}},
  \code{\link{getLearnerShortName}},
  \code{\link{getLearnerType}}, \code{\link{getParamSet}},
  \code{\link{helpLearnerParam}},
  \code{\link{helpLearner}}, \code{\link{makeLearner}},
  \code{\link{removeHyperPars}},
  \code{\link{setHyperPars}}, \code{\link{setId}},
  \code{\link{setLearnerId}},
  \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

