% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeModelMultiplexer}
\alias{ModelMultiplexer}
\alias{makeModelMultiplexer}
\title{Create a meta learner}
\usage{
makeModelMultiplexer(base.learners, id = "ModelMultiplexer")
}
\arguments{
\item{base.learners}{[\code{list} of \code{\link{Learner}}]\cr
List of Learners with unique IDs.}

\item{id}{[\code{character(1)}]\cr
Identifier for constructred ModelMultiplexer.
Default is \dQuote{ModelMultiplexer}.}
}
\value{
[\code{ModelMultiplexer}]. A \code{\link{Learner}} specialized as \code{ModelMultiplexer}.
}
\description{
A meta learner combines multiple base learners by dispatching
on the hyperparameter \dQuote{selected.learner}.
The parameter set is the union of all (unique) base learners.
In order to avoid name clashes all parameter names are prefixed
with the base learner id, i.e. \dQuote{[task.type].[learner.id].[parameter.name]}.
}

