% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleInstance.R
\name{makeResampleInstance}
\alias{makeResampleInstance}
\alias{ResampleInstance}
\title{Instantiates a resampling strategy object.}
\usage{
makeResampleInstance(desc, task, size, ...)
}
\arguments{
\item{desc}{(\link{ResampleDesc} | \code{character(1)})\cr
Resampling description object or name of resampling strategy.
In the latter case \link{makeResampleDesc} will be called internally on the string.}

\item{task}{(\link{Task})\cr
Data of task to resample from.
Prefer to pass this instead of \code{size}.}

\item{size}{(\link{integer})\cr
Size of the data set to resample.
Can be used instead of \code{task}.}

\item{...}{(any)\cr
Passed down to \link{makeResampleDesc} in case
you passed a string in \code{desc}.
Otherwise ignored.}
}
\value{
(\link{ResampleInstance}).
}
\description{
This class encapsulates training and test sets generated from the data set for a number
of iterations. It mainly stores a set of integer vectors indicating the training and
test examples for each iteration.
}
\details{
Object slots:
\describe{
\item{desc (\link{ResampleDesc})}{See argument.}
\item{size (\code{integer(1)})}{See argument.}
\item{train.inds (list of \link{integer})}{List of of training indices for all iterations.}
\item{test.inds (list of \link{integer})}{List of of test indices for all iterations.}
\item{group (\link{factor})}{Optional grouping of resampling iterations. This encodes whether
specific iterations 'belong together' (e.g. repeated CV), and it can later be used to
aggregate performance values accordingly. Default is 'factor()'.}
}
}
\examples{
rdesc = makeResampleDesc("Bootstrap", iters = 10)
rin = makeResampleInstance(rdesc, task = iris.task)

rdesc = makeResampleDesc("CV", iters = 50)
rin = makeResampleInstance(rdesc, size = nrow(iris))

rin = makeResampleInstance("CV", iters = 10, task = iris.task)
}
\seealso{
Other resample: 
\code{\link{ResamplePrediction}},
\code{\link{ResampleResult}},
\code{\link{addRRMeasure}()},
\code{\link{getRRPredictionList}()},
\code{\link{getRRPredictions}()},
\code{\link{getRRTaskDescription}()},
\code{\link{getRRTaskDesc}()},
\code{\link{makeResampleDesc}()},
\code{\link{resample}()}
}
\concept{resample}
