% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_helpers.R
\name{makeSuperLearnerTask}
\alias{makeSuperLearnerTask}
\title{Create a super learner task
FIXME: "save" version which rm constant features and features with NAs. BUT 
might not be useful owing to the fact that predictLearner does not know which 
features are removed}
\usage{
makeSuperLearnerTask(type, data, target)
}
\description{
Create a super learner task
FIXME: "save" version which rm constant features and features with NAs. BUT 
might not be useful owing to the fact that predictLearner does not know which 
features are removed
}

