% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskDesc.R
\name{makeTaskDescInternal}
\alias{makeTaskDescInternal}
\title{Exported for internal use.}
\usage{
makeTaskDescInternal(type, id, data, target, weights, blocking, coordinates)
}
\arguments{
\item{type}{(\code{character(1)})\cr
Task type.}

\item{id}{(\code{character(1)})\cr
task id}

\item{data}{(\link{data.frame})\cr
data}

\item{target}{(\link{character})\cr
target columns}

\item{weights}{(\link{numeric})\cr
weights}

\item{blocking}{(\link{numeric})\cr
task data blocking}

\item{coordinates}{(\code{logical(1)})\cr
whether spatial coordinates have been provided}
}
\description{
Exported for internal use.
}
\keyword{internal}
