% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeUndersampleWrapper}
\alias{makeOversampleWrapper}
\alias{makeUndersampleWrapper}
\title{Fuse learner with simple over/undersampling for binary classification.}
\usage{
makeUndersampleWrapper(learner, usw.rate)

makeOversampleWrapper(learner, osw.rate)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{usw.rate}{[\code{numeric(1)}]\cr
Factor to downsample the bigger class. Must be between 0 and 1,
where 1 means no downsampling, 0.5 implies reduction to 50 percent
and 0 would imply reduction to 0 observations.}

\item{osw.rate}{[\code{numeric(1)}]\cr
Factor to oversample the smaller class. Must be between 1 and \code{Inf},
where 1 means no oversampling and 2 would mean doubling the class size.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Creates a learner object, which can be
used like any other learner object.
Internally uses \code{\link{oversample}} or \code{\link{undersample}} before every model fit.
}

