% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeWrappedModel}
\alias{WrappedModel}
\alias{makeWrappedModel}
\title{Induced model of learner.}
\usage{
makeWrappedModel(learner, model, task.desc, subset, features, factor.levels,
  time)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{model}{[any]\cr
Underlying model.}

\item{task.desc}{[\code{\link{TaskDesc}}]\cr
Task description object.}

\item{subset}{[\code{integer}]\cr
Subset used for training.}

\item{features}{[\code{character}]\cr
Features used for training.}

\item{factor.levels}{[named \code{list} of \code{character}]\cr
Levels of factor variables (features and potentially target) in training data.
Named by variable name, non-factors do not occur in the list.}

\item{time}{[\code{numeric(1)}]\cr
Computation time for model fit in seconds.}
}
\value{
[\code{\link{WrappedModel}}].
}
\description{
Result from \code{\link{train}}. It internally stores the underlying fitted model,
the subset used for training, features used for training, levels of factors in the
data set and computation time that was spent for training.
}
\details{
The constructed \code{makeWrappedModel} is only for internal use.

Object members: See arguments.
}

