% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedStacking_helper.R
\name{makeXBestLearnersFromMMTuneResult}
\alias{makeXBestLearnersFromMMTuneResult}
\title{Creates x best learners from ModelMultiplexer TuneResult.}
\usage{
makeXBestLearnersFromMMTuneResult(tune.result, model.multiplexer,
  mm.ps = mm.ps, x.best = 5, measure = mmce)
}
\arguments{
\item{tune.result}{[\code{\link{TuneResult}}]\cr
\code{TuneResult} from \code{ModelMultiplexer}}

\item{x.best}{[\code{integer(1)}]
Number of best learners to extract.}

\item{measure}{[\code{\link{Measure}}]\cr
Measure to apply to "extract from best".}

\item{base.learners}{[\code{list of \link{TuneResult}}]\cr
Needed to exract fix parameters.}
}
\value{
[\code{list of x best learners}]
}
\description{
Creates x best learners from ModelMultiplexer TuneResult.
}

