% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_properties.R
\name{MeasureProperties}
\alias{MeasureProperties}
\alias{getMeasureProperties}
\alias{hasMeasureProperties}
\title{Query properties of measures.}
\usage{
getMeasureProperties(measure)

hasMeasureProperties(measure, props)
}
\arguments{
\item{measure}{(\link{Measure})\cr
Performance measure.
Default is the first measure used in the benchmark experiment.}

\item{props}{(\link{character})\cr
Vector of properties to query.}
}
\value{
\code{getMeasureProperties} returns a character vector with measure properties.
\code{hasMeasureProperties} returns a logical vector of the same length as \code{props}.
}
\description{
Properties can be accessed with \code{getMeasureProperties(measure)}, which returns a
character vector.

The measure properties are defined in \link{Measure}.
}
