% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{measures}
\alias{G1}
\alias{G2}
\alias{acc}
\alias{auc}
\alias{bac}
\alias{ber}
\alias{cindex}
\alias{db}
\alias{dunn}
\alias{f1}
\alias{fdr}
\alias{featperc}
\alias{fn}
\alias{fnr}
\alias{fp}
\alias{fpr}
\alias{gmean}
\alias{gpr}
\alias{mae}
\alias{mcc}
\alias{mcp}
\alias{meancosts}
\alias{measures}
\alias{medae}
\alias{medse}
\alias{mmce}
\alias{mse}
\alias{multiclass.auc}
\alias{npv}
\alias{ppv}
\alias{rmse}
\alias{sae}
\alias{silhouette}
\alias{sse}
\alias{timeboth}
\alias{timepredict}
\alias{timetrain}
\alias{tn}
\alias{tnr}
\alias{tp}
\alias{tpr}
\title{Performance measures.}
\format{none}
\usage{
featperc

timetrain

timepredict

timeboth

sse

mse

rmse

medse

sae

mae

medae

mmce

acc

ber

multiclass.auc

auc

bac

tp

tn

fp

fn

tpr

tnr

fpr

fnr

ppv

npv

fdr

mcc

f1

gmean

gpr

cindex

meancosts

mcp

db

dunn

G1

G2

silhouette
}
\description{
A performance measure is evaluated after a single train/predict step and returns a single number to assess the quality
of the prediction (or maybe only the model, think AIC).
The measure itself knows whether it wants to be minimized or maximized and for what tasks it is applicable.

All supported measures can be found by \code{\link{listMeasures}} or as a table
in the tutorial appendix: \url{http://berndbischl.github.io/mlr/tutorial/html/measures/}.

If you want a measure for a misclassification cost matrix, look at \code{\link{makeCostMeasure}}.
If you want to implement your own measure, look at \code{\link{makeMeasure}}.
}
\seealso{
Other performance: \code{\link{Measure}},
  \code{\link{makeMeasure}}; \code{\link{makeCostMeasure}};
  \code{\link{makeCustomResampledMeasure}};
  \code{\link{performance}}
}
\keyword{datasets}

