% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeSmallFactorLevels}
\alias{mergeSmallFactorLevels}
\title{Merges small levels of factors into new level.}
\usage{
mergeSmallFactorLevels(obj, target = NULL, cols = NULL, min.perc = 0.01,
  new.level = ".merged")
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{Task}}]\cr
Input data.}

\item{target}{[\code{character(1)} | \code{character(2)}]\cr
Name of the target variable(s).
Only used when \code{obj} is a data.frame, otherwise ignored.
If survival analysis is applicable, these are the names of the survival time and event columns,
so it has length 2.}

\item{cols}{[\code{character}]
Which columns to convert.
Default is all factor and character columns.}

\item{min.perc}{[\code{numeric(1)}]\cr
The smallest levels of a factor are merged until their combined proportion
w.r.t. the length of the factor exceeds \code{min.perc}.
Must be between 0 and 1.
Default is 0.01.}

\item{new.level}{[\code{character(1)}]\cr
New name of merged level.
Default is \dQuote{.merged}}
}
\value{
\code{factor}, where merged levels are combined into a new level of name \code{new.level}.
}
\description{
Merges small levels of factors into new level.
}
\seealso{
Other eda_and_preprocess: \code{\link{capLargeValues}};
  \code{\link{createDummyFeatures}};
  \code{\link{dropFeatures}};
  \code{\link{normalizeFeatures}};
  \code{\link{removeConstantFeatures}};
  \code{\link{summarizeColumns}}
}

