% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr-package}
\alias{mlr}
\alias{mlr-package}
\title{mlr: Machine Learning in R}
\description{
Interface to a large number of classification and
    regression techniques, including machine-readable parameter
    descriptions. There is also an experimental extension for survival
    analysis, clustering and general, example-specific cost-sensitive
    learning. Generic resampling, including cross-validation,
    bootstrapping and subsampling. Hyperparameter tuning with modern
    optimization techniques, for single- and multi-objective problems.
    Filter and wrapper methods for feature selection. Extension of basic
    learners with additional operations common in machine learning, also
    allowing for easy nested resampling. Most operations can be
    parallelized.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr}
  \item Report bugs at \url{https://github.com/mlr-org/mlr/issues}
}

}
\author{
\strong{Maintainer}: Bernd Bischl \email{bernd_bischl@gmx.net} (0000-0001-6002-6980)

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)
  \item Lars Kotthoff \email{larsko@uwyo.edu}
  \item Julia Schiffner \email{schiffner@math.uni-duesseldorf.de}
  \item Jakob Richter \email{code@jakob-r.de}
  \item Zachary Jones \email{zmj@zmjones.com}
  \item Giuseppe Casalicchio \email{giuseppe.casalicchio@stat.uni-muenchen.de} (0000-0001-5324-5966)
  \item Mason Gallo \email{masonagallo@gmail.com}
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)
}

Other contributors:
\itemize{
  \item Jakob Bossek \email{jakob.bossek@tu-dortmund.de} (0000-0002-4121-4668) [contributor]
  \item Erich Studerus \email{erich.studerus@upkbs.ch} (0000-0003-4233-0182) [contributor]
  \item Leonard Judt \email{leonard.judt@tu-dortmund.de} [contributor]
  \item Tobias Kuehn \email{tobi.kuehn@gmx.de} [contributor]
  \item Pascal Kerschke \email{kerschke@uni-muenster.de} (0000-0003-2862-1418) [contributor]
  \item Florian Fendt \email{flo_fendt@gmx.de} [contributor]
  \item Philipp Probst \email{philipp_probst@gmx.de} (0000-0001-8402-6790) [contributor]
  \item Xudong Sun \email{xudong.sun@stat.uni-muenchen.de} (0000-0003-3269-2307) [contributor]
  \item Janek Thomas \email{janek.thomas@stat.uni-muenchen.de} (0000-0003-4511-6245) [contributor]
  \item Bruno Vieira \email{bruno.hebling.vieira@usp.br} [contributor]
  \item Laura Beggel \email{laura.beggel@web.de} (0000-0002-8872-8535) [contributor]
  \item Quay Au \email{quay.au@stat.uni-muenchen.de} (0000-0002-5252-8902) [contributor]
  \item Martin Binder \email{ma.binder@campus.lmu.de} [contributor]
  \item Florian Pfisterer \email{pfistererf@googlemail.com} [contributor]
  \item Stefan Coors \email{stefan.coors@gmx.net} [contributor]
  \item Steve Bronder \email{sab2287@columbia.edu} [contributor]
  \item Alexander Engelhardt \email{alexander.w.engelhardt@gmail.com} [contributor]
  \item Christoph Molnar \email{christoph.molnar@stat.uni-muenchen.de} [contributor]
}

}
