% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.bh}
\alias{bh}
\alias{mlr.bh}
\title{Boston Housing regression task}
\description{
Regression task for the BostonHousing data set.
}
\references{
See \code{\link[mlbench]{BostonHousing}} in package mlbench.
}
\keyword{data}

