% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mlr.iris}
\alias{iris.rin}
\alias{iris.task}
\alias{mlr.iris}
\title{Iris classification task}
\description{
Contains the task (\code{iris.task}) and a resample instance
(\code{iris.rin}).
}
\references{
See \code{\link[datasets]{iris}}.
}
\keyword{data}

