% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.sonar}
\alias{mlr.sonar}
\alias{sonar.rin}
\alias{sonar.task}
\title{Sonar classification task}
\description{
Contains the task (\code{sonar.task}) and a matching resample instance
(\code{sonar.rin}).

Contains the task (\code{sonar.task}) and a matching resample instance
(\code{sonar.rin}).
}
\references{
See \code{\link[mlbench]{Sonar}}.

See \code{\link[mlbench]{Sonar}}.
}
\keyword{data}

