% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.wpbc}
\alias{mlr.wpbc}
\alias{wpbc}
\title{Wisonsin Prognostic Breast Cancer (WPBC) survival task}
\description{
Survival task for the WPBC data set.
}
\references{
See \code{\link[mboost]{wpbc}} in package mboost.
  Four incomplete cases have been removed from the task.
}
\keyword{data}

