% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mlr.wpbc}
\alias{mlr.wpbc}
\alias{wpbc.rin}
\alias{wpbc.task}
\title{Wisonsin Prognostic Breast Cancer (WPBC) survival task}
\description{
Contains the task (\code{wpbc.task}) and a matching resample instance
(\code{wpbc.rin}).
}
\references{
See \code{\link[TH.data]{wpbc}}.
 Incomplete cases have been removed from the task.
}
\keyword{data}

