% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mtcars.task}
\alias{mtcars.task}
\title{Motor Trend Car Road Tests clustering task.}
\description{
Contains the task (\code{mtcars.task}).
}
\references{
See \link[datasets:mtcars]{datasets::mtcars}.
}
\keyword{data}
