% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalizeFeatures}
\alias{normalizeFeatures}
\title{Normalize features}
\usage{
normalizeFeatures(obj, target = NULL, method = "standardize",
  exclude = NULL, range = c(0, 1))
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{SupervisedTask}}]\cr
Input data.}

\item{target}{[\code{character(1)}]\cr
Name of the column specifying the response.
Only used when \code{obj} is a data.frame, otherwise ignored.}

\item{method}{[\code{character(1)}]\cr
Normalizing method.\cr
Available are:\cr
\dQuote{center}: centering of each feature\cr
\dQuote{scale}: scaling of each feature\cr
\dQuote{standardize}: centering and scaling\cr
\dQuote{range}: Scale the data to a given range.\cr}

\item{exclude}{[\code{character}]\cr
Names of the columns to exclude.}

\item{range}{[\code{numeric(2)}]\cr
Range the features should be scaled to. Default is \code{c(0,1)}.}
}
\value{
[\code{data.frame} | \code{\link{SupervisedTask}}]. Same type as \code{obj}.
}
\description{
Normalize features by different methods. Internally \code{\link{normalize}} is used.
}
\seealso{
\code{\link{normalize}}
}

