\name{object-class}
\alias{object-class}
\alias{object}
\title{Base class for everything.}
\description{Base class for all S4 classes in this package.}
\section{Methods}{\describe{\item{\code{\link[=[,object-method]{[}}}{}\item{\code{\link[=to.string,object-method]{to.string}}}{}\item{\code{\link[=print,object-method]{print}}}{\code{signature( = "object")}: Generic print method.}\item{\code{\link[=show,object-method]{show}}}{\code{signature( = "object")}: Generic show method.}}}
\alias{[,object-method}

\alias{print,object-method}
\alias{print}
\alias{show,object-method}
\alias{show}
