% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pid.task}
\alias{pid.task}
\title{PimaIndiansDiabetes classification task.}
\description{
Contains the task (\code{pid.task}).
}
\references{
See \link[mlbench:PimaIndiansDiabetes]{mlbench::PimaIndiansDiabetes}.
Note that this is the uncorrected version from mlbench.
}
\keyword{data}
