% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBenchmarkResult.R
\name{plotBenchmarkResult}
\alias{plotBenchmarkResult}
\title{Create a Trellis-plot for a measure}
\usage{
plotBenchmarkResult(bmr, measure = NULL, style = "box", orderLrns = NULL,
  orderTsks = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{style}{[\code{character(1)]}]\cr
Type of Plot, can be \code{"box"} for a boxplot or \code{"violin"} for
a violin-plot.}

\item{orderLrns}{[\code{character(nLearners)}] \cr
                 [\code{integer(nLearners)}] \cr
Character vector with learner.ids in new order , or integer
vector refering to the positions in the new order.}

\item{orderLrns}{[\code{character(nTasks)}] \cr
                 [\code{integer(nTasks)}] \cr
Character vector with task.ids in new order , or integer
vector refering to the positions in the new order.}
}
\value{
[\link{ggplot2}] plot
}
\description{
Plots Boxplots for a selected measure accross all iterations of an algorithm
faceted by the task.id
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
plotBenchmarkResult(res,acc)
}

