% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBenchmarkSummary.R
\name{plotBenchmarkSummary}
\alias{plotBenchmarkSummary}
\title{plotBenchmarkSummary}
\usage{
plotBenchmarkSummary(bmr, measure = NULL, fill = "best", orderLrns = NULL,
  orderTsks = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc). \cr
Defaults to first.}

\item{fill}{[\code{character(0)}] \cr
Fill bars proportional to [\code{fill}] \cr
Can be \code{best} or \code{worst}. See \code{Details}.}

\item{orderLrns}{[\code{character(nLearners)}] \cr
                 [\code{integer(nLearners)}] \cr
Character vector with learner.ids in new order , or integer
vector refering to the positions in the new order.}

\item{orderLrns}{[\code{character(nTasks)}] \cr
                 [\code{integer(nTasks)}] \cr
Character vector with task.ids in new order , or integer
vector refering to the positions in the new order.}
}
\value{
\link{ggplot2}] plot
}
\description{
Plots a BenchmarkSummaryPlot for a selected measure. Full tiles correspond
to the worst performance accross all tasks and learners. The actuall fill
corresponds to the proportional performance of the learner in the task.
}
\details{
'best' compares all performances to the best performance accross all \cr
 classifiers. Bars are filled proportional to the best performance. \cr
'worst' compares all performances to the worst performance accross all \cr
 classifiers. Bars are filled proportional to the worst performance. \cr
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
plotBenchmarkSummary(res,acc,fill = "best")
}

