% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCritDifferences.R
\name{plotCritDifferences}
\alias{plotCritDifferences}
\title{plotCritDifferences}
\usage{
plotCritDifferences(bmr, measure = NULL, p.value = 0.05, base = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{p.value}{[\code{numeric}(1)]\cr
P-value for the critical difference. Default: 0.05}

\item{base[\code{character(1)}]:}{learner.id \cr
 Select a learner.id as baseline for the critical difference diagram,
 the critical difference will be positioned arround this learner.
 Defaults to best performing algorithm.}
}
\value{
[\link{ggplot2}] plot
}
\description{
Plots a Critical-Difference Diagram for all classifiers and a selected
measure. If a base is selected, the Critical Difference Interval will be
positioned arround the base. If not, the best performing algorithm will be
chosen as base.
}
\examples{
lrns = list(makeLearner("classif.lda"),makeLearner("classif.rpart"))
tasks = list(iris.task,sonar.task)
rdesc = makeResampleDesc("CV", iters = 3)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
plotCritDifferences(res,acc,0.05,"classif.lda")
}

