% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFilterValues.R
\name{plotFilterValuesGGVIS}
\alias{plotFilterValuesGGVIS}
\title{Plot filter values using ggvis.}
\usage{
plotFilterValuesGGVIS(fvalues, sort = "dec", n.show = 20L,
  feat.type.cols = FALSE)
}
\arguments{
\item{fvalues}{[\code{\link{FilterValues}}]\cr
Filter values.}

\item{sort}{[\code{character(1)}]\cr
Sort features like this.
\dQuote{dec} = decreasing, \dQuote{inc} = increasing, \dQuote{none} = no sorting.
Default is decreasing.}

\item{n.show}{[\code{integer(1)}]\cr
Number of features (maximal) to show.
Default is 20.}

\item{feat.type.cols}{[\code{logical(1)}]\cr
Color factor and numeric features.
\code{FALSE} means no colors.
Default is \code{FALSE}.}
}
\value{
a ggvis plot object.
}
\description{
Plot filter values using ggvis.
}
\examples{
fv = getFilterValues(iris.task, method = "chi.squared")
plotFilterValuesGGVIS(fv)
}

