% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateLearningCurve.R
\name{plotLearningCurve}
\alias{plotLearningCurve}
\title{Plot learning curve data.}
\usage{
plotLearningCurve(obj)
}
\arguments{
\item{obj}{[\code{LearningCurveData}]\cr
Result of \code{\link{generateLearningCurveData}}, with class \code{LearningCurveData}.}
}
\value{
ggplot2 plot object.
}
\description{
Visualizes data size (percentage used for model) vs. performance measure(s).
}

