% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateLearningCurve.R
\name{plotLearningCurve}
\alias{plotLearningCurve}
\title{Plot learning curve data using ggplot2.}
\usage{
plotLearningCurve(obj, color = "learner", facet = "measure")
}
\arguments{
\item{obj}{[\code{LearningCurveData}]\cr
Result of \code{\link{generateLearningCurveData}}, with class \code{LearningCurveData}.}

\item{color}{[\code{character(1)}]\cr
Selects \dQuote{measure} or \dQuote{learner} to be used for coloring the lines and points in the plot.
A vector cannot be mapped to the color aesthetic and used for facetting.
The vector mapped to \code{color} must have more than one unique value, otherwise it will
be ignored. The default is \dQuote{learner}.}

\item{facet}{[\code{character(1)}]\cr
Selects \dQuote{measure} or \dQuote{learner} to be the facetting variable.
A vector cannot be mapped to \code{color} and used for facetting.
The vector mapped to \code{facet} must have more than one unique value, otherwise it will
be ignored. The default is \dQuote{measure}.}
}
\value{
ggplot2 plot object.
}
\description{
Visualizes data size (percentage used for model) vs. performance measure(s).
}
\seealso{
Other learning_curve: \code{\link{generateLearningCurveData}};
  \code{\link{plotLearningCurveGGVIS}}

Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

