% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateLearningCurve.R
\name{plotLearningCurveGGVIS}
\alias{plotLearningCurveGGVIS}
\title{Plot learning curve data using ggvis.}
\usage{
plotLearningCurveGGVIS(obj, interactive = FALSE)
}
\arguments{
\item{obj}{[\code{LearningCurveData}]\cr
Result of \code{\link{generateLearningCurveData}}.}

\item{interactive}{[\code{logical(1)}]\cr
Whether to make the plot interactive with Shiny.
If \code{TRUE} then a sidebar menu is created that lets the user
select which learner to display.
Note that if there are multiple learners and multiple measures interactivity is
necessary as ggvis does not currently support facetting or subplots.
If \code{interactive} is \code{TRUE} but there are not multiple measures or learners then
the plot will be static.
Default is false.}
}
\value{
a ggvis plot object.
}
\description{
Visualizes data size (percentage used for model) vs. performance measure(s).
}
\seealso{
Other learning_curve: \code{\link{generateLearningCurveData}};
  \code{\link{plotLearningCurve}}

Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

