% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePartialPrediction.R
\name{plotPartialPrediction}
\alias{plotPartialPrediction}
\title{Plot a partial prediction with ggplot2}
\usage{
plotPartialPrediction(obj, facet = NULL)
}
\arguments{
\item{obj}{[\code{PartialPredictionData}]\cr
Generated by \code{\link{generatePartialPredictionData}}.}

\item{facet}{[\code{character(1)}]\cr
The name of a feature to be used for facetting.
This feature must have been an element of the \code{features} argument to
\code{\link{generatePartialPredictionData}} and is only applicable when said argument had length
greater than 1.
If \code{\link{generatePartialPredictionData}} is called with the \code{interaction} argument \code{FALSE}
(the default) with argument \code{features} of length greater than one, then \code{facet} is ignored and
each feature is plotted in its own facet.
Note that if any of the elements of the \code{features} argument of \code{\link{generatePartialPredictionData}}
are factors, they will be coerced to numerics.
Default is \code{NULL}.}
}
\value{
ggplot2 plot object.
}
\description{
Plot a partial prediction from \code{\link{generatePartialPredictionData}} using ggplot2.
}
\seealso{
Other partial_prediction: \code{\link{generatePartialPredictionData}};
  \code{\link{plotPartialPredictionGGVIS}}

Other plot: \code{\link{plotBenchmarkResult}};
  \code{\link{plotBenchmarkSummary}};
  \code{\link{plotCalibration}};
  \code{\link{plotCritDifferences}};
  \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotPartialPredictionGGVIS}};
  \code{\link{plotROCCurves}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotRankMatrixAsBar}};
  \code{\link{plotThreshVsPerfGGVIS}};
  \code{\link{plotThreshVsPerf}}
}

