% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHMHIWrapper.R
\name{plotPHMHI}
\alias{plotPHMHI}
\title{PHM HI model exploratory plots}
\usage{
plotPHMHI(learner, task, seq.ids = "all")
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{seq.ids}{Sequence id to be plotted. Default "all"}
}
\description{
Plots the created synthtic HI and prediction made by learner on a given task
}
