% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRankMatrixAsBar.R
\name{plotRankMatrixAsBar}
\alias{plotRankMatrixAsBar}
\title{Plot Rankmatrix as Barplot}
\usage{
plotRankMatrixAsBar(bmr, measure = NULL, pos = "tile", orderLrns = NULL,
  orderTsks = NULL)
}
\arguments{
\item{bmr}{\link[mlr]{BenchmarkResult}\cr
Output of a \link[mlr]{benchmark} function.}

\item{measure}{\link[mlr]{Measure} \cr
Measure for which ranks should be calculated (e.g: acc).
Defaults to first.}

\item{pos}{[\code{character(1)}]
Optionally set how the bars are positioned in ggplot2. \cr
See \code{Details} for explanation.}

\item{orderLrns}{[\code{character(nLearners)}] \cr
                 [\code{integer(nLearners)}] \cr
Character vector with learner.ids in new order , or integer
vector refering to the positions in the new order.}

\item{orderLrns}{[\code{character(nTasks)}] \cr
                 [\code{integer(nTasks)}] \cr
Character vector with task.ids in new order , or integer
vector refering to the positions in the new order.}
}
\value{
[\link{ggplot2}] plot
}
\description{
Plots a barchart from the ranks of algorithms. Alternatively
 tiles can be plotted for every rank-task combination, see \code{pos}
 for details. The x-axis accross all plots is the ranks of a learner.
 Areas are always coloured corresponding to the learner.
}
\details{
\code{'tile'} plots a heatmap with \link{task} as the y-axis.\cr
 It allows identification of the performance in a special task.\cr
 \code{'stack'} plots a stacked barplot. \cr
 It allows for comparison of learners within and and accross ranks.\cr
 \code{'dodge'} plots a barplot of with dodged instead of stacked bars.\cr
}
\examples{
lrns = list(makeLearner("classif.randomForest"),makeLearner("classif.rpart"),
            makeLearner("classif.nnet"), makeLearner("classif.svm"))
tasks = list(iris.task,sonar.task,pid.task)
rdesc = makeResampleDesc("CV", iters = 5)
meas = list(acc,mmce,ber,featperc)
res = benchmark(lrns, tasks, rdesc,meas)
plotRankMatrixAsBar(res,acc,"dodge")
plotRankMatrixAsBar(res,acc,orderLrns = c(1,2,4,3),orderTsks=c(3,2,1))
}

