% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotROCRCurves.R
\name{plotROCRCurves}
\alias{plotROCRCurves}
\title{Visualize binary classification predictions via ROC curves.}
\usage{
plotROCRCurves(obj, meas1 = "tpr", meas2 = "fpr", perf.args = list(),
  plot.args = list(), print.args = list(), interactive = FALSE,
  alpha = 0.05, ci = FALSE, task.id = NULL)
}
\arguments{
\item{obj}{[(list of) \code{\link{Prediction}} | (list of) \code{\link{ResampleResult}} | \code{\link{BenchmarkResult}}]\cr
Single prediction object, list of them, single resample result, list of them, or a benchmark result.
In case of a list probably produced by different learners you want to compare, then
name the list with the names you want to see in the plots, probably
learner shortnames or ids.}

\item{meas1}{[\code{character(1)}]\cr
Measure on x-axis. Note that this is a measure name from *ROCR* and not from mlr!
Default is \dQuote{tpr}.}

\item{meas2}{[\code{character(1)}]\cr
Measure on y-axis. Note that this is a measure name from *ROCR* and not from mlr!
Default is \dQuote{fpr}.}

\item{perf.args}{[named \code{list}]\cr
Further arguments passed to ROCR's \code{\link[ROCR]{performance}}.
Usually not needed and \code{meas1} and \code{meas2} are set internally.
Default is an empty list.}

\item{plot.args}{[named \code{list}]\cr
Further arguments to be passed to plotROC's \code{\link[plotROC]{ggroc}} or
\code{link[plotROC]{multi_ggroc}}.
Default is an empty list.}

\item{print.args}{[named \code{list}]\cr
Further arguments to be passed to plotROC's \code{\link[plotROC]{plot_journal_roc}} or
\code{\link[plotROC]{plot_interactive_roc}}.
Default is an empty list.}

\item{interactive}{[\code{logical(1)}]\cr
Whether to print the plot using \code{\link[plotROC]{plot_journal_roc}}, which occurs when
\code{interactive} is FALSE, or \code{\link[plotROC]{plot_interactive_roc}}, which occurs when
\code{interactive} is TRUE.
Default is FALSE.}

\item{alpha}{[\code{numeric(1)}]\cr
False positive rate for confidence intervals computed by \code{\link[plotROC]{calculate_roc}}.
Only applicable when \code{ci} is TRUE and a single Prediction or ResampleResult object is passed.
Default is .05.}

\item{ci}{[\code{logical(1)}]\cr
Whether to compute 1 - alpha confidence intervals using \code{\link[plotROC]{calculate_roc}}.
Only applicable when a single Prediction or Resampling object is passed.
Default is FALSE.}

\item{task.id}{[\code{character(1)}]\cr
  Selected task in \code{\link{BenchmarkResult}} to do plots for, ignored otherwise.
  Default is first task.}
}
\value{
if \code{interactive} TRUE, the output is a standalone HTML document with the plot embedded.
  By default (i.e. if the argument \code{file} is not passed to \code{\link[plotROC]{plot_interactive_roc}})
  then the document is put in a temporary file and opened in the default brower.
  If \code{interactive} FALSE then a \code{ggplot2} object is returned.

[\code{invisible(NULL)}].
}
\description{
Plot is generated by calling \code{\link{asROCRPrediction}},
ROCR's \code{\link[ROCR]{performance}},
then plotROC's \code{ggroc} or \code{multi_ggroc} function,
and then either plotROC's \code{plot_journal_roc} or \code{plot_interactive_roc}.

See these methods in ROCR and plotROC for further info.
}
\examples{
\dontrun{
lrn1 = makeLearner("classif.logreg", predict.type = "prob")
lrn2 = makeLearner("classif.rpart", predict.type = "prob")
b = benchmark(list(lrn1, lrn2), pid.task)
z = plotROCRCurves(b)
}
}
\seealso{
Other predict: \code{\link{asROCRPrediction}};
  \code{\link{getProbabilities}};
  \code{\link{plotViperCharts}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}

Other roc: \code{\link{asROCRPrediction}};
  \code{\link{plotViperCharts}}
}

