% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotROCRCurves}
\alias{plotROCRCurves}
\title{Visualize binary classification predictions via ROCR ROC curves.}
\usage{
plotROCRCurves(obj, meas1 = "tpr", meas2 = "fpr", avg = "threshold",
  legpos = "bottomright", task.id = NULL)
}
\arguments{
\item{obj}{[(list of) \code{\link{Prediction}} | \code{\link{BenchmarkResult}}]\cr
Single prediction object, list of them, or a benchmark result.
In case of a list probably produced by different learners you want to compare, then
name the list with the names you want to see in the plots, probably
learner shortnames or ids.}

\item{meas1}{[\code{character(1)}]\cr
Measure on x-axis. Note that this is a measure name from *ROCR* and not from mlr!
Default is \dQuote{tpr}.}

\item{meas2}{[\code{character(1)}]\cr
Measure on y-axis. Note that this is a measure name from *ROCR* and not from mlr!
Default is \dQuote{tpr}.}

\item{avg}{[\code{character(1)}]\cr
How to average results from resampling.
Default is \dQuote{threshold}.}

\item{task.id}{[\code{character(1)}]\cr
Selected task in \code{\link{BenchmarkResult}} to do plots for, ignored otherwise.
Default is first taks.}
}
\value{
[\code{character(1)}]. Invisibly returns the ViperCharts URL.
}
\description{
Please see \code{\link[ROCR]{ROCR}} for further info regarding plots.
}
\examples{
\dontrun{
lrn1 = makeLearner("classif.logreg", predict.type = "prob")
lrn2 = makeLearner("classif.rpart", predict.type = "prob")
b = benchmark(list(lrn1, lrn2), pid.task)
z = plotROCRCurves(b, chart = "lift", browse = TRUE)
}
}
\references{
Sluban and Lavrač - ViperCharts: Visual Performance Evaluation Platform,
ECML PKDD 2013, pp. 650-653, LNCS 8190, Springer, 2013.
}
\seealso{
Other predict: \code{\link{asROCRPrediction}};
  \code{\link{getProbabilities}};
  \code{\link{plotViperCharts}};
  \code{\link{predict.WrappedModel}};
  \code{\link{setPredictType}}

Other roc: \code{\link{asROCRPrediction}};
  \code{\link{plotViperCharts}}
}

