% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateThreshVsPerf.R
\name{plotThreshVsPerf}
\alias{plotThreshVsPerf}
\title{Plot threshold vs. performance(s) for 2-class classification using ggplot2.}
\usage{
plotThreshVsPerf(obj, mark.th = NA_real_)
}
\arguments{
\item{obj}{[\code{ThreshVsPerfData}]\cr
Result of \code{\link{generateThreshVsPerfData}}.}

\item{mark.th}{[\code{numeric(1)}]\cr
Mark given threshold with vertical line?
Default is \code{NA} which means not to do it.}
}
\value{
ggplot2 plot object.
}
\description{
Plot threshold vs. performance(s) for 2-class classification using ggplot2.
}
\examples{
\dontrun{
lrn = makeLearner("classif.rpart", predict.type = "prob")
mod = train(lrn, sonar.task)
pred = predict(mod, sonar.task)
pvs = generateThreshVsPerfData(pred, list(tpr, fpr))
plotThreshVsPerf(pvs)
}
}

