% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotThreshVsPerf}
\alias{plotThreshVsPerf}
\title{Plot threshold vs. performance(s) for 2-class classification.}
\usage{
plotThreshVsPerf(pred, measures, mark.th = NA_real_, gridsize = 100L,
  linesize = 1.5)
}
\arguments{
\item{pred}{[\code{\link{Prediction}}]\cr
Prediction object.}

\item{measures}{[\code{\link{Measure}} | list of \code{\link{Measure}}]\cr
Performance measure(s) to evaluate.}

\item{mark.th}{[\code{numeric(1)}]\cr
Mark given threshold with vertical line?
Default is \code{NA} which means not to do it.}

\item{gridsize}{[\code{integer(1)}]\cr
Grid resolution for x-axis (threshold).
Default is 100.}

\item{linesize}{[\code{numeric(1)}]\cr
Linesize for ggplot2 \code{\link[ggplot2]{geom_line}} for performance graphs.
Default is 1.5.}
}
\value{
ggplot2 plot object.
}
\description{
Plot threshold vs. performance(s) for 2-class classification.
}

