% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateThreshVsPerf.R
\name{plotThreshVsPerfGGVIS}
\alias{plotThreshVsPerfGGVIS}
\title{Plot threshold vs. performance(s) for 2-class classification using ggvis.}
\usage{
plotThreshVsPerfGGVIS(obj, color = "learner", interaction = "measure",
  mark.th = NA_real_)
}
\arguments{
\item{obj}{[\code{ThreshVsPerfData}]\cr
Result of \code{\link{generateThreshVsPerfData}}.}

\item{color}{[\code{character(1)}]\cr
Selects \dQuote{measure} or \dQuote{learner} to be used for coloring the lines and points in the plot.
A vector cannot be mapped to the color aesthetic and used for interaction.
The vector mapped to \code{color} must have more than one unique value, otherwise it will
be ignored. The default is \dQuote{learner}.}

\item{interaction}{[\code{character(1)}]\cr
Selects \dQuote{measure} or \dQuote{learner} to be used in a Shiny application
making the \code{interaction} variable selectable via a drop-down menu.
A vector cannot be mapped to \code{color} and used for interaction.
Note that if there are multiple learners and multiple measures interactivity is
necessary as ggvis does not currently support facetting or subplots.
The vector mapped to \code{interaction} must have more than one unique value, otherwise it will
be ignored. The default is \dQuote{measure}.}

\item{mark.th}{[\code{numeric(1)}]\cr
Mark given threshold with vertical line?
Default is \code{NA} which means not to do it.}
}
\value{
a ggvis plot object.
}
\description{
Plot threshold vs. performance(s) for 2-class classification using ggvis.
}
\examples{
\dontrun{
lrn = makeLearner("classif.rpart", predict.type = "prob")
mod = train(lrn, sonar.task)
pred = predict(mod, sonar.task)
pvs = generateThreshVsPerfData(pred, list(tpr, fpr))
plotThreshVsPerfGGVIS(pvs)
}
}
\seealso{
Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerf}}

Other thresh_vs_perf: \code{\link{generateThreshVsPerfData}};
  \code{\link{plotThreshVsPerf}}
}

