% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateThreshVsPerf.R
\name{plotThreshVsPerfGGVIS}
\alias{plotThreshVsPerfGGVIS}
\title{Plot threshold vs. performance(s) for 2-class classification using ggvis.}
\usage{
plotThreshVsPerfGGVIS(obj, mark.th = NA_real_, interactive = FALSE)
}
\arguments{
\item{obj}{[\code{ThreshVsPerfData}]\cr
Result of \code{\link{generateThreshVsPerfData}}.}

\item{mark.th}{[\code{numeric(1)}]\cr
Mark given threshold with vertical line?
Default is \code{NA} which means not to do it.}

\item{interactive}{[\code{logical(1)}]\cr
Whether to make the plot interactive with Shiny.
If \code{TRUE} then a sidebar menu is created that lets the user
select which learner to display.
Note that if there are multiple learners and multiple measures interactivity is
necessary as ggvis does not currently support facetting or subplots.
If \code{interactive} is \code{TRUE} but there are not multiple measures or learners then
the plot will be static.
Note that \code{mark.th} is not supported when \code{interactive = TRUE}.
Default is \code{FALSE}.}
}
\value{
a ggvis plot object.
}
\description{
Plot threshold vs. performance(s) for 2-class classification using ggvis.
}
\examples{
\dontrun{
lrn = makeLearner("classif.rpart", predict.type = "prob")
mod = train(lrn, sonar.task)
pred = predict(mod, sonar.task)
pvs = generateThreshVsPerfData(pred, list(tpr, fpr))
plotThreshVsPerfGGVIS(pvs)
}
}
\seealso{
Other plot: \code{\link{plotFilterValuesGGVIS}};
  \code{\link{plotFilterValues}};
  \code{\link{plotLearningCurveGGVIS}};
  \code{\link{plotLearningCurve}};
  \code{\link{plotROCRCurvesGGVIS}};
  \code{\link{plotROCRCurves}};
  \code{\link{plotThreshVsPerf}}

Other thresh_vs_perf: \code{\link{generateThreshVsPerfData}};
  \code{\link{plotThreshVsPerf}}
}

