% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedLearner_predict.R
\name{predictLearner.StackedLearner}
\alias{predictLearner.StackedLearner}
\title{Predict StackedLearner}
\usage{
\method{predictLearner}{StackedLearner}(.learner, .model, .newdata)
}
\arguments{
\item{.learner}{[\code{StackedLearner}]}

\item{.model}{[\code{BaseEnsembleModel}]}

\item{.newdata}{[\code{data.frame}]\cr Data to predict on}
}
\value{
Predictions are returned in matrix or vector.
}
\description{
Gets predictions from \code{getStackedBaseLearnerPredictions}. Then apply algo:
 \describe{
  \item{\code{average}}{Aggregate predictions using \code{\link{aggregatePredictions}}.}
  \item{\code{hill.climb}}{Expand prediction according to frequency chosen in training, then apply aggregation as for 'average'.}
  \item{\code{stack.cv}}{Convert list of \code{Prediction} to data.frame 
  (for classification with \code{predict.type = "prob"} the first feature 
   is removed to overcome mullticollinearity). Add original features if needed. 
   Apply meta model from training.}
 }
}

