\name{rda-class}
\alias{rda-class}
\alias{rda}
\title{rda}
\description{Wrapped learner for Regularized Discriminant Analysis from package \code{klaR} for classification problems.}
\details{\emph{Common hyperparameters:}
\describe{
\item{\code{gamma}}		
\item{\code{lambda}}	
\item{\code{startsimplex}}{(Optional) a starting simplex for the Nelder-Mead-minimization.}	
}}
\seealso{\code{\link[klaR]{rda}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,rda-method]{initialize}}}{\code{signature( = "rda")}: Constructor.}}}

