% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLearner_regr_featureless.R
\name{regr.featureless}
\alias{regr.featureless}
\title{Featureless regression learner.}
\description{
A very basic baseline method which is useful for model comparisons (if you
don't beat this, you very likely have a problem).
Does not consider any features of the task and only uses the target feature
of the training data to make predictions.
Using observation weights is currently not supported.

Methods \dQuote{mean} and \dQuote{median} always predict a constant value
for each new observation which corresponds to the observed mean or median of
the target feature in training data, respectively.

The default method is \dQuote{mean} which corresponds to the ZeroR algorithm
from WEKA, see \url{https://weka.wikispaces.com/ZeroR}.
}

