\name{regr.task-class}
\alias{regr.task-class}
\alias{regr.task}
\title{classif.task}
\description{General description object for a regression experiment.  
Instantiate it by using its factory method.}
\seealso{learn.task make.regr.task}
\section{Extends}{\code{\linkS4class{learn.task}}}
\section{Methods}{\describe{\item{\code{\link[=predict,regr.task-method]{predict}}}{\code{signature(object = "regr.task")}: Predicts the target values of a new data set based on 
an already fitted wrapped.model of a regression task.}\item{\code{\link[=initialize,regr.task-method]{initialize}}}{\code{signature( = "regr.task")}: Constructor.}}}
