\name{regr.task-class}
\alias{regr.task-class}
\alias{regr.task}
\title{Regression task.}
\description{General description object for a regression experiment.  
Instantiate it by using its factory method.}
\seealso{\code{\link{make.task}}}
\section{Extends}{\code{\linkS4class{learn.task}}}
\section{Methods}{\describe{\item{\code{\link[=to.string,regr.task-method]{to.string}}}{}}}
