% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{regression_fit}
\alias{regression_fit}
\title{regression_fit}
\usage{
regression_fit(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" or one that can be coerced by as.formula}

\item{data}{an object of class "data.frame" or one that can be coerced by as.data.frame.}
}
\value{
A list containing the model formula, fitted values, a table of coefficients with test results, and the results of an ANOVA test.
}
\description{
Fits multiple regression model
}
\examples{
data(mtcars)
model_fit <- regression_fit(formula = mpg ~ wt + hp + wt:hp, data = mtcars)


}
