% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeHyperPars.R
\name{removeHyperPars}
\alias{removeHyperPars}
\title{Remove hyperparameters settings of a learner.}
\usage{
removeHyperPars(learner, ids = character(0L))
}
\arguments{
\item{learner}{(\link{Learner} | \code{character(1)})\cr
The learner.
If you pass a string the learner will be created via \link{makeLearner}.}

\item{ids}{(\link{character})\cr
Parameter names to remove settings for.
Default is \code{character(0L)}.}
}
\value{
\link{Learner}.
}
\description{
Remove settings (previously set through mlr) for some parameters.
Which means that the default behavior for that param will now be used.
}
\seealso{
Other learner: 
\code{\link{LearnerProperties}},
\code{\link{getClassWeightParam}()},
\code{\link{getHyperPars}()},
\code{\link{getLearnerId}()},
\code{\link{getLearnerNote}()},
\code{\link{getLearnerPackages}()},
\code{\link{getLearnerParVals}()},
\code{\link{getLearnerParamSet}()},
\code{\link{getLearnerPredictType}()},
\code{\link{getLearnerShortName}()},
\code{\link{getLearnerType}()},
\code{\link{getParamSet}()},
\code{\link{helpLearnerParam}()},
\code{\link{helpLearner}()},
\code{\link{makeLearners}()},
\code{\link{makeLearner}()},
\code{\link{setHyperPars}()},
\code{\link{setId}()},
\code{\link{setLearnerId}()},
\code{\link{setPredictThreshold}()},
\code{\link{setPredictType}()}
}
\concept{learner}
