\name{resample.instance-class}
\alias{resample.instance-class}
\alias{resample.instance}
\title{resample.instance}
\description{Base class for specific resampling draws like cross-validation or bootstrapping.
This class encapsulates training and test sets generated from the data set for a number of iterations. 
It mainly stores a set of integer vectors indicating the training examples for each iteration.
Don't create objects from this class directly but use the corresponding subclasses.
For construction simply use the factory methods of the subclasses - e.g. for cross-validation 
use \code{\link{make.cv.instance}} to get a \code{\linkS4class{cv.instance}}.}
\note{If you want to add another resampling strategy, have a look at the web documentation.}
\seealso{\code{\linkS4class{resample.desc}}, \code{\link{make.cv.instance}}, \code{\link{make.bs.instance}}, \code{\link{make.subsample.instance}}, \code{\link{resample.fit}}}
\section{Slots}{\describe{\item{\code{desc}:}{(\code{\link{resample.desc}}) Description object for resampling strategy}\item{\code{size}:}{(\code{\link{integer}}) Number of observations in data}\item{\code{inds}:}{(\code{\link{list}}) List of integer vectors specifying the training cases for each iteration. Each vector might contain duplicated indices and the order matters for some classifiers.}}}
\section{Methods}{\describe{\item{\code{\link[=benchmark,wrapped.learner,learn.task,resample.instance-method]{benchmark}}}{\code{signature(learner = "wrapped.learner", task = "learn.task", resampling = "resample.instance")}: \code{benchmark} conducts a benchmark experiment for a single classifier on a single
data set. This consists of an inner stage and outer stage. At the outer stage a 
tuning set and a test set are repeatedly formed from the data through resampling 
(usually cross-validation or bootstrapping). The respective hyperparameters of the 
classifier are tuned on the tuning set again through an inner resampling process,
the classifier is trained on the complete tuning set with the best found 
hyperparameters and the performance is measured on the test set.}\item{\code{\link[=resample.fit,wrapped.learner,learn.task,resample.instance,list,character,character,function-method]{resample.fit}}}{\code{signature(learner = "wrapped.learner", task = "learn.task", resampling = "resample.instance", parset = "list", vars = "character", type = "character", extract = "function")}: Given the training and test indices (e.g. generated by cross-validation and generally specified by 
the \code{\linkS4class{resample.instance}} object) \code{resample.fit} 
fits the selected learner using the training sets and performs predictions for the test sets. These 
predictions are returned - encapsulated in a \code{\link{resample.result}} object. 
Optionally the fitted models are also stored.}\item{\code{\link[=initialize,resample.instance-method]{initialize}}}{\code{signature( = "resample.instance")}: This is mainly for internal use, you only need to use this, when you extend resample.instance to add another resampling strategy!}\item{\code{\link[=[,resample.instance-method]{[}}}{\code{signature( = "resample.instance")}: Getter.}}}
