% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplePrediction.R
\name{ResamplePrediction}
\alias{ResamplePrediction}
\title{Prediction from resampling.}
\description{
Contains predictions from resampling, returned (among other stuff) by function \link{resample}.
Can basically be used in the same way as \link{Prediction}, its super class.
The main differences are:
(a) The internal data.frame (member \code{data}) contains an additional column \code{iter}, specifying the iteration
of the resampling strategy, and and additional columns \code{set}, specifying whether the prediction
was from an observation in the \dQuote{train} or \dQuote{test} set. (b) The prediction \code{time} is
a numeric vector, its length equals the number of iterations.
}
\seealso{
Other resample: 
\code{\link{ResampleResult}},
\code{\link{addRRMeasure}()},
\code{\link{getRRPredictionList}()},
\code{\link{getRRPredictions}()},
\code{\link{getRRTaskDescription}()},
\code{\link{getRRTaskDesc}()},
\code{\link{makeResampleDesc}()},
\code{\link{makeResampleInstance}()},
\code{\link{resample}()}
}
\concept{resample}
