% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{retrafo<-}
\alias{retrafo<-}
\title{set an object's retransformation}
\usage{
retrafo(data) <- value
}
\description{
Set an object's retransformation function, as described
in \code{\link{retrafo}}. Set to \code{NULL} to delete.
}
