% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOAuxiliary.R
\name{retrafo<-}
\alias{retrafo<-}
\title{set an object's retransformation}
\usage{
retrafo(data) <- value
}
\arguments{
\item{data}{[\code{data.frame} | \code{\link{Task}}]\cr
The task of which to set the retrafo.}

\item{value}{[\code{function} | NULL]\cr
The retrafo function to set. This must either be a
function accepting a \code{data.frame} and returning
an object of the same kind, or NULL.
In most cases, you should use this only within
\code{CPOFunctionalConstructor} functions OR to
reset an object's retrafo to NULL.}
}
\description{
Set an object's retransformation function, as described
in \code{\link{retrafo}}. Set to \code{NULL} to delete.
}
\seealso{
Other CPO: \code{\link{CPO}}, \code{\link{NULLCPO}},
  \code{\link{\%>>\%}}, \code{\link{applyCPO}},
  \code{\link{attachCPO}}, \code{\link{chainCPO}},
  \code{\link{cpoPca}}, \code{\link{cpoScale}},
  \code{\link{getCPOName}}, \code{\link{getHyperPars}},
  \code{\link{getRetrafoState}},
  \code{\link{makeCPOTargetOp}}, \code{\link{makeCPO}},
  \code{\link{makeRetrafoFromState}},
  \code{\link{retrafo}}, \code{\link{setHyperPars}}
}
