% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{A multiple ridge regression model (RC)}
\arguments{
\item{formula}{Contains dependent and independent variables for linear regression}

\item{data}{A data.frame to conduct linear regression}
}
\description{
Class for the convenient multiple ridge regression
}
\section{Methods}{

\describe{
\item{\code{coef()}}{Return the coefficients as a named vector}

\item{\code{predict()}}{Return the predicted values y_hat}

\item{\code{print()}}{Print out the coefficients and coefficient names}
}}

