\name{rpart.classif-class}
\alias{rpart.classif-class}
\alias{rpart.classif}
\title{rpart.classif}
\description{Wrapped learner for Classification Trees from package \code{rpart}.}
\details{\emph{Common hyperparameters:}}
\seealso{\code{\link[rpart]{rpart}}}
\section{Extends}{\code{\linkS4class{wrapped.learner.classif}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,rpart.classif-method]{initialize}}}{\code{signature( = "rpart.classif")}: Constructor.}}}

