% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_operators.R
\name{setAggregation}
\alias{setAggregation}
\title{Set aggregation function of measure.}
\usage{
setAggregation(measure, aggr)
}
\arguments{
\item{measure}{(\link{Measure})\cr
Performance measure.}

\item{aggr}{(\link{Aggregation})\cr
Aggregation function.}
}
\value{
(\link{Measure}) with changed aggregation behaviour.
}
\description{
Set how this measure will be aggregated after resampling.
To see possible aggregation functions: \link{aggregations}.
}
\seealso{
Other performance: 
\code{\link{ConfusionMatrix}},
\code{\link{calculateConfusionMatrix}()},
\code{\link{calculateROCMeasures}()},
\code{\link{estimateRelativeOverfitting}()},
\code{\link{makeCostMeasure}()},
\code{\link{makeCustomResampledMeasure}()},
\code{\link{makeMeasure}()},
\code{\link{measures}},
\code{\link{performance}()},
\code{\link{setMeasurePars}()}
}
\concept{performance}
