% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHyperPars.R
\name{setHyperPars}
\alias{setHyperPars}
\title{Set the hyperparameters of a learner object.}
\usage{
setHyperPars(learner, ..., par.vals = list(), reset = "no")
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{...}{[any]\cr
Named (hyper)parameters with new setting. Alternatively these can be passed
using the \code{par.vals} argument.}

\item{par.vals}{[\code{list}]\cr
Optional list of named (hyper)parameter settings. The arguments in
\code{...} take precedence over values in this list.}

\item{reset}{[\code{character}]\cr
Can take values \code{soft} for setting the hyper parameters while keeping the \code{mlr.defaults} as far as they are feasible. 
\code{no} for updating the old parameter values with new ones without dropping old ones. A feasibility error then might occur later.
\code{hard} for completely dropping old parameter values and just unsing the new ones.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Set the hyperparameters of a learner object.
}
\note{
If a named (hyper)parameter can't be found for the given learner, the 3
closest (hyper)parameter names will be output in case the user mistyped.
}
\examples{
cl1 = makeLearner("classif.ksvm", sigma = 1)
cl2 = setHyperPars(cl1, sigma = 10, par.vals = list(C = 2))
print(cl1)
# note the now set and altered hyperparameters:
print(cl2)
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getClassWeightParam}},
  \code{\link{getHyperPars}}, \code{\link{getParamSet}},
  \code{\link{makeLearner}}, \code{\link{removeHyperPars}},
  \code{\link{setId}}, \code{\link{setPredictThreshold}},
  \code{\link{setPredictType}}
}

