% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHyperPars.R
\name{setHyperPars2}
\alias{setHyperPars2}
\title{Only exported for internal use.}
\usage{
setHyperPars2(learner, par.vals, show.info = getMlrOption("show.info"))
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{par.vals}{[\code{list}]\cr
List of named (hyper)parameter settings.}

\item{show.info}{[\code{logical(1)}]\cr
Print verbose output on console?
Default is set via \code{\link{configureMlr}}.}
}
\description{
Only exported for internal use.
}

