% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHyperPars.R
\name{setHyperPars2}
\alias{setHyperPars2}
\title{Only exported for internal use.}
\usage{
setHyperPars2(learner, par.vals, use.mlr.defaults = TRUE, update = TRUE)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}
}
\description{
Only exported for internal use.
}

