% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setHyperPars2}
\alias{setHyperPars2}
\title{Only exported for internal use.}
\usage{
setHyperPars2(learner, par.vals)
}
\arguments{
\item{learner}{[\code{\link{Learner}}]\cr
The learner.}

\item{par.vals}{[\code{list}]\cr
List of named (hyper)parameter settings.}
}
\description{
Only exported for internal use.
}

