% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHyperPars.R
\name{setHyperPars2}
\alias{setHyperPars2}
\title{Only exported for internal use.}
\usage{
setHyperPars2(learner, par.vals)
}
\arguments{
\item{learner}{(\link{Learner})\cr
The learner.}

\item{par.vals}{(\link{list})\cr
List of named (hyper)parameter settings.}
}
\description{
Only exported for internal use.
}
