% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setId.R
\name{setId}
\alias{setId}
\title{Set the id of a learner object.}
\usage{
setId(learner, id)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{id}{[\code{character(1)}]\cr
New id for learner.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Set the id of a learner object.
}
\seealso{
Other learner: \code{\link{LearnerProperties}},
  \code{\link{getLearnerProperties}},
  \code{\link{hasLearnerProperties}};
  \code{\link{Learner}}, \code{\link{makeLearner}};
  \code{\link{getClassWeightParam}};
  \code{\link{getHyperPars}}; \code{\link{getParamSet}};
  \code{\link{removeHyperPars}};
  \code{\link{setHyperPars}};
  \code{\link{setPredictThreshold}};
  \code{\link{setPredictType}}
}

