% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showHyperPars}
\alias{showHyperPars}
\title{Display all possible hyperparameter settings for a learner that mlr knows.}
\usage{
showHyperPars(learner)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Useful for a quick overview, also does not foce you to create the learner.
}
\seealso{
Other learner: \code{\link{getHyperPars}};
  \code{\link{getParamSet}}; \code{\link{setHyperPars}};
  \code{\link{setId}}; \code{\link{setPredictType}}
}

