% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Task_operators.R
\name{subsetTask}
\alias{subsetTask}
\title{Subset data in task.}
\usage{
subsetTask(task, subset, features)
}
\arguments{
\item{task}{[\code{\link{Task}}]\cr
The task.}

\item{subset}{[\code{integer} | \code{logical(n)}]\cr
Selected cases.
Default is all cases.}

\item{features}{[\code{character}]\cr
Selected inputs. Note that target feature is always included in the
resulting task, you should not pass it here.
Default is all features.}
}
\value{
[\code{\link{Task}}]. Task with subsetted data.
}
\description{
Subset data in task.
}
\examples{
task = makeClassifTask(data = iris, target = "Species")
subsetTask(task, subset = 1:100)
}
\seealso{
Other task: \code{\link{getTaskCosts}};
  \code{\link{getTaskData}};
  \code{\link{getTaskDescription}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFormula}},
  \code{\link{getTaskFormulaAsString}};
  \code{\link{getTaskId}}; \code{\link{getTaskNFeats}};
  \code{\link{getTaskSize}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTargets}}; \code{\link{getTaskType}}
}

