% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarizeLevels}
\alias{summarizeLevels}
\title{Summarize factors of a data.frame.}
\usage{
summarizeLevels(data, which)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data to summarize.
Characters and logicals will be treated as factors.}

\item{which}{[\code{character}]\cr
Restrict result to columns in \code{which}.
Default is all factor, character and logical columns of \code{data}.}
}
\value{
A named list of tables.
}
\description{
Summarizes factors of a data.frame by tabling them.
}

