% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeLevels.R
\name{summarizeLevels}
\alias{summarizeLevels}
\title{Summarizes factors of a data.frame by tabling them.}
\usage{
summarizeLevels(obj, cols = NULL)
}
\arguments{
\item{obj}{[\code{data.frame} | \code{\link{Task}}]\cr
Input data.}

\item{cols}{[\code{character}]\cr
Restrict result to columns in \code{cols}.
Default is all factor, character and logical columns of \code{obj}.}
}
\value{
[\code{list}]. Named list of tables.
}
\description{
Characters and logicals will be treated as factors.
}

