% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeClassifTask}
\alias{ClassifTask}
\alias{ClusterTask}
\alias{CostSensTask}
\alias{RegrTask}
\alias{SurvTask}
\alias{Task}
\alias{makeClassifTask}
\alias{makeClusterTask}
\alias{makeCostSensTask}
\alias{makeRegrTask}
\alias{makeSupervisedTask}
\alias{makeSurvTask}
\alias{makeTask}
\alias{makeUnsupervisedTask}
\title{Create a classification, regression, survival, cluster, or cost-sensitive classification task.}
\usage{
makeClassifTask(id = deparse(substitute(data)), data, target,
  weights = NULL, blocking = NULL, positive = NA_character_)

makeClusterTask(id = deparse(substitute(data)), data, weights = NULL,
  blocking = NULL)

makeCostSensTask(id = deparse(substitute(data)), data, costs,
  blocking = NULL)

makeRegrTask(id = deparse(substitute(data)), data, target, weights = NULL,
  blocking = NULL)

makeSupervisedTask(id = deparse(substitute(data)), data, target,
  weights = NULL, blocking = NULL)

makeSurvTask(id = deparse(substitute(data)), data, target,
  censoring = "rcens", weights = NULL, blocking = NULL)

makeTask(id = as.character(deparse(substitute(data))), data, weights = NULL,
  blocking = NULL)

makeUnsupervisedTask(id = deparse(substitute(data)), data, weights = NULL,
  blocking = NULL)
}
\arguments{
\item{id}{[\code{character(1)}]\cr
Id string for object.
Default is the name of R variable passed to \code{data}.}

\item{data}{[\code{data.frame}]\cr
A data frame containing the features and target variable(s).}

\item{target}{[\code{character}]\cr
Name of the target column(s) in \code{data}.}

\item{weights}{[\code{numeric}]\cr
Optional, non-negative case weight vector to be used during fitting.
Cannot be set for cost-sensitive learning.
Default is \code{NULL} which means no (= equal) weights.}

\item{blocking}{[\code{factor}]\cr
An optional factor of the same length as the number of observations.
Observations with the same blocking level \dQuote{belong together}.
Specifically, they are either put all in the training or the test set
during a resampling iteration.
Default is \code{NULL} which means no blocking.}

\item{costs}{[\code{data.frame}]\cr
A numeric matrix or data frame containing the costs of misclassification.
We assume the general case of observation specific costs.
This means we have n rows, corresponding to the observations, in the same order as \code{data}.
The columns correspond to classes and their names are the class labels
(if unnamed we use y1 to yk as labels).
Each entry (i,j) of the matrix specifies the cost of predicting class j
for observation i.}

\item{censoring}{[\code{character(1)}]\cr
Censoring type. Allowed choices are \dQuote{rcens} for right censored data (default),
\dQuote{lcens} for left censored and \dQuote{icens} for interval censored data using
the \dQuote{interval2} format.
See \code{\link[survival]{Surv}} for details.}
}
\value{
[\code{\link{Task}}].
}
\description{
The task encapsulates the data and specifies - through its subclasses -
the type of the task.


Object members:
\describe{
\item{env [\code{environment}]}{Environment where data for the task are stored.
  Use \code{\link{getTaskData}} in order to access it.}
\item{weights [\code{numeric}]}{See argument above. \code{NULL} if not present.}
\item{blocking [\code{factor}]}{See argument above. \code{NULL} if not present.}
\item{task.desc [\code{\link{TaskDesc}}]}{Encapsulates further information about the task.}
}
}
\examples{
library(mlbench)
data(BostonHousing)
data(Ionosphere)

makeClassifTask(data = iris, target = "Species")
makeRegrTask(data = BostonHousing, target = "medv")
# an example of a classification task with more than those standard arguments:
blocking = factor(c(rep(1, 51), rep(2, 300)))
makeClassifTask(id = "myIonosphere", data = Ionosphere, target = "Class",
  positive = "good", blocking = blocking)
makeClusterTask(data = iris[, -5L])
}
\seealso{
Other costsens: \code{\link{CostSensClassifModel}},
  \code{\link{CostSensClassifWrapper}},
  \code{\link{makeCostSensClassifWrapper}};
  \code{\link{CostSensRegrModel}},
  \code{\link{CostSensRegrWrapper}},
  \code{\link{makeCostSensRegrWrapper}};
  \code{\link{CostSensWeightedPairsModel}},
  \code{\link{CostSensWeightedPairsWrapper}},
  \code{\link{makeCostSensWeightedPairsWrapper}}

Other Task: \code{\link{copyTask}};
  \code{\link{getTaskFeatureNames}};
  \code{\link{getTaskFeatures}};
  \code{\link{getTaskTargetNames}};
  \code{\link{getTaskTarget}}
}

