% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneThreshold.R
\name{tuneThreshold}
\alias{tuneThreshold}
\title{Tune prediction threshold.}
\usage{
tuneThreshold(pred, measure, task, model, nsub = 20L, control = list())
}
\arguments{
\item{pred}{([Prediction])\cr
Prediction object.}

\item{measure}{([Measure])\cr
Performance measure to optimize.
Default is the default measure for the task.}

\item{task}{([Task])\cr
Learning task. Rarely neeeded,
only when required for the performance measure.}

\item{model}{([WrappedModel])\cr
Fitted model. Rarely neeeded,
only when required for the performance measure.}

\item{nsub}{(`integer(1)`)\cr
Passed to [BBmisc::optimizeSubInts] for 2class problems.
Default is 20.}

\item{control}{([list])\cr
Control object for [cmaes::cma_es] when used.
Default is empty list.}
}
\value{
([list]). A named list with with the following components:
  `th` is the optimal threshold, `perf` the performance value.
}
\description{
Optimizes the threshold of predictions based on probabilities.
Works for classification and multilabel tasks.
Uses [BBmisc::optimizeSubInts] for normal binary class problems and [cmaes::cma_es]
for multiclass and multilabel problems.
}
\seealso{
Other tune: \code{\link{TuneControl}},
  \code{\link{getNestedTuneResultsOptPathDf}},
  \code{\link{getNestedTuneResultsX}},
  \code{\link{getResamplingIndices}},
  \code{\link{getTuneResult}},
  \code{\link{makeModelMultiplexerParamSet}},
  \code{\link{makeModelMultiplexer}},
  \code{\link{makeTuneControlCMAES}},
  \code{\link{makeTuneControlDesign}},
  \code{\link{makeTuneControlGenSA}},
  \code{\link{makeTuneControlGrid}},
  \code{\link{makeTuneControlIrace}},
  \code{\link{makeTuneControlMBO}},
  \code{\link{makeTuneControlRandom}},
  \code{\link{makeTuneWrapper}}, \code{\link{tuneParams}}
}

