% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wpbc.task}
\alias{wpbc.task}
\title{Wisonsin Prognostic Breast Cancer (WPBC) survival task.}
\description{
Contains the task (\code{wpbc.task}).
}
\references{
See \link[TH.data:wpbc]{TH.data::wpbc}.
Incomplete cases have been removed from the task.
}
\keyword{data}
