\name{wrapped.learner.classif-class}
\alias{wrapped.learner.classif-class}
\alias{wrapped.learner.classif}
\title{wrapped.learner.classif}
\description{Wraps an already implemented classification method from R to make it accesible to mlr.}
\section{Extends}{\code{\linkS4class{wrapped.learner}}}
\section{Methods}{\describe{\item{\code{\link[=train,wrapped.learner.classif,classif.task,numeric,list,character-method]{train}}}{}}}
