\name{wrapped.learner.classif-class}
\alias{wrapped.learner.classif-class}
\alias{wrapped.learner.classif}
\title{wrapped.learner.classif}
\description{Wraps an already implemented classification method from R to make it accesible to mlr.}
\section{Slots}{\describe{\item{\code{train.par.for.classes}:}{(\code{\link{list}}) List of parameters to train a model for class predictions. Will be added to train.fct.pars in this case.}\item{\code{train.par.for.probs}:}{(\code{\link{list}}) List of parameters to train a model for probability predictions. Will be added to train.fct.pars in this case.}\item{\code{predict.par.for.classes}:}{(\code{\link{list}}) List of parameters to predict a model for class predictions. Will be added to predict.fct.pars in this case.}\item{\code{predict.par.for.probs}:}{(\code{\link{list}}) List of parameters to predict a model for probability predictions. Will be added to predict.fct.pars in this case.}\item{\code{trafo.for.classes}:}{(\code{\link{function}}) Function to transform the raw predictions to classes}\item{\code{trafo.for.probs}:}{(\code{\link{function}}) Function to transform the raw predictions to probabilities}\item{\code{dummy.classes}:}{(\code{\link{logical}}) Does the predict function need a class column in the newdata dataframe for prediction? If TRUE but no class column is avaible in the data a null column is internally generated (default is FALSE).}}}
\section{Extends}{\code{\linkS4class{wrapped.learner}}}
\section{Methods}{\describe{\item{\code{\link[=initialize,wrapped.learner.classif-method]{initialize}}}{\code{signature( = "wrapped.learner.classif")}: Constructor.}}}
