\name{wrapped.learner.regr-class}
\alias{wrapped.learner.regr-class}
\alias{wrapped.learner.regr}
\title{wrapped.learner.regr}
\description{Wrapped.learner.regr ---text!----}
\section{Slots}{\describe{\item{\code{train.par.for.classes}:}{}\item{\code{train.par.for.probs}:}{}\item{\code{predict.par.for.classes}:}{}\item{\code{predict.par.for.probs}:}{}\item{\code{trafo.for.classes}:}{}\item{\code{trafo.for.probs}:}{}\item{\code{dummy.classes}:}{}}}
\section{Extends}{\code{\linkS4class{wrapped.learner}}}

