\name{wrapped.model-class}
\alias{wrapped.model-class}
\alias{wrapped.model}
\title{wrapped.model}
\description{A classifier model is generated by \code{\link{train}}. It consists of the learn.task,
the classification model, the used hyperparameters and a vector specifying the cases of the training sample.}
\examples{see \link{train}}
\section{Slots}{\describe{\item{\code{task.class}:}{(\code{\link{character}}) String}\item{\code{learner.class}:}{(\code{\link{character}}) Specifies learner class}\item{\code{learner.name}:}{(\code{\link{character}}) Name of the learning method}\item{\code{learner.model}:}{(\code{\link{ANY}}) External model from existing R packages like lda, rpart, etc}\item{\code{subset}:}{(\code{\link{numeric}}) An index vector specifying the cases of the training sample that were used for the model.}\item{\code{vars}:}{(\code{\link{character}}) Vector specifying the variables which were used to build the model.}\item{\code{parset}:}{(\code{\link{list}}) Contains the hyperparameters of the train function. If empty no parameters were used.}}}
\section{Methods}{\describe{\item{\code{\link[=[,wrapped.model-method]{[}}}{\code{signature( = "wrapped.model")}: Getter.}}}
