\name{wrapped.model-class}
\alias{wrapped.model-class}
\alias{wrapped.model}
\title{wrapped.model}
\description{A classifier model is generated by \code{\link{train}}. It consists of the learn.task,
the classification model, the used hyperparameters and a vector specifying the cases of the training sample.}
\examples{see \link{train}}
\section{Slots}{\describe{\item{\code{task.class}:}{(\code{\link{character}}) [\code{\link{learn.task}}] Specifies learning task}\item{\code{learner.class}:}{(\code{\link{character}}) [character] Specifies learner class}\item{\code{learner.name}:}{(\code{\link{character}}) [character] Name of the learning method}\item{\code{learner.model}:}{(\code{\link{ANY}}) [external.model] External model from existing R packages like lda, rpart, etc}\item{\code{subset}:}{(\code{\link{numeric}}) [integer] An index vector specifying the cases of the training sample that were used for the model.}\item{\code{vars}:}{(\code{\link{character}}) [charactter] Vector specifying the variables which were used to build the model.}\item{\code{parset}:}{(\code{\link{list}}) [list] Contains the hyperparameters of the train function. If empty no parameters were used.}}}
\section{Methods}{\describe{\item{\code{\link[=as.character,wrapped.model-method]{as.character}}}{\code{signature( = "wrapped.model")}: Conversion to string.}\item{\code{\link[=print,wrapped.model-method]{print}}}{\code{signature( = "wrapped.model")}: Prints the object by calling as.character.}\item{\code{\link[=show,wrapped.model-method]{show}}}{\code{signature( = "wrapped.model")}: Shows the object by calling as.character.}\item{\code{\link[=[,wrapped.model-method]{[}}}{}}}

