\name{wrapped.model-class}
\alias{wrapped.model-class}
\alias{wrapped.model}
\title{Induced model of learner.}
\description{Result from \code{\link{train}}. It internally stores the underlying fitted model,
the IDs of the subset used for training, variables used for training and    
information about second-level optimization like tuned hyperparameters or selected variables.}
\details{\code{wrapped.model-class}: Getter.\cr
Note that all getters of \code{\linkS4class{task.desc}} and \code{\linkS4class{data.desc}} can also be used. 

\describe{
\item{learner [{\linkS4class{learner}}]}{Learner that was used to fit the model.}
\item{learner model [any]}{Underlying model from used R package.}
\item{subset [integer]}{Subset used for training.}
\item{fail [NULL | string]}{Generally NULL but if the training failed, the error message of the underlying train function.}
}

}
\section{Extends}{\code{\linkS4class{object}}}
\section{Methods}{\describe{\item{\code{\link[=pred.learner,classif.ada,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.adaboost.M1,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,base.wrapper,wrapped.model,data.frame,ANY-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.blackboost,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.blackboost,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.ctree,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.earth,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.gbm,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.gbm,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=resample.update,resample.instance,learn.task,wrapped.model,prediction-method]{resample.update}}}{}\item{\code{\link[=pred.learner,classif.glmboost,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.grplasso,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.J48,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.JRip,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.kknn,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.kknn,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.km,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.ksvm,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.ksvm,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.lasso,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.lda,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.lm,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.loclda,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.logreg,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.lssvm,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.lvq1,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.mars,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.mda,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,multiclass.wrapper,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.multinom,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.naiveBayes,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.nnet,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.nnet,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.OneR,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.PART,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.pcr,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.lpsvm,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=predict,wrapped.model-method]{predict}}}{\code{signature(object = "wrapped.model")}: Predict the target variable of new data using a fitted model. If the type is set to "prob" or "decision"
probabilities or decision values will be stored in the resulting object. The resulting class labels are 
the classes with the maximum values or thresholding can also be used.}\item{\code{\link[=pred.learner,preproc.wrapper,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.qda,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.randomForest,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.randomForest,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.rda,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.ridge,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.rpart,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.rpart,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.rsm,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,regr.rvm,wrapped.model,data.frame,missing-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.sda,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=pred.learner,classif.svm,wrapped.model,data.frame,character-method]{pred.learner}}}{}\item{\code{\link[=to.string,wrapped.model-method]{to.string}}}{}\item{\code{\link[=[,wrapped.model-method]{[}}}{\code{signature( = "wrapped.model")}: Getter.}}}
\alias{[,wrapped.model-method}

